% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{amicusCorpus}
\alias{amicusCorpus}
\title{Amicus briefs from  US Supreme court cases}
\format{corpus with 100 texts, and attribute for petitioner/respondent}
\description{
This is a corpus of petitioner and respondent briefs to two US Supreme court cases on affirmative action; Bakke(1978) and Bollinger (2008)
}
\references{
Evans, Michael, et al. "Recounting the courts? Applying automated content analysis to enhance empirical legal research." Journal of Empirical Legal Studies 4.4 (2007): 1007-1039.
}

