\docType{data}
\name{amicusTexts}
\alias{amicusTexts}
\title{Amicus briefs from  US Supreme court cases}
\format{character vector with 100 texts}
\description{
This is a text vector of petitioner and respondent briefs
to two US Supreme court cases on affirmative action;
Bakke(1978) and Bollinger (2008)
}
\references{
Evans, Michael, et al. "Recounting the courts? Applying
automated content analysis to enhance empirical legal
research." Journal of Empirical Legal Studies 4.4 (2007):
1007-1039.
}

