% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus-methods.R
\name{as.character.corpus}
\alias{as.character.corpus}
\alias{is.corpus}
\alias{as.corpus}
\title{Coercion and checking methods for corpus objects}
\usage{
\method{as.character}{corpus}(x, use.names = TRUE, ...)

is.corpus(x)

as.corpus(x)
}
\arguments{
\item{x}{object to be coerced or checked}

\item{use.names}{logical; preserve (document) names if \code{TRUE}}

\item{...}{additional arguments used by specific methods}
}
\value{
\code{as.character()} returns the corpus as a plain character vector, with
or without named elements.

\code{is.corpus} returns \code{TRUE} if the object is a corpus.

\code{as.corpus()} upgrades a corpus object to the newest format.
object.
}
\description{
Coercion functions to and from \link{corpus} objects, including conversion to a
plain \link{character} object; and checks for whether an object is a corpus.
}
\note{
\code{as.character(x)} where \code{x} is a corpus is equivalent to
calling the deprecated \code{texts(x)}.
}
