% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel-methods.R
\name{as.coefficients_textmodel}
\alias{as.coefficients_textmodel}
\title{Coerce various objects to coefficients_textmodel
This is a helper function used in \verb{summary.textmodel_*}.}
\usage{
as.coefficients_textmodel(x)
}
\arguments{
\item{x}{an object to be coerced}
}
\description{
Coerce various objects to coefficients_textmodel
This is a helper function used in \verb{summary.textmodel_*}.
}
\keyword{internal}
