% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpuszip.R
\name{as.corpus.corpuszip}
\alias{as.corpus.corpuszip}
\title{coerce a compressed corpus to a standard corpus}
\usage{
\method{as.corpus}{corpuszip}(x)
}
\arguments{
\item{x}{a compressed \link{corpus} object}
}
\description{
Recast a compressed corpus object into a standard (uncompressed) corpus object.
}
