% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus-methods-quanteda.R
\name{as.corpus}
\alias{as.corpus}
\title{Make sure a corpus is the current format}
\usage{
as.corpus(x)
}
\arguments{
\item{x}{a \link{corpus} or compressed \link{corpus} object}
}
\description{
Upgrades a corpus object to the current format.  For backwards compatibility,
converts a compressed corpus object into a standard (uncompressed) corpus
object.
}
\keyword{internal}
