% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfm-classes.R
\docType{methods}
\name{as.data.frame,dfm-method}
\alias{as.data.frame,dfm-method}
\alias{as.data.frame.dfm}
\title{coerce a dfm to a data.frame}
\usage{
\S4method{as.data.frame}{dfm}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{dfm to be coerced to a data.frame}

\item{row.names}{if \code{FALSE}, do not set the row names of the data.frame
to the docnames of the dfm (default); or a vector of values to which the
row names will be set.}

\item{optional}{not applicable to this method}

\item{...}{not used for this method}
}
\description{
Method for coercing a \link{dfm-class} object to a data.frame
}
\examples{
inaugDfm <- dfm(data_char_inaugural[1:5], verbose = FALSE)
as.data.frame(inaugDfm[, 1:10])
str(as.data.frame(inaugDfm))
as.data.frame(inaugDfm[, 1:10], row.names = FALSE)
}

