% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfm-classes.R
\name{as.data.frame.dfm}
\alias{as.data.frame.dfm}
\title{Convert a dfm to a data.frame}
\usage{
\method{as.data.frame}{dfm}(
  x,
  row.names = NULL,
  ...,
  document = docnames(x),
  docid_field = "doc_id",
  check.names = FALSE
)
}
\arguments{
\item{x}{any \R object.}

\item{row.names}{\code{NULL} or a character vector giving the row
    names for the data frame.  Missing values are not allowed.}

\item{...}{unused}

\item{document}{optional first column of mode \code{character} in the
data.frame, defaults \code{docnames(x)}.  Set to \code{NULL} to exclude.}

\item{docid_field}{character; the name of the column containing document
names used when \code{to = "data.frame"}.  Unused for other conversions.}

\item{check.names}{logical.  If \code{TRUE} then the names of the
    variables in the data frame are checked to ensure that they are
    syntactically valid variable names and are not duplicated.
    If necessary they are adjusted (by \code{\link[base]{make.names}})
    so that they are.}
}
\description{
Deprecated function to convert a dfm into a data.frame.
Recommended that you use \code{convert(x, to = "data.frame")} instead.
}
\seealso{
\code{\link[=convert]{convert()}}
}
\keyword{dfm}
\keyword{internal}
