% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textstat_dist_old.R
\name{as.dist.dist}
\alias{as.dist.dist}
\title{Coerce a dist into a dist}
\usage{
\method{as.dist}{dist}(m, diag = FALSE, upper = FALSE)
}
\description{
Internal function to guarantee that a dist remains a dist, if for some reason
a user wants to coerce a dist into a dist.
}
\keyword{internal}
\keyword{textstat}
