% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcm-classes.R
\name{as.fcm}
\alias{as.fcm}
\title{Coercion functions for fcm objects}
\usage{
as.fcm(x, slots)
}
\arguments{
\item{x}{an object coerced to \link{fcm}. Currently only support
\link{Matrix} objects.}

\item{slots}{a list of values to be assigned to slots}
}
\description{
Coercion functions for fcm objects
}
\keyword{internal}
