% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokens.R
\name{as.keras}
\alias{as.keras}
\alias{as.keras.tokens}
\title{coerce tokens to sequences for Keras}
\usage{
as.keras(x)

\method{as.keras}{tokens}(x)
}
\arguments{
\item{x}{object coerced to Keras sequences}
}
\value{
a Python list object
}
\description{
\code{as.keras} converts \link{tokens} into sequences of integers for
  \code{keras} models.
}
\details{
Conversion of R objects to Python objects is performed 
  using the \code{reticulate} package.
}
\examples{
\dontrun{
txt <- data_char_ukimmig2010

# use keras' tokenizer
keras_tokenizer <-
    keras::text_tokenizer() \%>\%
    keras::fit_text_tokenizer(txt)
seqs <- keras::texts_to_sequences(keras_tokenizer, txt)

# use quanteda's tokenizer
seqs <- as.keras(tokens(data_char_ukimmig2010))

}
}
