% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textstat_dist_old.R
\name{as.list.dist}
\alias{as.list.dist}
\title{Coerce a dist object into a list}
\usage{
\method{as.list}{dist}(x, sorted = TRUE, n = NULL, ...)
}
\arguments{
\item{x}{dist class object}

\item{sorted}{sort results in descending order if \code{TRUE}}

\item{n}{the top \code{n} highest-ranking items will be returned.  If n is 
\code{NULL}, return all items.}

\item{...}{unused}
}
\description{
Coerce a dist matrix into a list of selected target terms and similar terms,
in descending order of similarity.  Can be used after calling
\code{\link{textstat_simil}} or \code{\link{textstat_dist}}.
}
