% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textstat_dist.R
\name{as.list.dist_selection}
\alias{as.list.dist_selection}
\title{coerce a dist_selection object into a list}
\usage{
\method{as.list}{dist_selection}(x, sorted = TRUE, n = NULL, ...)
}
\arguments{
\item{x}{dist_selection class object}

\item{sorted}{sort results in descending order if \code{TRUE}}

\item{n}{the top \code{n} highest-ranking items will be returned.  If n is 
\code{NULL}, return all items.}

\item{...}{unused}
}
\description{
Coerce a dist_selection matrix into a list of selected terms and tarhet terms in
descending order.  Can be used after calling \code{\link{textstat_simil}} or
\code{\link{textstat_dist}} when selection is not NULL
}
\keyword{internal}
\keyword{textstat}
