% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textstat_dist.R
\name{as.matrix.dist_selection}
\alias{as.matrix.dist_selection}
\title{coerce a dist_selection object to a matrix}
\usage{
\method{as.matrix}{dist_selection}(x, ...)
}
\description{
Coerce a dist_selection object to a plain matrix.
}
\keyword{internal}
\keyword{textstat}
