% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textstat_simil.R
\name{as.Matrix}
\alias{as.Matrix}
\title{Coerce a textstat_simil or textstat_dist to a sparse Matrix format}
\usage{
as.Matrix(x)
}
\arguments{
\item{x}{input textstat object}
}
\description{
Coerce a \link{textstat_simil} or \link{textstat_dist} to a sparse \pkg{Matrix} format.
}
\keyword{internal}
\keyword{textstat}
