% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textstat_simil.R
\name{as.matrix.simil}
\alias{as.matrix.simil}
\title{Coerce a simil object into a matrix}
\usage{
\method{as.matrix}{simil}(x, diag = 1, ...)
}
\arguments{
\item{diag}{the value to use on the diagonal representing self-similarities}

\item{...}{unused}
}
\description{
\code{as.matrix.simil} coerces an object returned from
  `textstat_simil()` into a matrix
}
\note{
Because for the similarity methods implemented in  \pkg{quanteda}, the
  similarity of an object with itself will be 1.0, \code{diag} defaults to
  this value. This differs the default \code{diag = NA} in
  \link[proxy]{as.matrix.simil} in the \pkg{proxy} package.
}
\keyword{internal}
\keyword{textstat}
