% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textplot_network.R
\name{as.network}
\alias{as.network}
\title{redefinition of network::as.network()}
\usage{
as.network(x, ...)
}
\arguments{
\item{x}{input object}

\item{...}{additional arguments}
}
\description{
redefinition of network::as.network()
}
\keyword{internal}
