% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel-methods.R
\name{as.statistics_textmodel}
\alias{as.statistics_textmodel}
\title{Assign the textmodel_coefficients class to a data.frame}
\usage{
as.statistics_textmodel(x)
}
\arguments{
\item{x}{a data.frame}
}
\description{
Assign the textmodel_coefficients class to a data.frame
}
\keyword{internal}
