% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel-methods.R
\name{as.textmodel_coefficients}
\alias{as.textmodel_coefficients}
\title{Assign the textmodel_coefficients class to a numeric vector}
\usage{
as.textmodel_coefficients(x)
}
\arguments{
\item{x}{a numeric vector}
}
\description{
Assign the textmodel_coefficients class to a numeric vector
}
\keyword{internal}
