% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel-methods.R
\name{as.textmodel_statistics}
\alias{as.textmodel_statistics}
\title{Assign the textmodel_coefficients class to a data.frame}
\usage{
as.textmodel_statistics(x)
}
\arguments{
\item{x}{a data.frame}
}
\description{
Assign the textmodel_coefficients class to a data.frame
}
\keyword{internal}
