% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel-methods.R
\name{as.textmodel_summary}
\alias{as.textmodel_summary}
\title{Assign the textmodel_summary class to a list}
\usage{
as.textmodel_summary(x)
}
\arguments{
\item{x}{a named list}
}
\description{
Assign the textmodel_summary class to a list
}
\keyword{internal}
