% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokens_xptr.R
\name{is.tokens_xptr}
\alias{is.tokens_xptr}
\alias{as.tokens_xptr}
\alias{as.tokens_xptr.tokens}
\alias{as.tokens_xptr.tokens_xptr}
\title{Convert a tokens object to a tokens_xptr object}
\usage{
is.tokens_xptr(x)

as.tokens_xptr(x)

\method{as.tokens_xptr}{tokens}(x)

\method{as.tokens_xptr}{tokens_xptr}(x)
}
\arguments{
\item{x}{a \link{tokens} object to convert or a \code{tokens_xptr} object to deep copy.}
}
\description{
\code{tokens_xptr} is a external pointer-based tokens object designed to process a
large corpus efficiently.
}
