% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.weight.R
\name{as.weights}
\alias{as.weights}
\title{[Experimental] Create feature weight vector from weighted patterns}
\usage{
as.weights(x, y, valuetype = c("glob", "regex", "fixed"),
  case_insensitive = TRUE, nomatch_weight = NULL)
}
\arguments{
\item{x}{named numeric vector as weighted patterns}

\item{y}{dfm for which weight vector is created}
}
\description{
[Experimental] Create feature weight vector from weighted patterns
}
\examples{
sent = c("good" = 1, "great*" = 1, "bad" = -1, "wors*" = -1)
mt <- dfm(c("Good and great, greatest are positive", 
            "Bad, worse and worst are negative"), tolower = FALSE)
weight <- as.weights(sent, mt, nomatch_weight = 0)
dfm_weight(mt, weights = weight)
}
