% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{attributes<-}
\alias{attributes<-}
\title{function extending base::attributes()}
\usage{
attributes(x, overwrite = TRUE) <- value
}
\arguments{
\item{x}{an object}

\item{overwrite}{if \code{TRUE}, overwrite old attributes}

\item{value}{new attributes}
}
\description{
function extending base::attributes()
}
\keyword{internal}
