% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_corpus.R
\name{bootstrap_corpus}
\alias{bootstrap_corpus}
\alias{bootstrap_corpus.corpus}
\alias{bootstrap_character}
\alias{bootstrap_character.character}
\title{bootstrap a corpus or character object by resampling sentences}
\usage{
bootstrap_corpus(x, n = 10, verbose = quanteda_options("verbose"))

\method{bootstrap_corpus}{corpus}(x, n = 10,
  verbose = quanteda_options("verbose"))

bootstrap_character(x, n = 10, verbose = quanteda_options("verbose"))

\method{bootstrap_character}{character}(x, n = 10,
  verbose = quanteda_options("verbose"))
}
\arguments{
\item{x}{a \link{corpus} or character object}

\item{n}{number of resamples}

\item{verbose}{if \code{TRUE} print status messages}
}
\value{
A named list of character or \link{corpus} objects, where the first,
  the object name followed by \code{_0}, is the original item before
  resampling, and the subsequent elements are the sentence-resampled text
  objects.
}
\description{
Bootstrap a corpus or character object by resampling from sentences, with
replacement.
}
\details{
Function produces multiple, resampled character or \link{corpus} or
  objects, based on resampling sentences (wth replacement) from each
  document, recombining these into new "documents" and computing a dfm for
  each. Resampling of sentences is done strictly within document, so that
  every resampled document will contain at least some of its original tokens.
}
\examples{
set.seed(5)
# bootstrapping from a character
txt <- c(textone = "This is a sentence.  Another sentence.  Yet another.", 
         texttwo = "Premiere phrase.  Deuxieme phrase.")
bootstrap_character(txt, n = 3)

# bootstrapping from a corpus 
corp <- corpus(txt)
texts(bootstrap_corpus(corp, n = 3))
}
\author{
Kenneth Benoit
}
\keyword{bootstrap}
\keyword{corpus}
\keyword{experimental}
