% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_dfm.R
\name{bootstrap_dfm}
\alias{bootstrap_dfm}
\title{bootstrap a dfm}
\usage{
bootstrap_dfm(x, n = 10, ...)
}
\arguments{
\item{x}{the dfm}

\item{n}{number of resamples}

\item{...}{additional arguments passed to \link{\code{dfm}}}
}
\description{
Create an array of resampled dfms.
}
