% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quanteda-documentation.R
\name{case_insensitive}
\alias{case_insensitive}
\title{Case-insensitive pattern matching}
\arguments{
\item{case_insensitive}{logical; if \code{TRUE}, ignore case when matching a
\code{pattern} or \link{dictionary} values.}
}
\description{
Case-insensitive pattern matching
}
\keyword{internal}
