% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_class}
\alias{check_class}
\title{Print friendly object class not defined message}
\usage{
check_class(object_class, function_name)
}
\arguments{
\item{object_class}{character describing the object class}

\item{function_name}{character which is the function name}
}
\description{
Checks valid methods and issues a friendlier error message in case the method is
undefined for the supplied object type.
}
\examples{
# as.tokens.default <- function(x, concatenator = "", ...) {
#     stop(quanteda:::friendly_class_undefined_message(class(x), "as.tokens"))
# }
}
\keyword{internal}
