% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_dots}
\alias{check_dots}
\title{Check arguments passed via ...}
\usage{
check_dots(dots, permissible_args = NULL)
}
\arguments{
\item{...}{dots to check}

\item{class}{names of functions that the ...}
}
\description{
Check arguments passed via ...
}
\keyword{internal}
