% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_font}
\alias{check_font}
\title{Check if font is available on the system}
\usage{
check_font(font)
}
\arguments{
\item{font}{name of a font to be checked if available on the system.}
}
\value{
character string
}
\description{
This function checks if custom font is available to \pkg{ggplot} and
\pkg{graphics} APIs.
}
\keyword{internal}
