% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator.R
\name{check_integer}
\alias{check_integer}
\alias{check_double}
\alias{check_logical}
\alias{check_character}
\title{Validate input vectors}
\usage{
check_integer(x, min_len = 1, max_len = 1, min = -Inf, max = Inf)

check_double(x, min_len = 1, max_len = 1, min = -Inf, max = Inf)

check_logical(x, min_len = 1, max_len = 1)

check_character(x, min_len = 1, max_len = 1, min_nchar = 0, max_nchar = Inf)
}
\arguments{
\item{min_len}{minimum length of the vector}

\item{max_len}{maximum length of the vector}

\item{min}{minimum value in the vector}

\item{max}{maximum value in the vector}

\item{min_nchar}{minimum character length of values in the vector}

\item{max_nchar}{maximum character length of values in the vector}
}
\description{
Check the range of values and the length of input vectors
before used in control flow or passed to C++ functions.
}
\keyword{internal}
