% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{check_valid_meta_object}
\alias{check_valid_meta_object}
\title{Check if an object is eligible for meta functions}
\usage{
check_valid_meta_object(x)
}
\arguments{
\item{x}{an object to be checked for eligibility}
}
\value{
nothing; halts with an error if \code{x} is ineligible
}
\description{
Check the class of an object to see if it is eligible for \code{meta()}-type
functions.` This avoids having to make the meta functions object-oriented.
Even though \link{dfm} and \link{dictionary} functions are S4, the same code works
because the \code{\link[=attr]{attr()}} calls work for named slots (since they are just special
attributes).
}
\keyword{internal}
\keyword{meta}
