% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel_ca.R
\name{coef.textmodel_ca}
\alias{coef.textmodel_ca}
\alias{coefficients.textmodel_ca}
\title{Extract model coefficients from a fitted textmodel_ca object}
\usage{
\method{coef}{textmodel_ca}(object, doc_dim = 1, feat_dim = 1, ...)

coefficients.textmodel_ca(object, doc_dim = 1, feat_dim = 1, ...)
}
\arguments{
\item{object}{a fitted \link{textmodel_ca} object}

\item{doc_dim, feat_dim}{the document and feature dimension scores to be
extracted}

\item{...}{unused}
}
\description{
\code{coef()} extract model coefficients from a fitted \code{textmodel_ca}
object.  \code{coefficients()} is an alias.
}
\keyword{internal}
\keyword{textmodel}
