\name{collocates.flipIndex}
\alias{collocates.flipIndex}
\title{This function flips index to obtain words from column/row number of collocate matricis}
\usage{
collocates.flipIndex(index)
}
\arguments{
  \item{index}{Index created by collocates.createIndex()}
}
\value{
list List of words with column/row number as keys
}
\description{
This function flips index to obtain words from column/row
number of collocate matricis
}
\examples{
words <- collocates.flipIndex(index)
}
\author{
Kohei Watanabent
}

