% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textstat_collocations.R
\name{collocations}
\alias{collocations}
\alias{sequences}
\title{deprecated function names for textstat_collocations}
\usage{
collocations(x, ...)

sequences(x, ...)
}
\arguments{
\item{x}{a character, \link{corpus}, \link{tokens} object}

\item{...}{other arguments passed to  \code{\link{textstat_collocations}}}
}
\description{
Use \code{\link{textstat_collocations}} instead.
}
\seealso{
\link{textstat_collocations}
}
\keyword{collocations}
\keyword{deprecated}
\keyword{internal}
