% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collocations.R
\name{collocations_old}
\alias{collocations_old}
\title{detect collocations from text}
\usage{
collocations_old(x, method = c("lr", "chi2", "pmi", "dice", "all"),
  size = 2, n = NULL, toLower = TRUE, punctuation = c("dontspan",
  "ignore", "include"), ...)
}
\arguments{
\item{x}{a character, \link{corpus}, \link{tokens} object}

\item{method}{association measure for detecting collocations.  Let \eqn{i} 
index documents, and \eqn{j} index features, \eqn{n_{ij}} refers to 
observed counts, and \eqn{m_{ij}} the expected counts in a collocations 
frequency table of dimensions \eqn{(J - size + 1)^2}. Available measures 
are computed as: \describe{ \item{\code{"lr"}}{The likelihood ratio 
statistic \eqn{G^2}, computed as: \deqn{2 * \sum_i \sum_j ( n_{ij} * log 
\frac{n_{ij}}{m_{ij}} )} } \item{\code{"chi2"}}{Pearson's \eqn{\chi^2} 
statistic, computed as: \deqn{\sum_i \sum_j \frac{(n_{ij} - 
m_{ij})^2}{m_{ij}}} } \item{\code{"pmi"}}{point-wise mutual information 
score, computed as log \eqn{n_{11}/m_{11}}} \item{\code{"dice"}}{the Dice 
coefficient, computed as \eqn{n_{11}/n_{1.} + n_{.1}}} 
\item{\code{"all"}}{returns all of the above} }}

\item{size}{length of the collocation.  Only bigram (\code{n=2}) and trigram 
(\code{n=3}) collocations are currently implemented.  Can be \code{c(2,3)}
(or \code{2:3}) to return both bi- and tri-gram collocations.}

\item{n}{the number of collocations to return, sorted in descending order of 
the requested statistic, or \eqn{G^2} if none is specified.}

\item{toLower}{convert collocations to lower case if \code{TRUE} (default)}

\item{punctuation}{how to handle tokens separated by punctuation characters.  Options are:
\describe{
\item{\code{dontspan}}{do not form collocations from tokens separated by punctuation characters (default)}
\item{\code{ignore}}{ignore punctuation characters when forming collocations, meaning that collocations will 
include those separated by punctuation characters in the text.  The punctuation characters themselves are not
returned.}
\item{\code{include}}{do not treat punctuation characters specially, meaning that collocations will include
punctuation characters as tokens}
}}

\item{...}{additional parameters passed to \code{\link{tokens}}}
}
\value{
a collocations class object: a specially classed data.table consisting 
  of collocations, their frequencies, and the computed association measure(s).
}
\description{
Detects collocations from texts or a corpus, returning a data.frame of
collocations and their scores, sorted in descending order of the association
measure.  Words separated by punctuation delimiters are not counted by
default (\code{spanPunct = FALSE})  as adjacent and hence are not eligible to
be collocations.
}
\examples{
txt <- c("This is software testing: looking for (word) pairs!  
         This [is] a software testing again. For.",
         "Here: this is more Software Testing, looking again for word pairs.")
collocations_old(txt, punctuation = "dontspan") # default
collocations_old(txt, punctuation = "dontspan", removePunct = TRUE)  # includes "testing looking"
collocations_old(txt, punctuation = "ignore", removePunct = TRUE)    # same as previous 
collocations_old(txt, punctuation = "include", removePunct = FALSE)  # keep punctuation as tokens

collocations_old(txt, size = 2:3)
# removeFeatures(collocations_old(txt, size = 2:3), stopwords("english"))

collocations_old("@textasdata We really, really love the #quanteda package - thanks!!")
collocations_old("@textasdata We really, really love the #quanteda package - thanks!!",
              removeTwitter = TRUE)

collocations_old(data_char_inaugural[49:57], n = 10)
collocations_old(data_char_inaugural[49:57], method = "all", n = 10)
collocations_old(data_char_inaugural[49:57], method = "chi2", size = 3, n = 10)
collocations_old(corpus_subset(data_corpus_inaugural, Year>1980), method = "pmi", size = 3, n = 10)
}
\references{
McInnes, B T. 2004. "Extending the Log Likelihood Measure to 
  Improve Collocation Identification."  M.Sc. Thesis, University of 
  Minnesota.
}
\seealso{
\link{tokens_ngrams}
}
\author{
Kenneth Benoit
}
