% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textstat_collocationsdev.R
\name{collocationsdev}
\alias{collocationsdev}
\title{deprecated function names for textstat_collocationsdev}
\usage{
collocationsdev(x, ...)
}
\arguments{
\item{x}{a character, \link{corpus}, \link{tokens} object}

\item{...}{other arguments passed to  \code{\link{textstat_collocationsdev}}}
}
\description{
Use \code{\link{textstat_collocationsdev}} instead.
}
\seealso{
\link{textstat_collocationsdev}
}
\keyword{collocations}
\keyword{deprecated}
\keyword{internal}
