% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dfm.R
\docType{methods}
\name{colSums,dfmSparse-method}
\alias{colSums,dfmSparse-method}
\title{the calculations
myDfm <- dfm(inaugTexts, verbose=FALSE)}
\usage{
\S4method{colSums}{dfmSparse}(x, na.rm = FALSE, dims = 1L, ...)
}
\arguments{
\item{x}{a dfm, inheriting from \link[Matrix]{Matrix}}

\item{na.rm}{if \code{TRUE}, omit missing values (including \code{NaN}) from
the calculations}

\item{dims}{ignored}

\item{...}{additional arguments, for methods/generic compatibility}
}
\description{
the calculations
myDfm <- dfm(inaugTexts, verbose=FALSE)
}

