% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textstat_lexdiv.R
\name{compute_mattr}
\alias{compute_mattr}
\title{Computes the Moving-Average Type-Token Ratio for a Tokens Object}
\usage{
compute_mattr(x, window_size = NULL, all_windows = FALSE,
  mean_mattr = TRUE)
}
\arguments{
\item{x}{input \link{tokens}}

\item{window_size}{input: the size of the moving window for computation of TTR}

\item{all_windows}{default = FALSE. If TRUE, returns a vector with the MATTR for each window}

\item{mean_mattr}{default = TRUE. Returns the mean MATTR, given the window size, for the tokens object.}
}
\value{
return either a vector with the MATTR for each window, or the mean MATTR for the whole tokens object
}
\description{
Takes a tokens object which should have been preprocessed - removal of punctuation etc.
}
\keyword{dfm}
\keyword{internal}
