% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textstat_lexdiv.R
\name{compute_mattr2}
\alias{compute_mattr2}
\title{Computes the Moving-Average Type-Token Ratio}
\usage{
compute_mattr2(x, MATTR_window = 100L)
}
\arguments{
\item{x}{a \link{tokens} object}

\item{window_size}{integer; the size of the moving window for computation of
TTR, between 1 and the number of tokens of the document}
}
\description{
Takes a tokens object which should have been preprocessed - removal of
punctuation etc. and computes the Moving-Average Type-Token Ratio from
Covington & McFall (2010).
}
\details{
Ken's variation.
}
\keyword{internal}
\keyword{lexdiv}
\keyword{textstat}
