% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy.R
\name{copy}
\alias{copy}
\title{Copy an entire object}
\usage{
copy(x)
}
\arguments{
\item{x}{the input object to be copied}
}
\value{
an entire copy of \code{x}, by value
}
\description{
Creates a deep copy of an object, as an external pointer, for instance a
\link{tokens_xptr} class object.
}
\examples{
xtoks <- tokens(data_corpus_inaugural[1:3], xptr = TRUE)
print(xtoks, 0, 0)

# shallow copy
xtoks_shallow <- xtoks
print(xtoks_shallow, 0, 0)

# deep copy
xtoks_deep <- copy(xtoks)
print(xtoks_deep, 0, 0)
}
