\name{corpus.add.attributes}
\alias{corpus.add.attributes}
\title{This function adds a named list of attributes to an existing corpus}
\usage{
  corpus.add.attributes(corpus, newattribs,
    name = newattribs)
}
\arguments{
  \item{corpus}{Corpus to add attributes to}

  \item{newattribs}{A list of new attribues}

  \item{name}{A name for the new attribues}
}
\description{
  This function adds a named list of attributes to an
  existing corpus
}

