% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus_group.R
\name{corpus_group}
\alias{corpus_group}
\title{Recombine documents in corpus by a grouping variable}
\usage{
corpus_group(x, groups = NULL, fill = FALSE, concatenator = " ")
}
\arguments{
\item{x}{\link{corpus} object}

\item{groups}{grouping variable for sampling, equal in length to the number
of documents. This will be evaluated in the docvars data.frame, so that
docvars may be referred to by name without quoting. This also changes
previous behaviours for \code{groups}. See \code{?news(Version >= "2.9", package = "quanteda")} for details.}

\item{fill}{logical; if \code{TRUE} and \code{groups} is a factor, then use all levels
of the factor when forming the new "documents" of the grouped dfm.  This
will result in a new "document" with empty content, for levels not
observed.  Has no effect if the \code{groups} variable(s) are not factors.}

\item{concatenator}{the concatenation character that will connect the documents.}
}
\description{
Combine documents in a \link{corpus} object by a grouping variable, which can also
be one of the \link{docvars} attached to the dfm.
}
\examples{
corp <- corpus(c("a a b", "a b c c", "a c d d", "a c c d"),
               docvars = data.frame(grp = c("grp1", "grp1", "grp2", "grp2")))
corpus_group(corp, groups = "grp")
corpus_group(corp, groups = c(1, 1, 2, 2))
corpus_group(corp, groups = factor(c(1, 1, 2, 2), levels = 1:3))
}
\keyword{corpus}
