% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus_sample.R, R/dfm-methods.R
\name{corpus_sample}
\alias{corpus_sample}
\alias{sample.dfm}
\title{randomly sample documents from a corpus}
\usage{
corpus_sample(x, size = ndoc(x), replace = FALSE, prob = NULL, ...)

\method{sample}{dfm}(x, size = ndoc(x), replace = FALSE, prob = NULL,
  what = c("documents", "features"), ...)
}
\arguments{
\item{x}{a corpus object whose documents will be sampled}

\item{size}{a positive number, the number of documents to select}

\item{replace}{Should sampling be with replacement?}

\item{prob}{A vector of probability weights for obtaining the elements of the
vector being sampled.}

\item{...}{unused}

\item{what}{dimension (of a \link{dfm}) to sample: can be \code{documents} or
\code{features}}
}
\value{
A corpus object with number of documents equal to \code{size}, drawn 
  from the corpus \code{x}.  The returned corpus object will contain all of 
  the meta-data of the original corpus, and the same document variables for 
  the documents selected.

A dfm object with number of documents equal to \code{size}, drawn 
  from the corpus \code{x}.  The returned corpus object will contain all of 
  the meta-data of the original corpus, and the same document variables for 
  the documents selected.
}
\description{
Takes a random sample or documents or features of the specified size from a 
corpus or document-feature matrix, with or without replacement.  Works just 
as \code{\link{sample}} works for the documents and their associated
document-level variables.
}
\examples{
# sampling from a corpus
summary(corpus_sample(data_corpus_inaugural, 5)) 
summary(corpus_sample(data_corpus_inaugural, 10, replace=TRUE))
# sampling from a dfm
myDfm <- dfm(data_char_inaugural[1:10], verbose = FALSE)
sample(myDfm)[, 1:10]
sample(myDfm, replace = TRUE)[, 1:10]
sample(myDfm, what = "features")[1:10, ]
}
\seealso{
\code{\link{sample}}
}

