% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus_subset.R
\name{corpus_subset}
\alias{corpus_subset}
\title{extract a subset of a corpus}
\usage{
corpus_subset(x, subset, select, ...)
}
\arguments{
\item{x}{corpus object to be subsetted}

\item{subset}{logical expression indicating elements or rows to keep: missing
values are taken as false}

\item{select}{expression, indicating the attributes to select from the corpus}

\item{...}{not used}
}
\value{
corpus object
}
\description{
Returns subsets of a corpus that meet certain conditions, including direct 
logical operations on docvars (document-level variables).  `corpus_subset` 
functions identically to \link{subset} where non-standard evaluation is
used to evaluate conditions based on the \link{docvars} in the corpus.
}
\examples{
summary(corpus_subset(data_corpus_inaugural, Year>1980))
summary(corpus_subset(data_corpus_inaugural, Year>1930 & President=="Roosevelt", select=Year))
}
\seealso{
\code{\link{subset}}
}

