% Generated by roxygen2 (4.0.1): do not edit by hand
\name{corpusAppend}
\alias{corpusAppend}
\title{function to add new texts and attributes to an existing corpus
Accepts a list of texts and a list of associated attributes and
adds them to the corpus}
\usage{
corpusAppend(corpus1, newtexts, newattribs, ...)
}
\arguments{
\item{corpus1}{An existing corpus to add new texts and attributes to}

\item{newtexts}{New texts to be added to the corpus}

\item{newattribs}{New attribs associated with the new texts
text}
}
\description{
function to add new texts and attributes to an existing corpus
Accepts a list of texts and a list of associated attributes and
adds them to the corpus
}
\examples{
data(iebudgets)
data(ieAttribs)
data(ieTexts)
budgets <- corpusAppend(iebudgets, ieTexts, ieAttribs)
}

