% Generated by roxygen2 (4.0.1): do not edit by hand
\name{corpusCreate}
\alias{corpusCreate}
\title{Create a new corpus
This function creates a corpus from a character vector (of texts),
adds text-specific variables (which we term "attributes"), along
with optional meta-data and notes.}
\usage{
corpusCreate(texts, attribs = NULL, textnames = NULL, source = NULL,
  notes = NULL)
}
\arguments{
\item{texts}{A character vector containing the texts to be stored in the corpus.}

\item{textnames}{Names to be assigned to the texts, defaults to the names of the
character vector (if any), otherwise assigns "text1", "text2", etc.}

\item{attribs}{A data frame of attributes that is associated with each text.}

\item{source}{A string specifying the source of the texts, used for referencing.}

\item{notes}{A string containing notes about who created the text, warnings, To Dos, etc.}
}
\description{
Create a new corpus
This function creates a corpus from a character vector (of texts),
adds text-specific variables (which we term "attributes"), along
with optional meta-data and notes.
}
\examples{
data(ieTexts)
data(ieAttribs)
budgets <- corpusCreate(ieTexts, attribs=ieAttribs)
summary(budgets)
}

