% Generated by roxygen2 (4.0.1): do not edit by hand
\name{corpusFromFilenames}
\alias{corpusFromFilenames}
\title{create a new corpus with attribute-value pairs taken from filenames}
\usage{
corpusFromFilenames(directory, attNames, sep = "_", enc = NULL)
}
\arguments{
\item{directory}{Path to folder containing documents}

\item{attNames}{A vector naming the attribute types}

\item{sep}{A string by which the filename should be separated to get the values. Default is underscore.}

\item{enc}{Encoding for the texts (this is the "to" argument sent to iconv)}
}
\description{
This function takes a directory, reads in all the documents in that directory
and makes a new corpus where the attributes and values are created by splitting
the filename according to a separator.For example, a directory may contain files
with a naming scheme that identifies attribute values, e.g.:
"2010_BUDGET_05_Brian_Cowen_FF.txt".
}
\details{
To create a corpus object from texts named in this format, we can call this
function and specify the attribute types and separator, e.g:
new_corpus <- corpusFromFilenames(dirname, c("country", "electionType", "year", "language", "party"), sep='_')

Underscore is the default separator
}
\examples{
\dontrun{
new_corpus <- corpusFromFilenames(dirname, c("country", "electionType", "year", "language", "party"), sep='_')
}
}
\author{
Paul Nulty and Ken Benoit
}

