% Generated by roxygen2 (4.0.1): do not edit by hand
\name{corpusReshape}
\alias{corpusReshape}
\title{Transform a corpus by splitting texts into sentences}
\usage{
corpusReshape(corpus, to = c("sentence", "document"))
}
\arguments{
\item{corpus}{Corpus to transform}

\item{to}{target unit of analysis for the reshaping.  Currently only \code{sentence}
and \code{document} are supported.}
}
\description{
Each text in the corpus is split into sentences, and each
sentence becomes a standalone text, with attributes indicating
the text it is taken from and it's serial number in that text
}
\examples{
data(iebudgets)
ie2010document <- subset(iebudgets, year==2010)
summary(ie2010document)
ie2010sentence <- corpusReshape(ie2010document)  # reshape to sentence units
summary(ie2010sentence, 20)
}

