% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpusTagged-class.R
\docType{class}
\name{corpusTokenized}
\alias{as.corpus}
\alias{as.corpus,corpus-method}
\alias{as.corpus,corpusTokenized-method}
\alias{corpusTokenized}
\title{Virtual class "corpusTagged" for an indexed and tagged corpus object}
\usage{
as.corpus(object)

\S4method{as.corpus}{corpusTokenized}(object)

\S4method{as.corpus}{corpus}(object)
}
\description{
The \code{corpusTagged} object is a special type of 
  \link{corpus} that has been indexed by document, sentence, and token, and 
  tagged either minimally (distinguishing terms and numbers from white space 
  and punctuation characters) or fully using a part-of-speech tagger, if one 
  is installed and available (see Details).
}
\section{Slots}{

\describe{
\item{\code{documents}}{The original texts, and document vars, as a 
\link{data.frame}.}

\item{\code{documentsDT}}{The indexed corpus, as a \link{data.table}.}

\item{\code{tagger}}{character indicating which tagger was used.  Current permissible
values are \code{"none"} and \code{"spaCy"}.}

\item{\code{tagset}}{character indicating which tagset was applied to the corpus. 
Current permissible values are \code{"Penn"}, \code{"Google"}, and
\code{"none"} (which will be applied if no tagger was invoked). White space
characters, defined as a match with the Unicode property
\code{WHITE_SPACE}, are tagged as \code{spc}, and punctuation characters
not identified by a POS tagger are matched to the Unicode property
\code{PUNCTUATION} and marked as "punct".}
}}
\seealso{
\link{corpus}
}

