% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create}
\alias{create}
\title{Utility function to create a object with new set of attributes}
\usage{
create(x, what, attrs = NULL, overwrite_attributes = FALSE, ...)
}
\arguments{
\item{x}{an underlying R object of a new object}

\item{attrs}{attributes of a new object}

\item{overwrite_attributes}{overwrite attributes of the input object, if \code{TRUE}}
}
\description{
Utility function to create a object with new set of attributes
}
\keyword{internal}
