% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/textmodel-utils.R
\name{crossVal}
\alias{crossVal}
\title{Perform cross-validation}
\usage{
crossVal(dtm, values, k = nrow(dtm), model = "wordscores")
}
\arguments{
\item{dtm}{A dfm containing documents and features for training and testing.}

\item{values}{The 'true' target values for each document}

\item{k}{The number of 'folds' to split the data into. Default is leave-one-out.}

\item{model}{The model type to use for training and predictions}
}
\description{
Returns a numeric vector for use as target variable for a textmodel
}
\author{
Paul Nulty
}

