% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{data-relocated}
\alias{data-relocated}
\alias{data_corpus_dailnoconf1991}
\alias{data_corpus_irishbudget2010}
\alias{data_dictionary_LSD2015}
\title{Formerly included data objects}
\description{
The following corpus objects have been relocated to other (quanteda)
packages.
package:
\itemize{
\item \code{data_corpus_dailnoconf1991}
\item \code{data_corpus_irishbudget2010}
\item \code{data_dictionary_LSD2015}
}
}
\seealso{
\code{quanteda.textmodels::quanteda.textmodels-package},
\code{quanteda.sentiment::quanteda.sentiment-package}
}
