% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{data_char_encodedtexts}
\alias{data_char_encodedtexts}
\alias{encodedTexts}
\title{encoded texts for testing}
\format{An object of class \code{character} of length 10.}
\usage{
data_char_encodedtexts
}
\description{
\code{data_char_encodedtexts} is a 10-element character vector with 10
different encodings
}
\examples{
Encoding(data_char_encodedtexts)
data.frame(labelled = names(data_char_encodedtexts), 
           detected = encoding(data_char_encodedtexts)$all)
}
\keyword{datasets}

