% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{data_char_mobydick}
\alias{data_char_mobydick}
\alias{mobydickText}
\title{Project Gutenberg text of Herman Melville's \emph{Moby Dick}
Named character object of the ASCII text of Herman Melville's
 \emph{Moby Dick}, EBook no. 2701.}
\format{An object of class \code{character} of length 1.}
\source{
Project Gutenberg, \url{http://www.gutenberg.org}
}
\usage{
data_char_mobydick
}
\description{
Project Gutenberg text of Herman Melville's \emph{Moby Dick}
Named character object of the ASCII text of Herman Melville's
 \emph{Moby Dick}, EBook no. 2701.
}
\examples{
summary(data_char_mobydick)
}
\keyword{datasets}

