% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_download.R
\name{data_download}
\alias{data_download}
\title{donlowad RDS file from online locations}
\usage{
data_download(name = NULL, url = NULL, cache = TRUE, ...)
}
\arguments{
\item{name}{name of the RDS file}

\item{url}{user-specified location of the RDS file}

\item{cache}{if \code{TRUE}, save downloaded file to temporary folder}

\item{...}{extra argument passed to \code{download.file}}
}
\value{
an unserialized R object
}
\description{
donlowad RDS file from online locations
}
\details{
Currently only \code{data_corpus_foreignaffairscommittee} and
\code{data_corpus_guardian} are available.
}
\examples{
\\donotrun{
corp <- data_download('data_corpus_guardian')
corp <- data_download('data_corpus_foreignaffairscommittee')
}
}
