% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{deprecate_argument}
\alias{deprecate_argument}
\title{issue warning for deprecrated function arguments}
\usage{
deprecate_argument(old, new, args)
}
\description{
Sets a new argument name to an older one, and issues a deprecation warning.
}
\examples{
fn <- function(remove_numbers = TRUE, ...) {
    args <- as.list(match.call())
    remove_numbers <- quanteda:::deprecate_argument("removeNumbers", "remove_numbers", args)
    cat("remove_numbers =", remove_numbers, "\\n")
}
fn(removeNumbers = FALSE)
fn(remove_numbers = FALSE)
}
\keyword{internal}
