% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textstat-deprecated.R
\name{deprecated-textstat}
\alias{deprecated-textstat}
\alias{lexdiv}
\title{deprecated textstat names}
\usage{
lexdiv(x, ...)

lexdiv(x, ...)
}
\description{
Deprecated `textstat_` function names.
}
\details{
\code{lexdiv} is the deprecated; use \code{\link{textstat_lexdiv}} instead.

\code{readability} is the deprecated; use \code{\link{textstat_readability}} instead.
}
\keyword{deprecated}
\keyword{internal}

