% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfm-internal.R, R/dfm-methods.R
\docType{methods}
\name{dfm-internal}
\alias{dfm-internal}
\alias{Compare,dfmSparse,numeric-method}
\title{internal functions for dfm objects}
\usage{
\S4method{Compare}{dfmSparse,numeric}(e1, e2)
}
\arguments{
\item{e1}{a dfm}

\item{e2}{a numeric value to compare with values in a dfm}
}
\value{
a logical sparse matrix (lgCMatrix)
}
\description{
Internal function documentation for dfm objects.

enable relational operators to be use with dfm
}
