% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfm.R
\name{dfm}
\alias{dfm}
\title{Create a document-feature matrix}
\usage{
dfm(
  x,
  tolower = TRUE,
  remove = NULL,
  verbose = quanteda_options("verbose"),
  ...
)
}
\arguments{
\item{x}{a \link{tokens} or \link{dfm} object}

\item{tolower}{convert all features to lowercase}

\item{remove}{a \link{pattern} of user-supplied features to ignore, such as "stop
words".  To access one possible list (from any list you wish), use
\code{\link[=stopwords]{stopwords()}}.  The pattern matching type is fixed to "glob"; if you want
greater control, use \code{\link[=tokens_remove]{tokens_remove()}} or \code{\link[=dfm_remove]{dfm_remove()}}.  \code{remove} is
primarily useful for removing artifacts created during the \link{tokens}
creation process, such as pads.  See examples.}

\item{verbose}{display messages if \code{TRUE}}

\item{...}{not used directly}
}
\value{
a \linkS4class{dfm} object
}
\description{
Construct a sparse document-feature matrix, from a character, \link{corpus},
\link{tokens}, or even other \link{dfm} object.
}
\section{Changes to dfm() in version 3}{

In \pkg{quanteda} v3, many convenience functions formerly available in
\code{dfm()} were deprecated. Formerly, \code{dfm()} could be called directly on a
\code{character} or \code{corpus} object, but we now steer users to tokenise their
inputs first using \code{\link[=tokens]{tokens()}}.  Other convenience arguments to \code{dfm()} were
also removed, such as \code{select}, \code{dictionary}, \code{thesaurus}, and \code{groups}.  All
of these functions are available elsewhere, e.g. through \code{\link[=dfm_group]{dfm_group()}}.
See \link{quanteda-deprecations}.
}

\examples{
## for a corpus
toks <- data_corpus_inaugural \%>\%
  corpus_subset(Year > 1980) \%>\%
  tokens()
dfm(toks)

# removal options
toks <- tokens(c("a b c", "A B C D")) \%>\%
    tokens_remove("b", padding = TRUE)
toks
dfm(toks)                
dfm(toks, remove = "") # remove "pads"

# preserving case
dfm(toks, tolower = FALSE)
}
\seealso{
\code{\link[=dfm_select]{dfm_select()}}, \linkS4class{dfm}
}
\keyword{dfm}
