% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimental-functions.R
\name{dfm2}
\alias{dfm2}
\title{Experimental C++ indexing dfm}
\usage{
dfm2(x)
}
\arguments{
\item{x}{tokenizedTexts}
}
\description{
Experimental dfm function that takes tokenizedTexts object as an input.  For
testing only.
}
\examples{
\dontrun{## performance comparisons
data(SOTUCorpus, package = "quantedaData")
toks <- tokenize(tokenize(SOTUCorpus, what='sentence', simplify = TRUE), removePunct = TRUE)
microbenchmark::microbenchmark(dfm(toks2),
                               dfm2(toks2),
                               dfm3(toks2),
                               times=1, unit='relative')

toks2 <- rep(toks, 20)
attributes(toks2) <- attributes(toks)
microbenchmark::microbenchmark(dfm(toks2),
                               dfm2(toks2),
                               dfm3(toks2),
                               times=1, unit='relative')
}
}

