% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dfm2ldaformat}
\alias{dfm2ldaformat}
\title{Convert a \link{dfm} into the format needed by \link[pkg]{lda}}
\usage{
dfm2ldaformat(d)
}
\arguments{
\item{d}{A \code{\link{dfm}} object}
}
\value{
A list with components "documents" and "vocab" as needed by
  \link[pkg]{lda.collapsed.gibbs.sampler}
}
\description{
Convert a quanteda \link{dfm} object into the indexed format required by the
topic modelling package \link[pkg]{lda}.
}
\examples{
mycorpus <- subset(inaugCorpus, Year>1970)
d <- dfm(mycorpus, stopwords=TRUE)
d <- trimdfm(d, minCount=5, minDoc=3)
td <- dfm2ldaformat(d)
if (require(lda)) {
    tmodel.lda <- lda.collapsed.gibbs.sampler(documents=td$documents,
                                              K=10,
                                              vocab=td$vocab,
                                              num.iterations=50, alpha=0.1, eta=0.1)
    top.topic.words(tmodel.lda$topics, 10, by.score=TRUE) # top five words in each topic
}
}

