% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{dfm2lsa}
\alias{dfm2lsa}
\title{Convert a dfm to an lsa "textmatrix"}
\usage{
dfm2lsa(x)
}
\arguments{
\item{x}{dfm to be converted}
}
\description{
Converts a dfm to a textmatrix for use with the lsa package.
}
\examples{
\dontrun{
(dfmat <- dfm(c(d1 = "this is a first matrix",
                d2 = "this is second matrix as example")))
lsa::lsa(convert(dfmat, to = "lsa"))
}
}
\keyword{internal}
