% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/textmodel-lda.R
\name{dfm2stmformat}
\alias{dfm2stmformat}
\title{convert a dfm to stm's input document format}
\usage{
dfm2stmformat(data)
}
\arguments{
\item{data}{dfm object to be convereted}
}
\value{
A list containing the following elements:
\item{documents}{A list containing the documents in the stm format.}
\item{vocab}{Character vector of vocabulary.}
\item{meta}{NULL}
}
\description{
Convert a quanteda dfm object into the indexed format needed for estimating
a structural topic model from the \pkg{stm} package using \link[stm]{stm}.
}
\note{
Meta-data will need to be passed separately to \link[stm]{stm} as this
information is not included in a dfm object.
}

