% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfm_conform.R
\name{dfm_conform}
\alias{dfm_conform}
\title{Make features of a dfm identical}
\usage{
dfm_conform(x, features)
}
\arguments{
\item{x}{a dfm}

\item{features}{character vectors of features or a \code{dfm}}
}
\description{
Make features of a dfm identical
}
