% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfm_goodturing.R
\name{dfm_goodturing}
\alias{dfm_goodturing}
\title{Simple Good-Turing Frequency Estimation}
\usage{
dfm_goodturing(x, scheme = c("est_count", "est_prop"))
}
\arguments{
\item{x}{\link{dfm} object for which frequency estimates will be computed}

\item{scheme}{scheme for output; The default is \code{"est_count"}, the estimated
frequency that would have been observed if the sample were perfectly
representative of the population (sometimes known as r*). \code{"est_prop"}
gives the corresponding estimate of the relative frequency, correcting for
unobserved features in the sample.}
}
\description{
Simple Good-Turing Frequency Estimation
}
\details{
\code{dfm_goodturing} computes estimated term frequencies, correcting for
unobserved cases in the sample. This method was pioneered by Alan Turing
and I. J. Good for cracking German ciphers during World War II. The
particular algorithm implemented here was described by Gale & Sampson
(1995). Without this correction, observed token frequencies will
systematically overestimate population frequencies, since there are many
infrequent tokens that have not yet been observed in the sample.
}
\examples{
data_dfm_lbgexample
dfm_goodturing(data_dfm_lbgexample)
dfm_goodturing(data_dfm_lbgexample, scheme = "est_prop")
}
\references{
Gale, W. A., & Sampson, G. (1995). Good-Turing frequency
estimation without tears.
\emph{J. Quant. Linguistics}, 2, 217-237.
\url{https://doi.org/10.1080/09296179508590051}
}
\keyword{dfm}
\keyword{weighting}
