% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textstat_lexdiv.R
\name{dfm_split_hyphenated_features}
\alias{dfm_split_hyphenated_features}
\title{Split a dfm's hyphenated features into constituent parts}
\usage{
dfm_split_hyphenated_features(x)
}
\arguments{
\item{x}{input \link{dfm}}
}
\description{
Takes a dfm that contains features with hyphenated words, such as
"split-second" and turns them into features that split the elements
in the same was as \code{tokens(x, remove_hyphens = TRUE)} would have done.
}
\examples{
(dfmat <- dfm("One-two one two three."))
quanteda:::dfm_split_hyphenated_features(dfmat)
}
\keyword{dfm}
\keyword{internal}
