% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textstat_simil.R
\name{diag2na}
\alias{diag2na}
\title{convert same-value pairs to NA in a textstat_proxy object}
\usage{
diag2na(x)
}
\arguments{
\item{x}{the return from \code{\link[=textstat_simil]{textstat_simil()}} or \code{\link[=textstat_dist]{textstat_dist()}}}
}
\value{
sparse Matrix format with same-pair values replaced with \code{NA}
}
\description{
Converts the diagonal, or the same-pair equivalent in an object
where the columns have been selected, to NA.
}
\keyword{internal}
\keyword{textstat}
