% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textstat_simil.R
\name{diag_to_NA}
\alias{diag_to_NA}
\title{convert same-value pairs to NA in a textstat_simildist object}
\usage{
diag_to_NA(x)
}
\arguments{
\item{x}{the return from \code{\link{texstat_simil}} or \code{\link{texstat_dist}}}
}
\value{
sparse Matrix format with same-pair values replaced with \code{NA}
}
\description{
Converts the diagonal, or the same-pair equivalent in an object 
where the columns have been selected, to NA.
}
\keyword{internal}
\keyword{textstat}
