% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R
\name{dictionary}
\alias{dictionary}
\title{Create a dictionary object}
\usage{
dictionary(
  x,
  file = NULL,
  format = NULL,
  separator = " ",
  tolower = TRUE,
  tokenize = FALSE,
  levels = 1:100,
  encoding = "utf-8"
)
}
\arguments{
\item{x}{a named list of \link{valuetype} patterns or an existing dictionary
object. See examples. This argument should be omitted if \code{file} is speficied.}

\item{file}{file identifier for a foreign dictionary.}

\item{format}{character identifier for the format of the foreign dictionary.
If not supplied, the format is guessed from the dictionary file's
extension. Available options are: \describe{
\item{\code{"wordstat"}}{format used by Provalis Research's WordStat
software} \item{\code{"LIWC"}}{format used by the Linguistic Inquiry and
Word Count software} \item{\code{"yoshikoder"}}{ format used by Yoshikoder
software} \item{\code{"lexicoder"}}{format used by Lexicoder}
\item{\code{"YAML"}}{the standard YAML format}}}

\item{separator}{the character in between multi-word dictionary values. This
defaults to \code{" "}.}

\item{tolower}{if \code{TRUE}, convert all dictionary values to lowercase.}

\item{tokenize}{if \code{TRUE} segment dictionary values by separators to using
the built-in tokenizer. Useful for Japanese and Chinese dictionaries.}

\item{levels}{integers specifying the levels of entries in \code{x} or \code{file} to be
included in the object.}

\item{encoding}{additional optional encoding value for reading in imported
dictionaries. This uses the \link{iconv} labels for encoding.  See the
"Encoding" section of the help for \link{file}.}
}
\value{
A dictionary class object, essentially a specially classed named list
of characters.
}
\description{
Create a \pkg{quanteda} dictionary object to perform pattern matching on \link{tokens},
\link{dfm} and \link{fcm}.
}
\details{
A dictionary object can include multi-word expressions segmented by
\code{separator}. When it is applied to tokens object, they match both sequences
of separate tokens and compounded tokens.

Dictionary objects can be subsetted using
\code{\link[=dictionary2-class]{[}} and
\code{\link[=dictionary2-class]{[[}}, operating the same as the equivalent
\link[=dictionary2-class]{list} operators. If \code{dictionary()} is applied to existing
objects, it is possible to select \code{levels}.

Dictionary objects can be coerced from and to lists using \code{\link[=as.dictionary]{as.dictionary()}}
and \code{\link[=dictionary2-class]{as.list()}}, and checked using \code{\link[=is.dictionary]{is.dictionary()}}.

Currently supported input file formats are
the WordStat, LIWC, Lexicoder v2 and v3, and Yoshikoder formats.  The import
using the LIWC format works with all currently available dictionary files
supplied as part of the LIWC 2001, 2007, and 2015 software (see References).
}
\examples{
corp <- corpus_subset(data_corpus_inaugural, Year > 2000)
toks <- tokens(corp)

dict <- dictionary(list(
   tax = c("tax", "taxes", "taxing"),         # fixed patterns
   economy = list("econom*",                  # glob patterns
                  job = c("work*", "job*")),  # nested keys
   health = c("health care", "public health") # multi-word expressions
))

# compound tokens
tokens_compound(toks, pattern = dict) |>
    dfm() |>
    dfm_select(dict) 
                          
tokens_lookup(toks, dictionary = dict, levels = 1) |>
    dfm()

# subset a dictionary
dict[1:2]
dict[c("economy")]

# update a dictionary
dictionary(dict, levels = 2)

\dontrun{
dfmat <- dfm(tokens(data_corpus_inaugural))

# import the Laver-Garry dictionary from Provalis Research
download.file("https://provalisresearch.com/Download/LaverGarry.zip",
              tf <- tempfile(), mode = "wb")
unzip(tf, exdir = (td <- tempdir()))
dict_lg <- dictionary(file = paste(td, "LaverGarry.cat", sep = "/"))
dfm_lookup(dfmat, dict_lg)

# import a LIWC formatted dictionary from http://www.moralfoundations.org
download.file("http://bit.ly/37cV95h", tf <- tempfile())
dict_liwc <- dictionary(file = tf, format = "LIWC")
dfm_lookup(dfmat, dict_liwc)
}
}
\references{
WordStat dictionaries page, from Provalis Research
\url{https://provalisresearch.com/products/content-analysis-software/wordstat-dictionary/}.

Pennebaker, J.W., Chung, C.K., Ireland, M., Gonzales, A., & Booth, R.J.
(2007). The development and psychometric properties of LIWC2007. [Software
manual]. Austin, TX (\url{https://www.liwc.app/}).

Yoshikoder page, from Will Lowe
\url{https://conjugateprior.org/software/yoshikoder/}.

Lexicoder format, \url{https://www.snsoroka.com/data-lexicoder/}
}
\seealso{
\code{\link[=as.dictionary]{as.dictionary()}},
\code{\link[=dictionary2-class]{as.list()}}, \code{\link[=is.dictionary]{is.dictionary()}}
}
