% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R
\name{dictionary_tokenize}
\alias{dictionary_tokenize}
\title{Separate dictionary values using tokenizer}
\usage{
dictionary_tokenize(dictionary, ...)
}
\arguments{
\item{dictionary}{a \link{dictionary} object.}

\item{...}{passed to \code{\link[=tokens]{tokens()}}.}
}
\description{
Apply \code{\link[=tokens]{tokens()}} to dictionary values to improve pattern matching.
}
