% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimnames.R
\name{dimnames<-.dfm}
\alias{dimnames<-.dfm}
\title{Internal functions to set dimnames. Default dimnames() converts empty character
vectors to NULL, leading to disfunctioning of subsetting functions}
\usage{
\method{dimnames}{dfm}(x) <- value
}
\description{
Internal functions to set dimnames. Default dimnames() converts empty character
vectors to NULL, leading to disfunctioning of subsetting functions
}
