% Generated by roxygen2 (4.0.1): do not edit by hand
\name{directory}
\alias{directory}
\title{Function to declare a connection to a directory (containing files)}
\usage{
directory(path)
}
\arguments{
\item{path}{String describing the full path of the directory}
}
\description{
Function to declare a connection to a directory, although unlike \link{file} it does not require closing.
If the directory does not exist, the function will return an error.
}
\examples{
\dontrun{mydir <- directory("~/Dropbox/QUANTESS/corpora/ukManRenamed")
corpus(mydir)}
}

