% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textstat_dist.R
\docType{class}
\name{dist-class}
\alias{dist-class}
\title{Distance matrix between documents and/or features}
\description{
These functions compute distance matrix between documents and/or features from a 
\code{\link{dfm}} and return a standard \code{\link[stats]{dist}} object.
}
\section{Slots}{

\describe{
\item{\code{selection}}{character or character vector of document names or feature 
labels from the dfm}

\item{\code{n}}{the top \code{n} most similar items will be returned.  If n is \code{NULL}, return all items.}

\item{\code{margin}}{identifies the margin of the dfm on which similarity will be 
computed:  \code{documents} for documents or \code{features} for word/term
features.}

\item{\code{method}}{the distance measure to be used, options are "euclidean", "hamming",
"Chisquared","Chisquared2" and "kullback", default "euclidean". More options are avaible 
in \code{\link{textstat_simil}}}

\item{\code{normalize}}{a deprecated argument retained (temporarily) for legacy 
reasons.  If you want to compute similarity on a "normalized" dfm objects 
(e.g. \code{x}), wrap it in \code{\link{weight}(x, "relFreq")}.}

\item{\code{digits}}{decimal places to display similarity values}

\item{\code{tri}}{whether the upper triangle of the symmetric \eqn{V \times V} matrix is recorded}

\item{\code{diag}}{whether the diagonal of the distance matrix should be recorded}

\item{\code{dist}}{whether the distance matrix should be converted into an object of class "dist". Distance matrix 
created from some methods, such as "kullback", is not symmetric.}
}}
\seealso{
\link{dfm}
}
\keyword{internal}

