% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop0.R
\name{drop0}
\alias{drop0}
\title{Drop small values from an object}
\usage{
drop0(x, ...)
}
\arguments{
\item{x}{object to threshold.}

\item{...}{arguments passed to methods.}
}
\description{
Generic for dropping small values (hard thresholding).
}
