% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocs.R
\docType{data}
\name{encodedTexts}
\alias{encodedTexts}
\title{encoded texts for testing}
\description{
\code{encodedTexts} is a 10-element character vector with 10 different encodings
}
\examples{
Encoding(encodedTexts)
data.frame(labelled = names(encodedTexts), detected = encoding(encodedTexts)$all)
}

