% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dataDocs.R
\docType{data}
\name{englishSyllables}
\alias{englishSyllables}
\title{syllable counts of English words}
\source{
\url{http://www.speech.cs.cmu.edu/cgi-bin/cmudict}
}
\description{
A named numeric vector of syllable counts for the words used as names.
Taken from the freely available CMU pronunciation dictionary.
}
\examples{
data(englishSyllables)
englishSyllables["sixths"]
englishSyllables["onomatopeia"]
}

