% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{escape_regex}
\alias{escape_regex}
\title{Internal function for \code{select_types()} to escape regular expressions}
\usage{
escape_regex(x)
}
\arguments{
\item{x}{character vector to be escaped}
}
\description{
This function escapes glob patterns before \code{utils:glob2rx()}, therefore * and ?
are unescaped.
}
\keyword{internal}
