% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{exampleString}
\alias{exampleString}
\title{exampleString for testing functions}
\format{character vector with one element (nchar 2914)}
\description{
This is a long paragraph (2914 characters) of text taken from an Irish budget speech by Joe Higgins
}

