% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/corpus-sources.R
\name{excel}
\alias{excel}
\title{Function to declare a connection to an excel file}
\usage{
excel(path = NULL, sheetIndex = 1)
}
\arguments{
\item{path}{String describing the full path to the excel file or NULL to use
a GUI to choose a directory from disk}

\item{sheetIndex}{The index of the sheet of the excel file to read (as passed
to read.xlsx2)}
}
\description{
Function to declare a connection to a excel file.
}

