% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex2fixed.R
\name{expand}
\alias{expand}
\title{simpler and faster version of expand.grid() in base package}
\usage{
expand(elem)
}
\arguments{
\item{elem}{list of elements to be combined}
}
\description{
simpler and faster version of expand.grid() in base package
}
\examples{
quanteda:::expand(list(c('a', 'b', 'c'), c('x', 'y')))
}
\keyword{internal}
