% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcm.R
\name{fcm}
\alias{fcm}
\alias{fcm.character}
\alias{fcm.corpus}
\alias{fcm.tokenizedTexts}
\title{create a feature co-occurrence matrix}
\usage{
fcm(x, ...)

\method{fcm}{tokenizedTexts}(x, context = c("document", "window"),
  window = 5L, weights = rep(1, length(window)), spanSentence = TRUE,
  tri = TRUE, ...)

\method{fcm}{corpus}(x, ...)

\method{fcm}{character}(x, ...)
}
\arguments{
\item{x}{character vector, corpus, or tokenized texts from which to generate
the context-feature co-occurrence matrix.}

\item{...}{additional arguments passed to \link{tokenize}, which can include
for instance \code{ngrams} and \code{concatenator} for tokenizing
multi-token sequences}

\item{context}{the context in which to consider term co-occurrence: 
\code{"document"} for co-occurrence counts within document; \code{"window"}
for co-occurrence within a defined window of words, which requires a 
postive integer value for \code{window}}

\item{window}{positive integer value for the size of a window on either side 
of the target feature, default is 5, meaning 5 words before and after the 
target feature}

\item{weights}{a vector of weights applied to each distance from
\code{1:window}, strictly decreasing and of the same length as
\code{length(weights)}}

\item{spanSentence}{if \code{FALSE}, then word windows will not span
sentences}

\item{tri}{if \code{TRUE} return only upper triangle (including diagonal)}
}
\description{
Create a sparse feature co-occurrence matrix within a user-defined context. 
The context can be defined as a document or a window within a collection of
documents, with an optional vector of weights applied to the co-occurrence
counts.
}
\examples{
(txts <- c(paste(letters[c(1, 1:3)], collapse = " "), 
           paste(letters[c(1, 3, 5)], collapse = " "), 
           paste(letters[c(5, 6, 7)], collapse = " ")))
toks <- tokenize(toLower(txts), removePunct = TRUE)
fcm(toks, context = "document")

txt <- c("The quick brown fox jumped over the lazy dog.",
         "The dog jumped and ate the fox.")
toks <- tokenize(toLower(txt), removePunct = TRUE)
fcm(toks, context = "window", window = 3)
fcm(toks, context = "window", window = 2)
}

