% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfm_weight.R
\name{featfreq}
\alias{featfreq}
\title{Compute the frequencies of features}
\usage{
featfreq(x)
}
\arguments{
\item{x}{a \link{dfm}, \link{tokens}, or \link{tokens_xptr} objects.}
}
\value{
a (named) numeric vector of feature frequencies.
}
\description{
Returns a frequency for each feature, computed across all documents.
This is equivalent to \code{colSums(x)} when \code{x} is a dfm.
}
\examples{
dfmat <- dfm(tokens(data_char_sampletext))
featfreq(dfmat)
}
\seealso{
\code{\link[=dfm_tfidf]{dfm_tfidf()}}, \code{\link[=dfm_weight]{dfm_weight()}}
}
\keyword{dfm}
\keyword{weighting}
