% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfm-methods.R
\name{features}
\alias{features}
\title{deprecated function for featnames}
\usage{
features(x)
}
\arguments{
\item{x}{a \link{dfm}}
}
\description{
Deprecated function equivalent to \code{\link{featnames}}.  Soon to be eliminated.
}
\seealso{
\code{\link{featnames}}
}
\keyword{deprecated}
\keyword{internal}
