% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{features2vector}
\alias{features2vector}
\title{convert various input as features to a vector}
\usage{
features2vector(features)
}
\arguments{
\item{features}{the input features, one of: \itemize{ \item{character,
}{whose elements will be split on whitespace;} \item{list of characters,
}{consisting of a list of token patterns, where sequences are separate elements}; 
\item{\link{phrases} object;} \item{\link{dictionary} object}{;} 
\item{\link{dictionary} object}{;} 
\item{\link{collocations} object.}{} }}
}
\value{
an unnamed vector of features
}
\description{
Convert various input as features to a vector for functions that 
do not support multi-word features e.g.
\code{\link{dfm_select}, \link{sequences}, \link{collocations}}.
}
\keyword{internal}
\keyword{utilities}
