% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{field_system}
\alias{field_system}
\alias{field_system<-}
\alias{field_object}
\alias{field_object<-}
\alias{field_user}
\alias{field_user<-}
\title{Internal functions to access or assign metadata}
\usage{
field_system(x, field = NULL)

field_system(x, field = NULL) <- value

field_object(x, field = NULL)

field_object(x, field = NULL) <- value

field_user(x, field = NULL)

field_user(x, field = NULL) <- value
}
\arguments{
\item{x}{an object for which the metadata will be read or set}

\item{field}{metadata field name(s); if \code{NULL} (default), return all
metadata names}

\item{value}{new value of the metadata field}
}
\description{
Internal functions to access or replace an object metadata field
from a list of attributes.
}
\keyword{internal}
