% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textstat-methods.R
\name{filter.textstat}
\alias{filter.textstat}
\title{Return rows with matching conditions including feature matches}
\usage{
\method{filter}{textstat}(.data, ..., pattern = NULL, valuetype = c("glob",
  "regex", "fixed"), case_insensitive = TRUE)
}
\arguments{
\item{.data}{a \code{textstat} object returned from one of the applicable
\code{textstat_*} functions in \pkg{quanteda}}

\item{pattern}{a character vector, list of character vectors, \link{dictionary},
\link{collocations}, or \link{dfm}. See \link{pattern} for details.}

\item{valuetype}{the type of pattern matching: \code{"glob"} for 
"glob"-style wildcard expressions; \code{"regex"} for regular expressions;
or \code{"fixed"} for exact matching. See \link{valuetype} for details.}

\item{case_insensitive}{ignore case when matching, if \code{TRUE}}
}
\description{
Use filters to find rows of the return objects from many \code{textstat_*()}
functions where conditions are true, including matches to features using
\code{"glob"}, \code{"regex"} or \code{"fixed"} patterns.
}
\examples{
period <- ifelse(docvars(data_corpus_inaugural, "Year") < 1945, "pre-war", "post-war")
mydfm <- dfm(data_corpus_inaugural, groups = period)
keyness <- textstat_keyness(mydfm)
filter(keyness, pattern = "america*")
filter(keyness, p < .00001) \%>\% head()
filter(keyness, p < .00001, pattern = "america*") 
}
\seealso{
\code{\link{textstat_collocations}},
\code{\link{textstat_keyness}}, \code{\link{textstat_frequency}}
}
\keyword{internal}
\keyword{textstat}
