% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fitmodel}
\alias{fitmodel}
\alias{fitmodel.dfm}
\title{fit a word model}
\usage{
fitmodel(x, ...)

\method{fitmodel}{dfm}(x, method = c("wordscores", "NB"), ...)
}
\arguments{
\item{x}{a dfm object}

\item{...}{additional parameters to be passed to other functions}

\item{method}{the model type to be fit}
}
\description{
Fit a word model to a dfm.
}
\examples{
require(quantedaData)
data(ie2010Corpus)
ieDfm <- dfm(ie2010Corpus)
ws <- fitmodel(ieDfm, train=c("2010_BUDGET_06_Enda_Kenny_FG", "2010_BUDGET_05_Brian_Cowen_FF"),
               train.scores=c(-1,1), smooth=1)
bs <- fitmodel(ieDfm, train=c("2010_BUDGET_06_Enda_Kenny_FG", "2010_BUDGET_05_Brian_Cowen_FF"),
               train.scores=c(-1,1), scale="logit", smooth=1)
plot(ws$featureScores, bs$featureScores, xlim=c(-1, 1),
     xlab="Linear word score", ylab="Logit word score")

# prediction method for wordscores
predict(ws, ieDfm, rescaling="mv")
}

