% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern2fixed.R
\name{flatten_id}
\alias{flatten_id}
\title{Flatten ids keeping no matches}
\usage{
flatten_id(ids, keep_nomatch = FALSE)
}
\arguments{
\item{ids}{nested list of token ID from \code{pattern2id}}

\item{keep_nomatch}{keep list element for no-matches}
}
\description{
Flatten ids keeping no matches
}
\keyword{internal}
