% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R
\name{flatten_list}
\alias{flatten_list}
\title{Internal function to flatten a nested list}
\usage{
flatten_list(
  lis,
  levels = 1:100,
  level = 1,
  key_parent = "",
  lis_flat = list()
)
}
\arguments{
\item{lis}{a nested list}

\item{levels}{an integer vector indicating levels in the list}

\item{level}{an internal argument to pass current levels}

\item{key_parent}{an internal argument to pass for parent keys}

\item{lis_flat}{an internal argument to pass the flattened list}
}
\description{
Internal function to flatten a nested list
}
\examples{
lis <- list("A" = list("B" = c("b", "B"), c("a", "A", "aa")))
quanteda:::flatten_list(lis, 1:2)
quanteda:::flatten_list(lis, 1)
}
\keyword{internal}
