% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel_nb.R
\name{gain}
\alias{gain}
\title{Compute the information gain of features}
\usage{
gain(object, base = 2)
}
\arguments{
\item{object}{a \link{textmodel_nb} object}

\item{base}{the base of the logarithmic function to be used. Defaults to 2.}
}
\value{
\code{gain} returns a named vector of \eqn{G(w)} values, with the
  feature as the element name.
}
\description{
Computes the information gain of features from a fitted \link{textmodel_nb}
object.
\describe{
  \item{Formula:}{
  \deqn{G(w) = \sum_{i=1}^{C}{P(c_i) \log{P(c_i)}}
  + P(w) \sum_{i=1}^{C}{P(c_i | w) \log{P(c_i | w)}}
  + P(w^{c}) \sum_{i=1}^{C}{P(c_i | w^{c}) \log{P(c_i | w^{c})}}
  }
  where \itemize{
  \item \eqn{G(w)} is the information gain from word \eqn{w},
  \item \eqn{P(c_i)} is the prior probability of class \eqn{i},
  \item \eqn{P(w)} is the probability of word \eqn{w} being present in a
  document ,
  \item \eqn{P(w^c)} is the probability of word \eqn{w} being absent in a
  document,
  \item \eqn{P(c_i | w)} is the posterior probability of class \eqn{i} given
  that word \eqn{w}, and
  \item \eqn{P(c_i | w^c)} is the posterior probability  of class\eqn{i} of
  the complement set from \eqn{w}.
  }
 }
}
}
\examples{
dfmat <- data_corpus_inaugural[c("1865-Lincoln", "2017-Trump")] \%>\% 
    dfm(remove_punct = TRUE) \%>\%
    dfm_remove(stopwords("en")) \%>\%
    dfm_trim(min_docfreq = 2)
tmod <- textmodel_nb(dfmat, y = c("Lincoln", "Trump"))
gw <- gain(tmod)
sort(gw, decreasing = TRUE)
}
\references{
Jurafsky, Daniel and James H. Martin.  (2018).
\href{https://web.stanford.edu/~jurafsky/slp3/4.pdf}{\emph{Speech and
Language Processing}}.   Stanford University typescript.  Draft of September
23.

Yang, Y. and Pedersen, J. O. (1997).
\href{http://www.surdeanu.info/mihai/teaching/ista555-spring15/readings/yang97comparative.pdf}{A
comparative study on feature selection in text categorization.} In
\emph{ICML} 97(412-420, July).
}
\keyword{textmodel}
\keyword{textmodel_nb}
