% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfm_group.R
\name{generate_groups}
\alias{generate_groups}
\title{Generate a grouping vector from docvars}
\usage{
generate_groups(x, groups, drop = FALSE)
}
\arguments{
\item{x}{corpus, tokens or dfm}

\item{groups}{names of docvars or vector that can be coerced to a factor}
}
\value{
a factor
}
\description{
Internal function to generate a grouping vector from docvars used in
dfm.corpus, dfm.tokens, dfm.group, and tokens_group
}
\keyword{internal}
