% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docvars.R
\name{get_docvars}
\alias{get_docvars}
\title{internal function to return all the user docvars}
\usage{
get_docvars(x, field = NULL, drop = FALSE, system = FALSE)
}
\description{
internal function to return all the user docvars
}
