% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_object_version}
\alias{get_object_version}
\title{Get the object version}
\usage{
get_object_version(x)
}
\description{
Return the the \pkg{quanteda} 
package version in which a \link{dfm}, \link{tokens}, or \link{corpus}
object was created.
}
\keyword{internal}
\keyword{utils}
