% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/corpus-sources.R
\name{getFBpage}
\alias{getFBpage}
\title{Extract list of posts from a public Facebook page}
\usage{
getFBpage(page, token, n = 100, since = NULL, until = NULL,
  feed = FALSE)
}
\arguments{
\item{page}{A page ID or page name.}

\item{token}{An access token created at
\url{https://developers.facebook.com/tools/explorer}.}

\item{n}{Number of posts of page to return. Note that number can be sometimes
higher or lower, depending on status of API.}

\item{since}{A UNIX timestamp or strtotime data value that points to
the start of the time range to be searched. For more information on the
accepted values, see: \url{http://php.net/manual/en/function.strtotime.php}}

\item{until}{A UNIX timestamp or strtotime data value that points to
the end of the time range to be searched. For more information on the
accepted values, see: \url{http://php.net/manual/en/function.strtotime.php}}

\item{feed}{If \code{TRUE}, the function will also return posts on the page
that were made by others (not only the admin of the page).}
}
\description{
\code{getPage} retrieves information from a public Facebook page. Note that
information about users that have turned on the "follow" option on their
profile can also be retrieved with this function. See \code{Rfacebook} package
for additional methods to query the Facebook Graph API.
}
\examples{
\dontrun{
# scraping the 100 most recent posts on Barack Obama's page
token <- 'YOUR_FB_TOKEN_HERE'
pg <- getFBpage('barackobama', token, n=100)
# creating corpus object
fbcorpus <- corpus(pg)
summary(fbcorpus)
# viewing the DFM using a word cloud
fbDfm <- dfm(fbcorpus, ignoredFeatures = stopwordsGet())
plot(fbDfm)
}
}
\author{
Pablo Barbera
}

