% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get-texts.R
\name{getRootFileNames}
\alias{getRootFileNames}
\title{Truncate absolute filepaths to root filenames}
\usage{
getRootFileNames(longFilenames)
}
\arguments{
\item{longFilenames}{Absolute filenames including a full path with directory}
}
\value{
character vector of filenames withouth directory path
}
\description{
This function takes an absolute filepath and returns just the
document name
}
\examples{
\dontrun{
getRootFileNames('/home/paul/documents/libdem09.txt')
}
}
\author{
Paul Nulty
}

