% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get-texts.R
\name{getTextDir}
\alias{getTextDir}
\title{loads all text files from a given directory}
\usage{
getTextDir(dirname, pattern = "\\\\.txt$", enc = "unknown")
}
\arguments{
\item{dirname}{A directory path}

\item{pattern}{a \link[=regex]{regular expression} pattern match for the input
file names}

\item{enc}{a value for encoding that is a legal value for \link{Encoding}}
}
\value{
character vector of texts read from disk
}
\description{
given a directory name, get a list of all files in that directory and load
them into a character vector using getTextFiles
}
\examples{
\dontrun{
getTextDir('/home/paul/documents/')
}
}
\author{
Paul Nulty
}

