% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get-texts.R
\name{getTextDirGui}
\alias{getTextDirGui}
\title{provides a gui interface to choose a gui to load texts from}
\usage{
getTextDirGui()
}
\value{
character vector of texts read from disk
}
\description{
launches a GUI to allow the user to choose a directory from
which to load all files.
}
\examples{
\dontrun{
getTextFiles('/home/paul/documents/libdem09.txt')
}
}
\author{
Paul Nulty
}

