% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getTextFiles}
\alias{getTextFiles}
\title{load text files from disk into a vector of character vectors
points to files, reads them into a character vector of the texts
with optional names, default being filenames
returns a named vector of complete, unedited texts}
\usage{
getTextFiles(filenames, textnames = NULL, enc = "unknown",
  verbose = FALSE)
}
\arguments{
\item{filenames}{a vector of paths to text files}

\item{textnames}{names to assign to the texts}

\item{verbose}{If TRUE, print out names of files being read. Default is FALSE}
}
\value{
character vector of texts read from disk
}
\description{
load text files from disk into a vector of character vectors
points to files, reads them into a character vector of the texts
with optional names, default being filenames
returns a named vector of complete, unedited texts
}
\examples{
\dontrun{
getTextFiles('/home/paul/documents/libdem09.txt')
}
}
\author{
Paul Nulty
}

