% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/corpus-sources.R
\name{getTimeline}
\alias{getTimeline}
\title{return a time-line of most recent Tweets from a given user}
\usage{
getTimeline(screen_name, numResults = 200, filename = "default", key,
  cons_secret, token, access_secret, df = TRUE)
}
\arguments{
\item{screen_name}{user name of the Twitter user for which tweets
will be downloaded}

\item{numResults}{number of tweets to be downloaded (maximum is 3,200)}

\item{filename}{file where tweets will be stored (in json format).
If "default", they will be stored in a file whose name is the
screen name of the queried user. If \code{NA} or \code{NULL},
tweets will be stored in a temporary file that will be deleted.}

\item{key}{Key for twitter API authentication}

\item{cons_secret}{for twitter API authentication}

\item{token}{String for twitter API authentication}

\item{access_secret}{for twitter API authentication}

\item{df}{If \code{TRUE}, will return tweets in data frame format.
If \code{FALSE}, will only store tweets in json format in disk.}
}
\description{
Connect to the REST API of Twitter and returns up to
3,200 recent tweets sent by this user.
}
\examples{
\dontrun{
 key = 'your consumer key here'
 cons_secret = 'your consumer secret here'
 token = 'your access token here'
 access_secret = 'your access secret here'

 # download recent tweets by user "p_barbera"
 tweets <- getTimeline(screen_name="p_barbera", numResults=600,
     filename='p_barbera.json', key, cons_secret, token, access_secret)

 # creating corpus object
 twcorpus <- corpus(tweets)
 summary(twcorpus)

 # viewing the DFM using a word cloud
 twDfm <- dfm(twcorpus, gnoredFeatures = c(stopwordsGet(), "rt"))
 plot(twDfm)
}
}
\author{
Pablo Barbera
}

