% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getTweets}
\alias{getTweets}
\title{Function to declare a twitter search}
\usage{
getTweets(query, numResults = 50, key, cons_secret, token, access_secret)
}
\arguments{
\item{query}{String describing the search query terms}

\item{numResults}{Number of tweets to return. Maximum of approximately 1500}

\item{key}{Key for twitter API authentication}

\item{cons_secret}{for twitter API authentication}

\item{token}{String for twitter API authentication}

\item{access_secret}{for twitter API authentication}
}
\value{
The search results marked as a 'twitter' object for use by
  corpus.twitter()
}
\description{
Function to declare a connection to a twitter search
}

