% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getWordStatCSV}
\alias{getWordStatCSV}
\title{Imports a Wordstat corpus from a CSV file}
\usage{
getWordStatCSV(filename = NULL)
}
\arguments{
\item{filename}{Path to wordstat CSV file}
}
\description{
Reads in a wordstat CSV file and creates a corpus object
with the document as text and variables as attributes
}

