% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quanteda-documentation.R
\name{groups}
\alias{groups}
\title{Grouping variable(s) for various functions}
\arguments{
\item{groups}{either: a character vector containing the names of document
variables to be used for grouping; or a factor or object that can be
coerced into a factor equal in length or rows to the number of documents.
\code{NA} values of the grouping value are dropped.
See \link{groups} for details.}

\item{fill}{logical; if \code{TRUE} and \code{groups} is a factor, then use
all levels of the factor when forming the new "documents" of the grouped
object.  This will result in empty documents for levels not
observed.  Has no effect if the \code{groups} variable(s) are not factors.}
}
\description{
Groups for aggregation by various functions that take grouping options.
Groups can be the name(s) of document variables (as a character vector), or
variables whose length or number of rows (if a data.frame) equal the number
of documents.
}
\keyword{internal}
