% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokens_select_pos.R
\name{has_annotation}
\alias{has_annotation}
\title{#' @rdname as.tokens
#' @export
as.tokens.spacyr_parsed <- function(x, concatenator = '_') {
    result <- as.tokens(base::split(x$token, factor(x$doc_id, levels = unique(x$doc_id))))
    pos <- as.tokens(base::split(x$pos, factor(x$doc_id, levels = unique(x$doc_id))))
    attr(result, 'annotation') <- as.annotation(pos, attr(pos, 'types'))
    return(result)
}}
\usage{
has_annotation(x)
}
\description{
#' @rdname as.tokens
#' @export
as.tokens.spacyr_parsed <- function(x, concatenator = '_') {
    result <- as.tokens(base::split(x$token, factor(x$doc_id, levels = unique(x$doc_id))))
    pos <- as.tokens(base::split(x$pos, factor(x$doc_id, levels = unique(x$doc_id))))
    attr(result, 'annotation') <- as.annotation(pos, attr(pos, 'types'))
    return(result)
}
}
\keyword{internal}
