% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hashingTokens.R
\name{hashingTokens}
\alias{as.tokenizedTexts.tokenizedTextsHashed}
\alias{hashingTokens}
\alias{hashingtokens.tokenizedTexts}
\alias{is.tokenizedTextsHashed}
\title{constructor for tokenizedTextsHashed objects}
\usage{
hashingTokens(x, ...)

hashingtokens.tokenizedTexts(x)

as.tokenizedTexts.tokenizedTextsHashed(x)

is.tokenizedTextsHashed(x)
}
\arguments{
\item{x}{a source of tokenizedText}

\item{...}{additional arguments}
}
\description{
Creates a hashed version of tokenizedText.
}
\details{
\code{as.tokenizedTexts} coerces tokenizedTextsHashed to a tokenizedText class object, 
making the methods available for this object type available to this object.
}
\seealso{
\link{tokenize}
}

