% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hashTokens.R
\name{hashTokens}
\alias{as.list.tokenizedTextsHashed}
\alias{as.tokenizedTexts.tokenizedTextsHashed}
\alias{hashTokens}
\alias{hashTokens.tokenizedTexts}
\alias{is.tokenizedTextsHashed}
\title{Constructor for tokenizedTextsHashed objects}
\usage{
hashTokens(x, ...)

\method{hashTokens}{tokenizedTexts}(x, ...)

\method{as.tokenizedTexts}{tokenizedTextsHashed}(x, ...)

\method{as.list}{tokenizedTextsHashed}(x, ...)

is.tokenizedTextsHashed(x)
}
\arguments{
\item{x}{a source of tokenizedText}

\item{...}{additional arguments}
}
\value{
A list the hashed tokens found in each text
}
\description{
Creates a hashed object of tokenizedTexts.
}
\details{
The hashTokens is designed to hash the tokenizedTexts object
and create the tokenizedTextsHashed object for functions like \code{\link{dfm}},
\code{\link{cfm}}, \code{\link{selectFeatures}} etc. to improve the performance
of these functions in terms of speed, especially for large dataset.

\code{as.tokenizedTexts} coerces tokenizedTextsHashed to a tokenizedText class object, 
making the methods available for this object type available to this object.
}
\examples{
txt <- c("The quick brown fox jumped over the lazy dog.",
         "The dog jumped and ate the fox.")
toks <- tokenize(toLower(txt), removePunct = TRUE)
toksHashed <- hashTokens(toks)
# returned as a list
as.list(toksHashed)
# returned as a tokenized Text
as.tokenizedTexts(toksHashed)
}
\author{
Kenneth Benoit and Haiyan Wang
}
\seealso{
\code{\link{tokenize}}
}

