% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfm-print.R
\name{head.dfm}
\alias{head.dfm}
\alias{tail.dfm}
\title{return the first or last part of a dfm}
\usage{
\method{head}{dfm}(x, n = 6L, nfeature = 6L, ...)

\method{tail}{dfm}(x, n = 6L, nfeature = 6L, ...)
}
\arguments{
\item{x}{a dfm object}

\item{n}{a single, positive integer.  If positive, size for the resulting object: 
number of first/last documents for the dfm. If negative, all but the n 
last/first number of documents of x.}

\item{nfeature}{the number of features to return, where the resulting object 
will contain the first \code{ncol} features}

\item{...}{additional arguments passed to other functions}
}
\value{
A \link{dfm} class object corresponding to the subset defined 
  by \code{n} and \code{nfeature}.
}
\description{
For a \link{dfm} object, returns the first or last \code{n} documents 
and first \code{nfeature} features.
}
\examples{
head(data_dfm_lbgexample, 3, nfeature = 5)
head(data_dfm_lbgexample, -4)

tail(data_dfm_lbgexample)
tail(data_dfm_lbgexample, n = 3, nfeature = 4)
}
\keyword{dfm}
