% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textstat_simil.R
\name{head.textstat_proxy}
\alias{head.textstat_proxy}
\alias{tail.textstat_proxy}
\title{Return the first or last part of a textstat_proxy object}
\usage{
\method{head}{textstat_proxy}(x, n = 6L, ...)

\method{tail}{textstat_proxy}(x, n = 6L, ...)
}
\arguments{
\item{x}{a textstat_simil/textstat_dist object}

\item{n}{a single, positive integer.  If positive, size for the resulting
object: number of first/last documents for the dfm. If negative, all but
the n last/first number of documents of x.}

\item{...}{unused}
}
\value{
A \link{matrix} corresponding to the subset defined
by \code{n}.
}
\description{
For a similarity or distance object computed via \link{textstat_simil} or
\link{textstat_dist}, returns the first or last \code{n} rows.
}
\keyword{internal}
\keyword{textstat}
