% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textstat_simil.R
\name{head.textstat_simildist}
\alias{head.textstat_simildist}
\alias{tail.textstat_simildist}
\title{Return the first or last part of a textstat_simildist object}
\usage{
\method{head}{textstat_simildist}(x, n = 6L, ...)

\method{tail}{textstat_simildist}(x, n = 6L, ...)
}
\arguments{
\item{x}{a textstat_simildist object}

\item{n}{a single, positive integer.  If positive, size for the resulting
object: number of first/last documents for the dfm. If negative, all but
the n last/first number of documents of x.}

\item{...}{unused}
}
\value{
A \link{matrix} corresponding to the subset defined 
  by \code{n}.
}
\description{
For a similarity or distance object computed via \link{textstat_simil} or
\link{textstat_dist}, returns the first or last \code{n} rows.
}
\keyword{internal}
\keyword{textstat}
