% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findSequences.R
\name{head.tokenSequences}
\alias{head.tokenSequences}
\title{print a tokenSequences objects}
\usage{
\method{head}{tokenSequences}(x, ...)
}
\arguments{
\item{x}{a tokenSequences object created by \link{findSequences}}

\item{...}{further arguments passed to base print method}
}
\description{
print method for a tokenSequences object
}

