% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocs.R
\docType{data}
\name{ie30corpus}
\alias{ie30corpus}
\title{Irish Dail speeches from 2007-2011}
\format{The corpus object for the 30th Dail speeches, with document-level 
  variables for speaker, debate, and speaker's party.}
\source{
Benjamin E Lauderdale and Alexander Herzog.  2016. "A Scaling Model 
  for Estimating Time-Series Party Positions from Texts." \emph{Political Analysis}.
}
\description{
Speeches and document-level variables from debates in the 30th Dail.
}
\examples{
summary(ie30corpus)
}

