% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{ieAttribs}
\alias{ieAttribs}
\title{A vector of attributes to match ieBudget documents}
\format{a dataframe of attributes, 5 obs. of 6 variables}
\description{
This is a small vector of attributes for use in examples with ieBudgets
}

