\docType{data}
\name{ieTexts}
\alias{ieTexts}
\title{A vector of texts from Irish budget debates for use with examples}
\description{
This is a small vector of texts from the ieBudget corpus
for use with testing examples

This is a small vector of texts from the ieBudget corpus
for use with testing examples
}

