\docType{data}
\name{inaugCorpus}
\alias{inaugCorpus}
\title{A corpus of US inaugural addresses from 1789-2013}
\description{
A \ref{quanteda} corpus of US presidents' inaugural
addresses since 1789. Meta-data contains the year of the
address and the last name of the president.
}
\examples{
data(inaugCorpus)
summary(inaugCorpus)
}
\references{
\url{https://archive.org/details/Inaugural-Address-Corpus-1789-2009}
and \ref{http://www.presidency.ucsb.edu/inaugurals.php}.
}

