% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpusTagged-class.R
\docType{methods}
\name{index}
\alias{index}
\alias{index,corpus-method}
\title{Index (and tag) a corpus into documents, sentences, and tokens}
\usage{
index(x, tagPOS, tagset = c("Penn", "Google"), tagger = c("none", "spaCy"),
  verbose = TRUE)

\S4method{index}{corpus}(x, tagPOS, tagset = c("Penn", "Google"),
  tagger = c("none", "spaCy"), verbose = TRUE)
}
\value{
An S4 object of class \link{corpusTokenized} that inherits from class \link{corpus}.
}
\description{
Implements indexing and tagging for corpus objects.  This can include
applying a part-of-speech tagger, if one is installed and available.
Indexing corpus objects can be time-consuming, but means that subsequent operations such as tokenizing, 
grouping, or resampling are performed extremely efficiently.
}
\examples{
oldCorpus <- corpus(c("This, is document 1.  It's Bill's\\nsecond sentence.", 
                      "Isn't this\\tthe 2nd document?  Document\\rtwo, second sentence!", 
                      "And: voila!"))
newCorpus <- index(oldCorpus, verbose = TRUE)
\dontrun{# with part of speech tagging
newCorpusTagged <- index(oldCorpus, tagger = "spaCy")
}
}

