% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex2fixed.R
\name{index_types}
\alias{index_types}
\title{index types for fastest regular expression matches}
\usage{
index_types(types, valuetype, case_insensitive, max_len = NULL)
}
\arguments{
\item{valuetype}{the type of pattern matching: \code{"glob"} for 
"glob"-style wildcard expressions; \code{"regex"} for regular expressions;
or \code{"fixed"} for exact matching. See \link{valuetype} for details.}

\item{case_insensitive}{ignore case when matching, if \code{TRUE}}

\item{max_len}{maximum length of types to be indexed}
}
\value{
a list of integer vectors containing type IDs with index keys as an
  attribute
}
\description{
An internal function for \code{\link{regex2id}} that constructs an index of
regex patterns (e.g. \code{^xxxx}, \code{xxxx$} and \code{^xxxx$}) to avoid
expensive sequential search by \link[stringi]{stri_detect_regex}.
}
\examples{
types <- c('xxx', 'yyyy', 'ZZZ')
index <- quanteda:::index_types(types, 'regex', FALSE, 3)
type_id <- quanteda:::search_index('^y', index)
types[type_id]
}
\keyword{internal}
