% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message.R
\name{inflect}
\alias{inflect}
\title{Return inflected forms of words}
\usage{
inflect(word, n)
}
\arguments{
\item{word}{singular form of the word.}

\item{n}{the number of elements.}
}
\description{
Return inflected forms of words
}
\keyword{development}
\keyword{internal}
