% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel_affinity.R
\name{influence.predict.textmodel_affinity}
\alias{influence.predict.textmodel_affinity}
\title{Compute feature influence from a predicted textmodel_affinity object}
\usage{
\method{influence}{predict.textmodel_affinity}(model, subset = !train,
  ...)
}
\arguments{
\item{model}{a predicted
\link[=predict.textmodel_affinity]{textmodel_affinity} object}

\item{subset}{whether to use all data or a subset (for instance, exclude the
training set)}

\item{...}{unused}
}
\description{
Computes the influence of features on scaled \link{textmodel_affinity}
applications.
}
\examples{
tmot <- textmodel_affinity(data_dfm_lbgexample, y = c("L", NA, NA, NA, "R", NA))
pred <- predict(tmot) 
influence(pred)
}
\seealso{
\code{\link{influence.lm}}
}
\keyword{internal}
\keyword{textmodel}
