% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel_affinity.R
\name{influence.textmodel_affinity_predicted}
\alias{influence.textmodel_affinity_predicted}
\title{Compute feature influence from a predicted textmodel_affinity object}
\usage{
\method{influence}{textmodel_affinity_predicted}(model, subset = !train, ...)
}
\arguments{
\item{model}{a fitted \link{textmodel_affinity} object}

\item{subset}{whether to use all data or a subset (for instance, exclude the
training set)}

\item{...}{unused}
}
\description{
Computes the influence of features on scaled \link{textmodel_affinity}
applications.
}
\seealso{
\code{\link{influence.lm}}
}
\keyword{textmodel_affinity}
