% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message.R
\name{info}
\alias{info}
\title{Conditionally format messages}
\usage{
info(x, values, indices = NULL)
}
\arguments{
\item{x}{message template to be passed to \code{\link[=sprintf]{sprintf()}}.}

\item{values}{list of values to be used in the template. Coerced to list if vector is given.}

\item{indices}{list of integer to specify which value to be used.}
}
\description{
Conditionally format messages
}
\examples{
\dontrun{
quanteda:::info("you cannot delete \%s", 
                c("a document", "documents"), indices = TRUE)
quanteda:::info("tokens has \%s", 
                c("sentences", "paragraphs", "documents"), indices = 2)

dfmat <- data_dfm_lbgexample
quanteda:::info("dfm has \%d \%s and \%d \%s", 
     list(ndoc(dfmat), c("document", "documents"),
          nfeat(dfmat), c("feature", "features")), 
     list(1, ndoc(dfmat) > 1, 
          1, nfeat(dfmat) > 1))
}      
}
