% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokens_segment.R
\name{insert_values}
\alias{insert_values}
\title{Insert integers into a vector at defined intervals}
\usage{
insert_values(x, n, val, truncate = TRUE)
}
\arguments{
\item{x}{a numeric vector}

\item{n}{the length of the partitions of \code{x} to be divided by \code{val}}

\item{val}{the numeric value to be inserted}

\item{truncate}{logical; if \code{TRUE}, discard any remaining sequence less
  than \code{n} at the end of the chunked tokens.
max(unlist(unclass(x)))}
}
\description{
Inserts \code{val} into a numeric vector \code{x} every \code{n} values.
}
