% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collocations.R, R/textstat_collocations.R
\name{is.collocations}
\alias{is.collocations}
\alias{is.collocations}
\title{check if an object is collocations type}
\usage{
is.collocations(x)

is.collocations(x)
}
\arguments{
\item{x}{any object}
}
\value{
\code{is.collocation} returns \code{TRUE} if the object is of class
  collocation, \code{FALSE} otherwise.
}
\description{
Return \code{TRUE} if an object was constructed by \link{collocations}.
}
\keyword{collocations}
