% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfm-methods.R
\name{is.dfm}
\alias{is.dfm}
\alias{as.dfm}
\title{coercion and checking functions for dfm objects}
\usage{
is.dfm(x)

as.dfm(x)
}
\arguments{
\item{x}{a \link{dfm} object}
}
\value{
\code{is.dfm} returns \code{TRUE} if and only if its argument is a \link{dfm}.

\code{as.dfm} coerces a matrix or data.frame to a dfm
}
\description{
Check for a dfm, or convert
a matrix into a dfm.
}
\seealso{
\code{\link{as.data.frame.dfm}}, \code{\link{as.matrix.dfm}}
}
