% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R
\name{is.dictionary}
\alias{is.dictionary}
\title{check if an object is a dictionary}
\usage{
is.dictionary(x)
}
\arguments{
\item{x}{any object}
}
\description{
Return \code{TRUE} if an object is a \pkg{quanteda} \link{dictionary}.
}

