% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern2fixed.R
\name{is_glob}
\alias{is_glob}
\title{Check if patterns contains glob wildcard}
\usage{
is_glob(pattern)
}
\arguments{
\item{pattern}{a glob pattern to be tested}
}
\description{
Check if patterns contains glob wildcard
}
\keyword{internal}
