% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex2fixed.R
\name{is_indexed}
\alias{is_indexed}
\title{internal function for select_types() to check if regex a pattern is indexed by /code{index_types()}}
\usage{
is_indexed(x)
}
\arguments{
\item{x}{a regex pattern to be tested}
}
\description{
internal function for select_types() to check if regex a pattern is indexed by /code{index_types()}
}
\keyword{internal}
