% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinTokens2.R
\name{joinTokens2}
\alias{joinTokens2}
\title{join tokens function}
\usage{
joinTokens2(x, ...)
}
\arguments{
\item{x}{tokens or tokenizedTexts object}

\item{...}{additional arguments passed to other methods}
}
\description{
For a set of tokens, given a set of token sequences, join the tokens matching the sequences.
}

