% Generated by roxygen2 (4.0.2): do not edit by hand
\name{json}
\alias{json}
\title{Function to read files with tweets in JSON format}
\usage{
json(path = NULL, source = "twitter", ...)
}
\arguments{
\item{path}{String describing the full path of a directory that
 contains files in json format, or a vector of file names in
 in json format}

\item{source}{Source of data in JSON format.}
}
\description{
Function to read files with tweets in JSON format
}
\examples{
\dontrun{
# name a directory of files in json format
tweets <- json("~/Dropbox/QUANTESS/corpora/tweets")
corpus(tweets)

# read a single file in json format
tweets <- json("~/Dropbox/QUANTESS/corpora/tweets/BarackObama.json")
corpus(tweets)
}
}

