% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textstat_keyness.R
\name{keyness}
\alias{keyness}
\alias{keyness_chi2_dt}
\alias{keyness_chi2_stats}
\alias{keyness_exact}
\title{compute keyness (internal functions)}
\usage{
keyness_chi2_dt(x)

keyness_chi2_stats(x)

keyness_exact(x)
}
\value{
named numeric vector containing the chi2 value for each feature
}
\description{
Internal function used in textstat_keyness.  Computes Chi^2 with Yates'
continuity correction for 2x2 tables.
}
\details{
\code{keyness_chi2_dt} uses vectorized computation from data.table
objects.

\code{keyness_chi2_stats} uses element-by-element application of
\link[stats]{chisq.test}.

\code{keyness_exact} computes Fisher's exact test
}
\examples{
mydfm <- dfm(c(d1 = "a a a b b c c c c c c d e f g h h",
               d2 = "a a b c c d d d d e f h"))
quanteda:::keyness_chi2_dt(mydfm)
quanteda:::keyness_chi2_stats(mydfm)
quanteda:::keyness_exact(mydfm)
}
\references{
\url{https://en.wikipedia.org/wiki/Yates's_correction_for_continuity}
}
\keyword{internal}
\keyword{textstat}

