% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kwic2.R
\name{kwic_split}
\alias{kwic_split}
\title{split kwic results}
\usage{
kwic_split(char, mask, window)
}
\arguments{
\item{char}{Kohei to explain}

\item{mask}{Kohei to explain}

\item{window}{Kohei to explain}
}
\description{
Helper function for kwic whose purpose KW will explain.
}
\examples{
kwic_split(letters[1:5], c(0, 1, 1, 0, 1), window = 1)
kwic_split(letters[1:5], c(1, 1, 1, 0, 1), window = 1)
}
\keyword{internal}

