% Generated by roxygen2 (4.0.2): do not edit by hand
\name{language}
\alias{language}
\alias{language<-}
\title{get or set the language of corpus documents}
\usage{
language(corp, drop = TRUE)

language(corp) <- value
}
\arguments{
\item{corp}{a corpus object}

\item{drop}{return as a vector if \code{TRUE}, otherwise return a \code{data.frame}}

\item{value}{the new value for the language meta-data field, a string or
character vector equal in length to \code{ndoc(corp)}}
}
\description{
Get or set the \code{_language} document-level metadata field in a corpus.
}
\details{
This function modifies the \code{_language} value set by
  \code{\link{metadoc}}.  It is a wrapper for \code{metadoc(corp, "language")}.
}

