% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{lbg}
\alias{lbg}
\title{Example data from Laver Benoit and Garry (2003)}
\description{
Example data to demonstrate the Wordscores algorithm, from Laver Benoit and Garry (2003)
}
\details{
his is the example word count data from Laver,
 Benoit and Garry's (2000) article
 on Wordscores.  Documents R1 to R5 are assumed to have known
 positions: -1.5, -0.75, 0, 0.75, 1.5.  Document V1 is assumed
 unknown, and will have a raw text score of approximately -0.45 when computed
 as per LBG (2003).

 \link{lbg} is a document-feature matrix.
}
\examples{
data(inaugCorpus)
summary(inaugCorpus)
}
\references{
Laver, Benoit and Garry (2003) `Estimating policy positions from
political text using words as data' American Political Science Review 97(2).
}

