% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataDocs.R
\docType{data}
\name{LBGexample}
\alias{LBGexample}
\title{dfm containing example data from Table 1 of Laver Benoit and Garry
  (2003)}
\format{A \link{dfm} object with 6 documents and 37 features}
\description{
Example data to demonstrate the Wordscores algorithm, from Laver
  Benoit and Garry (2003)
}
\details{
This is the example word count data from Laver, Benoit and Garry's
  (2003) Table 1. Documents R1 to R5 are assumed to have known positions:
  -1.5, -0.75, 0, 0.75, 1.5.  Document V1 is assumed unknown, and will have a
  raw text score of approximately -0.45 when computed as per LBG (2003).
}
\references{
Laver, Michael, Kenneth Benoit, and John Garry.  2003.
  "\href{http://www.kenbenoit.net/pdfs/WORDSCORESAPSR.pdf}{Estimating policy
  positions from political text using words as data.}" American Political
  Science Review 97(2): 311-331.
}

