% Generated by roxygen2 (4.0.2): do not edit by hand
\name{likelihood.test}
\alias{likelihood.test}
\title{likelihood test for contingency tables}
\usage{
likelihood.test(x)
}
\arguments{
\item{x}{a contingency table or matrix object}
}
\value{
A list of return values
}
\description{
returns a list of values
}
\author{
Kenneth Benoit
}

