% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R
\name{list2dictionary}
\alias{list2dictionary}
\title{Internal function to convert a list to a dictionary}
\usage{
list2dictionary(dict)
}
\arguments{
\item{dict}{list of object}
}
\description{
A dictionary is internally a list of list to keys and values to coexist in
the same level.
}
\keyword{internal}
