% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R
\name{lowercase_dictionary_values}
\alias{lowercase_dictionary_values}
\title{Internal function to lowercase dictionary values}
\usage{
lowercase_dictionary_values(dict)
}
\arguments{
\item{dict}{the dictionary whose values will be lowercased}
}
\description{
Internal function to lowercase dictionary values
}
\examples{
dict <- list(KEY1 = list(SUBKEY1 = c("A", "B"),
                          SUBKEY2 = c("C", "D")),
              KEY2 = list(SUBKEY3 = c("E", "F"),
                          SUBKEY4 = c("G", "F", "I")),
              KEY3 = list(SUBKEY5 = list(SUBKEY7 = c("J", "K")),
                          SUBKEY6 = list(SUBKEY8 = c("L"))))
quanteda:::lowercase_dictionary_values(dict)
}
\keyword{dictionary}
\keyword{internal}
