% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docvars.R
\name{make_docvars}
\alias{make_docvars}
\title{Internal function to create docvars}
\usage{
make_docvars(n, docname = NULL, unique = TRUE)
}
\arguments{
\item{n}{the number of documents}

\item{docname}{a character vector for the names of documents. Must be the
same length as \code{n} or \code{NULL}. If NULL, names are generated automatically.}

\item{unique}{if \code{TRUE}, names must be all unique. If \code{FALSE}, documents with the same
names are treated as segments from the same document and given serial number.}
}
\description{
Internal function to create docvars
}
\keyword{internal}
