% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfm-methods.R
\name{matrix2dfm}
\alias{matrix2dfm}
\title{Conversts a Matrix to a dfm}
\usage{
matrix2dfm(x, slots = NULL)
}
\arguments{
\item{x}{a Matrix}

\item{slots}{slots a list of values to be assigned to slots}
}
\description{
Conversts a Matrix to a dfm
}
\keyword{internal}
