% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcm-methods.R
\name{matrix2fcm}
\alias{matrix2fcm}
\title{Converts a Matrix to a fcm}
\usage{
matrix2fcm(x, slots = NULL)
}
\arguments{
\item{x}{a Matrix}

\item{slots}{slots a list of values to be assigned to slots}
}
\description{
Converts a Matrix to a fcm
}
\keyword{internal}
