% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R
\name{merge_dictionary_values}
\alias{merge_dictionary_values}
\title{Internal function to merge values of duplicated keys}
\usage{
merge_dictionary_values(dict)
}
\arguments{
\item{dict}{a dictionary object}
}
\description{
Internal function to merge values of duplicated keys
}
\examples{
dict <- list("A" = list(AA = list("aaaaa"), "a"),
             "B" = list("b"),
             "C" = list("c"),
             "A" = list("aa"))
quanteda:::merge_dictionary_values(dict)
}
\keyword{internal}
