% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message.R
\name{message_corpus}
\alias{message_corpus}
\title{Print messages in corpus methods}
\usage{
message_corpus(operation, stat)
}
\arguments{
\item{stat}{object statistics after the operation.}
}
\description{
Print messages in corpus methods
}
\keyword{internal}
\keyword{message}
