% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{message_error}
\alias{message_error}
\title{Return an error message}
\usage{
message_error(key = NULL)
}
\arguments{
\item{key}{type of error message}
}
\description{
Return an error message
}
\keyword{internal}
