% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message.R
\name{message_fcm}
\alias{message_fcm}
\title{Print messages in fcm methods}
\usage{
message_fcm(operation, before, after)
}
\arguments{
\item{before, after}{object statistics before and after the operation.}
}
\description{
Print messages in fcm methods
}
\keyword{internal}
\keyword{message}
