% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message.R
\name{messages}
\alias{messages}
\title{Message parameter documentation}
\arguments{
\item{verbose}{if \code{TRUE} print a summary of the object returned by the
function: the number of documents and, depending on the object, the number
of characters, tokens and types, or features and occurrences. The number of
tokens includes paddings, but the number of features excludes them.}

\item{stat}{object statistics after the operation.}
}
\description{
Used in printing verbose messages for message_corpus(), message_tokens(),
message_dfm() and message_fcm().
}
\seealso{
\code{\link[=message_corpus]{message_corpus()}}, \code{\link[=message_tokens]{message_tokens()}}, \code{\link[=message_dfm]{message_dfm()}},
\code{\link[=message_fcm]{message_fcm()}}
}
\keyword{internal}
