% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{meta}
\alias{meta}
\alias{metacorpus}
\alias{meta<-}
\alias{"metacorpus<-"}
\alias{metacorpus<-}
\title{Get or set object metadata}
\usage{
meta(x, field = NULL, type = c("user", "system", "all"))

meta(x, field = NULL) <- value

metacorpus(x, field = NULL, type = c("user", "system", "all"))

metacorpus(x, field = NULL) <- value
}
\arguments{
\item{x}{an object for which the metadata will be read or set}

\item{field}{metadata field name(s); if \code{NULL} (default), return all
metadata names}

\item{type}{\code{"user"} for user-supplied metadata;
\code{"system"} for metadata set automatically when the object is created;
or \code{"all"} for all metadata.}

\item{value}{new value of the metadata field}
}
\value{
For \code{meta}, a named list of the metadata fields.

For \verb{meta <-}, the object with the updated user-level metadata.  Only
user-level metadata may be assigned.
}
\description{
Get or set metadata in a \link{corpus}, \link{tokens}, \link{dfm}, or \link{dictionary} object.
}
\details{
\code{metacorpus} and \verb{metacorpus<-} are synonyms for \code{meta()} and \verb{meta()<-} for
\link{corpus} objects, but are deprecated.
}
\examples{
meta(data_corpus_inaugural)
meta(data_corpus_inaugural, "source")
meta(data_corpus_inaugural, "citation") <- "Presidential Speeches Online Project (2014)."
meta(data_corpus_inaugural, "citation")

dict <- dictionary(list(neg = c("bad", "awful"), pos = "good"))
meta(dict, "source") <- "The example from meta()."
meta(dict)
}
\keyword{meta}
