% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{meta_init}
\alias{meta_init}
\alias{meta_init.corpus}
\alias{meta_init.tokens}
\title{Internal function to initialize meta-data}
\usage{
meta_init(x, source, user = list())

\method{meta_init}{corpus}(x, source, user = list())

\method{meta_init}{tokens}(x, source, user = list())
}
\arguments{
\item{x}{input object for which the metadata will be set}

\item{source}{character; the input object class}

\item{list;}{key-value set of user metadata}
}
\description{
Initializes meta-data for new objects.  The "system" element of this list is
for automatically set system data; the "user" element of this list is set by
the user and replaces the \link{metacorpus} function.
}
\keyword{corpus}
\keyword{internal}
