% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{meta_system<-}
\alias{meta_system<-}
\title{Internal function to set or initialize system meta-data}
\usage{
meta_system(x, field = NULL, initialize = TRUE) <- value
}
\arguments{
\item{x}{input object for which the metadata will be set}

\item{source}{character; the input object class}

\item{list;}{key-value set of user metadata}
}
\description{
Sets or initializes system meta-data for new objects.
}
\keyword{internal}
