% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocs.R
\docType{data}
\name{mobydickText}
\alias{mobydickText}
\title{Project Gutenberg text of Herman Melville's \emph{Moby Dick}}
\source{
Project Gutenberg, \url{http://www.gutenberg.org}
}
\description{
Named character object of the ASCII text of Herman Melville's
  \emph{Moby Dick}, EBook no. 2701.
}
\examples{
summary(mobydickText)
}

