% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quanteda-documentation.R
\name{modify_if}
\alias{modify_if}
\title{Modify only documents matching a logical condition}
\arguments{
\item{modify_if}{logical vector of length \code{ndoc(x)}; documents are modified
only when corresponding values are \code{TRUE}, others are left unchanged.}
}
\description{
Applies the modification only to documents matching a condition.
}
\keyword{internal}
