% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{movies}
\alias{movies}
\title{A corpus object containing 2000 movie reviews}
\description{
A corpus object containing 2000 movie reviews classified by positive or negative sentiment
}
\references{
\url{http://dl.acm.org/citation.cfm?id=1118704}
}

