% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message.R
\name{msg}
\alias{msg}
\title{Conditionally format messages}
\usage{
msg(format, ..., prepend = "", append = "")
}
\arguments{
\item{format}{character vector of format strings}

\item{...}{vectors (coercible to integer, real, or character)}

\item{prepend}{text is added before the message}

\item{append}{text is added after the message}
}
\description{
Conditionally format messages
}
\examples{
quanteda:::msg("you cannot delete \%s \%s", 2000, "documents")
}
\seealso{
\link[stringi:stri_sprintf]{stringi::stri_sprintf}
}
\keyword{development}
\keyword{internal}
