% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docnames.R
\name{names.corpus}
\alias{names.corpus}
\alias{names<-.corpus}
\alias{names<-.tokens}
\alias{rownames<-.dfm}
\title{Special handling for names of quanteda objects}
\usage{
\method{names}{corpus}(x) <- value

names.tokens(x) <- value

rownames.dfm(x) <- value
}
\arguments{
\item{x}{an \R object.}

\item{value}{a character vector of up to the same length as \code{x}, or
    \code{NULL}.}
}
\description{
Keeps the element names and rownames in sync with the system docvar
\code{docname_}.
}
\keyword{corpus}
\keyword{internal}
