% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfunctions.R
\name{ndocument}
\alias{ndocument}
\title{count the number of documents or features}
\usage{
ndocument(x)
}
\arguments{
\item{x}{a \pkg{quanteda} object: a \link{corpus}, \link{dfm}, or
\link{tokens} object, or a readtext object from the \pkg{readtext} package.}
}
\value{
an integer (count) of the number of documents or features
}
\description{
Get the number of documents or features in an object.
}
\details{
\code{ndocument} returns the number of documents in a  \link{corpus},
  \link{dfm}, or \link{tokens} object, or a readtext object from the
  \pkg{readtext} package
  
  \code{nfeature} returns the number of features in a \link{dfm}
}
\examples{
# number of documents
ndoc(data_corpus_inaugural)
ndoc(corpus_subset(data_corpus_inaugural, Year > 1980))
ndoc(tokens(data_corpus_inaugural))
ndoc(dfm(corpus_subset(data_corpus_inaugural, Year > 1980)))

}
