% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R
\name{nest_dictionary}
\alias{nest_dictionary}
\title{Utility function to generate a nested list}
\usage{
nest_dictionary(dict, depth)
}
\arguments{
\item{dict}{a flat dictionary}

\item{depth}{depths of nested element}
}
\description{
Utility function to generate a nested list
}
\examples{
lis <- list("A" = c("a", "aa", "aaa"), "B" = c("b", "bb"), "C" = c("c", "cc"), "D" = c("ddd"))
dict <- quanteda:::list2dictionary(lis)
quanteda:::nest_dictionary(dict, c(1, 1, 2, 2))
quanteda:::nest_dictionary(dict, c(1, 2, 1, 2))

}
\keyword{internal}
