% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfunctions.R
\name{nfeature}
\alias{nfeature}
\title{Defunct form of nfeat}
\usage{
nfeature(x)
}
\description{
Defunct form of nfeat
}
\details{
\code{nfeature} is the defunct form of \code{\link{nfeat}}.
}
\keyword{defunct}
\keyword{internal}
