% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/resample.R
\name{nresample}
\alias{nresample}
\alias{nresample.corpus}
\alias{nresample.dfm}
\title{get the number of resamples}
\usage{
nresample(x)

\method{nresample}{corpus}(x)

\method{nresample}{dfm}(x)
}
\arguments{
\item{x}{corpus object containing the texts to be resampled}
}
\value{
an integer as the number of resampled texts
}
\description{
Get the number of resamples from a corpus or dfm object
}

