% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsyllable.R
\name{nsyllable}
\alias{nsyllable}
\title{count syllables in a text}
\usage{
nsyllable(x, syllable_dictionary = quanteda::data_int_syllables, ...)
}
\arguments{
\item{x}{character vector or \code{tokens} object  whose 
syllables will be counted}

\item{syllable_dictionary}{optional named integer vector of syllable counts where 
the names are lower case tokens.  When set to \code{NULL} (default), then 
the function will use the quanteda data object \code{data_int_syllables}, an 
English pronunciation dictionary from CMU.}

\item{...}{additional arguments passed to \code{\link{tokens}}}
}
\value{
If \code{x} is a character vector, a named numeric vector of the 
  counts of the syllables in each text, without tokenization.  If \code{x} 
  consists of (a list of) tokenized texts, then return a list of syllable 
  counts corresponding to the tokenized texts.
}
\description{
Returns a count of the number of syllables in texts. For English
  words, the syllable count is exact and looked up from the CMU pronunciation
  dictionary, from the default syllable dictionary \code{data_int_syllables}. 
  For any word not in the dictionary, the syllable count is estimated by 
  counting vowel clusters.
  
  \code{data_int_syllables} is a quanteda-supplied data object consisting of a 
  named numeric vector of syllable counts for the words used as names.  This 
  is the default object used to count English syllables.  This object that 
  can be accessed directly, but we strongly encourage you to access it only 
  through the \code{nsyllable()} wrapper function.
}
\note{
All tokens are automatically converted to lowercase to perform the
  matching with the syllable dictionary, so there is no need to perform this
  step prior to calling \code{nsyllable()}.
}
\examples{
nsyllable("This is an example sentence.")
nsyllable(tokenize("This is an example sentence.", simplify=TRUE))
myTexts <- c(text1 = "Text one.", 
             text2 = "Superduper text number two.", 
             text3 = "One more for the road.")
nsyllable(myTexts)
nsyllable(tokenize(myTexts, removePunct = TRUE))
nsyllable("supercalifragilisticexpialidocious")
}

