% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/corpus.R
\name{ntoken}
\alias{ntoken}
\alias{ntoken.character}
\alias{ntoken.corpus}
\title{count the number of tokens}
\usage{
ntoken(x, block.size = 200, verbose = TRUE)

\method{ntoken}{corpus}(x, block.size = 200, verbose = TRUE)

\method{ntoken}{character}(x, block.size = 200, verbose = TRUE)
}
\arguments{
\item{x}{texts or corpus whose tokens will be counted}

\item{block.size}{how many texts to process at a time; experimentation
indicates that for bery large collections of texts, 200 seems fastest}

\item{verbose}{if \code{TRUE} print progress indicator and time elapsed}
}
\value{
scalar count of the total tokens
}
\description{
Return the count of tokens in a text or corpus.  "tokens" here
means all words, not unique words, and these are not cleaned
prior to counting.
}
\examples{
ntokens(inaugTexts, verbose=FALSE)
ntokens(inaugCorpus, verbose=FALSE)
if (require(quantedaData)) {
    data(ukManifestosCorpus, package="quantedaData")
    ntoken(ukManifestosCorpus, block.size=5)
}
}

