% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object-builder.R
\name{build_dfm}
\alias{build_dfm}
\alias{rebuild_dfm}
\alias{build_tokens}
\alias{rebuild_tokens}
\alias{build_corpus}
\alias{rebuild_corpus}
\alias{build_dictionary2}
\alias{rebuild_dictionary2}
\title{Object compilers}
\usage{
build_dfm(x, features, docvars = data.frame(), meta = list(), ...)

rebuild_dfm(x, attrs)

build_tokens(
  x,
  types,
  padding = FALSE,
  docvars = data.frame(),
  meta = list(),
  ...
)

rebuild_tokens(x, attrs)

build_corpus(x, docvars = data.frame(), meta = list(), ...)

rebuild_corpus(x, attrs)

build_dictionary2(x, meta = list(), ...)

rebuild_dictionary2(x, attrs)
}
\arguments{
\item{docvars}{data.frame for document level variables}

\item{meta}{list for meta fields}

\item{...}{added to object meta fields}

\item{attrs}{a list of attributes to be reassigned}

\item{types}{character for types of resulting `tokens`` object}

\item{padding}{logical indicating if the \code{tokens} object contains paddings}

\item{source}{character that indicating source object}

\item{feature}{character for feature of resulting \code{dfm}}
}
\description{
Object compilers
}
\keyword{internal}
