% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object-compiler.R
\name{compile_dfm}
\alias{compile_dfm}
\alias{compile_tokens}
\alias{compile_corpus}
\title{Object compilers}
\usage{
compile_dfm(x, source, features, docvars = data.frame(), meta = list(), ...)

compile_tokens(
  x,
  source,
  types,
  padding = FALSE,
  docvars = data.frame(),
  meta = list(),
  ...
)

compile_corpus(x, source, docvars = data.frame(), meta = list(), ...)
}
\arguments{
\item{source}{character that indicating source object}

\item{docvars}{data.frame for docment level variables}

\item{meta}{list for meta fields}

\item{...}{added to object meta fields}

\item{types}{character for types of resulting `tokens`` object}

\item{padding}{logical indicating if the \code{tokens} object contains paddings}

\item{feature}{character for feature of resulting \code{dfm}}
}
\description{
Object compilers
}
\keyword{internal}
