% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object2fixed.R
\name{object2id}
\alias{object2id}
\alias{object2fixed}
\title{Convert quanteda's objects (list, dictionary, collocations) to type IDs or
fixed patterns.}
\usage{
object2id(
  x,
  types,
  valuetype = c("glob", "fixed", "regex"),
  case_insensitive = TRUE,
  concatenator = "_",
  levels = 1,
  remove_unigram = FALSE,
  keep_nomatch = FALSE
)

object2fixed(
  x,
  types,
  valuetype = c("glob", "fixed", "regex"),
  case_insensitive = TRUE,
  concatenator = "_",
  levels = 1,
  remove_unigram = FALSE,
  keep_nomatch = FALSE
)
}
\arguments{
\item{valuetype}{the type of pattern matching: \code{"glob"} for "glob"-style
wildcard expressions; \code{"regex"} for regular expressions; or \code{"fixed"} for
exact matching. See \link{valuetype} for details.}

\item{case_insensitive}{logical; if \code{TRUE}, ignore case when matching a
\code{pattern} or \link{dictionary} values}

\item{concatenator}{concatenator that join multi-word expression in tokens
object}

\item{levels}{only used when pattern is a dictionary}

\item{remove_unigram}{ignore single-word patterns if \code{TRUE}}
}
\description{
Convert quanteda's objects (list, dictionary, collocations) to type IDs or
fixed patterns.
}
\examples{
types <- c("A", "AA", "B", "BB", "B_B", "C", "C-C")

dict <- dictionary(list(A = c("a", "aa"), 
                        B = c("BB", "B B"),
                        C = c("C", "C-C")))
object2fixed(dict, types)
object2fixed(dict, types, remove_unigram = TRUE)

pat <- phrase(c("a", "aa", "zz", "bb", "b b"))
object2fixed(pat, types)
object2fixed(pat, types, keep_nomatch = TRUE)
}
\seealso{
\code{\link[=pattern2id]{pattern2id()}}
}
\keyword{development}
