% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pattern2list}
\alias{pattern2list}
\title{Function to assign multiple slots to a S4 object}
\usage{
pattern2list(
  x,
  types,
  valuetype,
  case_insensitive,
  concatenator = "_",
  levels = 1,
  remove_unigram = FALSE,
  keep_nomatch = FALSE
)
}
\arguments{
\item{x}{an S4 object}

\item{valuetype}{the type of pattern matching: \code{"glob"} for "glob"-style
wildcard expressions; \code{"regex"} for regular expressions; or \code{"fixed"} for
exact matching. See \link{valuetype} for details.}

\item{case_insensitive}{logical; if \code{TRUE}, ignore case when matching a
\code{pattern} or \link{dictionary} values}

\item{concatenator}{concatenator that join multi-word expression in tokens object}

\item{levels}{only used when pattern is a dictionary}

\item{remove_unigram}{ignore single-word patterns if \code{TRUE}}

\item{exceptions}{slots to ignore}

\item{value}{a list of attributes extracted by attributes()}
}
\description{
Function to assign multiple slots to a S4 object
}
\seealso{
\code{\link[=pattern2id]{pattern2id()}}
}
\keyword{Convert}
\keyword{a}
\keyword{and}
\keyword{as}
\keyword{in}
\keyword{input}
\keyword{internal}
\keyword{kwic.}
\keyword{pattern}
\keyword{to}
\keyword{tokens_compound}
\keyword{tokens_select,}
\keyword{used}
\keyword{various}
\keyword{vector}
