% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-deprecated.R
\name{plot-deprecated}
\alias{plot-deprecated}
\alias{plot.dfm}
\alias{plot.kwic}
\alias{plot.textmodel_wordfish_fitted}
\title{deprecated plotting functions}
\usage{
\method{plot}{dfm}(x, ...)

\method{plot}{kwic}(x, ...)

\method{plot}{textmodel_wordfish_fitted}(x, ...)
}
\arguments{
\item{x}{object to be plotted}

\item{...}{additional arguments passed to the new function}
}
\description{
Functions with deprecated names for plotting quanteda objects.
}
\details{
\code{plot.dfm} is the deprecated method to produce a word cloud from 
a dfm; use \code{\link{textplot_wordcloud}} instead.

\code{plot.kwic} is the deprecated method to produce a dispersion
plot from a \link{kwic} object; use \code{\link{textplot_xray}} instead.

\code{plot.textmodel_wordfish_fitted} is the deprecated method to
  produce a scale plot with error bars for a fitted wordfish text model object;
  \code{\link{textplot_scale1d}} instead.
}
\keyword{deprecated}
\keyword{internal}
\keyword{plot}
