% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.textmodel_wordfish_fitted}
\alias{plot.textmodel_wordfish_fitted}
\title{plot a fitted wordfish model}
\usage{
\method{plot}{textmodel_wordfish_fitted}(x, type = c("docs", "feats"),
  doclabels = NULL, sort = TRUE, left.mar = 8, pickout = NULL,
  alpha = 0.5, ...)
}
\arguments{
\item{x}{for plot method, the object to be plotted}

\item{type}{\code{docs} to plot document scores (the default) or \code{feats} to 
plot psi against beta parameters}

\item{doclabels}{a vector of names for document. If left NULL (the default), ordinary 
document names will be used.}

\item{sort}{If TRUE (the default), order points from low to high score. 
If a vector, order according to these values from low to high.
Only applies when \code{type} is 'docs'}

\item{left.mar}{an overridden left margin, passed to \code{par} (default 8.1). This
overrides R's default 4.1, which is typically too cramped for document names.}

\item{pickout}{A vector of feature names to draw attention to in a feature plot. Only 
applies if \code{type} is 'feats'}

\item{alpha}{A number between 0 and 1 (default 0.5) representing the level of alpha
transparency used to overplot feature names in a feature plot. 
Only applies if \code{type} is 'feats'.}

\item{...}{additional arguments passed to \code{\link{plot}}}
}
\description{
either as an ideal point-style plot (theta plus confidence interval on the 
x-axis, document labels on the y) with optional renaming and sorting, or as
'eiffel tower' plots (beta on the x, psi on the y, feature names over-plotted
with alpha transparency, optionally some highlighted) as in
Slapin and Proksch, 2008.
}
\examples{
postwar <- trim(dfm(inaugCorpus[41:57]), minCount = 5, minDoc = 2)
mod <- textmodel(postwar, model = "wordfish")
plot(mod, sort=FALSE)
plot(mod, type="feats", pickout = c("government", "countries", "children", 
     "the", "nuclear", "federal"))
}

