% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textstat_sentiment.R
\name{polarity}
\alias{polarity}
\alias{polarity<-}
\title{Set or get the sentiment polarity of a dictionary}
\usage{
polarity(x)

polarity(x) <- value
}
\arguments{
\item{x}{a \link{dictionary} object}

\item{value}{list consisting of named character vectors \code{pos}, \code{neg}, and
(optionally) \code{neut} corresponding to positive, negative, and neutral
sentiment categories respectively.  Each element may contain multiple
key names.  The \code{neut} category is optional but \code{pos} and \code{neg} must be
supplied.}
}
\value{
\code{polarity()} returns the polarity as a list.

\verb{polarity<-} sets the dictionary's polarity.
}
\description{
Set or retrieve the polarity of a \link{dictionary} object for the purposes of
sentiment analysis.  Polarity consists of a set of dictionary keys that are
associated with positive, negative, and (optionally) neutral categories for
use in \code{\link[=textstat_sentiment]{textstat_sentiment()}}.
}
\details{
A dictionary may have only one set of polarities at a time, but may be
changed as needed.
}
\examples{
simpledict <- dictionary(list(
    happy = c("happy", "jubilant", "exuberant"),
    sad = c("sad", "morose", "down")
))
polarity(simpledict)
polarity(simpledict) <- list(pos = "happy", neg = "sad")
polarity(simpledict)

# can list multiple keys
polarity(data_dictionary_LSD2015) <- list(
    pos = c("positive", "neg_negative"),
    neg = c("negative", "neg_positive")
)
polarity(data_dictionary_LSD2015)
}
\keyword{dictionary}
\keyword{textstat}
\keyword{utility}
