% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postag.R
\name{postag}
\alias{postag}
\alias{postag.character}
\alias{postag.corpus}
\title{do part-of-speech tagging}
\usage{
postag(x, ...)

\method{postag}{character}(x, model = NULL, method = c("perceptron", "hmm"))

\method{postag}{corpus}(x, method = c("perceptron", "hmm"))
}
\description{
Tag text(s) with part-of-speech tagger
}

