% Generated by roxygen2 (4.0.2): do not edit by hand
\name{predict.naivebayes}
\alias{predict.naivebayes}
\title{prediction method for Naive Bayes classifiers}
\usage{
\method{predict}{naivebayes}(object, newdata = NULL, scores = c(-1, 1))
}
\arguments{
\item{object}{a naivebayes class object}

\item{newdata}{new data on which to perform classification}

\item{scores}{"reference" values when the wordscores equivalent implementation
of Naive Bayes prediction is used.  Default is \code{c(-1, 1)}.}
}
\value{
A list of two data frames, named \code{docs} and \code{words} corresponding
to word- and document-level predicted quantities

\item{docs}{data frame with document-level predictive quantities:
nb.predicted, ws.predicted, bs.predicted, PcGw, wordscore.doc, bayesscore.doc,
posterior.diff, posterior.logdiff.  Note that the diff quantities are currently
implemented only for two-class solutions.}

\item{words}{data-frame with word-level predictive quantities:
wordscore.word, bayesscore.word}
}
\description{
prediction method for Naive Bayes classifier objects
}
\details{
implements class predictions using trained Naive Bayes examples
(from \code{naiveBayesText()})
}
\author{
Kenneth Benoit
}

