% Generated by roxygen2 (4.0.2): do not edit by hand
\name{predict.textmodel}
\alias{predict.textmodel}
\alias{predict.wordscores}
\title{predict a text model on a test set}
\usage{
\method{predict}{textmodel}(object, ...)

\method{predict}{wordscores}(object, newdata, se.fit = FALSE,
  rescaling = c("none", "lbg", "mv"), ...)
}
\arguments{
\item{object}{a fitted textmodel object (from \code{\link{fitmodel}})}
}
\description{
Apply a fitted text model to make predictions on test data.
}
\references{
LBG (2003); Martin and Vanberg (2007)
}

