% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel_affinity.R
\name{predict.textmodel_affinity}
\alias{predict.textmodel_affinity}
\title{Prediction for a fitted affinity textmodel}
\usage{
\method{predict}{textmodel_affinity}(object, newdata = NULL, level = 0.95,
  verbose = FALSE, ...)
}
\arguments{
\item{object}{a fitted affinity textmodel}

\item{newdata}{dfm on which prediction should be made}

\item{level}{probability level for confidence interval width}

\item{verbose}{If \code{TRUE}, output status messages}

\item{...}{additional arguments passed to other functions}
}
\value{
The \code{predict} method for a affinity fitted object returns a
  list of format...
}
\description{
Estimate \eqn{\theta_i} for each document, from a fitted
\link{textmodel_affinity} object.
}
\keyword{internal}
\keyword{textmodel}
