% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel-methods.R
\name{print.coef.textmodel}
\alias{print.coef.textmodel}
\title{Print methods for textmodel features estimates}
\usage{
\method{print}{coef.textmodel}(x, digits = max(3L, getOption("digits") - 3L),
  ...)
}
\arguments{
\item{x}{a textmodel_features object}

\item{digits}{minimal number of \emph{significant digits}, see
\code{\link{print.default}}}

\item{...}{additional arguments not used}
}
\description{
Print methods for textmodel features estimates
}
