% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel-methods.R
\name{print.coefficients_textmodel}
\alias{print.coefficients_textmodel}
\title{Print methods for textmodel features estimates
This is a helper function used in \code{print.summary.textmodel}.}
\usage{
\method{print}{coefficients_textmodel}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{a coefficients_textmodel object}

\item{digits}{minimal number of \emph{significant digits}, see
\code{\link[=print.default]{print.default()}}}

\item{...}{additional arguments not used}
}
\description{
Print methods for textmodel features estimates
This is a helper function used in \code{print.summary.textmodel}.
}
\keyword{internal}
\keyword{textmodel}
