% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phrases.R
\name{print.phrases}
\alias{print.phrases}
\title{Print a phrase object}
\usage{
\method{print}{phrases}(x, ...)
}
\arguments{
\item{x}{a phrases (constructed by \code{\link[=phrase]{phrase()}} object to be printed}

\item{...}{further arguments passed to or from other methods}
}
\description{
prints a phrase object in a way that looks like a standard list.
}
\keyword{internal}
