% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus-methods-base.R, R/dfm-print.R,
%   R/fcm-print.R, R/tokens-methods-base.R
\name{print.corpus}
\alias{print.corpus}
\alias{print,dfm-method}
\alias{print,fcm-method}
\alias{print.tokens}
\title{print a tokens objects print method for a tokens object}
\usage{
\method{print}{corpus}(
  x,
  max_ndoc = quanteda_options("print_corpus_max_ndoc"),
  max_nchar = quanteda_options("print_corpus_max_nchar"),
  show.summary = quanteda_options("print_corpus_summary"),
  ...
)

\S4method{print}{dfm}(
  x,
  max_ndoc = quanteda_options("print_dfm_max_ndoc"),
  max_nfeat = quanteda_options("print_dfm_max_nfeat"),
  show.summary = TRUE,
  ...
)

\S4method{print}{fcm}(
  x,
  max_nfeat = quanteda_options("print_dfm_max_nfeat"),
  show.summary = TRUE,
  ...
)

\method{print}{tokens}(
  x,
  max_ndoc = quanteda_options("print_tokens_max_ndoc"),
  max_ntoken = quanteda_options("print_tokens_max_ntoken"),
  show.summary = quanteda_options("print_tokens_summary"),
  ...
)
}
\arguments{
\item{x}{a tokens object created by \code{\link[=tokens]{tokens()}}}

\item{max_ndoc}{max number of documents to print; default is from the
\code{print_dfm_max_ndoc} setting of \code{\link[=quanteda_options]{quanteda_options()}}}

\item{max_nchar}{integer; number of characters to use when showing the first
part of the of text; 0 shows none, and -1 shows all}

\item{show.summary}{print a brief summary indicating the number of documents
and features}

\item{...}{further arguments passed to base print method}

\item{max_nfeat}{max number of features to print; default is from the
\code{print_dfm_max_nfeat} setting of \code{\link[=quanteda_options]{quanteda_options()}}}

\item{max_ntoken}{integer; number of tokens to use when showing the first
part of the of text; 0 shows none, and -1 shows all}
}
\description{
print a tokens objects print method for a tokens object
}
\seealso{
\code{\link[=quanteda_options]{quanteda_options()}}
}
\keyword{dfm}
