% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel-methods.R
\name{print.summary.textmodel}
\alias{print.summary.textmodel}
\title{print method for summary.textmodel}
\usage{
\method{print}{summary.textmodel}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{a \code{summary.textmodel} object}

\item{digits}{minimal number of \emph{significant digits}, see
\code{\link[=print.default]{print.default()}}}

\item{...}{additional arguments not used}
}
\description{
print method for summary.textmodel
}
\keyword{internal}
\keyword{textmodel}
