% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.textmodel}
\alias{print.naivebayes}
\alias{print.textmodel}
\alias{print.textmodelfitted}
\alias{print.textmodelpredicted}
\alias{print.wordscores}
\title{print a textmodel object}
\usage{
\method{print}{textmodel}(x, ...)

\method{print}{textmodelfitted}(x, ...)

\method{print}{textmodelpredicted}(x, ...)

\method{print}{naivebayes}(x, n = 30L, ...)

\method{print}{wordscores}(x, n = 30L, ...)
}
\arguments{
\item{x}{textmodel object to be printed}

\item{...}{additional arguments passed to \code{print}}

\item{n}{max rows of dfm to print}
}
\description{
print a class of textmodel object
}

