% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel-methods.R
\name{print.textmodel_coefficients}
\alias{print.textmodel_coefficients}
\title{Impliments print methods for textmodel_features}
\usage{
print.textmodel_coefficients(x, digits = 2, ...)
}
\arguments{
\item{x}{a textmodel_features object}

\item{...}{additional arguments not used}
}
\description{
Impliments print methods for textmodel_features
}
