% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel-methods.R
\name{print.textmodel_statistics}
\alias{print.textmodel_statistics}
\title{Impliments print methods for textmodel_statistics}
\usage{
print.textmodel_statistics(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{a textmodel_wordscore_statistics object}

\item{...}{additional arguments not used}
}
\description{
Impliments print methods for textmodel_statistics
}
