% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel-methods.R
\name{print.textmodel_summary}
\alias{print.textmodel_summary}
\title{Impliments print methods for textmodel_summary}
\usage{
print.textmodel_summary(x, ...)
}
\arguments{
\item{x}{a textmodel_summary object}

\item{...}{additional arguments not used}
}
\description{
Impliments print methods for textmodel_summary
}
