% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel_wordfish.R
\name{print.textmodel_wordfish}
\alias{print.textmodel_wordfish}
\title{print method for a wordfish model}
\usage{
\method{print}{textmodel_wordfish}(x, ...)
}
\arguments{
\item{x}{for print method, the object to be printed}

\item{...}{unused}
}
\description{
print method for a wordfish model
}
\keyword{internal}
\keyword{textmodel}
