% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel_wordscores.R
\name{print.textmodel_wordscore_statistics}
\alias{print.textmodel_wordscore_statistics}
\title{Impliments print methods for textmodel_wordscore_statistics}
\usage{
print.textmodel_wordscore_statistics(x, ...)
}
\arguments{
\item{x}{a textmodel_wordscore_statistics object}

\item{...}{additional arguments not used}
}
\description{
Impliments print methods for textmodel_wordscore_statistics
}
