% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tokenize.R
\name{print.tokenizedTexts}
\alias{print.tokenizedTexts}
\title{print a tokenizedTexts objects}
\usage{
\method{print}{tokenizedTexts}(x, ...)
}
\arguments{
\item{x}{a tokenizedText object created by \link{tokenize}}

\item{...}{further arguments passed to base print method}
}
\description{
print method for a \link{tokenize}dText object
}

