% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hashingTokens.R
\name{print.tokenizedTextsHashed}
\alias{print.tokenizedTextsHashed}
\title{print a tokenizedTextsHashed objects}
\usage{
\method{print}{tokenizedTextsHashed}(x, ...)
}
\arguments{
\item{x}{a tokenizedTextHashed object created by \link{hashingTokens}}

\item{...}{further arguments passed to base print method}
}
\description{
print method for a tokenizedTextsHashed object
}

