% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokens.R
\name{print.tokens}
\alias{print.tokens}
\title{print a tokens objects
print method for a tokenizedTextsHashed object}
\usage{
\method{print}{tokens}(x, ...)
}
\arguments{
\item{x}{a tokens object created by \code{\link{tokens}}}

\item{...}{further arguments passed to base print method}
}
\description{
print a tokens objects
print method for a tokenizedTextsHashed object
}

