% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quanteda-package.R
\docType{package}
\name{quanteda-package}
\alias{quanteda-package}
\title{An R package for the quantitative analysis of textual data.}
\description{
A set of functions for creating and managing text corpora, extracting features from 
text corpora, and analyzing those features using quantitative methods.

More detailed description, and some examples, to go here.
}
\author{
Ken Benoit and Paul Nulty
}

