% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.corpus}
\alias{summary.corpus}
\title{Summarize a corpus}
\usage{
\method{summary}{corpus}(object, cache = TRUE, ...)
}
\arguments{
\item{object}{corpus to be summarized}

\item{...}{additional arguments passed through to \code{\link[=tokens]{tokens()}}}

\item{n}{maximum number of texts to describe, default=100}

\item{showmeta}{set to \code{TRUE} to include document-level
meta-data}

\item{tolower}{convert texts to lower case before counting types}
}
\description{
Displays information about a corpus, including attributes and metadata such
as date of number of texts, creation and source.
}
\examples{
corp <- data_corpus_inaugural
summary(corp, cache = TRUE)
toks <- tokens(corp)
summary(toks, cache = TRUE)
dfmat <- dfm(toks)
summary(dfmat, cache = TRUE)
}
\keyword{corpus}
\keyword{internal}
