\name{quantedaRefresh}
\alias{quantedaRefresh}
\title{Re-install \package{quanteda} from github}
\usage{
quantedaRefresh(branch = c("dev", "master"))
}
\arguments{
  \item{branch}{default is \code{"dev"}}
}
\value{
Nothing
}
\description{
Refresh the installation from the github repository for the
package.  Useful if you need to pull the latest changes.
}
\author{
Kenneth Benoit
}

