% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R
\name{read_dict_lexicoder}
\alias{read_dict_lexicoder}
\title{Import a Lexicoder dictionary}
\usage{
read_dict_lexicoder(path)
}
\arguments{
\item{path}{the path from which to read in the Lexicoder dictionary}
}
\description{
Import a Lexicoder dictionary
}
\examples{
\dontrun{
# dict <- read_dict_lexicoder("/home/kohei/Documents/Dictionary/Lexicoder/LSDaug2015/LSD2015.lc3")
}
}
\keyword{dictionary}
\keyword{internal}
