% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R
\name{read_dict_wordstat}
\alias{read_dict_wordstat}
\title{Import a Wordstat dictionary}
\usage{
read_dict_wordstat(path, encoding = "auto")
}
\arguments{
\item{path}{path to file to be imported}

\item{encoding}{the encoding of the file to be imported}
}
\description{
Import a Wordstat dictionary
}
\examples{
\dontrun{
# dict <- read_dict_wordstat("/home/kohei/Documents/Dictionary/LaverGarry.txt", "utf-8")
# dict <- read_dict_wordstat("/home/kohei/Documents/Dictionary/Wordstat/ROGET.cat", "utf-8")
# dict <- read_dict_wordstat("/home/kohei/Documents/Dictionary/Wordstat/WordStat Sentiments.cat", 
#                            encoding = "iso-8859-1")
}
}
\keyword{dictionary}
\keyword{internal}
