% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R
\name{read_dict_yoshikoder}
\alias{read_dict_yoshikoder}
\title{Import a Yoshikoder dictionary file.}
\usage{
read_dict_yoshikoder(path)
}
\arguments{
\item{path}{to filename to be imported}
}
\description{
Import a Yoshikoder dictionary file.
}
\examples{
# dict <- read_dict_yoshikoder("/home/kohei/Documents/Dictionary/Yoshikoder/laver-garry-ajps.ykd")
}
\keyword{dictionary}
\keyword{internal}
