\name{readWStatDict}
\alias{readWStatDict}
\title{Make a flattened list from a hierarchical wordstat dictionary}
\usage{
readWStatDict(path)
}
\arguments{
  \item{path}{path to the wordstat dictionary file}
}
\value{
flattened dictionary as a list
}
\description{
Make a flattened list from a hierarchical wordstat
dictionary
}

