% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reassign_attributes}
\alias{reassign_attributes}
\title{copy the attributes from one S3 object to another}
\usage{
reassign_attributes(x_new, x_from, exceptions = NULL, attr_only = FALSE)
}
\arguments{
\item{x_new}{the object to which the attributes will be copied}

\item{x_from}{the object from which the attributes will be copied, or a list
of attrbiutes if \code{attr_only = TRUE}}

\item{attr_only}{logical; if \code{TRUE}, then \code{x_orig} is a list of
attributes rather than an object with attributes.}

\item{exception}{a character vector of attribute names NOT to be copied}
}
\description{
Copy the attributes from one S3 object to another.  Necessary when some
operation defined for the base class obliterates them.
}
\author{
Ken Benoit
}
\keyword{internal}
