% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectFeatures2.R
\name{regex2fixed}
\alias{regex2fixed}
\title{Convert regex to fixed}
\usage{
regex2fixed(regex, types, case_insensitive = TRUE, ...)
}
\arguments{
\item{regex}{regular expression}

\item{types}{unique types of tokens}

\item{case_insensitive}{case sensitivity}
}
\description{
Convert regex to fixed
}

