% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R
\name{replace_dictionary_values}
\alias{replace_dictionary_values}
\title{Internal function to replace dictionary values}
\usage{
replace_dictionary_values(dict, from, to)
}
\arguments{
\item{dict}{a \link{dictionary} object}
}
\description{
Internal function to replace dictionary values
}
\examples{
dict <- list(KEY1 = list(SUBKEY1 = list("A_B"),
                          SUBKEY2 = list("C_D")),
              KEY2 = list(SUBKEY3 = list("E_F"),
                          SUBKEY4 = list("G_F_I")),
              KEY3 = list(SUBKEY5 = list(SUBKEY7 = list("J_K")),
                          SUBKEY6 = list(SUBKEY8 = list("L"))))
quanteda:::replace_dictionary_values(dict, "_", " ")
}
\keyword{internal}
