% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus-deprecated.R, R/dfm-deprecated.R
\name{sample}
\alias{sample}
\alias{sample.corpus}
\alias{sample.default}
\alias{sample.dfm}
\title{randomly sample documents or features}
\usage{
sample(x, size, replace = FALSE, prob = NULL, ...)

\method{sample}{default}(x, size, replace = FALSE, prob = NULL, ...)

\method{sample}{corpus}(x, ...)

\method{sample}{dfm}(x, ...)
}
\arguments{
\item{x}{a corpus or dfm object whose documents or features will be sampled}

\item{size}{a positive number, the number of documents to select}

\item{replace}{Should sampling be with replacement?}

\item{prob}{A vector of probability weights for obtaining the elements of the
vector being sampled.}

\item{...}{unused}
}
\description{
Deprecated function framework for randomly sampling from a corpus or dfm object.  Use
\code{\link{corpus_sample}} or \code{dfm_sample} instead
}
\seealso{
\code{\link{corpus_sample}}, \code{\link{dfm_sample}}
}
\keyword{deprecated}
\keyword{internal}

