% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern2fixed.R
\name{search_glob}
\alias{search_glob}
\alias{search_regex}
\alias{search_fixed}
\title{Select types without performing slow regex search}
\usage{
search_glob(patterns, types_search, index)

search_regex(patterns, types_search, case_insensitive)

search_fixed(patterns, types_search, index)
}
\arguments{
\item{patterns}{a list of "glob", "fixed" or "regex" patterns}

\item{types_search}{lowercased types when \code{case_insensitive=TRUE}, but
not used in glob and fixed matching as types are in the index.}

\item{index}{index object created by \code{index_types}}

\item{case_insensitive}{ignore case when matching, if \code{TRUE}, but not
used in glob and fixed matching as types are lowercased in the index.}
}
\description{
This is an internal function for \code{pattern2id} that select types using
keys in index when available.
}
\keyword{internal}
