% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex2fixed.R
\name{search_index}
\alias{search_index}
\title{internal function for \code{select_types()} to seach the index using fastmatch.}
\usage{
search_index(regex, index)
}
\arguments{
\item{regex}{a regular expression to search}

\item{index}{an index object created by \code{index_types()}}
}
\description{
internal function for \code{select_types()} to seach the index using fastmatch.
}
\seealso{
index_types
}
\keyword{internal}
