% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R
\name{select_dictionary_levels}
\alias{select_dictionary_levels}
\title{Internal function to select dictionary nested levels}
\usage{
select_dictionary_levels(dict, levels = 1:100, level = 1)
}
\arguments{
\item{dict}{a dictionary object}

\item{levels}{an integer vector indicating levels to select}

\item{level}{an internal argument to pass current levels}
}
\description{
Internal function to select dictionary nested levels
}
\examples{
dict <- list("A" = list("B" = c("b", "B"), c("a", "A", "aa")))
quanteda:::select_dictionary_levels(dict, 1:2)
quanteda:::select_dictionary_levels(dict, 1)
quanteda:::select_dictionary_levels(dict, 2)
}
\keyword{internal}
