\name{sentenceSeg}
\alias{sentenceSeg}
\title{split a text into sentences
This function takes a text and splits it into sentences.}
\usage{
sentenceSeg(text, pat = "[\\\\.\\\\?\\\\!][\\\\n* ]|\\\\n\\\\n*",
  abbreviations = NULL, parag = TRUE)
}
\arguments{
  \item{text}{Text to be segmented}
}
\description{
split a text into sentences This function takes a text and
splits it into sentences.
}

