% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sequence2list}
\alias{sequence2list}
\title{convert sequences to a simple list}
\usage{
sequence2list(sequences)
}
\arguments{
\item{sequences}{the input sequence, one of: \itemize{ \item{character vector,
}{whose elements will be split on whitespace;} \item{list of characters,
}{consisting of a list of token patterns, separated by white space}; 
\item{\link{tokens} object;} \item{\link{dictionary} object}{;} 
\item{\link{collocations} object.}{} }}
}
\value{
an unnamed list of sequences, with each element of the list a
  character vector with the split sequence.
}
\description{
Convert a sequence into a simple list, for input as a sequence in e.g. 
\code{\link{tokens_compound}}.
}
\keyword{internal}
\keyword{utilities}
