% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimnames.R
\name{set_dfm_dimnames<-}
\alias{set_dfm_dimnames<-}
\alias{set_dfm_docnames<-}
\alias{set_dfm_featnames<-}
\alias{set_fcm_dimnames<-}
\alias{set_fcm_featnames<-}
\title{Internal functions to set dimnames}
\usage{
set_dfm_dimnames(x) <- value

set_dfm_docnames(x) <- value

set_dfm_featnames(x) <- value

set_fcm_dimnames(x) <- value

set_fcm_featnames(x) <- value
}
\arguments{
\item{x}{a dfm or fcm}

\item{value}{character a vector for docnames or featmanes or a list of them
for dimnames}
}
\description{
Default dimnames() converts a zero-length character vector to NULL, leading
to disfunctioning of subsetting functions. These are safer methods to set
dimnames of a dfm or fcm.
}
\keyword{internal}
