% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfm-methods.R
\name{set_dfm_slots<-}
\alias{set_dfm_slots<-}
\alias{get_dfm_slots}
\title{Set values to a dfm's S4 slots}
\usage{
set_dfm_slots(x, exceptions = NULL) <- value

get_dfm_slots(x)
}
\arguments{
\item{x}{a dfm}

\item{exceptions}{names of slots to be ignored}

\item{value}{a list of values extracted using \code{attributes} and to be assigned to slots}
}
\description{
Set values to a dfm's S4 slots
}
\keyword{internal}
