% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfm-methods.R
\name{set_dfm_slots}
\alias{set_dfm_slots}
\title{Set values to a dfm's S4 slots}
\usage{
set_dfm_slots(x, slots = NULL, exceptions = NULL)
}
\arguments{
\item{x}{a dfm}

\item{slots}{a list of values extracted using \code{attributes} and to be assigned to slots}

\item{exceptions}{names of slots to be ignored}
}
\description{
Set values to a dfm's S4 slots
}
\keyword{internal}
