% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcm-methods.R
\name{set_fcm_slots<-}
\alias{set_fcm_slots<-}
\alias{get_fcm_slots}
\title{Set values to a fcm's S4 slots}
\usage{
set_fcm_slots(x, exceptions = NULL) <- value

get_fcm_slots(x)
}
\arguments{
\item{x}{a fcm}

\item{exceptions}{names of slots to be ignored}

\item{value}{a list of values extracted using \code{attributes} and to be assigned to slots}
}
\description{
Set values to a fcm's S4 slots
}
\keyword{internal}
