% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcm-methods.R
\name{set_fcm_slots}
\alias{set_fcm_slots}
\title{Set values to a fcm's S4 slots}
\usage{
set_fcm_slots(x, slots = NULL, exceptions = NULL)
}
\arguments{
\item{x}{a fcm}

\item{slots}{a list of values extracted using \code{attributes} and to be assigned to slots}

\item{exceptions}{names of slots to be ignored}
}
\description{
Set values to a fcm's S4 slots
}
\keyword{internal}
