% Generated by roxygen2 (4.0.2): do not edit by hand
\name{settingsInitialize}
\alias{settingsInitialize}
\title{\code{settingsInitialize} returns a list of legal settings, set to their default values}
\usage{
settingsInitialize()
}
\description{
\code{settingsInitialize} returns a list of legal settings, set to their default values
}

