% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dictionaries.R
\docType{methods}
\name{show,dictionary-method}
\alias{show,dictionary-method}
\title{print a dictionary object}
\usage{
\S4method{show}{dictionary}(object)
}
\arguments{
\item{object}{the dictionary to be printed}
}
\description{
Print/show method for dictionary objects.
}

