% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textstat_simil.R
\docType{methods}
\name{textstat_simil}
\alias{textstat_simil}
\alias{textstat_simil,dfm-method}
\title{Similarity matrix between documents and/or features}
\usage{
textstat_simil(x, selection = character(0), n = NULL,
  margin = c("documents", "features"), method = "correlation",
  normalize = FALSE, tri = FALSE, diag = FALSE)

\S4method{textstat_simil}{dfm}(x, selection = character(0), n = NULL,
  margin = c("documents", "features"), method = "correlation",
  normalize = FALSE, tri = TRUE, diag = FALSE)
}
\arguments{
\item{x}{a \link{dfm} object}

\item{selection}{character or character vector of document names or feature 
labels from the dfm}

\item{n}{the top \code{n} most similar items will be returned.  If n is \code{NULL}, return all items.}

\item{margin}{identifies the margin of the dfm on which similarity will be 
computed:  \code{documents} for documents or \code{features} for word/term
features.}

\item{method}{method the distance measure to be used, options are "cosine", "correlation", 
"jaccard", "eJaccard", "dice", "eDice", "simple matching", "hamann", "faith", default "correlation". More options are avaible 
 in \code{\link{textstat_dist}}}

\item{normalize}{a deprecated argument retained (temporarily) for legacy 
reasons.  If you want to compute similarity on a "normalized" dfm objects 
(e.g. \code{x}), wrap it in \code{\link{weight}(x, "relFreq")}.}

\item{tri}{whether the upper triangle of the symmetric \eqn{V \times V} matrix is recorded}

\item{diag}{whether the diagonal of the distance matrix should be recorded}
}
\description{
These functions compute similarity matrix between documents and/or features from a 
\code{\link{dfm}} and return a standard \code{\link[stats]{dist}} object.
}
\examples{
# create a dfm from inaugural addresses from Reagan onwards
presDfm <- dfm(corpus_subset(inaugCorpus, Year > 1980), remove = stopwords("english"),
               stem = TRUE)

# compute some document similarities
(tmp <- textstat_simil(presDfm, margin = "documents"))
# output as a list
#as.list(tmp)
# for specific comparisons
textstat_simil(presDfm, "1985-Reagan", n = 5, margin = "documents")
textstat_simil(presDfm, c("2009-Obama" , "2013-Obama"), n = 5, margin = "documents")
textstat_simil(presDfm, c("2009-Obama" , "2013-Obama"), margin = "documents")
textstat_simil(presDfm, c("2009-Obama" , "2013-Obama"), margin = "documents", method = "cosine")
# compute some term similarities
textstat_simil(presDfm, c("fair", "health", "terror"), method="cosine", margin = "features", 20)

}

