% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{slots<-}
\alias{slots<-}
\title{function to assign multiple slots to a S4 object}
\usage{
slots(x, exceptions = c("Dim", "Dimnames", "i", "p", "x", "factors"),
  values) <- value
}
\arguments{
\item{x}{an S4 object}

\item{exceptions}{slots to ignore}

\item{values}{list of slots extracted by attributes()}
}
\description{
function to assign multiple slots to a S4 object
}
\keyword{internal}
