% Generated by roxygen2 (4.0.2): do not edit by hand
\name{smoothdfm}
\alias{smoothdfm}
\title{Additive smoothing of feature frequencies in a dfm}
\usage{
smoothdfm(x, alpha = 0.5)
}
\arguments{
\item{x}{document-feature matrix created by \link{dfm}}

\item{alpha}{The value to add to all counts. Default is 0.5}
}
\value{
The original dfm, with values weighted according to type function.
}
\description{
Smooths the feature counts by adding a small value (default 0.5) to remove
zero counts. Zero counts are problematic for probability-based models.
}
\examples{
dtm <- dfm(inaugCorpus)
dtm[1:5,1:10]
smDtm <- smoothdfm(dtm)
smDtm[1:5,1:10]
}
\author{
Paul Nulty
}

