% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfm-deprecated.R
\name{sort.dfm}
\alias{sort.dfm}
\title{sort a dfm by one or more margins}
\usage{
\method{sort}{dfm}(x, decreasing = TRUE, margin = c("features", "docs",
  "both"), ...)
}
\arguments{
\item{x}{Document-feature matrix created by \code{\link{dfm}}}

\item{decreasing}{TRUE (default) if sort will be in descending order,
otherwise sort in increasing order}

\item{margin}{which margin to sort on \code{features} to sort by frequency of
features, \code{docs} to sort by total feature counts in documents, and
\code{both} to sort by both}

\item{...}{additional arguments passed to base method \code{sort.int}}
}
\value{
A sorted \link{dfm} matrix object
}
\description{
Sorts a \link{dfm} by frequency of total features, total features in
documents, or both
}
\note{
Deprecated; use \code{\link{dfm_sort}} instead.
}
\seealso{
dfm_sort
}
\keyword{deprecated}
\keyword{internal}

