% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocs.R
\docType{data}
\name{SOTUCorpus}
\alias{SOTUCorpus}
\alias{sotuCorp}
\title{U.S. State of the Union addresses from 1790 to 2016}
\source{
The American Presidency Project, 
  \url{http://www.presidency.ucsb.edu/sou.php}.
}
\description{
A corpus object 230 US State-of-the-Union addresses, from 1790 
  to 2016. The corpus includes the following document variables: \describe{
  \item{FirstName}{President's first and middle names.} 
  \item{President}{President's last name.}
  
  \item{Date}{Date of the delivery of the speech or document.}
  
  \item{delivery}{Either \code{written} or \code{spoken}, depending on the
  format.  See Source.}
  
  \item{type}{Either \code{SOTU} for an official State of the Union Address,
  or \code{other}, for a different form of speech or message. See Source.}
   
  \item{party}{President's party.}
  }
}

