% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfm-methods.R
\name{sparsity}
\alias{sparsity}
\title{compute the sparsity of a document-feature matrix}
\usage{
sparsity(x)
}
\arguments{
\item{x}{the document-feature matrix}
}
\description{
Return the proportion of sparseness of a document-feature matrix, equal
to the proportion of cells that have zero counts.
}
\examples{
inaug_dfm <- dfm(data_corpus_inaugural, verbose = FALSE)
sparsity(inaug_dfm)
sparsity(dfm_trim(inaug_dfm, min_count = 5))
}

