% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R
\name{split_values}
\alias{split_values}
\title{Internal function for special handling of multi-word dictionary values}
\usage{
split_values(dict, separator, concatenator)
}
\arguments{
\item{dict}{a flatten dictionary}

\item{separator}{separator from a dictionary object}

\item{concatenator}{concatenator from a tokens object}
}
\description{
Internal function for special handling of multi-word dictionary values
}
\keyword{internal}
