% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/remove-stopwords.R
\name{stopwordsGet}
\alias{stopwordsGet}
\title{access stopwords}
\usage{
stopwordsGet(kind = "english")
}
\arguments{
\item{kind}{The pre-set kind of stopwords (as a character string)}
}
\value{
a character vector or dfm with stopwords removed
}
\description{
This function retrieves stopwords from the type specified in the \code{kind}
argument and returns the stopword list as a character vector The default is
English. See \link{stopwords} for information about the list.
}
\examples{
stopwordsGet()
stopwordsGet("italian")
}

