% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus_subset.R
\name{subset.corpus}
\alias{subset.corpus}
\title{deprecated name for corpus_subset}
\usage{
\method{subset}{corpus}(x, ...)
}
\arguments{
\item{x}{corpus to be subset}

\item{...}{subsetting arguments, see \code{\link{subset.data.frame}}}
}
\description{
The deprecated function name for what is now \code{\link{corpus_subset}}. 
Please use that instead.
}
\keyword{deprecated}
\keyword{internal}
