% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character-methods.R
\name{summary.character}
\alias{summary.character}
\title{summary statistics on a character vector}
\usage{
\method{summary}{character}(object, n = 100, tolower = FALSE, ...)
}
\arguments{
\item{object}{corpus to be summarized}

\item{n}{maximum number of texts to describe, default=100}

\item{tolower}{convert texts to lower case before counting types}

\item{...}{additional arguments passed through to \code{\link{tokens}}}
}
\description{
Internal-only function to compute summary statistics on a character object.
}
\examples{
# summarize texts
summary(c("Testing this text.  Second sentence.", "And this one."))
summary(data_char_ukimmig2010)
myTextSummaryDF <- summary(data_char_ukimmig2010)
head(myTextSummaryDF)
}
\keyword{internal}
