% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.corpus}
\alias{summary.corpus}
\alias{summary.tokens}
\alias{summary.dfm}
\alias{summary.fcm}
\title{Summarize a corpus}
\usage{
\method{summary}{corpus}(object, ...)

\method{summary}{tokens}(object, ...)

\method{summary}{dfm}(object, ...)

\method{summary}{fcm}(object, ...)
}
\arguments{
\item{object}{a object to be summarized.}

\item{...}{not used.}
}
\value{
a invisible list of summary statistics.
}
\description{
Displays information about a corpus, including attributes and metadata such
as date of number of texts, creation and source.
}
\examples{
summary(data_corpus_inaugural)
}
