% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/textmodel-S4generics.R
\name{summary.textmodel}
\alias{summary.textmodel}
\alias{summary.textmodelfitted}
\alias{summary.textmodelpredicted}
\title{summary of a textmodel object}
\usage{
\method{summary}{textmodel}(object, ...)

\method{summary}{textmodelfitted}(object, ...)

\method{summary}{textmodelpredicted}(object, ...)
}
\arguments{
\item{object}{textmodel object to be summarized}

\item{...}{additional arguments to \link{print}}
}
\description{
Summarize the results of a fitted or predicted textmodel object.
}

