% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel_nb.R
\name{summary.textmodel_nb}
\alias{summary.textmodel_nb}
\title{summary method for textmodel_nb objects}
\usage{
\method{summary}{textmodel_nb}(object, n = 30, ...)
}
\arguments{
\item{object}{output from \code{\link[=textmodel_nb]{textmodel_nb()}}}

\item{n}{how many coefficients to print before truncating}

\item{...}{additional arguments not used}
}
\description{
summary method for textmodel_nb objects
}
\keyword{internal}
\keyword{textmodel}
