\docType{data}
\name{swissdebates}
\alias{swissdebates}
\title{Swiss speeches from Daniel Schwartz's data project}
\format{A corpus class object from the quanteda package.}
\description{
A dataset containing all speech from the Swiss parliament
from 2000 to 2010 (or something like that - Daniel will
need to correct this!)  This data contains the following
variables
}
\details{
\itemize{ \item[billID] A numeric code that identifies the
bill, using the xxx system \item[speakerID] The name of the
speaker \item[type] An identifier of the type of the speech
act, including... }
}
\keyword{datasets}

