\docType{data}
\name{sylCounts}
\alias{sylCounts}
\title{A named list mapping words to counts of their syllables}
\description{
A named list mapping words to counts of their syllables,
generated from the CMU pronunciation dictionary
}
\references{
\url{http://www.speech.cs.cmu.edu/cgi-bin/cmudict}
}

