% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/corpus-sources-S4.R
\docType{methods}
\name{textfile}
\alias{textfile}
\alias{textfile,character,index,missing-method}
\title{read a text corpus source from a file}
\usage{
textfile(file, textField, directory, ...)

\S4method{textfile}{character,index,missing}(file, textField,
  directory = NULL, ...)
}
\arguments{
\item{file}{the complete filename to be read}

\item{textField}{a variable (column) name or column number indicating where
to find the texts that form the documents for the corpus}

\item{directory}{not used yet, and may be removed (if I move this to a new
method called \code{textfiles})}
}
\value{
an object of class \link{corpusSource-class} that can be read by
  \libk{corpus} to construct a corpus
}
\description{
Read a text corpus from a source file, where the single file will consist of
a set of texts in columns and document variables and document-level meta-data
in additional columns.  For spreadsheet-like files, the first row must be a
header.
}
\details{
The constructor does not store a copy of the texts, but rather reads
in the texts and associated data, and saves them to a temporary R object
whose location is specified in the \link{corpusSource-class} object.  This
prevents a complete copy of the object from cluttering the global environment
and consuming additional space.  This does mean however that the state of the
file containing the source data will not be cross-platform and may not be
persistent across sessions.  So the recommended usage is to load the data
into a corpus in the same session in which \code{textfile} is called.
}

