% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel-generics.R, R/textmodel_affinity.R,
%   R/textmodel_ca.R, R/textmodel_lsa.R, R/textmodel_wordfish.R
\docType{class}
\name{textmodel-internal}
\alias{textmodel-internal}
\alias{textmodel_affinity_fitted-class}
\alias{textmodel_affinity_predicted-class}
\alias{show,textmodel_affinity_fitted-method}
\alias{show,textmodel_affinity_predicted-method}
\alias{coef,textmodel_affinity_predicted-method}
\alias{coefficients,textmodel_affinity_predicted-method}
\alias{residuals,textmodel_affinity_predicted-method}
\alias{resid,textmodel_affinity_predicted-method}
\alias{rstandard.textmodel_affinity_predicted}
\alias{print.affinity_influence}
\alias{summary.affinity_influence}
\alias{print.summary_affinity_influence}
\alias{textmodel_ca_fitted-class}
\alias{coef,textmodel_ca_fitted-method}
\alias{coefficients,textmodel_ca_fitted-method}
\alias{textmodel_lsa_fitted-class}
\alias{textmodel_lsa_predicted-class}
\alias{coef,textmodel_lsa_fitted-method}
\alias{coefficients,textmodel_lsa_fitted-method}
\alias{print.textmodel_wordfish}
\alias{coef.textmodel_wordfish}
\alias{coefficient.textmodel_wordfish}
\title{Internal functions for textmodel objects}
\usage{
\S4method{show}{textmodel_affinity_fitted}(object)

\S4method{show}{textmodel_affinity_predicted}(object)

\S4method{coef}{textmodel_affinity_predicted}(object, ...)

\S4method{coefficients}{textmodel_affinity_predicted}(object, ...)

\S4method{residuals}{textmodel_affinity_predicted}(object,
  type = c("response", "pearson"), ...)

\S4method{resid}{textmodel_affinity_predicted}(object, ...)

\method{rstandard}{textmodel_affinity_predicted}(model, ...)

\method{print}{affinity_influence}(x, n = 30, ...)

\method{summary}{affinity_influence}(object, ...)

\method{print}{summary_affinity_influence}(x, n = 30, ...)

\S4method{coef}{textmodel_ca_fitted}(object, doc_dim = 1, feat_dim = 1, ...)

\S4method{coefficients}{textmodel_ca_fitted}(object, ...)

\S4method{coef}{textmodel_lsa_fitted}(object, doc_dim = 1, feat_dim = 1,
  ...)

\S4method{coefficients}{textmodel_lsa_fitted}(object, ...)

\method{print}{textmodel_wordfish}(x, ...)

coef.textmodel_wordfish(object, ...)

coefficient.textmodel_wordfish(object, ...)
}
\arguments{
\item{...}{additional arguments passed to \code{\link{print}}}

\item{type}{see \link{residuals.lm}}

\item{x}{for print method, the object to be printed}

\item{n}{max rows of dfm to print}

\item{doc_dim, feat_dim}{the document and feature dimension scores to be 
extracted as coefficients}

\item{doc_dim, feat_dim}{the document and feature dimension scores to be 
extracted as coefficients}
}
\description{
Internal function documentation for textmodel objects.
}
\keyword{internal}
\keyword{textmodel}
