% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/textmodel-wordscoresS4.R
\name{textmodel2_wordscores}
\alias{predict.textmodel_wordscores_fitted}
\alias{print.textmodel_wordscores_fitted}
\alias{print.textmodel_wordscores_predicted}
\alias{show.textmodel_wordscores_fitted}
\alias{textmodel2_wordscores}
\title{Wordscores text model}
\usage{
textmodel2_wordscores(data, scores, scale = c("linear", "logit"),
  smooth = 0)

\method{predict}{textmodel_wordscores_fitted}(object, newdata = NULL,
  rescaling = "none", level = 0.95, verbose = TRUE, ...)

\method{print}{textmodel_wordscores_fitted}(x, n = 30L, ...)

show.textmodel_wordscores_fitted(x, n = 30L, ...)

\method{print}{textmodel_wordscores_predicted}(x, ...)
}
\arguments{
\item{data}{the dfm on which the model will be fit.  Does not need to
contain only the training documents, since the index of these will be
matched automatically.}

\item{scores}{vector of training scores associated with each document
identified in \code{refData}}

\item{scale}{classic LBG linear posterior weighted word class differences, or
logit scale of log posterior differences}

\item{smooth}{a smoothing parameter for word counts; defaults to zero for the
to match the LBG (2003) method.}

\item{object}{the fitted wordscores textmodel on which prediction will be made}

\item{newdata}{dfm on which prediction should be made}

\item{rescaling}{\code{none} for "raw" scores; \code{lbg} for LBG (2003)
rescaling; or \code{mv} for the rescaling proposed by Martin and Vanberg
(2007).  (Note to authors: Provide full details here in documentation.)}

\item{level}{probability level for confidence interval width}

\item{verbose}{If \code{TRUE}, output status messages}

\item{...}{additional argumennts passed to other functions}

\item{x}{for print method, the object to be printed}

\item{n}{max rows of dfm to print}
}
\value{
\code{predict.wordscores} returns a data.frame whose rows are the
  documents fitted and whose columns contain the scored textvalues, with the
  number of columns depending on the options called (for instance, how many
  rescaled scores, and whether standard errors were requested.)  (Note: We
  may very well change this soon so that it is a list similar to other
  existing fitted objects.)
}
\description{
\code{textmodel_wordscores} implements Laver, Benoit and Garry's (2003)
wordscores method for scaling of a single dimension.  This can be called directly, but the
recommended method is through \code{\link{textmodel}}.
}
\examples{
data(LBGexample, package="quantedaData")
ws <- textmodel2(LBGexample, c(seq(-1.5, 1.5, .75), NA), model="wordscores")
ws
predict(ws)
# same as:
ws2 <- textmodel2_wordscores(LBGexample, c(seq(-1.5, 1.5, .75), NA))
print(ws2)
wsp2 <- predict(ws2)
print(wsp2)
}
\author{
Kenneth Benoit

Ken Benoit, borrowed in places from Will Lowe, who probably borrowed
  from me at some early stage.
}
\references{
Laver, Michael, Kenneth R Benoit, and John Garry. 2003.
  "Extracting Policy Positions From Political Texts Using Words as Data."
  American Political Science Review 97(02): 311-31; Beauchamp, N. 2012.
  "Using Text to Scale Legislatures with Uninformative Voting." New York
  University Mimeo.; Martin, L W, and G Vanberg. 2007. "A Robust
  Transformation Procedure for Interpreting Political Text." Political
  Analysis 16(1): 93-100.

LBG (2003); Martin and Vanberg (2007)
}

