% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel_affinity.R
\name{textmodel_affinity-internal}
\alias{textmodel_affinity-internal}
\alias{print.influence.predict.textmodel_affinity}
\alias{summary.influence.predict.textmodel_affinity}
\alias{print.summary.influence.predict.textmodel_affinity}
\title{Internal methods for textmodel_affinity}
\usage{
\method{print}{influence.predict.textmodel_affinity}(x, n = 30, ...)

\method{summary}{influence.predict.textmodel_affinity}(object, ...)

\method{print}{summary.influence.predict.textmodel_affinity}(x, n = 30, ...)
}
\arguments{
\item{n}{how many coefficients to print before truncating}
}
\description{
Internal print and summary methods for derivative
\link{textmodel_affinity} objects.
}
\keyword{internal}
\keyword{textmodel}
