% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel-ca.R
\name{textmodel_ca}
\alias{textmodel_ca}
\title{correspondence analysis of a document-feature matrix}
\usage{
textmodel_ca(data, smooth = 0, ...)
}
\arguments{
\item{data}{the dfm on which the model will be fit}

\item{smooth}{a smoothing parameter for word counts; defaults to zero.}

\item{...}{additional arguments passed to \link[ca]{ca.matrix}}
}
\description{
\code{textmodel_ca} implements correspondence analysis scaling on a
\link{dfm}.  Currently the method is a wrapper to \link[ca]{ca.matrix} in the \pkg{ca} package.
}
\examples{
ieDfm <- dfm(ie2010Corpus)
wca <- textmodel_ca(ieDfm)
summary(wca) 
}
\author{
Kenneth Benoit
}

