% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{textmodel_fitted-class}
\alias{textmodel_fitted-class}
\alias{textmodel_wordscores_fitted-class}
\alias{textmodel_wordscores_predicted-class}
\title{the fitted textmodel classes}
\description{
The \code{textmodel} virtual class is a parent class for more specific fitted text models,
which are the result of a quantitative text analysis applied to a document-feature
matrix.
}
\details{
Available types currently include...
}
\section{Slots}{

\describe{
\item{\code{dfm}}{a \link{dfm-class} document-feature matrix}
}}

