% Generated by roxygen2 (4.0.2): do not edit by hand
\name{textmodel_lda}
\alias{textmodel_lda}
\title{latent Dirichlet allocation text model}
\usage{
textmodel_lda(x, model = c("lda", "ctm", "stm"), k, smooth = 0,
  meta = NULL, ...)
}
\arguments{
\item{model}{the class of lda-type model to fit.  \code{lda} for classic
latent Dirichlet allocation; \code{ctm} for Blei and Rafferty's (2007)
correlated topic model.}

\item{k}{number of topics}

\item{smooth}{a smoothing parameter for word counts, defaults to 0}

\item{meta}{metadata passed to \link[stm]{stm} if fitting a structural topic
model}

\item{...}{additional arguments passed to \link[topicmodels]{LDA}}

\item{data}{the dfm on which the model will be fit}
}
\description{
\code{textmodel_lda} estimates the parameters of Blei et. al. (2003)Slapin
and Proksch's (2008) Poisson scaling model, also known as "wordfish", for a
single dimension.   This function is a wrapper to \link[topicmodels]{LDA} and
\link[topicmodels]{CTM} in \pkg{topicmodels}.
}
\examples{
library(quantedaData)
data(sotuCorp)
SOTUCorpus <- sotuCorp
presDfm <- dfm(subset(SOTUCorpus, year>1960), stopwords=TRUE, stem=TRUE)
presDfm <- trimdfm(presDfm, minCount=5, minDoc=3)
presLDA <- textmodel_lda(presDfm, k=10)
require(topicmodels)  # need this to access methods below
terms(presLDA, k=10)  # top 10 terms in each topic
topics(presLDA)       # dominant topics for each document
presCTM <- textmodel_lda(presDfm, model="ctm", k=10)
terms(presCTM, k=10)  # top 10 terms in each topic
topics(presCTM)       # dominant topics for each document

# fit a structural topic model
if (require(stm)) {
    gadarianCorpus <- corpus(gadarian$open.ended.response, docvars=gadarian[, 1:3],
                             source="From stm package, data from Gadarian and Albertson (forthcoming)")
    gadarianDfm <- dfm(gadarianCorpus, stopwords=TRUE, stem=TRUE)
    gadarianSTM <- textmodel_lda(gadarianDfm, "stm", k=3,
                                 prevalence = ~treatment + s(pid_rep),
                                 data = docvars(gadarianCorpus))
    summary(gadarianSTM)
}
}
\author{
Kenneth Benoit
}
\references{
Blei, David M., Andrew Y. Ng, and Michael I. Jordan. 2003. "Latent
  Dirichlet Allocation." \emph{The Journal of Machine Learning Research} 3:
  993–1022.

  Blei, David M, and John D. Lafferty. 2007. "A Correlated Topic Model of
  Science." \emph{The Annals of Applied Statistics} 1(1): 17-35.

  Roberts, M., Stewart, B., Tingley, D., and Airoldi, E. (2013) "The
  structural topic model and ap- plied social science." In \emph{Advances in Neural
  Information Processing Systems Workshop on Topic Models: Computation,
  Application, and Evaluation}. \url{http://goo.gl/uHkXAQ}
}

