% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel_lsa.R
\name{textmodel_lsa-postestimation}
\alias{textmodel_lsa-postestimation}
\alias{predict.textmodel_lsa}
\alias{as.dfm.textmodel_lsa}
\alias{coef.textmodel_lsa}
\alias{coefficients.textmodel_lsa}
\title{Post-estimations methods for textmodel_lsa}
\usage{
\method{predict}{textmodel_lsa}(object, newdata = NULL, ...)

\method{as.dfm}{textmodel_lsa}(x)

\method{coef}{textmodel_lsa}(object, doc_dim = 1, feat_dim = 1, ...)

coefficients.textmodel_lsa(object, doc_dim = 1, feat_dim = 1, ...)
}
\arguments{
\item{object, x}{previously fitted \link{textmodel_lsa} object}

\item{newdata}{new matrix to be transformed into the lsa space}

\item{...}{unused}

\item{doc_dim, feat_dim}{the document and feature dimension scores to be
extracted}
}
\value{
\code{predict()} returns a predicted \link{textmodel_lsa} object, projecting the patterns onto
new data.

\code{coef.textmodel_lsa} extracts model coefficients from a fitted
\link{textmodel_ca} object.
}
\description{
Post-estimation methods for fitted \link{textmodel_lsa} objects.
}
\keyword{internal}
\keyword{textmodel}
