% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel_wordscores.R
\name{textmodel_wordscores}
\alias{textmodel_wordscores}
\alias{predict.textmodel_wordscores}
\title{Wordscores text model}
\usage{
textmodel_wordscores(x, y, scale = c("linear", "logit"), smooth = 0)

\method{predict}{textmodel_wordscores}(object, newdata = NULL,
  se.fit = FALSE, interval = c("none", "confidence"), level = 0.95,
  rescaling = c("none", "lbg", "mv"), ...)
}
\arguments{
\item{x}{the \link{dfm} on which the model will be trained}

\item{y}{vector of training scores associated with each document 
in \code{x}}

\item{scale}{scale on which to score the words; \code{"linear"} for classic 
LBG linear posterior weighted word class differences, or \code{"logit"}
for log posterior differences}

\item{smooth}{a smoothing parameter for word counts; defaults to zero for the
to match the LBG (2003) method.}

\item{object}{a fitted Wordscores textmodel}

\item{newdata}{dfm on which prediction should be made}

\item{se.fit}{if \code{TRUE}, return standard errors as well}

\item{interval}{type of confidence interval calculation}

\item{level}{tolerance/confidence level for intervals}

\item{rescaling}{\code{"none"} for "raw" scores; \code{"lbg"} for LBG (2003) 
rescaling; or \code{"mv"} for the rescaling proposed by Martin and Vanberg 
(2007).  See References.}

\item{...}{not used}
}
\value{
\code{textmodel_wordscores()} returns a list that is also classed as a
  \code{textmodel_wordscores} object, containing the following elements:

\item{wordscores}{the scores computed for each word in the training set
(\eqn{S_{wd}} from Laver, Benoit and Garry 2003)}

\item{scale}{either \code{linear} or \code{logit}, according to the value of \code{scale}}

\item{x}{the dfm on which the wordscores model was called}

\item{y}{the vector of document reference values}

\item{call}{the function call that fitted the model}

\code{predict.textmodel_wordscores()} returns a named vector of predicted
document scores ("text scores" \eqn{S_{vd}} in LBG 2003), or a named list if
\code{se.fit = TRUE} consisting of the predicted scores (\code{$fit}) and the
associated standard errors (\code{$se.fit}). When \code{interval =
"confidence"}, the predicted values will be a matrix.  This behaviour matches
that of \code{\link[stats]{predict.lm}}.
}
\description{
\code{textmodel_wordscores} implements Laver, Benoit and Garry's (2003)
"Wordscores" method for scaling texts on a single dimension, given a set of
anchoring or \emph{reference} texts whose values are set through reference
scores. This scale can be fitted in the linear space (as per LBG 2003) or in
the logit space (as per Beauchamp 2012).  Estimates of \emph{virgin} or
unknown texts are obtained using the \code{predict()} method to score
documents from a fitted \code{textmodel_wordscores} object.
}
\details{
The \code{textmodel_wordscores()} function and the associated
  \code{predict()} method are designed to function in the same manner as
  \code{\link[stats]{predict.lm}}.  \code{coef()} can also be used to extract
  the word coefficients from the fitted \code{textmodel_wordscore} object,
  and \code{summary()} will print a nice summary of the fitted object.
}
\examples{
(ws <- textmodel_wordscores(data_dfm_lbgexample, c(seq(-1.5, 1.5, .75), NA)))
summary(ws)
coef(ws)
predict(ws)
predict(ws, rescaling = "mv")
predict(ws, rescaling = "lbg")
predict(ws, se.fit = TRUE)
predict(ws, se.fit = TRUE, interval = "confidence")
predict(ws, se.fit = TRUE, interval = "confidence", rescaling = "mv")
}
\references{
Laver, Michael, Kenneth R Benoit, and John Garry. 2003. 
  "\href{http://www.kenbenoit.net/pdfs/WORDSCORESAPSR.pdf}{Extracting Policy Positions From Political Texts Using Words as Data.}" 
  \emph{American Political Science Review} 97(02): 311-31
  
  Beauchamp, N. 2012. "Using Text to Scale Legislatures with Uninformative 
  Voting." New York University Mimeo.
  
  Martin, L W, and G Vanberg. 2007. "A Robust Transformation Procedure for 
  Interpreting Political Text." \emph{Political Analysis} 16(1): 93-100.
}
\author{
Kenneth Benoit
}
\keyword{textmodel}
