% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct-functions.R
\name{textmodel_wordshoal}
\alias{textmodel_wordshoal}
\title{Wordshoal text model (redirect)}
\description{
This model has been moved to the \href{https://github.com/kbenoit/wordshoal}{\pkg{wordshoal}} package.
}
\references{
Benjamin E. Lauderdale and Alexander Herzog.  2016.
  "\href{https://www.cambridge.org/core/journals/political-analysis/article/measuring-political-positions-from-legislative-speech/35D8B53C4B7367185325C25BBE5F42B4}{Measuring
   Political Positions from Legislative Speech}." \emph{Political Analysis}
  24 (3, July): 374-394.
}
\keyword{internal}
\keyword{textmodel}
