% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel.R, R/textstat.R
\name{textmodels}
\alias{textmodels}
\title{Models for scaling and classification of textual data}
\description{
The \verb{textmodel_*()} functions formerly in \pkg{quanteda} have now been moved
to the \pkg{quanteda.textmodels} package.

The \verb{textstat_*()} functions formerly in \pkg{quanteda} have now been moved
to the \pkg{quanteda.textstats} package.
}
\seealso{
\code{quanteda.textmodels::quanteda.textmodels-package}

\code{quanteda.textstats::quanteda.textstats-package}
}
