% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textplot_frequency.R
\name{textplot_frequency}
\alias{textplot_frequency}
\title{plot the frequency of a feature across documents}
\usage{
textplot_frequency(x, feature, type = "frequency", sort = TRUE,
  doclabels = docnames(x))
}
\arguments{
\item{x}{a dfm object}

\item{feature}{the feature whose frequency will be plotted}

\item{type}{the frequency type to be plotted. Options are "frequency", 
"relFreq", "relMaxFreq", "logFreq", "tfidf". See \code{\link{dfm_weight}} 
for legal values.}

\item{sort}{logical; if \code{TRUE}, sort the documents on the x-axis
according to the frequency of the feature}

\item{doclabels}{a vector of names for document; if left \code{NULL} (the
default), \code{docnames(x)} will be used}
}
\value{
a \pkg{ggplot2} object
}
\description{
Plots the absolute or relative frequency of a feature in all documents of a 
document-feature matrix. Given that this returns a ggplot object, you can 
modify the plot by adding ggplot layers (see example).
}
\examples{
\dontrun{
doclab <- apply(docvars(data_corpus_irishbudget2010, c("name", "party")), 
                1, paste, collapse = " ")
textplot_frequency(x = dfm(data_corpus_irishbudget2010),
                   feature = "tax", sort = TRUE, doclabels = doclab)
textplot_frequency(x = dfm(data_corpus_irishbudget2010),
                   feature = "tax", sort = FALSE, doclabels = doclab)
textplot_frequency(x = dfm(data_corpus_irishbudget2010), type = "logFreq",
                   feature = "tax", doclabels = doclab)
                  
data_corpus_inauguralPost80 <- corpus_subset(data_corpus_inaugural, Year > 1980)
textplot_frequency(dfm(data_corpus_inauguralPost80), "america", sort = TRUE)
}
}
\seealso{
\code{\link{dfm}}, \code{\link{dfm_weight}}
}
\author{
Stefan Müller
}
\keyword{plot}
