% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textplot_keyness.R
\name{textplot_keyness}
\alias{textplot_keyness}
\title{plot word keyness}
\usage{
textplot_keyness(x, sort = TRUE, show_reference = FALSE, n = 20)
}
\arguments{
\item{x}{a return object from \code{\link{textstat_keyness}}}

\item{sort}{logical; if \code{TRUE} (the default), order points from low to high 
score.}

\item{show_reference}{logical; if \code{TRUE}, show key reference features in addition to key target features}

\item{n}{number of terms to plot}
}
\value{
a \pkg{ggplot2} object
}
\description{
Plot the results of a keyword analysis using \code{\link{textstat_keyness}}.
}
\examples{
\dontrun{
# compare pre- v. post-war terms using grouping
period <- ifelse(docvars(data_corpus_inaugural, "Year") < 1945, "pre-war", "post-war")
mydfm <- dfm(data_corpus_inaugural, groups = period)
head(mydfm) # make sure 'post-war' is in the first row
top20 <- head(result <- textstat_keyness(mydfm), 20)
tail10 <- tail(result, 10)

# plot estimated word keyness
textplot_keyness(top20) 
textplot_keyness(result, show_reference = TRUE)
}
}
\seealso{
\code{\link{textstat_keyness}}
}
\author{
Haiyan Wang
}
\keyword{textplot}
