% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textplot_network.R
\name{textplot_network}
\alias{textplot_network}
\title{plot a semantic network of feature co-occurrences}
\usage{
textplot_network(x, color = "sky blue", ignore = 0.5,
  omit_isolated = TRUE)
}
\arguments{
\item{x}{a \link{dfm} object}

\item{color}{needs description}

\item{ignore}{needs description}

\item{omit_isolated}{needs description}
}
\description{
Plot an \link{fcm} object as a network, where vertices are features and edges
are co-occurrences.
}
\examples{
\dontrun{
toks <- corpus_subset(data_corpus_irishbudget2010) \%>\%
    tokens(remove_punct = TRUE) \%>\%
    tokens_tolower() \%>\%
    tokens_remove(stopwords("english"), padding = FALSE)
myfcm <- fcm(toks, context = "window", tri = FALSE)
feat <- names(topfeatures(myfcm, 30))
fcm_select(myfcm, feat, verbose = FALSE) \%>\% textplot_network(ignore = 0.5)
fcm_select(myfcm, feat, verbose = FALSE) \%>\% textplot_network(ignore = 0.8)
}
}
\seealso{
\code{\link{fcm}}
}
\author{
Kohei Watanabe and Stefan Müller
}
\keyword{textplot}
