% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textplot_topfeatures.R
\name{textplot_topfeatures}
\alias{textplot_topfeatures}
\title{plot the most frequent features in a dfm}
\usage{
textplot_topfeatures(x, n, sort = TRUE, angle = 90)
}
\arguments{
\item{x}{a dfm object}

\item{n}{how many top features should be returned}

\item{sort}{whether to sort the documents on the x-axis according to the frequency}

\item{angle}{The angle of the document labels on the x-axis}
}
\value{
a **ggplot2** object
}
\description{
Plots the most frequent features in a dfm in descending order. Given that this returns a ggplot object, you can modify the plot by adding ggplot layers (see example).
}
\examples{
\dontrun{
dfm_inaugural <- dfm(data_corpus_inaugural, remove = stopwords("english"), 
                    remove_punct = TRUE)
textplot_topfeatures(x = dfm_inaugural, n = 50, angle = 45)
}
}
\seealso{
\code{\link{dfm}}, \code{\link{dfm_weight}}
}
\author{
Stefan Müller
}
\keyword{plot}
