% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textplot.R
\name{textplots}
\alias{textplots}
\title{Plots for textual data}
\description{
The \verb{textplot_*()} functions formerly in \pkg{quanteda} have now been moved
to the \pkg{quanteda.textplots} package.
}
\seealso{
\code{quanteda.textplots::quanteda.textplots-package}
}
