% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textstat_entropy.R
\name{textstat_entropy}
\alias{textstat_entropy}
\title{Calculate entropy
Calculate entropy for a dfm or fcm object.}
\usage{
textstat_entropy(x, selection = NULL, margin = c("documents",
  "features"), method = c("ml", "mm", "jeffreys", "laplace", "sg",
  "minimax", "cs", "nsb", "shrink"), unit = c("nat", "log", "log2",
  "log10"))
}
\arguments{
\item{x}{an input \link{dfm} object}

\item{selection}{a subset of \link{dfm} on which entropy is supposed to be computed}

\item{margin}{a character indicating on which margin to compress a dfm, either "documents" or "features"}

\item{method}{a character defining the entropy measure to be calculated}

\item{unit}{a character vector defining the unit in which entropy is measured. The default is "nats" (natural units). For computing entropy in "bits" set unit="log2".}
}
\description{
Calculate entropy
Calculate entropy for a dfm or fcm object.
}
