% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textstat-keyness.R
\name{textstat_keyness}
\alias{textstat_keyness}
\title{calculate keyness statistics}
\usage{
textstat_keyness(x, ...)
}
\arguments{
\item{x}{a dfm in which the first row is the target and the second is the reference document.}

\item{...}{not used}
}
\value{
a named vector of keyness scores
}
\description{
calculate keyness statistics
}
\examples{
period <- as.factor(ifelse(docvars(data_corpus_inaugural)$Year < 1945, 'post-war', 'pre-war'))
mydfm <- dfm(data_corpus_inaugural, groups = period)
mydfm <- dfm_select(mydfm, min_count = 2)
mydfm <- dfm_smooth(mydfm, smoothing = 1)
rownames(mydfm) # make sure 'post-war' is in the first row
textstat_keyness(mydfm)
}
\references{
Bondi, Marina, and Mike Scott, eds. Keyness in Texts. Amsterdam, 
  Philadelphia: John Benjamins, 2010.
  Stubbs, Michael. ‘Three Concepts of Keywords’. In Keyness in Texts, edited 
  by Marina Bondi and Mike Scott, 21–42. Amsterdam, Philadelphia: John Benjamins, 2010.
}

