% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textstat_simil.R
\docType{class}
\name{textstat_proxy-class}
\alias{textstat_proxy-class}
\alias{textstat_dist-class}
\alias{textstat_dist_symm-class}
\alias{textstat_simil-class}
\alias{textstat_simil_symm-class}
\alias{textstat_simil_sparse-class}
\alias{textstat_simil_symm_sparse-class}
\alias{show,textstat_proxy-method}
\title{textstat_simil/dist classes}
\usage{
\S4method{show}{textstat_proxy}(object)
}
\arguments{
\item{object}{the textstat_proxy object to be printed}
}
\description{
Sparse classes for similarity and distance matrices created by
\code{\link[=textstat_simil]{textstat_simil()}} and \code{\link[=textstat_dist]{textstat_dist()}}.

Print/show method for objects created by \code{textstat_simil} and
\code{textstat_dist}.
}
\section{Slots}{

\describe{
\item{\code{.Data}}{a sparse \pkg{Matrix} object, symmetric if selection is
\code{NULL}}

\item{\code{method}}{the method used for computing similarity or distance}

\item{\code{min_simil}}{numeric; a threshold for the similarity values below which similarity
values are not computed}

\item{\code{margin}}{identifies the margin of the dfm on which similarity or
difference was computed:  \code{"documents"} for documents or
\code{"features"} for word/term features.}

\item{\code{type}}{either \code{"textstat_simil"} or \code{"textstat_dist"}}

\item{\code{selection}}{target units, if any}
}}

\seealso{
\code{\link[=textstat_simil]{textstat_simil()}}
}
\keyword{internal}
\keyword{textstat}
