% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textstat_simil.R
\name{textstat_proxy}
\alias{textstat_proxy}
\title{[Experimental] Compute document/feature proximity}
\usage{
textstat_proxy(x, selection = NULL, margin = c("documents",
  "features"), method = c("cosine", "correlation", "jaccard", "ejaccard",
  "dice", "edice", "hamman", "simple matching", "faith", "euclidean",
  "chisquared", "hamming", "kullback", "manhattan", "maximum", "canberra",
  "minkowski"), p = 2, min_proxy = NULL, rank = NULL)
}
\arguments{
\item{x}{a \link{dfm} object}

\item{selection}{a valid index for document or feature names from \code{x},
to be selected for comparison}

\item{margin}{identifies the margin of the dfm on which similarity or
difference will be computed:  \code{"documents"} for documents or
\code{"features"} for word/term features.}

\item{method}{method the similarity or distance measure to be used; see
Details.}

\item{p}{The power of the Minkowski distance.}

\item{min_proxy}{the minimum proximity value to be recoded.}

\item{rank}{an integer value specifying top-n most proximity values to be
recorded.}
}
\description{
This is an underlying function for \code{textstat_dist} and
\code{textstat_simil} but returns \code{TsparseMatrix}.
}
\seealso{
\code{\link{textstat_dist}}, \code{\link{textstat_simil}}
}
\keyword{internal}
