% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textstat_sentiment.R
\name{textstat_sentiment}
\alias{textstat_sentiment}
\title{Compute sentiment using a dictionary method}
\usage{
textstat_sentiment(
  x,
  pos = "positive",
  neg = "negative",
  neut = NULL,
  fun = sent_logit,
  ...
)
}
\arguments{
\item{x}{a \link{dfm} or \link{tokens} object containing sentiment types or features}

\item{pos, neg}{character; names of the dictionary keys or features
for positive and negative sentiment, respectively}

\item{neut}{character; (optional) name of the neutral sentiment feature or
key}

\item{fun}{function; the formula for computing sentiment, which must refer to
\code{pos}, \code{neg}, and (optionally) \code{neut}.  Default is the Lowe et al (2001)
"logit" scale which is the log of (positive / negative) counts.  See
\link{sentiment-functions} for details and for additional available functions,
as well as details on how to supply custom functions.}

\item{...}{additional arguments passed to \code{fun}}
}
\value{
a data.frame of sentiment scores
}
\description{
Compute sentiment scores from tokens or document-feature matrices, based on
assigned categories (types or features) of positive, negative, and neutral
sentiment. Several formulas are available, or the user can supply a new one.
}
\examples{
toks <- tail(data_corpus_inaugural, n = 5) \%>\%
    tokens(remove_punct = TRUE, remove_numbers = TRUE) \%>\%
    tokens_lookup(dictionary = data_dictionary_LSD2015, nomatch = "_other_")
textstat_sentiment(toks)
textstat_sentiment(toks, smooth = 1)
textstat_sentiment(toks, fun = sent_relpropdiff)

dfmat <- dfm(toks)
textstat_sentiment(dfmat)

# works with multiple types associated with each sentiment category
textstat_sentiment(toks, pos = c("positive", "neg_negative"),
                   neg = c("negative", "neg_positive"))

# with a user-supplied function
sent_fn <- function(x) (x[, "pos"] - x[, "neg"]) / rowSums(x) * 100
textstat_sentiment(toks, fun = sent_fn)
}
\references{
Lowe, W., Benoit, K. R., Mikhaylov, S., & Laver, M. (2011).
Scaling Policy Preferences from Coded Political Texts. \emph{Legislative Studies
Quarterly}, 36(1), 123–155.
\url{http://doi.org/10.1111/j.1939-9162.2010.00006.x}
}
