% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textstat_summary.R
\name{textstat_summary}
\alias{textstat_summary}
\title{Summarize a corpus}
\usage{
textstat_summary(x, cache = TRUE, ...)
}
\arguments{
\item{cache}{if \code{TRUE}, use internal cache from the second time}

\item{...}{additional arguments passed through to \code{\link[=dfm]{dfm()}}}

\item{object}{corpus to be summarized}
}
\description{
Displays information about a corpus, including attributes and metadata such
as date of number of texts, creation and source.
}
\examples{
corp <- data_corpus_inaugural
textstat_summary(corp, cache = TRUE)
toks <- tokens(corp)
textstat_summary(toks, cache = TRUE)
dfmat <- dfm(toks)
textstat_summary(dfmat, cache = TRUE)
}
\keyword{textstat}
