% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textplot.R, R/textstat.R
\name{textstats}
\alias{textstats}
\title{Plots for textual data}
\description{
The \verb{textplot_*()} functions formerly in \pkg{quanteda} have now been moved
to the \pkg{quanteda.textplots} package.

The \verb{textstat_*()} functions formerly in \pkg{quanteda} have now been moved
to the \pkg{quanteda.textstats} package.
}
\seealso{
\code{quanteda.textplots::quanteda.textplots-package}

\code{quanteda.textstats::quanteda.textstats-package}
}
