% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfm_weight.R
\name{tf}
\alias{tf}
\title{deprecated name for dfm_weight}
\usage{
tf(...)
}
\arguments{
\item{...}{arguments passed to \code{\link{dfm_weight}}}
}
\description{
Deprecated function name for \code{\link{dfm_weight}}.
}
\keyword{deprecated}
\keyword{internal}
