% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfm_weight.R
\name{tfidf}
\alias{tfidf}
\title{Deprecated form of \code{\link{dfm_tfidf}}}
\usage{
tfidf(...)
}
\arguments{
\item{...}{arguments passed to \code{\link{dfm_tfidf}}}
}
\description{
Deprecated function name for \emph{tf-idf} weighting of a
document-feature matrix.
}
\keyword{deprecated}
\keyword{dfm}
\keyword{internal}
