% Generated by roxygen2 (4.0.1): do not edit by hand
\name{tokenise}
\alias{tokenise}
\alias{tokenise.corpus}
\title{tokenise a corpus and store the results}
\usage{
tokenise(x, ...)

\method{tokenise}{corpus}(corp)
}
\arguments{
\item{corp}{Corpus to be tokenized}
}
\value{
\item{tokens}{A list consisting of the following:}

\item{$dfm}{A \link{dfm} document-feature matrix object created with \link{settings}.}

\item{$nwords}{A vector of token counts for each document.}

\item{$ntypes}{A vector of type counts for each document.}

\item{$nsents}{A vector of sentence counts for each document.}

\item{$nparagr}{A vector of paragraph counts for each document.}
}
\description{
Tokenize the corpus and store the results as a dfm, along with counts for the syllables of each token,
the number of sentences per text, and the number of paragraphs per text.
}
\examples{
mycorpus <- corpus(uk2010immig)
mycorpus
tokenise(mycorpus)
mycorpus
}

