% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tokenize_c}
\alias{tokenize_c}
\title{c tokenizing function}
\usage{
tokenize_c(x, sep = " ", simplify = FALSE, minLength = 1,
  toLower = TRUE, removeDigits = TRUE, removePunct = TRUE,
  removeTwitter = TRUE, removeURL = TRUE, removeAdditional = "")
}
\arguments{
\item{x}{string to be tokenized}

\item{sep}{by default, tokenize expects a "white-space" delimiter between
tokens. Alternatively, \code{sep} can be used to specify another character
which delimits fields.}
}
\description{
c tokenizing function
}
\author{
Kohei Watanabe
}

