% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R
\name{tokenize_dictionary}
\alias{tokenize_dictionary}
\title{Separate dictionary values using tokenizer}
\usage{
tokenize_dictionary(dictionary, ...)
}
\arguments{
\item{dictionary}{a \link{dictionary} object.}

\item{...}{passed to \code{\link[=tokens]{tokens()}}.}
}
\description{
Apply \code{\link[=tokens]{tokens()}} to dictionary values to improve pattern matching.
}
\keyword{development}
\keyword{dictionary}
\keyword{internal}
