% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R
\name{tokenize_dictionary_values}
\alias{tokenize_dictionary_values}
\title{Internal function to tokenize dictionary values}
\usage{
tokenize_dictionary_values(dict, separator, ...)
}
\arguments{
\item{dict}{list of object}
}
\description{
Internal function to tokenize dictionary values
}
\examples{
dict <- dictionary(list(ASIA = list("IN" = "\u5370\u5ea6",
                                    "ID" = "\u5370\u5ea6\u5c3c\u897f\u4e9a")))
quanteda:::tokenize_dictionary_values(dict, " ")
}
\keyword{internal}
