% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokens-methods-base.R
\name{tokens-class}
\alias{tokens-class}
\alias{print.tokens}
\alias{[.tokens}
\alias{+.tokens}
\alias{c.tokens}
\title{Base method extensions for tokens objects}
\usage{
\method{print}{tokens}(
  x,
  max_ndoc = quanteda_options("print_tokens_max_ndoc"),
  max_ntoken = quanteda_options("print_tokens_max_ntoken"),
  show_summary = quanteda_options("print_tokens_summary"),
  ...
)

\method{[}{tokens}(x, i, drop_docid = TRUE)

\method{+}{tokens}(t1, t2)

\method{c}{tokens}(...)
}
\arguments{
\item{x}{a tokens object}

\item{max_ntoken}{max number of tokens to print; default is from the
\code{print_tokens_max_ntoken} setting of \code{\link[=quanteda_options]{quanteda_options()}}}

\item{...}{not used}

\item{i}{document names or indices for documents to extract.}

\item{drop_docid}{if \code{TRUE}, \code{docid} for documents are removed as the result
of extraction.}

\item{t1}{tokens one to be added}

\item{t2}{tokens two to be added}
}
\value{
\code{c(...)} and \code{+} return a tokens object whose documents
have been added as a single sequence of documents.
}
\description{
Extensions of base R functions for tokens objects.
}
\examples{
toks <- tokens(c(d1 = "one two three", d2 = "four five six", d3 = "seven eight"))
str(toks)
toks[c(1,3)]
# combining tokens
toks1 <- tokens(c(doc1 = "a b c d e", doc2 = "f g h"))
toks2 <- tokens(c(doc3 = "1 2 3"))
toks1 + toks2
c(toks1, toks2)

}
\seealso{
\code{\link[=summary.tokens]{summary.tokens()}}
}
\keyword{internal}
\keyword{tokens}
