% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokens_chunk.R
\name{tokens_chunk2}
\alias{tokens_chunk2}
\alias{tokens_chunk2.tokens}
\title{Segment tokens object by chunks of a given size (alternative)}
\usage{
tokens_chunk2(x, size, discard_remainder = TRUE)

\method{tokens_chunk2}{tokens}(x, size, discard_remainder = TRUE)
}
\arguments{
\item{x}{\link{tokens} object whose token elements will be segmented into
chunks}

\item{size}{integer; the size of the chunks in tokens}

\item{discard_remainder}{logical; if \code{TRUE}, discard any ending chunk
whose length is less than \code{size}}
}
\value{
A \link{tokens} object whose documents have been split into chunks,
  similar to \code{\link{tokens_segment}}.
}
\description{
Segment tokens by splitting into equally sized chunks.
}
\examples{
txts <- c(doc1 = "Fellow citizens, I am again called upon by the voice of 
                  my country to execute the functions of its Chief Magistrate.",
          doc2 = "When the occasion proper for it shall arrive, I shall 
                  endeavor to express the high sense I entertain of this
                  distinguished honor.")
toks <- tokens(txts)
tokens_chunk(toks, size = 5)
}
\keyword{internal}
\keyword{tokens}
