% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hashTokens.R
\name{tokens_hashed_recompile}
\alias{tokens_hashed_recompile}
\title{recompile a hashed tokens object}
\usage{
tokens_hashed_recompile(x)
}
\arguments{
\item{x}{the \link[=hashTokens]{tokenizedTexts} object to be recompiled}
}
\description{
This function recompiles a hashed tokens object when the vocabulary has been changed in 
a way that makes some of its types identical, such as lowercasing when a lowercased 
version of the type already exists in the hash table.
}
\examples{
toksh <- hashTokens(tokenize(c(one = "a b c d A B C D",
                               two = "A B C d")))
vocabulary(toksh) <- toLower(vocabulary(toksh))
tokens_hashed_recompile(toksh)
}

