% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokens_select.R
\name{tokens_remove_padding}
\alias{tokens_remove_padding}
\title{remove padding from a tokens object}
\usage{
tokens_remove_padding(x)
}
\description{
Remove the 0s from a tokens object, if they exist.
}
\keyword{internal}
\keyword{tokens}

