% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokens_select_pos.R
\name{tokens_select_pos}
\alias{tokens_select_pos}
\title{select or remove tokens from a tokens object}
\usage{
tokens_select_pos(x, pattern, selection = c("keep", "remove"),
  valuetype = c("glob", "regex", "fixed"), case_insensitive = TRUE,
  padding = FALSE, verbose = quanteda_options("verbose"))
}
\arguments{
\item{x}{\link{tokens} object whose token elements will be selected}

\item{pattern}{a character vector, list of character vectors, \link{dictionary},
\link{collocations}, or \link{dfm}. See \link{pattern} for details.}

\item{selection}{whether to \code{"keep"} or \code{"remove"} the tokens 
matching \code{pattern}}

\item{valuetype}{the type of pattern matching: \code{"glob"} for 
"glob"-style wildcard expressions; \code{"regex"} for regular expressions;
or \code{"fixed"} for exact matching. See \link{valuetype} for details.}

\item{case_insensitive}{ignore case when matching, if \code{TRUE}}

\item{padding}{if \code{TRUE}, leave an empty string where the removed tokens
previously existed.  This is useful if a positional match is needed between
the pre- and post-selected tokens, for instance if a window of adjacency 
needs to be computed.}

\item{verbose}{if \code{TRUE} print messages about how many tokens were 
selected or removed}
}
\value{
a \link{tokens} object with tokens selected or removed based on their
  match to \code{pattern}
}
\description{
This function selects or discards tokens from a \link{tokens} objects, with 
the shortcut \code{tokens_remove(x, pattern)} defined as a shortcut for 
\code{tokens_select_pos(x, pattern, selection = "remove")}.  The most common 
usage for \code{tokens_remove} will be to eliminate stop words from a text or
text-based object, while the most common use of \code{tokens_select} will be 
to select tokens with only positive pattern matches from a list of regular 
expressions, including a dictionary.
}
