% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokens.R
\name{tokens_serialize}
\alias{tokens_serialize}
\title{Function to serialized list-of-character tokens}
\usage{
tokens_serialize(x, types_reserved, ...)
}
\arguments{
\item{x}{a list of character vectors}

\item{types_reserved}{optional pre-existing types for mapping of tokens}

\item{...}{additional arguments}
}
\value{
a list the serialized tokens found in each text
}
\description{
Creates a serialized object of tokens, called by \code{\link{tokens}}.
}
\keyword{internal}
\keyword{tokens}
