% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casechange-functions.R
\name{tokens_tolower}
\alias{tokens_tolower}
\alias{tokens_toupper}
\title{Convert the case of tokens}
\usage{
tokens_tolower(x, keep_acronyms = FALSE, verbose = quanteda_options("verbose"))

tokens_toupper(x, verbose = quanteda_options("verbose"))
}
\arguments{
\item{x}{the input object whose character/tokens/feature elements will be
case-converted}

\item{keep_acronyms}{logical; if \code{TRUE}, do not lowercase any
all-uppercase words (applies only to \verb{*_tolower()} functions)}

\item{verbose}{if \code{TRUE} print a summary of the object returned by the
function: the number of documents and, depending on the object, the number
of characters, tokens and types, or features and occurrences. The number of
tokens includes paddings, but the number of features excludes them.}
}
\description{
\code{tokens_tolower()} and \code{tokens_toupper()} convert the features of a
\link{tokens} object and re-index the types.
}
\examples{
# for a document-feature matrix
toks <- tokens(c(txt1 = "b A A", txt2 = "C C a b B"))
tokens_tolower(toks)
tokens_toupper(toks)
}
