% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directionchange-functions.R
\name{tokens_tortl}
\alias{tokens_tortl}
\alias{char_tortl}
\title{[Experimental] Change direction of words in tokens}
\usage{
tokens_tortl(x)

char_tortl(x)
}
\arguments{
\item{x}{the input object whose punctuation marks will be modified by the
direction mark}
}
\description{
This function adds a Unicode direction mark to tokens types for punctuations
and symbols to correct how right-to-left languages (e.g. Arabic, Hebrew,
Persian, and Urdu) are printed in HTML-based consoles (e.g. R Studio). This
is an experimental function subject to future change.
}
\keyword{character}
\keyword{experimental}
\keyword{tokens}
