% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokens_xptr.R
\name{is.tokens_xptr}
\alias{is.tokens_xptr}
\alias{as.tokens_xptr}
\alias{as.tokens_xptr.tokens}
\alias{as.tokens_xptr.tokens_xptr}
\title{Convert a tokens object to a tokenx_xptr object}
\usage{
is.tokens_xptr(x)

as.tokens_xptr(x)

\method{as.tokens_xptr}{tokens}(x)

\method{as.tokens_xptr}{tokens_xptr}(x)
}
\arguments{
\item{x}{a \link{tokens} object to convert to \link{tokens_xptr} or a \link{tokens_xptr} to
to deep-copy.}
}
\description{
\link{tokenx_xptr} is a external pointer object to process a large number of
documents efficiently.
}
