% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toUpper.R
\name{toUpper}
\alias{toUpper}
\alias{toUpper.NULL}
\alias{toUpper.character}
\alias{toUpper.corpus}
\alias{toUpper.tokenizedTexts}
\title{Convert texts to upper case}
\usage{
toUpper(x, keepAcronyms = FALSE, ...)

\method{toUpper}{character}(x, keepAcronyms = FALSE, ...)

\method{toUpper}{NULL}(x, ...)

\method{toUpper}{tokenizedTexts}(x, keepAcronyms = FALSE, ...)

\method{toUpper}{corpus}(x, keepAcronyms = FALSE, ...)
}
\arguments{
\item{x}{texts to be upper-cased}

\item{keepAcronyms}{if \code{TRUE}, do not uppercase any all-uppercase words}

\item{...}{additional arguments passed to \pkg{stringi} functions, (e.g. 
\code{\link{stri_trans_toupper}}), such as \code{locale}}
}
\value{
Texts tranformed into their uppercased versions.  If \code{x} is a 
  character vector or a corpus, return a uppercased character vector.  If 
  \code{x} is a list of tokenized texts, then return a list of upper-cased 
  tokenized texts.
}
\description{
Convert texts or tokens to upper case
}
\examples{
test1 <- c(text1 = "England and France are members of NATO and UNESCO", 
           text2 = "NASA sent a rocket into space.")
toUpper(test1)
toUpper(test1, keepAcronyms = TRUE)

test2 <- tokenize(test1, removePunct=TRUE)
toUpper(test2)
toUpper(test2, keepAcronyms = TRUE)
}

