% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postag.R
\name{trainHMM}
\alias{trainHMM}
\title{train hidden Markov model(HMM)}
\usage{
trainHMM(labeled, alpha = 0.1)
}
\arguments{
\item{labeled}{a list of sentences where each element of a sentence is a list
consisting of token and tag}

\item{alpha}{adjustable parameter used for smoothing (alpha must be between 0 
and 1)}
}
\value{
A list of two data.tables with token observation and tag transition 
probabilities
}
\description{
Train hidden Markov model(HMM) which maximises the joint probability of
token and tag sequence. This is done by aggregating the frequencies of tags
being assigned to a specific token and the frequencies of transitions between
tags. The function takes labelled list of sentences as an input, where each
element of a sentence is a list consisting of token and tag.
}

