% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/textmodel-utils.R
\name{transformOrdinal}
\alias{transformOrdinal}
\title{Transform a list of ordinal target labels to a numeric vector}
\usage{
transformOrdinal(categories, values = NULL)
}
\arguments{
\item{categories}{A character vector of ordinal target values}

\item{values}{An optional list of numbers to map the ordinal values on to}
}
\value{
A numeric vector of the same length as categories
}
\description{
Returns a numeric vector for use as target variable for a textmodel that
requires numeric response variables. This is appropriate for ordinal or
two-class classification, but not when the targets are non-transitive
}
\examples{
cats <- docvars(inaugCorpus, 'Year')
}
\author{
Paul Nulty
}

