% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfm-deprecated.R
\name{trim}
\alias{trim}
\alias{trim.dfm}
\alias{trimdfm}
\title{deprecated name for dfm_trim}
\usage{
trim(x, ...)

\method{trim}{dfm}(x, ...)

trimdfm(x, ...)
}
\arguments{
\item{x}{dfm whose features or documents will be trimmed}

\item{...}{arguments passed to \code{dfm_trim}}
}
\description{
Deprecated function name for \code{\link{dfm_trim}}.
}
\note{
Deprecated; use \code{\link{dfm_trim}} instead.
}
\keyword{deprecated}
\keyword{internal}
