% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{ukimmigrTexts}
\alias{ukimmigrTexts}
\title{Immigration-related sections of 2010 UK party manifestos}
\format{A named character vector of plain ASCII texts}
\description{
Extracts from the election manifestos of 9 UK political parties from 2010, related
to immigration or asylum-seekers.
}
\examples{
data(ukimmigrTexts)
ukimmigrCorpus <- corpus(ukimmigrTexts, docvars=list(party=names(ukimmigrTexts)))
language(ukimmigrCorpus) <- "english"
encoding(ukimmigrCorpus) <- "UTF-8"
summary(ukimmigrCorpus)
}

