% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocs.R
\docType{data}
\name{ukimmigTexts}
\alias{ukimmigTexts}
\title{Immigration-related sections of 2010 UK party manifestos}
\format{A named character vector of plain ASCII texts}
\description{
Extracts from the election manifestos of 9 UK political parties from 2010, related
to immigration or asylum-seekers.
}
\examples{
ukimmigCorpus <- corpus(ukimmigTexts, docvars=data.frame(party=names(ukimmigTexts)))
metadoc(ukimmigCorpus, "language") <- "english"
summary(ukimmigCorpus, showmeta = TRUE)
}

