\docType{data}
\name{UKManifestos}
\alias{UKManifestos}
\title{A corpus object containing 105 UK Manifestos}
\description{
A corpus object containing 105 UK Manifestos from
1945-2005, with party and year attributes
}
\references{
As used in Laver, Michael. 1998a. \'Party Policy in
Britain, 1997: Results from an Expert Survey.\' Political
Studies 46: 336–47.
}

