% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern2fixed.R
\name{unlist_character}
\alias{unlist_character}
\title{Unlist a list of character vectors safely}
\usage{
unlist_character(x, unique = FALSE, ...)
}
\arguments{
\item{x}{a list of integers}

\item{unique}{if \code{TURE} remove duplicated elements}

\item{...}{passed to \code{unlist}}
}
\value{
character vector
}
\description{
Unlist a list of character vectors safely
}
\keyword{internal}
