% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern2fixed.R
\name{unlist_integer}
\alias{unlist_integer}
\title{Unlist a list of integer vectors safely}
\usage{
unlist_integer(x, unique = FALSE, ...)
}
\arguments{
\item{x}{a list of integers}

\item{unique}{if \code{TURE} remove duplicated elements}

\item{...}{passed to \code{unlist}}
}
\value{
integer vector
}
\description{
Unlist a list of integer vectors safely
}
\keyword{internal}
