% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unused_dots}
\alias{unused_dots}
\title{Raise warning of unused dots}
\usage{
unused_dots(...)
}
\arguments{
\item{...}{dots to check}

\item{fun_name}{name of the function in which dots are not used}
}
\description{
Raise warning of unused dots
}
\keyword{internal}
