% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textstat_sentiment.R
\name{valence}
\alias{valence}
\alias{valence<-}
\title{Set or get the valences of dictionary values or keys}
\usage{
valence(x)

valence(x) <- value
}
\arguments{
\item{x}{a \link{dictionary} object}

\item{value}{named list consisting of numerical value.  The names of the
elements must correspond to a dictionary key. Each element must be:
\itemize{
\item a single numeric value that will be applied to all of the dictionary
values in that key; or
\item a vector of numeric values that matches the length and order of the
dictionary values in that key; or
\item a named numeric vector where each element name matches dictionary values
in the key.
}}
}
\value{
\code{valences()} returns the valences as a list named numeric vectors,
where each list element corresponds to a key in the dictionary, and each
numeric element matches a value within that key.

\verb{valence<-} sets the dictionary's valences.
}
\description{
Set or retrieve the valences of a \link{dictionary} object for the purposes of
sentiment analysis.  Valences consist of numerical values attached to each
dictionary "value".  For dictionaries with a more "polarity"-based approach,
a key may have the same valence for every value in that key -- for instance,
-1 for all values in a "negative" key.
}
\details{
Valences are used only in \code{\link[=textstat_sentiment]{textstat_sentiment()}}.

A dictionary may have only one set of valences at a time, but may be
changed as needed.
}
\examples{
# setting valences
dict <- dictionary(list(
    happiness = c("happy", "jubilant", "exuberant", "content"),
    anger = c("mad", "peeved", "irate", "furious", "livid")
))
valence(dict)
# using a 5-point scale
valence(dict) <- list(happiness = c(3, 4, 5, 2),
                      anger = c(3.1, 2.4, 2.9, 4.1, 5.0))
valence(dict)
}
\seealso{
\code{\link[=textstat_sentiment]{textstat_sentiment()}}, \code{\link[=polarity]{polarity()}}
}
\keyword{dictionary}
\keyword{textstat}
\keyword{utility}
