% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex2fixed5.R
\name{vector2list}
\alias{vector2list}
\title{convert a vector to a list}
\usage{
vector2list(x)
}
\arguments{
\item{x}{character (vector)}
}
\description{
Utility function to convert a vector to a list, used by \code{link{kwic}}.
}
\examples{

}

