% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/View.R
\name{View}
\alias{View}
\alias{View.default}
\alias{View.kwic}
\alias{View.dfm}
\title{View methods for quanteda}
\usage{
View(x, title)

\method{View}{default}(x, title)

\method{View}{kwic}(x, title)

\method{View}{dfm}(x, title)
}
\arguments{
\item{x}{an R object which can be coerced to a data frame with non-zero numbers of rows and columns}

\item{title}{for viewer window. Defaults to name of x prefixed by \verb{Data:}}
}
\description{
Redefines utils::View() as a generic, so that we can define methods for
kwic to produce a nice viewer.
}
\examples{
\dontrun{
## for kwic
View(kwic(data_char_ukimmig2010, "illeg*"))

## for data.frame
View(dfm(data_char_ukimmig2010))
}
}
\seealso{
\link[utils:View]{View}
}
\keyword{View}
\keyword{internal}
