% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postag.R
\name{viterbi}
\alias{viterbi}
\title{tag text using Viterbi algorithm}
\usage{
viterbi(txt, transition = NULL, observation = NULL)
}
\arguments{
\item{txt}{character string}

\item{transition}{a table of tag transition probabilities created with 
\code{train()} function.}

\item{observation}{a table of state observation likelihoods created with
\code{train()} funcion}
}
\value{
A vector of part-of-speech tags
}
\description{
Tag text using hidden Markov model(HMM) with Viterbi algorithm for decoding. 
Use either pre-trained models from quantedaData package or train your own model
with \code{train()} function.
}
\references{
Jurafsky, D. & Martin J. (2009). Speech and Language Processing.
}

