% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textplot_wordcloud.R
\name{wordcloud}
\alias{wordcloud}
\title{Internal function for textplot_wordcloud}
\usage{
wordcloud(x, scale = c(4, 0.5), max_words = Inf, random_order = FALSE,
  random_color = FALSE, rot_per = 0.1, colors = "black",
  ordered_colors = FALSE, fixed_asp = TRUE, min.freq, max.words,
  random.order, random.color, rot.per, ordered.colors, use.r.layout, fixed.asp,
  ...)
}
\arguments{
\item{x}{a dfm object}

\item{scale}{vector of length 2 indicating the range of the size of the
words}

\item{max_words}{maximum number of words to be plotted. least frequent terms
dropped}

\item{random_order}{plot words in random order. If \code{FALSE}, they will be
plotted in decreasing frequency}

\item{random_color}{choose colors randomly from the colors. If \code{FALSE},
the color is chosen based on the frequency}

\item{rot_per}{proportion words with 90 degree rotation colors}

\item{colors}{words from least to most frequent}

\item{ordered_colors}{if \code{TRUE}, then colors are assigned to words in
order}

\item{min.freq}{deprecated argument}

\item{max.words}{deprecated argument}

\item{random.order}{deprecated argument}

\item{random.color}{deprecated argument}

\item{rot.per}{deprecated argument}

\item{ordered.colors}{deprecated argument}

\item{use.r.layout}{deprecated argument}

\item{fixed.asp}{deprecated argument}

\item{...}{additional parameters passed to \link{text} (and \link{strheight},
\link{strwidth})}
}
\description{
This function impliments wordcloud without dependecies. Code is adopted from 
\code{\link[wordcloud]{wordcloud}}.
}
\author{
Ian Fellows
}
\keyword{internal}
