% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textplot_wordcloud.R
\name{wordcloud_comparison}
\alias{wordcloud_comparison}
\title{Internal function for textplot_wordcloud}
\usage{
wordcloud_comparison(x, scale = c(4, 0.5), max_words = Inf,
  random_order = FALSE, rot_per = 0.1, colors = NULL, title_size = 3,
  min.freq, max.words, random.order, rot.per, use.r.layout, title.size, ...)
}
\arguments{
\item{x}{a dfm object}

\item{scale}{vector of length 2 indicating the range of the size of the
words}

\item{max_words}{maximum number of words to be plotted. least frequent terms
dropped}

\item{random_order}{plot words in random order. If \code{FALSE}, they will be
plotted in decreasing frequency}

\item{rot_per}{proportion words with 90 degree rotation colors}

\item{min.freq}{deprecated argument}

\item{max.words}{deprecated argument}

\item{random.order}{deprecated argument}

\item{rot.per}{deprecated argument}

\item{use.r.layout}{deprecated argument}

\item{title.size}{deprecated argument}

\item{...}{additional parameters passed to \link{text} (and \link{strheight},
\link{strwidth})}

\item{random.color}{deprecated argument}

\item{ordered.colors}{deprecated argument}
}
\description{
This function impliments wordcloud that compares documents. Code is adopted from
\code{\link[wordcloud]{comparison.cloud}}.
}
\author{
Ian Fellows
}
\keyword{internal}
