% Generated by roxygen2 (4.0.1): do not edit by hand
\name{wordcloudDfm}
\alias{wordcloudDfm}
\title{Plot a word cloud for a \link{dfm}}
\usage{
wordcloudDfm(dfm, doc.index, ...)
}
\arguments{
\item{dfm}{document-feature matrix created in quanteda}

\item{document}{index of the document whose words will be plotted}

\item{...}{additional arguments to pass to \link[wordcloud]{wordcloud}}
}
\value{
None
}
\description{
plots a document as a wordcloud of its features
}
\examples{
data(iebudgets)
iebudgets2010 <- subset(iebudgets, year==2010)
wfm <- dfm(iebudgets2010, stopwords=TRUE)
wordcloudDfm(wfm, 1)  # plot the finance minister's speech as a wordcloud
}
\author{
Kenneth Benoit
}

