% Generated by roxygen2 (4.0.2): do not edit by hand
\name{wordscores_model}
\alias{wordscores_model}
\title{Wordscores model}
\usage{
wordscores_model(x, train, train.scores, scale = c("linear", "logit"),
  smooth = 0)
}
\arguments{
\item{train}{integer index of documents, or character vector of document
names.  For \code{scale="logit"}, only two documents can be specified}

\item{scale}{classic or logit scale of log posterior differences}

\item{smooth}{a smoothing parameter for word counts}
}
\description{
\code{wordscores_model} implements Laver, Benoit and Garry's (2003)
wordscores method for scaling of a single dimension
}
\author{
Kenneth Benoit
}
\references{
Laver, Benoit and Garry (2003)
}

