% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message.R
\name{wrap}
\alias{wrap}
\title{Wrap and print long lines}
\usage{
wrap(x, ...)
}
\arguments{
\item{x}{a character string to wrap.}

\item{...}{extra arguments passed to \link[stringi:stri_wrap]{stringi::stri_wrap}`.}
}
\description{
Wrap and print long lines
}
\keyword{development}
\keyword{internal}
