% Generated by roxygen2 (4.0.2): do not edit by hand
\name{zipfiles}
\alias{zipfiles}
\title{unzip a zipped collection of text files and return the directory}
\usage{
zipfiles(zfile = NULL, ...)
}
\arguments{
\item{zfile}{a character string specifying the name (including path) of the
zipped file, or a URL naming the file (see example); or NULL to use a GUI
to choose a file from disk}

\item{...}{additional arguments passed to \link{unzip}}
}
\value{
a \link{directory} class object containing the unzipped files
}
\description{
\code{zipfiles} extracts a set of text files in a zip archives, and returns
the name of the temporary directory where they are stored.  It can be passed
to \link{corpus.directory} for import.
}
\examples{
\dontrun{
# from a zip file on the web
myzipcorp <- corpus(zipfiles("http://kenbenoit.net/files/EUcoalsubsidies.zip"),
                    notes="From some EP debate about coal mine subsidies")
docvars(myzipcorp, speakername=docnames(myzipcorp))
summary(myzipcorp)

# call up interactive user input
myzipcorp <- corpus(zipfiles())
}
}

