% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus-methods-quanteda.R
\name{as.corpus}
\alias{as.corpus}
\title{coerce a compressed corpus to a standard corpus}
\usage{
as.corpus(x)
}
\arguments{
\item{x}{a compressed \link{corpus} object}
}
\description{
Recast a compressed corpus object into a standard (uncompressed) corpus
object.
}
\keyword{internal}
