% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcm-methods.R
\name{as.fcm}
\alias{as.fcm}
\title{Coercion functions for fcm objects}
\usage{
as.fcm(x)
}
\arguments{
\item{x}{an object coerced to \link{fcm}. Currently only support
\link{Matrix} objects.}
}
\description{
Coercion functions for fcm objects
}
\keyword{internal}
