% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textplot_network.R
\name{as.igraph}
\alias{as.igraph}
\title{Convert an fcm to an igraph object}
\usage{
as.igraph(x, ...)
}
\description{
Convert an \link{fcm} object to an \pkg{igraph} graph object.
}
\keyword{internal}
