% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfm-classes.R
\name{as.matrix.dfm}
\alias{as.matrix.dfm}
\title{Coerce a dfm to a matrix or data.frame}
\usage{
\method{as.matrix}{dfm}(x, ...)
}
\arguments{
\item{x}{dfm to be coerced}

\item{...}{unused}
}
\description{
Methods for coercing a \link{dfm} object to a matrix or data.frame object.
}
\examples{
# coercion to matrix
as.matrix(data_dfm_lbgexample[, 1:10])

}
\keyword{dfm}
