% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textstat_simil.R
\name{as.matrix.simil}
\alias{as.matrix.simil}
\title{Coerce a simil object into a matrix}
\usage{
\method{as.matrix}{simil}(x, Diag = 1L, ...)
}
\arguments{
\item{x}{a simil object from \code{\link{textstat_simil}}}

\item{Diag}{sets the value for matrix diagonal}

\item{...}{unused}
}
\description{
Coerces a simil object, which is a specially classed dist object, into a matrix.
}
\keyword{internal}
\keyword{textstat}
