% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel-methods.R
\name{as.statistics_textmodel}
\alias{as.statistics_textmodel}
\title{Coerce various objects to statistics_textmodel
This is a helper function used in \code{summary.textmodel_*}.}
\usage{
as.statistics_textmodel(x)
}
\arguments{
\item{x}{an object to be coerced}
}
\description{
Coerce various objects to statistics_textmodel
This is a helper function used in \code{summary.textmodel_*}.
}
\keyword{internal}
\keyword{textmodel}
