% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokens.R
\name{as.tokens}
\alias{as.tokens}
\alias{as.tokens.list}
\alias{as.list.tokens}
\alias{as.character.tokens}
\alias{is.tokens}
\title{coercion and checking functions for tokens objects}
\usage{
as.tokens(x)

\method{as.tokens}{list}(x)

as.list.tokens(x, ...)

\method{as.character}{tokens}(x, use.names = FALSE, ...)

is.tokens(x)
}
\arguments{
\item{x}{list of character elements}

\item{...}{unused}

\item{use.names}{logical; preserve names if \code{TRUE}.  For
\code{as.character} only.}
}
\value{
\code{as.tokens} returns a quanteda \link{tokens} object

\code{as.list} returns a simple list of characters from a
  \link{tokens} object

\code{as.character} returns a character vector from a 
  \link{tokens} object

\code{is.tokens} returns \code{TRUE} if the object is of class
  tokens, \code{FALSE} otherwise.
}
\description{
Coerce a list of character elements to a quanteda \link{tokens} object, or check whether 
an object is a \link{tokens} object.
}
