% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{attributes<-}
\alias{attributes<-}
\title{R-like alternative to reassign_attributes()}
\usage{
attributes(x, overwrite = TRUE) <- value
}
\arguments{
\item{x}{an object}

\item{overwrite}{if \code{TRUE}, overwrite old attributes}

\item{value}{new attributes}
}
\description{
R-like alternative to reassign_attributes()
}
\author{
Kohei Watanabe
}
\keyword{internal}
