% Generated by roxygen2 (4.0.2): do not edit by hand
\name{bigrams}
\alias{bigrams}
\title{Create bigrams}
\usage{
bigrams(text, window = 1, concatenator = "_", include.unigrams = FALSE,
  ...)
}
\arguments{
\item{text}{character vector containing the texts from which bigrams will be constructed}

\item{window}{how many words to be counted for adjacency.  Default is 1 for only immediately
neighbouring words.  This is only available for bigrams, not for \link{ngram}.}

\item{concatenator}{character for combining words, default is \code{_} (underscore) character}

\item{include.unigrams}{if \code{TRUE}, return unigrams as well}

\item{...}{additional arguments passed to \link{tokenize}}
}
\value{
a character vector of bigrams
}
\description{
Create bigrams
}
\examples{
bigrams("The quick brown fox jumped over the lazy dog.")
bigrams(c("The quick brown fox", "jumped over the lazy dog."))
bigrams(c("The quick brown fox", "jumped over the lazy dog."), window=2)
}
\author{
Ken Benoit and Kohei Watanabe
}

