% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfm.R
\docType{class}
\name{cfm-class}
\alias{cfm-class}
\title{Virtual class "cfm" for a context-feature matrix}
\description{
The cfm class of object is a special type of \link{dfm-class}.
}
\section{Slots}{

\describe{
\item{\code{context}}{the context definition}

\item{\code{window}}{the size of the window, if \code{context = "window"}}

\item{\code{weights}}{context weighting for distance from target feature, equal in length to \code{window}}
}}
\seealso{
\link{cfm}
}

