% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcm.R
\name{cfm}
\alias{cfm}
\alias{cfm.character}
\alias{cfm.corpus}
\alias{cfm.tokenizedTexts}
\title{create a context-feature co-occurrence matrix}
\usage{
cfm(x, ...)

\method{cfm}{tokenizedTexts}(x, context = c("document", "window"),
  window = 5L, weights = rep(1, length(window)), span_sentence = TRUE,
  tri = TRUE, ...)

\method{cfm}{corpus}(x, ...)

\method{cfm}{character}(x, ...)
}
\arguments{
\item{x}{character vector, corpus, or tokenized texts from which to generate 
the context-feature co-occurrence matrix.}

\item{...}{additional arguments passed to \link{tokenize}, which can include 
for instance \code{ngrams} and \code{concatenator} for tokenizing 
multi-token sequences}

\item{context}{the context in which to consider term co-occurrence: 
\code{"document"} for co-occurrence counts within document; \code{"window"}
for co-occurrence within a defined window of words, which requires a 
postive integer value for \code{window}}

\item{window}{positive integer value for the size of a window on either side 
of the target feature, default is 5, meaning 5 words before and after the 
target feature}

\item{weights}{a vector of weights applied to each distance from
\code{1:window}, strictly decreasing and of the same length as
\code{length(weights)}}

\item{span_sentence}{if \code{FALSE}, then word windows will not span
sentences}

\item{tri}{if \code{TRUE} return only upper triangle (including diagonal)}

\item{verbose}{if \code{TRUE} print status messages to the console}
}
\description{
Create a sparse context-feature co-occurrence matrix, measuring
co-occurrences of features within a user-defined context. The context can be
defined as a document or a window within a collection of documents, with an
optional vector of weights applied to the co-occurrence counts.
}
\examples{
# see http://bit.ly/29b2zOA
txt <- "A D C E A D F E B A C E D"
cfm(txt, context = "window", window = 2)

txts <- c("a a b b c", "a c e", "e f g")
cfm(txts, context = "document")

txt <- c("The quick brown fox jumped over the lazy dog.",
         "The dog jumped and ate the fox.")
cfm(toks, context = "document")
cfm(toks, context = "window", window = 3)
cfm(toks, context = "window", window = 2)
}
\author{
Kenneth Benoit (R) and Kohei Watanabe (C++)
}

