% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus_reshape.R
\name{changeunits}
\alias{changeunits}
\title{deprecated name for corpus_reshape}
\usage{
changeunits(...)
}
\arguments{
\item{...}{all arguments}
}
\description{
The deprecated function name for what is now \code{\link{corpus_reshape}}. 
Please use that instead.
}
\keyword{deprecated}
\keyword{internal}
