% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator.R
\name{check_class}
\alias{check_class}
\title{Check object class for functions}
\usage{
check_class(class, method, defunct_methods = NULL)
}
\arguments{
\item{class}{the object class to check}

\item{method}{the name of functions to be called}
}
\description{
Checks if the method is defined for the class.
}
\examples{
\dontrun{
quanteda:::check_class("tokens", "dfm_select")
}
}
\keyword{development}
\keyword{internal}
