% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_dots}
\alias{check_dots}
\title{Check arguments passed to other functions via ...}
\usage{
check_dots(..., method = NULL)
}
\arguments{
\item{...}{dots to check}

\item{method}{the ames of functions \code{...} is passed to}
}
\description{
Check arguments passed to other functions via ...
}
\keyword{internal}
