% Generated by roxygen2 (4.0.2): do not edit by hand
\name{clean}
\alias{clean}
\alias{clean.character}
\alias{clean.corpus}
\title{clean: simple pre-processing cleanup}
\usage{
clean(x, ...)

\method{clean}{character}(x, removeDigits = TRUE, removePunct = TRUE,
  lower = TRUE, ...)

\method{clean}{corpus}(x, removeDigits = TRUE, removePunct = TRUE,
  lower = TRUE, ...)
}
\arguments{
\item{x}{The object to be cleaned. Can be either a character vector or a
corpus object. If x is a corpus, \code{clean} returns a copy of the x with
the texts cleaned.}

\item{removeDigits}{Should digits be removed? TRUE or FALSE}

\item{removePunct}{Should punctuation be removed? TRUE or FALSE}

\item{lower}{Convert to lower case? TRUE or FALSE}
}
\description{
\code{clean} removes punctuation and digits from text, using the regex
character classes for punctuation and digits. \code{clean} uses the standard R
function \code{tolower} to convert the text to lower case. Each of these
steps is optional, but switched on by default, so for example, to remove
punctuation and convert to lower, but keep digits, the command would be:
\code{clean(mytexts, removeDigits=FALSE)}
}
\examples{
#convert a set of texts to lower case and remove
punctuation, keeping digits

clean(inaugTexts, removeDigits=FALSE)

# remove digits from a corpus
clean(inaugTexts, removePunct=FALSE, lower=FALSE)
}

