% Generated by roxygen2 (4.0.2): do not edit by hand
\name{clean}
\alias{clean}
\alias{clean.character}
\alias{clean.corpus}
\title{simple cleaning of text before processing}
\usage{
clean(x, ...)

\method{clean}{character}(x, removeDigits = TRUE, removePunct = TRUE,
  lower = TRUE, additional = NULL, twitter = TRUE, removeURL = TRUE,
  ...)

\method{clean}{corpus}(x, removeDigits = TRUE, removePunct = TRUE,
  lower = TRUE, additional = NULL, twitter = TRUE, ...)
}
\arguments{
\item{x}{The object to be cleaned. Can be either a character vector or a
corpus object. If x is a corpus, \code{clean} returns a copy of the x with
the texts cleaned.}

\item{...}{additional parameters}

\item{removeDigits}{remove numbers if \code{TRUE}}

\item{removePunct}{remove punctuation if \code{TRUE}}

\item{lower}{convert text to lower case \code{TRUE}}

\item{additional}{additional characters to remove (\link[=regex]{regular expression})}

\item{twitter}{if \code{TRUE}, do not remove \code{@} or \code{#}}

\item{removeURL}{removes URLs (web addresses starting with \code{http:} or \code{https:}), based
on a regular expression from \url{http://daringfireball.net/2010/07/improved_regex_for_matching_urls}}
}
\value{
A character vector equal in length to the original texts, after cleaning.
}
\description{
\code{clean} removes punctuation and digits from text, using the regex
character classes for punctuation and digits. \code{clean} uses the standard R
function \code{tolower} to convert the text to lower case. Each of these
steps is optional, but switched on by default, so for example, to remove
punctuation and convert to lower, but keep digits, the command would be:
\code{clean(mytexts, removeDigits=FALSE)}
}
\examples{
clean("This is 1 sentence with 2.0 numbers in it, and one comma.", removeDigits=FALSE)
clean("This is 1 sentence with 2.0 numbers in it, and one comma.", lower=FALSE)
clean("We are his Beliebers, and him is #ourjustin @justinbieber we love u", twitter=TRUE)
clean("Collocations can be represented as inheritance_tax using the _ character.")
clean("But under_scores can be removed using the additional argument.", additional="[_]")
clean("This is a $5m watch and €20bn budget and $100,000 in cash plus a ¢50 cigar.")
clean("This is a $5m watch and €20bn budget and $100,000 in cash plus a ¢50 cigar.", removeDigits=FALSE)

# for a vector of texts
clean(c("This is 1 sentence with 2.0 numbers in it, and one comma.",
        "$1.2 billion was spent on text analysis in 2014."))
}

