% Generated by roxygen2 (4.0.2): do not edit by hand
\name{clean}
\alias{clean}
\alias{clean.character}
\alias{clean.corpus}
\title{simple cleaning of text before processing}
\usage{
clean(x, ...)

\method{clean}{character}(x, removeDigits = TRUE, removePunct = TRUE,
  lower = TRUE, ...)

\method{clean}{corpus}(x, removeDigits = TRUE, removePunct = TRUE,
  lower = TRUE, ...)
}
\arguments{
\item{x}{The object to be cleaned. Can be either a character vector or a
corpus object. If x is a corpus, \code{clean} returns a copy of the x with
the texts cleaned.}

\item{...}{additional parameters}

\item{removeDigits}{remove numbers if \code{TRUE}}

\item{removePunct}{remove punctuation if \code{TRUE}}

\item{lower}{convert text to lower case \code{TRUE}}
}
\value{
A character vector equal in length to the original texts, after cleaning.
}
\description{
\code{clean} removes punctuation and digits from text, using the regex
character classes for punctuation and digits. \code{clean} uses the standard R
function \code{tolower} to convert the text to lower case. Each of these
steps is optional, but switched on by default, so for example, to remove
punctuation and convert to lower, but keep digits, the command would be:
\code{clean(mytexts, removeDigits=FALSE)}
}
\examples{
clean("This is 1 sentence with 1.9 numbers in it, and one comma.", removeDigits=FALSE)
clean("This is 1 sentence with 1.9 numbers in it, and one comma.", lower=FALSE)

# for a vector of texts
clean(c("This is 1 sentence with 1.9 numbers in it, and one comma.",
        "€1.2 billion was spent on text analysis in 2014."))
}

