% Generated by roxygen2 (4.0.2): do not edit by hand
\name{collocations}
\alias{collocations}
\title{Detect collocations in a text}
\usage{
collocations(text = NULL, n = 2, method = c("lr", "chi2", "mi"))

collocations(text = NULL, n = 2, method = c("lr", "chi2", "mi"))
}
\arguments{
\item{text}{a text or vector of texts}

\item{n}{Only bigrams (n=2) implemented so far.}

\item{method}{association measure for detecting collocations}

\item{file}{a filename containing a text}

\item{top}{threshold number for number of collocations to be returned (in descending order of association value)}

\item{distance}{distance between pairs of collocations}

\item{text}{a text or vector of texts}

\item{method}{association measure for detecting collocations}

\item{n}{Only bigrams (n=2) implemented so far.}
}
\value{
A list of collocations, their frequencies, and their test statistics

A list of collocations, their frequencies, and the G^2 likelihood ratio, and Chi^2
statistic.
}
\description{
returns a list of collocations.  Note: Currently works only for pairs (bigram collocations).

Detects collocation bigrams using a fast vectorized/indexed method.
}
\note{
We will eventually add options to remove collocations containing stopwords, and
additional measures of association (mutual information, for instance).
}
\examples{
collocations(texts(inaugCorpus)[1], top=50)
collocations(texts(inaugCorpus)[1], top=50, method="chi2")
collocations(texts(inaugCorpus)[1], top=50)
collocations(texts(inaugCorpus)[1], top=50, method="chi2")
}
\author{
Kenneth Benoit

Kenneth Benoit
}

