% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textstat_lexdiv_replaced.R
\name{compute_mattr}
\alias{compute_mattr}
\title{Computes the Moving-Average Type-Token Ratio}
\usage{
compute_mattr(x, window_size = NULL, all_windows = FALSE,
  mean_mattr = TRUE)
}
\arguments{
\item{x}{a \link{tokens} object}

\item{window_size}{integer; the size of the moving window for computation of
TTR, between 1 and the number of tokens of the document}

\item{all_windows}{default = FALSE. If TRUE, returns a vector with the MATTR
for each window}

\item{mean_mattr}{default = TRUE. Returns the mean MATTR, given the window
size, for the tokens object.}
}
\value{
return either a vector with the MATTR for each window, or the mean
  MATTR for the whole tokens object
}
\description{
Takes a tokens object which should have been preprocessed - removal of
punctuation etc. and computes the Moving-Average Type-Token Ratio from
Covington & McFall (2010).
}
\keyword{internal}
\keyword{tokens}
