% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textstat_lexdiv.R
\name{compute_mattr}
\alias{compute_mattr}
\title{Computes the Moving-Average Type-Token Ratio (Covington & McFall, 2010) for a Tokens Object}
\usage{
compute_mattr(x, window_size = NULL, all_windows = FALSE,
  mean_mattr = TRUE)
}
\arguments{
\item{x}{input \link{tokens}}

\item{window_size}{input: the size of the moving window for computation of TTR}

\item{all_windows}{default = FALSE. If TRUE, returns a vector with the MATTR for each window}

\item{mean_mattr}{default = TRUE. Returns the mean MATTR, given the window size, for the tokens object.}
}
\value{
return either a vector with the MATTR for each window, or the mean MATTR for the whole tokens object
}
\description{
Takes a tokens object which should have been preprocessed - removal of punctuation etc.
}
\examples{
txt <- "There's shrimp-kabobs, shrimp creole, shrimp gumbo. Pan fried, deep fried, stir-fried. There's 
pineapple shrimp, lemon shrimp, coconut shrimp, pepper shrimp, shrimp soup,
shrimp stew, shrimp salad, shrimp and potatoes, shrimp burger, shrimp
sandwich."
# Return only Mean MATTR
compute_mattr(tokens(txt[1]), window_size = 20)
[1] 0.52
# Return MATTR for each moving window
compute_mattr(tokens(txt[1]), window_size = 20, all_windows = TRUE, mean_mattr = FALSE)
tokens1_20  tokens2_21  tokens3_22  tokens4_23  tokens5_24  tokens6_25  tokens7_26  tokens8_27  tokens9_28 
0.60        0.60        0.60        0.60        0.60        0.60        0.60        0.60        0.60 
tokens10_29 tokens11_30 tokens12_31 tokens13_32 tokens14_33 tokens15_34 tokens16_35 tokens17_36 tokens18_37 
0.60        0.55        0.55        0.60        0.55        0.50        0.55        0.50        0.45 
tokens19_38 tokens20_39 tokens21_40 tokens22_41 tokens23_42 tokens24_43 tokens25_44 tokens26_45 tokens27_46 
0.45        0.40        0.40        0.45        0.45        0.45        0.45        0.45        0.45 
tokens28_47 tokens29_48 tokens30_49 
0.45        0.45        0.50 
}
\keyword{internal}
\keyword{tokens}
