% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textstat_lexdiv.R
\name{compute_mattr}
\alias{compute_mattr}
\title{Compute the Moving-Average Type-Token Ratio (MATTR)}
\usage{
compute_mattr(x, MATTR_window = 100L)
}
\arguments{
\item{x}{a \link{tokens} object}

\item{MATTR_window}{integer; the size of the moving window for computation of
TTR, between 1 and the number of tokens of the document}
}
\description{
From a tokens object, computes the Moving-Average Type-Token Ratio (MATTR)
from Covington & McFall (2010), averaging all of the sequential moving
windows of tokens of size \code{MATTR_window} across the text, returning the
average as the MATTR.
}
\keyword{internal}
\keyword{lexdiv}
\keyword{textstat}
