% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textstat_lexdiv_replaced.R
\name{compute_msttr}
\alias{compute_msttr}
\title{Computes the Mean Segmental Type-Token Ratio}
\usage{
compute_msttr(x, segment_size = NULL, discard_remainder = TRUE,
  all_segments = FALSE, mean_sttr = TRUE)
}
\arguments{
\item{x}{a \link{tokens} object}

\item{segment_size}{input: the size of the segment.}

\item{discard_remainder}{default = TRUE. If TRUE, the final segment of the document not divisible by segment_size is dropped.}

\item{all_segments}{default = FALSE. If TRUE, returns a vector with the TTR for each segment.}

\item{mean_sttr}{default = TRUE. Returns the Mean TTR across Segments for the tokens object.}
}
\value{
returns a vector with the MSSTR for each segment
}
\description{
Computes the Mean Segmental Type-Token Ratio (Johnson 1944) for a tokens
object.
}
\keyword{internal}
\keyword{tokens}
