% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textstat_lexdiv.R
\name{compute_msttr}
\alias{compute_msttr}
\title{Computes the Mean Segmental Type-Token Ratio (Johnson, 1944) for a Tokens Object}
\usage{
compute_msttr(x, segment_size = NULL)
}
\arguments{
\item{x}{input \link{tokens}}

\item{segment_size}{input: the size of the segment.}
}
\value{
returns a vector with the MSSTR for each segment
}
\description{
Takes a tokens object which should have been preprocessed - removal of punctuation etc.
}
\examples{
txt <- "There's shrimp-kabobs, shrimp creole, shrimp gumbo. Pan fried, deep fried, stir-fried. There's 
pineapple shrimp, lemon shrimp, coconut shrimp, pepper shrimp, shrimp soup,
shrimp stew, shrimp salad, shrimp and potatoes, shrimp burger, shrimp
sandwich."
compute_msttr(tokens(txt[1]), segment_size = 7)
segment1_7  segment8_14 segment15_21 segment22_28 segment29_35 segment36_42 segment43_49 
0.7142857    0.8571429    0.8571429    0.7142857    0.5714286    0.7142857    0.7142857 
}
\keyword{internal}
\keyword{tokens}
