% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textstat_lexdiv.R
\name{compute_msttr}
\alias{compute_msttr}
\title{Compute the Mean Segmental Type-Token Ratio (MSTTR)}
\usage{
compute_msttr(x, MSTTR_segment)
}
\arguments{
\item{x}{input \link{tokens}}

\item{segment_size}{the size of the segment}
}
\description{
Compute the Mean Segmental Type-Token Ratio (Johnson 1944) for a tokens input.
}
\keyword{internal}
\keyword{lexdiv}
\keyword{textstat}
