% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/toTdm.R
\name{convert}
\alias{convert}
\alias{convert.dfm}
\title{convert a dfm to a non-quanteda format}
\usage{
convert(object, ...)

\method{convert}{dfm}(object, to = c("lda", "tm", "stm", "austin"),
  weighting = weightTf)
}
\arguments{
\item{object}{dfm to be converted}

\item{to}{target conversion format, one of
\describe{
\item{"lda"}{a list with components "documents" and "vocab" as needed by
  \link[lda]{lda.collapsed.gibbs.sampler}}
\item{"tm"}{a \link[tm]{DocumentTermMatrix} from the \pkg{tm} package}
\item{"stm"}{the data format for the \pkg{\link{stm}} package}
\item{"austin"}{the \link[austin]{wfm} format from the \pkg{\link{austin}} package}
}}
}
\value{
a converted object determined by the value of \code{to} (see above)
}
\description{
Convert a quanteda \link{dfm-class} object to a format useable by
other text analysis packages.
}

