% Generated by roxygen2 (4.0.1): do not edit by hand
\name{corpus}
\alias{corpus}
\alias{corpus.character}
\alias{corpus.default}
\alias{corpus.directory}
\title{Constructor for corpus objects}
\usage{
corpus(texts, ...)

\method{corpus}{default}()

\method{corpus}{directory}(path, enc = NULL, docnames = NULL,
  docvarsfrom = c("filenames", "headers"), docvarnames = NULL, sep = "_",
  source = NULL, notes = NULL, citation = NULL)

\method{corpus}{character}(texts, enc = NULL, docnames = NULL,
  docvars = NULL, source = NULL, notes = NULL, citation = NULL, ...)
}
\arguments{
\item{texts}{A character vector of texts, or a filepath to a directory containing text documents.}

\item{docvarsfrom}{Argument to specify where docvars are to be taken, from parsing the filenames (\link{filenames}) separated
by \code{sep} or from meta-data embedded in the text file header (\code{headers}).}

\item{docvarnames}{Character vector of variable names for \code{docvars}}

\item{sep}{Separator if \link{docvar} names are taken from the filenames.}

\item{texts}{A character vector containing the texts}

\item{docnames}{Names to be assigned to the texts, defaults to the names of the
character vector (if any), otherwise assigns "text1", "text2", etc.}

\item{docvars}{A data frame of attributes that is associated with each text.}

\item{source}{A string specifying the source of the texts, used for referencing.}

\item{notes}{A string containing notes about who created the text, warnings, To Dos, etc.}
}
\description{
Creates a corpus from a character vector (of texts), or an object such as a
directory containing text files.  Corpus-level meta-data can be specified
at creation, containing (for example) citation information and notes.
}
\examples{
\dontrun{corpus()  # pop up a file selection GUI interface
}
\dontrun{
tempcorpus <- corpus(directory("~/Dropbox/QUANTESS/corpora/ukManRenamed"), enc="UTF-8", source="Ken's manifesto archive")
summary(tempcorpus, n=10)
}
corpus(inaugTexts)
uk2010immigCorpus <- corpus(uk2010immig, docvars=data.frame(party=names(uk2010immig)), enc="UTF-8")
}

