% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus_extract.R
\name{corpus_extract}
\alias{corpus_extract}
\alias{char_extract}
\title{extract and store values in corpus}
\usage{
corpus_extract(x, pattern, field, remove = TRUE, ...)

char_extract(x, pattern, ...)
}
\arguments{
\item{x}{character or \link{corpus} object whose texts will be segmented}

\item{pattern}{a regular expression pattern that match characters in
documents}

\item{field}{a name of a \link{docvars} field to store extracted values}

\item{remove}{remove matched characters from documents if TRUE}
}
\value{
\code{corpus_extract} returns a corpus of texts, with extracted 
  values in \link{docvars}.

\code{char_extract} returns a character vector of matched characters
}
\description{
Extract characters from douments with regular expression pattern and store
them in \link{corpus} in corpus.
}
\examples{
txts <- c("##INTRO This is the introduction.",
          "##DOC1 This is the first document.",
          "##DOC2 This is the second document.",
          "##DOC3 This is the third document.")
corp <- corpus(txts)
corp <- corpus_extract(corp, '##[A-Z0-9]+', 'tag', remove = TRUE)
head(corp)
head(docvars(corp))

## segmenting a character object

txts <- c("##INTRO This is the introduction.",
          "##DOC1 This is the first document.",
          "##DOC2 This is the second document.",
          "##DOC3 This is the third document.")
char_extract(txts, '##[A-Z0-9]+')
}
\author{
Kenneth Benoit
}
\keyword{character}
