\name{corpusFromFilenames}
\alias{corpusFromFilenames}
\title{create a new corpus with attribute-value pairs taken from document filenames
Work in progress}
\usage{
corpusFromFilenames(directory, attNames, sep = "_")
}
\arguments{
  \item{directory}{}
}
\description{
create a new corpus with attribute-value pairs taken from
document filenames Work in progress
}
\examples{
\dontrun{
budgets <- corpusFromHeaders("~/Dropbox/QUANTESS/corpora/withHeader")
}
}

