% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/corpus-sources-S4.R
\docType{class}
\name{corpusSource-class}
\alias{corpusSource-class}
\title{corpus source classes}
\description{
The \code{corpusSource} virtual class is a parent class for more specific
corpus source objects.
}
\section{Slots}{

\describe{
\item{\code{texts}}{the texts that form the core of the corpus}

\item{\code{docvars}}{document variables in a data.frame}

\item{\code{source}}{source recorded for the corpus, based on type of source}

\item{\code{created}}{a time stamp}
}}

