% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textfile.R
\docType{class}
\name{corpusSource-class}
\alias{corpusSource-class}
\alias{show,corpusSource-method}
\title{corpus source classes}
\usage{
\S4method{show}{corpusSource}(object)
}
\arguments{
\item{object}{corpusSource object to be printed}
}
\description{
The \code{corpusSource} virtual class is a parent class for more specific 
corpus source objects.
}
\section{Slots}{

\describe{
\item{\code{texts}}{the texts that form the core of the corpus}

\item{\code{docvars}}{document variables in a data.frame}

\item{\code{source}}{source recorded for the corpus, based on type of source}

\item{\code{created}}{a time stamp}

\item{\code{cachedfile}}{if read to a temporary file, a string containing the
location of the temporary file}
}}

