% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-deprecated.R
\docType{data}
\name{data-deprecated}
\alias{LBGexample}
\alias{data-deprecated}
\alias{exampleString}
\alias{ie2010Corpus}
\alias{inaugCorpus}
\alias{inaugTexts}
\alias{ukimmigTexts}
\title{datasets with deprecated or defunct names}
\format{See the specific object documentation for the renamed objects.}
\usage{
ukimmigTexts

ie2010Corpus

inaugTexts

inaugCorpus
}
\description{
These are datasets that have been renamed but that have been
  temporarily retained for backward compatibility, or that have been renamed
  and the old names removed. Please change your code to use the new names, as
  the deprecated functions will become defunct in future versions.
}
\details{
\code{ukimmigTexts} is deprecated; please use 
  \code{data_char_ukimmigration} instead.

\code{ie2010Corpus} is deprecated; please use 
  \code{data_corpus_irishbudget2010}

\code{inaugTexts} is deprecated; please use
  \code{data_char_inaugural} instead.

\code{inaugCorpus} is deprecated; please use 
  \code{data_corpus_inaugural} instead.

\code{LBGexample} is defunct; please use 
  \code{data_dfm_LBGexample} instead.

\code{exampleString} is defunct; please use 
  \code{data_char_sampletext} instead.
}
\keyword{datasets}
\keyword{deprecated}
\keyword{internal}

