% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-deprecated.R
\docType{data}
\name{data-deprecated}
\alias{data-deprecated}
\alias{ie2010Corpus}
\alias{inaugCorpus}
\alias{inaugTexts}
\alias{ukimmigTexts}
\title{datasets with deprecated names}
\format{See the specific object documentation for the renamed objects.}
\usage{
ukimmigTexts

ie2010Corpus

inaugTexts

inaugCorpus
}
\description{
Datasets that have been renamed but that have been temporarily retained
for backward compatibility.  Please change your code to use the new names,
as these functions will be removed in future versions.
}
\details{
\code{ukimmigTexts} is deprecated; please use 
  \code{data_char_ukimmigration} instead.

\code{ie2010Corpus} is deprecated; please use 
  \code{data_corpus_irishbudget2010}

\code{inaugTexts} is deprecated; please use
  \code{data_char_inaugural} instead.

\code{inaugCorpus} is deprecated; please use 
  \code{data_corpus_inaugural} instead.
}
\keyword{datasets}

