% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R, R/readability.R, R/stopwords.R, R/syllables.R
\docType{data}
\name{data-internal}
\alias{data-internal}
\alias{data_char_stopwords}
\alias{data_char_wordlists}
\alias{data_int_syllables}
\title{internal data sets}
\format{An object of class \code{list} of length 2.}
\source{
The English stopwords are taken from the SMART information retrieval
system (obtained from 
\url{http://jmlr.csail.mit.edu/papers/volume5/lewis04a/a11-smart-stop-list/english.stop}).

Additional stopword lists are taken from the Snowball stemmer project in
different languages (see \url{http://snowballstem.org/projects.html}).

The Greek stopwords were supplied by Carsten Schwemmer (see
\href{https://github.com/kbenoit/quanteda/issues/282}{GitHub issue #282}).
}
\usage{
data_char_wordlists

data_char_stopwords

data_int_syllables
}
\description{
Data sets used for mainly internal purposes by the \pkg{quanteda} package.
}
\details{
\code{data_char_wordlists} provides word lists used in some readability indexes; 
it is a named list of character vectors where each list element 
corresponds to a different readability index.  

These are:
\describe{
\item{\code{DaleChall}}{The long Dale-Chall list of 3,000 familiar (English) words needed to compute the Dale-Chall Readability Formula.}
\item{\code{Spache}}{The revised Spache word list (see Klare 1975, 73) needed to compute the Spache Revised Formula of readability (Spache 1974.}
}

\code{data_char_stopwords} provides stopword lists in multiple
languages; it is a named list of characters with the lowercase language name
(in English) as the name of each list element.

Supported languages are Arabic, Danish, Dutch, English, Finnish, French,
German, Greek, Hungarian, Italian, Norwegian, Portuguese, Russian, Spanish,
and Swedish.

\code{data_int_syllables} provides an English-language syllables dictionary; it is
an integer vector whose element names correspond to English words. Built from
the freely available CMU pronunciation dictionary at 
\code{http://www.speech.cs.cmu.edu/cgi-bin/cmudict}.
}
\references{
Chall, J. S., & Dale, E.  1995. \emph{Readability Revisited: The New Dale-Chall Readability Formula}. Brookline Books.

Klare, G. R. 1975. "Assessing readability." \emph{Reading Research Quarterly} 10(1): 62–102.

Spache, G. 1953. "A new readability formula for primary-grade reading materials." \emph{The Elementary School Journal} 53: 410-413.
}
\keyword{data}
\keyword{datasets}
\keyword{internal}

