% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{data-internal}
\alias{data-internal}
\title{Internal data sets}
\description{
Data sets used for mainly internal purposes by the \pkg{quanteda} package.
}
\keyword{data}
\keyword{internal}
