% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{data_char_sampletext}
\alias{data_char_sampletext}
\title{A paragraph of text for testing various text-based functions}
\format{
character vector with one element
}
\source{
Dáil Éireann Debate,
Financial Resolution No. 13: General (Resumed).
\verb{https://www.oireachtas.ie/en/debates/debate/dail/2011-12-07/6/}. 7 December
2011.  vol. 749, no. 1.
}
\usage{
data(data_char_sampletext)
}
\description{
This is a long paragraph (2,914 characters) of text taken from a debate on
Joe Higgins, delivered December 8, 2011.
}
\examples{
tokens(data_char_sampletext, remove_punct = TRUE)
}
\keyword{data}
