% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{data_char_sampletext}
\alias{data_char_sampletext}
\title{a paragraph of text for testing various text-based functions}
\format{character vector with one element}
\usage{
data_char_sampletext
}
\description{
This is a long paragraph (2,914 characters) of text taken from an Irish budget speech by Joe Higgins.
}
\examples{
tokenize(data_char_sampletext, removePunct = TRUE)
}
\keyword{data}
