% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{data_char_sampletext}
\alias{data_char_sampletext}
\title{A paragraph of text for testing various text-based functions}
\format{
character vector with one element
}
\source{
Dáil Éireann Debate,
\href{http://oireachtasdebates.oireachtas.ie/debates\%20authoring/debateswebpack.nsf/takes/dail2011120700006?opendocument}{Financial Resolution No. 13: General (Resumed).}
7 December 2011.  vol. 749, no. 1.
}
\usage{
data_char_sampletext
}
\description{
This is a long paragraph (2,914 characters) of text taken from a debate on
Joe Higgins, delivered December 8, 2011.
}
\examples{
tokens(data_char_sampletext, remove_punct = TRUE)
}
\keyword{data}
