% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-deprecated.R, R/data-documentation.R
\docType{data}
\name{ukimmigTexts}
\alias{data_char_ukimmig2010}
\alias{ukimmigTexts}
\title{immigration-related sections of 2010 UK party manifestos}
\format{An object of class \code{character} of length 9.}
\usage{
ukimmigTexts

data_char_ukimmig2010
}
\description{
Extracts from the election manifestos of 9 UK political parties from 2010, related
to immigration or asylum-seekers.
}
\note{
\code{ukimmigTexts} is deprecated; please use
  \code{data_char_ukimmigration} instead.
}
\examples{
data_corpus_ukimmig2010 <- 
    corpus(data_char_ukimmig2010, 
           docvars = data.frame(party = names(ukimmigTexts)))
metadoc(data_corpus_ukimmig2010, "language") <- "english"
summary(data_corpus_ukimmig2010, showmeta = TRUE)
}
\keyword{datasets}

