% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{data_char_ukimmig2010}
\alias{data_char_ukimmig2010}
\title{Immigration-related sections of 2010 UK party manifestos}
\format{
A named character vector of plain ASCII texts
}
\usage{
data_char_ukimmig2010
}
\description{
Extracts from the election manifestos of 9 UK political parties from 2010, related
to immigration or asylum-seekers.
}
\examples{
data_corpus_ukimmig2010 <- 
    corpus(data_char_ukimmig2010, 
           docvars = data.frame(party = names(data_char_ukimmig2010)))
summary(data_corpus_ukimmig2010, showmeta = TRUE)
}
\keyword{data}
