% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-deprecated.R, R/data-documentation.R
\docType{data}
\name{data_corpus_irishbudget2010}
\alias{data_corpus_irishbudget2010}
\alias{ie2010Corpus}
\alias{iebudgets}
\title{Irish budget speeches from 2010}
\format{An object of class \code{corpus} (inherits from \code{list}) of length 4.}
\source{
Lowe, Will, and Kenneth R Benoit. 2013. "Validating Estimates of 
  Latent Traits From Textual Data Using Human Judgment as a Benchmark." 
  \emph{Political Analysis} 21: 298-313.
}
\usage{
data_corpus_irishbudget2010

data_corpus_irishbudget2010
}
\description{
Speeches and document-level variables from the debate over the Irish budget
of 2010.
}
\note{
\code{data_corpus_irishbudget2010} is deprecated; please use
  \code{data_corpus_irishbudget2010} instead.
}
\examples{
summary(data_corpus_irishbudget2010)
}
\keyword{datasets}

