\name{describeTexts}
\alias{describeTexts}
\title{print a summary of texts
Prints to the console a desription of the texts, including
number of types, tokens, and sentences}
\usage{
  describeTexts(texts)
}
\arguments{
  \item{texts}{}
}
\description{
  print a summary of texts Prints to the console a
  desription of the texts, including number of types,
  tokens, and sentences
}
\examples{
describeTexts(texts)
}

