\name{describeTexts}
\alias{describeTexts}
\title{print a summary of texts
Prints to the console a desription of the texts, including
number of types, tokens, and sentences}
\usage{
describeTexts(texts, output = TRUE)
}
\arguments{
  \item{texts}{The texts to be described}
}
\description{
print a summary of texts Prints to the console a desription
of the texts, including number of types, tokens, and
sentences
}
\examples{
texts <- "This is an example text"
describeTexts(texts)
}

