% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfm-classes.R, R/dfm-subsetting.R
\docType{class}
\name{dfm-class}
\alias{dfm-class}
\alias{t,dfm-method}
\alias{colSums,dfm-method}
\alias{rowSums,dfm-method}
\alias{colMeans,dfm-method}
\alias{rowMeans,dfm-method}
\alias{+,dfm,numeric-method}
\alias{+,numeric,dfm-method}
\alias{[,dfm,index_logical_missing,index_logical_missing,missing-method}
\title{Virtual class "dfm" for a document-feature matrix}
\usage{
\S4method{t}{dfm}(x)

\S4method{colSums}{dfm}(x, na.rm = FALSE, dims = 1L, ...)

\S4method{rowSums}{dfm}(x, na.rm = FALSE, dims = 1L, ...)

\S4method{colMeans}{dfm}(x, na.rm = FALSE, dims = 1L, ...)

\S4method{rowMeans}{dfm}(x, na.rm = FALSE, dims = 1L, ...)

\S4method{+}{dfm,numeric}(e1, e2)

\S4method{+}{numeric,dfm}(e1, e2)

\S4method{[}{dfm,index_logical_missing,index_logical_missing,missing}(x, i, j,
  drop)
}
\arguments{
\item{x}{the dfm object}

\item{na.rm}{if \code{TRUE}, omit missing values (including \code{NaN}) from
the calculations}

\item{dims}{ignored}

\item{...}{additional arguments not used here}

\item{e1}{first quantity in "+" operation for dfm}

\item{e2}{second quantity in "+" operation for dfm}

\item{i}{index for documents}

\item{j}{index for features}

\item{drop}{always set to \code{FALSE}}
}
\description{
The dfm class of object is a type of \link[Matrix]{Matrix-class}
  object with additional slots, described below.
}
\section{Slots}{

\describe{
\item{\code{settings}}{settings that govern corpus handling and subsequent downstream
operations, including the settings used to clean and tokenize the texts, 
and to create the dfm.  See \code{\link{settings}}.}

\item{\code{weighting}}{the feature weighting applied to the dfm.  Default is
\code{"frequency"}, indicating that the values in the cells of the dfm are
simple feature counts.  To change this, use the \code{\link{weight}}
method.}

\item{\code{smooth}}{a smoothing parameter, defaults to zero.  Can be changed using 
either the \code{\link{smooth}} or the \code{\link{weight}} methods.}

\item{\code{Dimnames}}{These are inherited from \link[Matrix]{Matrix-class} but are 
named \code{docs} and \code{features} respectively.}
}}

\examples{
# dfm subsetting
x <- dfm(tokens(c("this contains lots of stopwords",
                  "no if, and, or but about it: lots",
                  "and a third document is it"),
                remove_punct = TRUE))
x[1:2, ]
x[1:2, 1:5]
x[c(TRUE, TRUE, FALSE),]
x[c(TRUE, TRUE, FALSE), c(rep(TRUE, 5), rep(FALSE, 11))]

# fcm subsetting
y <- fcm(tokens(c("this contains lots of stopwords",
                  "no if, and, or but about it: lots"),
                remove_punct = TRUE))
y[1:3, ]
y[4:5, 1:5]
}
\seealso{
\link{dfm}
}
\keyword{dfm}
\keyword{internal}
