% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfm-methods.R
\docType{methods}
\name{dfm-internal}
\alias{dfm-internal}
\alias{Compare,dfm,numeric-method}
\title{internal functions for dfm objects}
\usage{
\S4method{Compare}{dfm,numeric}(e1, e2)
}
\arguments{
\item{e1}{a \link{dfm}}

\item{e2}{a numeric value to compare with values in a dfm}
}
\description{
Internal function documentation for \link{dfm} objects.

The \code{Compare} methods enable relational operators to be use with dfm. 
Relational operations on a dfm with a numeric will return a
\link[Matrix]{dgCMatrix-class} object.
}
\seealso{
\link{Comparison} operators
}
\keyword{dfm}
\keyword{internal}
