% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dfm}
\alias{dfm}
\alias{dfm.character}
\alias{dfm.corpus}
\alias{is.dfm}
\title{Create a document-feature matrix from a corpus object}
\usage{
dfm(x, ...)

\method{dfm}{corpus}(x, feature = c("word"), stem = FALSE,
  stopwords = NULL, bigram = FALSE, groups = NULL, verbose = TRUE,
  dictionary = NULL, dictionary_regex = FALSE, bootstrap = FALSE,
  addto = NULL, ...)

\method{dfm}{character}(x, feature = c("word"), stem = FALSE,
  stopwords = NULL, bigram = FALSE, verbose = TRUE, dictionary = NULL,
  dictionary_regex = FALSE, addto = NULL, ...)

is.dfm(x)
}
\arguments{
\item{x}{Corpus or character vector from which to generate the
document-feature matrix}

\item{...}{additional arguments passed to \code{\link{clean}}}

\item{feature}{Feature to count (e.g. words)}

\item{stem}{Stem the words}

\item{stopwords}{A character vector of stopwords that will be removed from
the text when constructing the \link{dfm}.  If \code{NULL} (default) then
no stopwords will be applied.  If "TRUE" then it currently defaults to
\code{\link{stopwords}}.}

\item{bigram}{include bigrams as well as unigram features, if \code{TRUE}}

\item{groups}{Grouping variable for aggregating documents}

\item{verbose}{Get info to screen on the progress}

\item{dictionary}{A list of character vector dictionary entries, including
regular expressions (see examples)}

\item{dictionary_regex}{\code{TRUE} means the dictionary is already in
regular expression format, otherwise it will be converted from "wildcard"
format}

\item{bootstrap}{if \code{TRUE}, compute multiple \code{dfm}'s from resampled
texts in the corpus.  Requires a resampled corpus.  See
\code{\link{resample}}.}

\item{addto}{\code{NULL} by default, but if an existing dfm object is
specified, then the new dfm will be added to the one named. If both
\link{dfm}'s are built from dictionaries, the combined dfm will have its
\code{Non_Dictionary} total adjusted.}
}
\value{
A specially classed matrix object with row names equal to the
  document names and column names equal to the feature labels.  Additional
  information is attached to this object as \code{\link{attributes}}, such as
  \link{settings}.
}
\description{
Returns a document by feature matrix with additional meta-information
(settings, identification of training texts for supervised models, resampling
information, etc.) that is useful in other quanteda functions.  A typical
usage would be to produce a word-frequency matrix where the cells are counts
of words by document, but the definition of "features" is entirely general.
}
\details{
\code{is.dfm} returns \code{TRUE} if and only if its argument is a \link{dfm}.
}
\examples{
data(inaugCorpus)
wfm <- dfm(inaugCorpus)

## by president, after 1960
wfmByPresfrom1900 <- dfm(subset(inaugCorpus, Year>1900), groups="President")
docnames(wfmByPresfrom1900)

## with dictionaries
data(inaugCorpus)
mycorpus <- subset(inaugCorpus, Year>1900)
mydict <- list(christmas=c("Christmas", "Santa", "holiday"),
               opposition=c("Opposition", "reject", "notincorpus"),
               taxing="taxing",
               taxation="taxation",
               taxregex="tax*")
dictDfm <- dfm(mycorpus, dictionary=mydict)
dictDfm

## removing stopwords
testText <- "The quick brown fox named Seamus jumps over the lazy dog also named Seamus, with
             the newspaper from a a boy named Seamus, in his mouth."
testCorpus <- corpus(testText)
settings(testCorpus, "stopwords")
dfm(testCorpus, stopwords=TRUE)
}
\author{
Kenneth Benoit
}

