% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dfm2tmformat}
\alias{dfm2tmformat}
\title{Convert a \link{dfm} into a \pkg{tm} \link[tm]{DocumentTermMatrix}}
\usage{
dfm2tmformat(d, weighting = weightTf, ...)
}
\arguments{
\item{d}{A \link[quanteda]{dfm} object}

\item{weighting}{weight function  arguments passed to
\code{as.TermDocumentMatrix}, defaults to term frequency (see
\code{\link[tm]{as.DocumentTermMatrix}} for a list of options, such as
tf-idf).}
}
\value{
A simple triplet matrix of class \link[tm]{as.DocumentTermMatrix}
}
\description{
\pkg{tm} represents sparse document-feature matrixes in the
\link[slam:matrix]{simple triplet matrix} format of the package \pkg{slam}.
This function converts a \code{dfm} into a
\code{\link[tm]{DocumentTermMatrix}}, enabling a dfm to be used with other
packages that expect this format, such as \pkg{topicmodels}.
}
\examples{
mycorpus <- subset(inaugCorpus, Year>1970)
d <- trimdfm(dfm(mycorpus), minCount=5, minDoc=3)
dim(d)
td <- dfm2tmformat(d)
length(td$v)
if (require(topicmodels)) (tmodel.lda <- LDA(td, control = list(alpha = 0.1), k = 5))
}

