% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfm_match.R
\name{dfm_match}
\alias{dfm_match}
\title{Match the dfm columns with given features}
\usage{
dfm_match(x, features, verbose = quanteda_options("verbose"))
}
\arguments{
\item{x}{the \link{dfm} object.}

\item{features}{character vector for the feature names to be matched in the
resulting \link{dfm}. Columns not included in \code{features} are removed.}

\item{verbose}{if \code{TRUE} print the number of tokens and documents before and
after the function is applied. The number of tokens does not include paddings.}
}
\value{
A \link{dfm} whose features are identical to those specified in
\code{features}.
}
\description{
Match the columns of multiple \link{dfm} objects using a character vector.
}
\details{
Matching the dfm columns is necessary when you split a corpus into two: you fit
a model on the test set and evaluation it on the test set whose features must be
identical. It is also used in \code{\link[=bootstrap_dfm]{bootstrap_dfm()}}.
}
\note{
Unlike \code{\link[=dfm_select]{dfm_select()}}, this function will add feature names
not already present in \code{x}. It also provides only fixed,
case-sensitive matches.  For more flexible feature selection, see
\code{\link[=dfm_select]{dfm_select()}}.
}
\examples{
# matching a dfm to a feature vector
dfm_match(dfm(tokens("")), letters[1:5])
dfm_match(data_dfm_lbgexample, c("A", "B", "Z"))
dfm_match(data_dfm_lbgexample, c("B", "newfeat1", "A", "newfeat2"))

# matching one dfm to another
txt <- c("This is text one", "The text two", "This is text three")
(dfmt1 <- dfm(tokens(txt[1:2])))
(dfmt2 <- dfm(tokens(txt[2:3])))
(dfmt3 <- dfm(dfm_match(dfmt1, featnames(dfmt2))))
identical(featnames(dfmt2), featnames(dfmt3))
}
\seealso{
\code{\link[=dfm_select]{dfm_select()}}
}
\keyword{dfm}
