% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfm_match.R
\name{dfm_match}
\alias{dfm_match}
\title{Make features of a dfm identical}
\usage{
dfm_match(x, features)
}
\arguments{
\item{x}{a dfm}

\item{features}{character vectors of features}
}
\description{
Make features of a dfm identical
}
