% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfm_subset.R
\name{dfm_subset}
\alias{dfm_subset}
\title{extract a subset of a dfm}
\usage{
dfm_subset(x, subset, select, ...)
}
\arguments{
\item{x}{\link{dfm} object to be subsetted}

\item{subset}{logical expression indicating the documents to keep: missing
values are taken as false}

\item{select}{expression, indicating the docvars to select from the dfm}

\item{...}{not used}
}
\value{
\link{dfm} object, with a subset of documents (and docvars) selected according to arguments
}
\description{
Returns \emph{document} subsets of a dfm that meet certain conditions, including direct 
logical operations on docvars (document-level variables).  \code{dfm_subset}
functions identically to \code{\link{subset.data.frame}}, using non-standard
evaluation to evaluate conditions based on the \link{docvars} in the dfm.
}
\details{
To select or subset \emph{features}, see \code{\link{dfm_select}} instead.
}
\examples{
testcorp <- corpus(c(d1 = "a b c d", d2 = "a a b e",
                     d3 = "b b c e", d4 = "e e f a b"),
                   docvars = data.frame(grp = c(1, 1, 2, 3)))
testdfm <- dfm(testcorp)
dfm_subset(testdfm, grp > 1)
dfm_subset(testdfm, c(TRUE, FALSE, TRUE, FALSE))
}
\seealso{
\code{\link{subset.data.frame}}
}
\keyword{corpus}
