% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfm_trim.R
\name{dfm_trim}
\alias{dfm_trim}
\title{trim a dfm using threshold-based or random feature selection}
\usage{
dfm_trim(x, min_count = 1, min_docfreq = 1, sparsity = NULL,
  verbose = TRUE)
}
\arguments{
\item{x}{document-feature matrix of \link{dfm-class}}

\item{min_count}{minimum count or fraction of features in across all documents}

\item{min_docfreq}{minimum number or fraction of documents in which a feature appears}

\item{sparsity}{equivalent to 1 - min_docfreq, included for comparison with tm}

\item{verbose}{print messages}
}
\value{
A \link{dfm-class} object reduced in features (with the same number 
  of documents)
}
\description{
Returns a document by feature matrix reduced in size based on document and 
term frequency, and/or subsampling.
}
\note{
Trimming a \link{dfm-class} object is an operation based on the values 
  in the document-feature \emph{matrix}.  To select subsets of a dfm based on
  attributes of the features themselves -- such as selecting features 
  matching a regular expression, or removing features matching a stopword 
  list, use \link{dfm_select}.
}
\examples{
(myDfm <- dfm(data_corpus_inaugural, verbose = FALSE))

# only words occuring >=10 times and in >=2 docs
dfm_trim(myDfm, min_count = 10, min_docfreq = 2) 

# only words occuring >=10 times and in at least 0.4 of the documents
dfm_trim(myDfm, min_count = 10, min_docfreq = 0.4)

# only words occuring at least 0.01 times and in >=2 documents
dfm_trim(myDfm, min_count = .01, min_docfreq = 2)

# only words occuring 5 times in 1000
dfm_trim(myDfm, min_docfreq = 0.2, min_count = 0.005)

\dontrun{
# compare to removeSpareTerms from the tm package 
if (require(tm)) {
    (tmdtm <- convert(myDfm, "tm"))
    removeSparseTerms(tmdtm, 0.7)
    dfm_trim(td, min_docfreq = 0.3)
    dfm_trim(td, sparsity = 0.7)
}
}
}
\author{
Paul Nulty and Ken Benoit, with some inspiration from Will Lowe's (see \code{trim} from the 
  \code{austin} package)
}
\seealso{
\code{\link{dfm_select}}, \code{\link{dfm_sample}}
}

