% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfm_weight.R
\name{dfm_weight}
\alias{dfm_weight}
\alias{dfm_smooth}
\alias{dfm_tfidf}
\title{Weight the feature frequencies in a dfm}
\usage{
dfm_weight(
  x,
  scheme = c("count", "prop", "propmax", "logcount", "boolean", "augmented", "logave"),
  weights = NULL,
  base = 10,
  k = 0.5,
  force = FALSE
)

dfm_smooth(x, smoothing = 1)

dfm_tfidf(
  x,
  scheme_tf = "count",
  scheme_df = "inverse",
  base = 10,
  k = 0,
  threshold = 0,
  force = FALSE
)
}
\arguments{
\item{x}{object for which idf or tf-idf will be computed (a document-feature
matrix)}

\item{scheme}{a label of the weight type:
\describe{
\item{\code{count}}{\eqn{tf_{ij}}, an integer feature count (default when a dfm is created)}
\item{\code{prop}}{the proportion of the feature counts of total feature
counts (aka relative frequency), calculated as \eqn{tf_{ij} / \sum_j tf_{ij}}}
\item{\code{propmax}}{the proportion of the feature counts of the highest
feature count in a document, \eqn{tf_{ij} / \textrm{max}_j tf_{ij}}}
\item{\code{logcount}}{take the 1 + the logarithm of each count, for the
given base, or 0 if the count was zero: \eqn{1 +
  \textrm{log}_{base}(tf_{ij})} if \eqn{tf_{ij} > 0}, or 0 otherwise.}
\item{\code{boolean}}{recode all non-zero counts as 1}
\item{\code{augmented}}{equivalent to \eqn{k + (1 - k) *} \code{dfm_weight(x, "propmax")}}
\item{\code{logave}}{1 + the log of the counts) / (1 + log of the counts / the average count within document), or
\deqn{\frac{1 + \textrm{log}_{base} tf_{ij}}{1 + \textrm{log}_{base}(\sum_j tf_{ij} / N_i)}}}
}}

\item{weights}{if \code{scheme} is unused, then \code{weights} can be a named
numeric vector of weights to be applied to the dfm, where the names of the
vector correspond to feature labels of the dfm, and the weights will be
applied as multipliers to the existing feature counts for the corresponding
named features.  Any features not named will be assigned a weight of 1.0
(meaning they will be unchanged).}

\item{base}{the base for the logarithms in the \code{\link[=dfm_weight]{dfm_weight()}} and
\code{\link[=docfreq]{docfreq()}} calls; default is 10}

\item{k}{the k for the augmentation when \code{scheme = "augmented"}}

\item{force}{logical; if \code{TRUE}, apply weighting scheme even if the dfm
has been weighted before.  This can result in invalid weights, such as as
weighting by \code{"prop"} after applying \code{"logcount"}, or after
having grouped a dfm using \code{\link[=dfm_group]{dfm_group()}}.}

\item{smoothing}{constant added to the dfm cells for smoothing, default is 1}

\item{scheme_tf}{scheme for \code{\link[=dfm_weight]{dfm_weight()}}; defaults to \code{"count"}}

\item{scheme_df}{scheme for \code{\link[=docfreq]{docfreq()}}; defaults to
\code{"inverse"}.}

\item{threshold}{numeric value of the threshold \emph{above which} a feature
will considered in the computation of document frequency.  The default is
0, meaning that a feature's document frequency will be the number of
documents in which it occurs greater than zero times.}
}
\value{
\code{dfm_weight} returns the dfm with weighted values.  Note the
because the default weighting scheme is \code{"count"}, simply calling this
function on an unweighted dfm will return the same object.  Many users will
want the normalized dfm consisting of the proportions of the feature counts
within each document, which requires setting \code{scheme = "prop"}.

\code{dfm_smooth} returns a dfm whose values have been smoothed by
adding the \code{smoothing} amount. Note that this effectively converts a
matrix from sparse to dense format, so may exceed memory requirements
depending on the size of your input matrix.
}
\description{
Weight a dfm by term frequency-inverse document frequency (\emph{tf-idf}),
with full control over options.  Uses fully sparse methods for efficiency.
}
\details{
\code{dfm_tfidf} computes term frequency-inverse document frequency
weighting.  The default is to use counts instead of normalized term
frequency (the relative term frequency within document), but this
can be overridden using \code{scheme_tf = "prop"}.
}
\examples{
dfmat1 <- dfm(data_corpus_inaugural)

dfmat2 <- dfm_weight(dfmat1, scheme = "prop")
topfeatures(dfmat2)
dfmat3 <- dfm_weight(dfmat1)
topfeatures(dfmat3)
dfmat4 <- dfm_weight(dfmat1, scheme = "logcount")
topfeatures(dfmat4)
dfmat5 <- dfm_weight(dfmat1, scheme = "logave")
topfeatures(dfmat5)

# combine these methods for more complex dfm_weightings, e.g. as in Section 6.4
# of Introduction to Information Retrieval
head(dfm_tfidf(dfmat1, scheme_tf = "logcount"))

# apply numeric weights
str <- c("apple is better than banana", "banana banana apple much better")
(dfmat6 <- dfm(str, remove = stopwords("english")))
dfm_weight(dfmat6, weights = c(apple = 5, banana = 3, much = 0.5))

# smooth the dfm
dfmat <- dfm(data_corpus_inaugural)
dfm_smooth(dfmat, 0.5)
dfmat1 <- as.dfm(data_dfm_lbgexample)
head(dfmat1[, 5:10])
head(dfm_tfidf(dfmat1)[, 5:10])
docfreq(dfmat1)[5:15]
head(dfm_weight(dfmat1)[, 5:10])

# replication of worked example from
# https://en.wikipedia.org/wiki/Tf-idf#Example_of_tf.E2.80.93idf
dfmat2 <-
    matrix(c(1,1,2,1,0,0, 1,1,0,0,2,3),
           byrow = TRUE, nrow = 2,
           dimnames = list(docs = c("document1", "document2"),
                           features = c("this", "is", "a", "sample",
                                        "another", "example"))) \%>\%
    as.dfm()
dfmat2
docfreq(dfmat2)
dfm_tfidf(dfmat2, scheme_tf = "prop") \%>\% round(digits = 2)

\dontrun{
# comparison with tm
if (requireNamespace("tm")) {
    convert(dfmat2, to = "tm") \%>\% tm::weightTfIdf() \%>\% as.matrix()
    # same as:
    dfm_tfidf(dfmat2, base = 2, scheme_tf = "prop")
}
}
}
\references{
Manning, C.D., Raghavan, P., & Schütze, H. (2008).
\emph{An Introduction to Information Retrieval}. Cambridge: Cambridge University Press.
\url{https://nlp.stanford.edu/IR-book/pdf/irbookonlinereading.pdf}

Manning, C. D., Raghavan, P., & Schütze, H. (2008).
\emph{Introduction to Information Retrieval}. Cambridge: Cambridge University Press.
\url{https://nlp.stanford.edu/IR-book/pdf/irbookonlinereading.pdf}
}
\seealso{
\code{\link[=docfreq]{docfreq()}}
}
\keyword{dfm}
\keyword{weighting}
