% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dfms}
\alias{dfms}
\title{create a sparse matrix dfm}
\usage{
dfms(x, verbose = TRUE, clean = TRUE, ...)
}
\arguments{
\item{x}{Character vector from which to generate the document-feature matrix}

\item{...}{additional arguments passed to \code{\link{clean}}}
}
\value{
A specially classed \link[Matrix]{Matrix} object with row names equal to the
  document names and column names equal to the feature labels.
}
\description{
Create a sparse matrix dfm from a vector of texts.
}
\examples{
# with inaugural texts
dfmsInaug <- dfms(inaugTexts)
(size1 <- object.size(dfmsInaug))
(size2 <- object.size(dfm(inaugTexts)))
cat("Compacted by ", round(as.numeric((1-size1/size2)*100), 1), "\%.\\n", sep="")

\dontrun{
# try it with approx 35,000 court documents from Lauderdale and Clark (200?)
load('~/Dropbox/QUANTESS/Manuscripts/Collocations/Corpora/lauderdaleClark/Opinion_files.RData')
txts <- unlist(Opinion_files[1])
names(txts) <- NULL

# dfms without cleaning
require(Matrix)
system.time(dfmsBig <- dfms(txts, clean=FALSE, verbose=FALSE))
object.size(dfmsBig)
dim(dfmsBig)
# compare with tm
require(tm)
tmcorp <- VCorpus(VectorSource(txts))
system.time(tmDTM <- DocumentTermMatrix(tmcorp))
object.size(tmDTM)
dim(tmDTM)

# with cleaning - the gsub() calls in clean() take a long time
system.time(dfmsBig <- dfms(txts, clean=TRUE, additional="[-_—]"))
object.size(dfmsBig)
dim(dfmsBig)
# 100 top features
topf <- colSums(dfmsBig)
names(topf) <- colnames(dfmsBig)
head(sort(topf, decreasing=TRUE), 100)
}
}
\author{
Kenneth Benoit
}

