% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R
\docType{class}
\name{dictionary-class}
\alias{dictionary-class}
\alias{show,dictionary-method}
\alias{[,dictionary,index,ANY,ANY-method}
\alias{[[,dictionary,index-method}
\alias{as.list,dictionary-method}
\title{print a dictionary object}
\usage{
\S4method{show}{dictionary}(object)

\S4method{[}{dictionary,index,ANY,ANY}(x, i)

\S4method{[[}{dictionary,index}(x, i)

\S4method{as.list}{dictionary}(x)
}
\arguments{
\item{object}{the dictionary to be printed}

\item{i}{index for entries}

\item{object}{the dictionary to be extracted}

\item{object}{the dictionary to be extracted}

\item{i}{index for entries}

\item{object}{the dictionary to be coerced}
}
\description{
Print/show method for dictionary objects.

Extractor for dictionary objects

Extractor for dictionary objects

Coerce a dictionary object into a list
}
\section{Slots}{

\describe{
\item{\code{.Data}}{named list of mode character, where each element name is a
dictionary "key" and each element is one or more dictionary entry "values"
consisting of a pattern match}

\item{\code{concatenator}}{character object specifying space between multi-word
values}

\item{\code{format}}{dictionary format (if imported)}

\item{\code{file}}{file from which a dictionary was read (if imported)}
}}

\keyword{dictionary}
\keyword{internal}
