% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R
\docType{class}
\name{dictionary2-class}
\alias{dictionary2-class}
\alias{as.list,dictionary2-method}
\alias{show,dictionary2-method}
\alias{[,dictionary2,index,ANY,ANY-method}
\alias{[[,dictionary2,index-method}
\alias{$.dictionary2}
\alias{c,dictionary2-method}
\title{Coerce a dictionary object into a list}
\usage{
\S4method{as.list}{dictionary2}(x)

\S4method{show}{dictionary2}(object)

\S4method{[}{dictionary2,index,ANY,ANY}(x, i)

\S4method{[[}{dictionary2,index}(x, i)

\method{$}{dictionary2}(x, name)

\S4method{c}{dictionary2}(x, ...)
}
\arguments{
\item{object}{the dictionary to be coerced}

\item{i}{index for entries}

\item{name}{the dictionary key}

\item{...}{\link{dictionary} objects to be concatenated}

\item{object}{the dictionary to be printed}

\item{object}{the dictionary to be extracted}

\item{object}{the dictionary to be extracted}

\item{i}{index for entries}
}
\description{
Print/show method for dictionary objects.
}
\section{Slots}{

\describe{
\item{\code{.Data}}{named list of mode character, where each element name is a
dictionary "key" and each element is one or more dictionary entry "values"
consisting of a pattern match}

\item{\code{concatenator}}{character object specifying space between multi-word
values}
}}

\keyword{dictionary}
\keyword{internal}
