% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R
\docType{class}
\name{dictionary2-class}
\alias{dictionary2-class}
\alias{as.list,dictionary2-method}
\alias{[,dictionary2,index,ANY,ANY-method}
\alias{[[,dictionary2,index-method}
\alias{$.dictionary2}
\alias{c,dictionary2-method}
\title{dictionary class objects and functions}
\usage{
\S4method{as.list}{dictionary2}(x, flatten = FALSE, levels = 1:100)

\S4method{[}{dictionary2,index,ANY,ANY}(x, i)

\S4method{[[}{dictionary2,index}(x, i)

\method{$}{dictionary2}(x, name)

\S4method{c}{dictionary2}(x, ...)
}
\arguments{
\item{flatten}{flatten the nested structure if \code{TRUE}}

\item{levels}{integer vector indicating levels in the dictionary. Used only
when \code{flatten = TRUE}.}

\item{i}{index for entries}

\item{name}{the dictionary key}

\item{...}{\link{dictionary} objects to be concatenated}

\item{object}{the dictionary to be extracted}
}
\description{
The \code{dictionary2} class constructed by \code{\link[=dictionary]{dictionary()}}, and associated core
class functions.
}
\section{Slots}{

\describe{
\item{\code{.Data}}{named list of mode character, where each element name is a
dictionary "key" and each element is one or more dictionary entry "values"
consisting of a pattern match}

\item{\code{meta}}{list of object metadata}
}}

\keyword{dictionary}
\keyword{internal}
