% Generated by roxygen2 (4.0.2): do not edit by hand
\name{directory}
\alias{directory}
\title{Function to declare a connection to a directory (containing files)}
\usage{
directory(path = NULL)
}
\arguments{
\item{path}{String describing the full path of the directory or NULL to use a GUI
to choose a directory from disk}
}
\description{
Function to declare a connection to a directory, although unlike \link{file} it does not require closing.
If the directory does not exist, the function will return an error.
}
\examples{
\dontrun{
# name a directory of files
mydir <- directory("~/Dropbox/QUANTESS/corpora/ukManRenamed")
corpus(mydir)

# choose a directory using a GUI
corpus(directory())}
}

