% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textstat_dist.R
\docType{class}
\name{dist-class}
\alias{as.list.dist}
\alias{dist-class}
\alias{textstat_dist}
\alias{textstat_dist,dfm-method}
\title{Distance matrix between documents and/or features}
\usage{
textstat_dist(x, selection = character(0), n = NULL,
  margin = c("documents", "features"), method = "euclidean",
  normalize = FALSE, tri = FALSE, diag = FALSE)

\S4method{textstat_dist}{dfm}(x, selection = character(0), n = NULL,
  margin = c("documents", "features"), method = "euclidean",
  normalize = FALSE, tri = TRUE, diag = FALSE)

\method{as.list}{dist}(x, sorted = TRUE, n = NULL, ...)
}
\arguments{
\item{n}{the top \code{n} most similar items will be returned, sorted in 
descending order.  If n is \code{NULL}, return all items.}

\item{sorted}{sort results in descending order if \code{TRUE}}

\item{...}{unused}
}
\description{
These functions compute distance matrix between documents and/or features from a 
\code{\link{dfm}} and return a standard \code{\link[stats]{dist}} object.
}
\section{Slots}{

\describe{
\item{\code{selection}}{character or character vector of document names or feature 
labels from the dfm}

\item{\code{n}}{the top \code{n} most similar items will be returned.  If n is \code{NULL}, return all items.}

\item{\code{margin}}{identifies the margin of the dfm on which similarity will be 
computed:  \code{documents} for documents or \code{features} for word/term
features.}

\item{\code{method}}{a valid method for computing similarity from 
\code{\link[proxy]{pr_DB}}, default "euclidean".}

\item{\code{normalize}}{a deprecated argument retained (temporarily) for legacy 
reasons.  If you want to compute similarity on a "normalized" dfm objects 
(e.g. \code{x}), wrap it in \code{\link{weight}(x, "relFreq")}.}

\item{\code{digits}}{decimal places to display similarity values}

\item{\code{tri}}{whether the upper triangle of the symmetric \eqn{V \times V} matrix is recorded}

\item{\code{diag}}{whether the diagonal of the distance matrix should be recorded}
}}
\examples{
# create a dfm from inaugural addresses from Reagan onwards
presDfm <- dfm(corpus_subset(inaugCorpus, Year > 1980), ignoredFeatures = stopwords("english"),
               stem = TRUE)

# compute some document similarities
(tmp <- textstat_dist(presDfm, margin = "documents"))
# output as a list
#as.list(tmp)
# for specific comparisons
textstat_dist(presDfm, "1985-Reagan", n = 5, margin = "documents")
textstat_dist(presDfm, c("2009-Obama" , "2013-Obama"), n = 5, margin = "documents")
textstat_dist(presDfm, c("2009-Obama" , "2013-Obama"), margin = "documents")
#textstat_dist(presDfm, "2005-Bush", margin = "documents", method = "eJaccard")

}
\seealso{
\link{dfm}
}
\keyword{internal}

