% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfm_weight.R
\name{docfreq}
\alias{docfreq}
\title{Compute the (weighted) document frequency of a feature}
\usage{
docfreq(
  x,
  scheme = c("count", "inverse", "inversemax", "inverseprob", "unary"),
  smoothing = 0,
  k = 0,
  base = 10,
  threshold = 0
)
}
\arguments{
\item{x}{object for which idf or tf-idf will be computed (a document-feature
matrix)}

\item{scheme}{a label of the weight type:
\describe{
\item{\code{count}}{\eqn{tf_{ij}}, an integer feature count (default when a dfm is created)}
\item{\code{prop}}{the proportion of the feature counts of total feature
counts (aka relative frequency), calculated as \eqn{tf_{ij} / \sum_j tf_{ij}}}
\item{\code{propmax}}{the proportion of the feature counts of the highest
feature count in a document, \eqn{tf_{ij} / \textrm{max}_j tf_{ij}}}
\item{\code{logcount}}{take the 1 + the logarithm of each count, for the
given base, or 0 if the count was zero: \eqn{1 +
  \textrm{log}_{base}(tf_{ij})} if \eqn{tf_{ij} > 0}, or 0 otherwise.}
\item{\code{boolean}}{recode all non-zero counts as 1}
\item{\code{augmented}}{equivalent to \eqn{k + (1 - k) *} \code{dfm_weight(x, "propmax")}}
\item{\code{logave}}{1 + the log of the counts) / (1 + log of the counts / the average count within document), or
\deqn{\frac{1 + \textrm{log}_{base} tf_{ij}}{1 + \textrm{log}_{base}(\sum_j tf_{ij} / N_i)}}}
}}

\item{smoothing}{constant added to the dfm cells for smoothing, default is 1}

\item{k}{the k for the augmentation when \code{scheme = "augmented"}}

\item{base}{the base for the logarithms in the \code{\link[=dfm_weight]{dfm_weight()}} and
\code{\link[=docfreq]{docfreq()}} calls; default is 10}

\item{threshold}{numeric value of the threshold \emph{above which} a feature
will considered in the computation of document frequency.  The default is
0, meaning that a feature's document frequency will be the number of
documents in which it occurs greater than zero times.}
}
\value{
a numeric vector of document frequencies for each feature
}
\description{
For a \link{dfm} object, returns a (weighted) document frequency for each
term.  The default is a simple count of the number of documents in which a
feature occurs more than a given frequency threshold.  (The default threshold
is  zero, meaning that any feature occurring at least once in a document will
be counted.)
}
\examples{
dfmat1 <- dfm(data_corpus_inaugural[1:2])
docfreq(dfmat1[, 1:20])

# replication of worked example from
# https://en.wikipedia.org/wiki/Tf-idf#Example_of_tf.E2.80.93idf
dfmat2 <- 
    matrix(c(1,1,2,1,0,0, 1,1,0,0,2,3),
           byrow = TRUE, nrow = 2,
           dimnames = list(docs = c("document1", "document2"),
                           features = c("this", "is", "a", "sample",
                                        "another", "example"))) \%>\%
    as.dfm()
dfmat2
docfreq(dfmat2)
docfreq(dfmat2, scheme = "inverse")
docfreq(dfmat2, scheme = "inverse", k = 1, smoothing = 1)
docfreq(dfmat2, scheme = "unary")
docfreq(dfmat2, scheme = "inversemax")
docfreq(dfmat2, scheme = "inverseprob")
}
\references{
Manning, C. D., Raghavan, P., & Schütze, H. (2008).
\emph{Introduction to Information Retrieval}. Cambridge: Cambridge University Press.
\url{https://nlp.stanford.edu/IR-book/pdf/irbookonlinereading.pdf}
}
\keyword{dfm}
\keyword{weighting}
