% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docnames.R
\name{docnames}
\alias{docnames}
\alias{docnames<-}
\alias{docid}
\alias{segid}
\title{Get or set document names}
\usage{
docnames(x)

docnames(x) <- value

docid(x)

segid(x)
}
\arguments{
\item{x}{the object with docnames}

\item{value}{a character vector of the same length as \code{x}}
}
\value{
\code{docnames} returns a character vector of the document names

\verb{docnames <-} assigns new values to the document names of an object.
docnames can only be character, so any non-character value assigned to be a
docname will be coerced to mode \code{character}.

\code{docid} returns an internal variable denoting the original "docname"
from which a document came.  If an object has been reshaped (e.g.
\code{\link[=corpus_reshape]{corpus_reshape()}} or segmented (e.g. \code{\link[=corpus_segment]{corpus_segment()}}), \code{docid(x)} returns
the original docnames but \code{segid(x)} does the serial number of those segments
within the original document.
}
\description{
Get or set the document names of a \link{corpus}, \link{tokens}, or \link{dfm} object.
}
\note{
\code{docid} and \code{segid} are designed primarily for developers, not for end users.  In
most cases, you will want \code{docnames} instead.  It is, however, the
default for \link{groups}, so that documents that have been previously reshaped
(e.g. \code{\link[=corpus_reshape]{corpus_reshape()}} or segmented (e.g.
\code{\link[=corpus_segment]{corpus_segment()}}) will be regrouped into their original \code{docnames} when
\code{groups = docid(x)}.
}
\examples{
# get and set doument names to a corpus
corp <- data_corpus_inaugural
docnames(corp) <- char_tolower(docnames(corp))

# get and set doument names to a tokens
toks <- tokens(corp)
docnames(toks) <- char_tolower(docnames(toks))

# get and set doument names to a dfm
dfmat <- dfm(tokens(corp))
docnames(dfmat) <- char_tolower(docnames(dfmat))

# reassign the document names of the inaugural speech corpus
corp <- data_corpus_inaugural
docnames(corp) <- paste0("Speech", seq_len(ndoc(corp)))


corp <- corpus(c(textone = "This is a sentence.  Another sentence.  Yet another.",
                 textwo = "Sentence 1. Sentence 2."))
corp_sent <- corp |>
    corpus_reshape(to = "sentences")
docnames(corp_sent)

# docid
docid(corp_sent)
docid(tokens(corp_sent))
docid(dfm(tokens(corp_sent)))

# segid
segid(corp_sent)
segid(tokens(corp_sent))
segid(dfm(tokens(corp_sent)))
}
\seealso{
\code{\link[=featnames]{featnames()}}
}
\keyword{corpus}
\keyword{dfm}
