% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus.R, R/textfile.R
\name{docvars}
\alias{docvars}
\alias{docvars.corpus}
\alias{docvars.corpusSource}
\alias{docvars<-}
\alias{docvars<-.corpus}
\title{get or set for document-level variables}
\usage{
docvars(x, field = NULL, drop = TRUE)

\method{docvars}{corpus}(x, field = NULL, drop = TRUE)

docvars(x, field = NULL) <- value

\method{docvars}{corpus}(x, field = NULL) <- value

\method{docvars}{corpusSource}(x, field = NULL)
}
\arguments{
\item{x}{corpus whose document-level variables will be read or set}

\item{field}{string containing the document-level variable name}

\item{drop}{If TRUE the result is coerced to a vector, if only one column is
selected. This only works for extracting elements, not for the replacement.
See \code{\link{drop}} for further details.}

\item{value}{the new values of the document-level variable}
}
\value{
\code{docvars} returns a data.frame of the document-level variables

\code{docvars<-} assigns \code{value} to the named \code{field}
}
\description{
Get or set document-level variables in a corpus.
}
\examples{
# get docvars
head(docvars(ie2010Corpus))
head(docvars(ie2010Corpus, "party"))
head(docvars(ie2010Corpus, "party", drop = FALSE))
 
# replacement
docvars(inaugCorpus, "President") <- paste("prez", 1:ndoc(inaugCorpus), sep="")
head(docvars(inaugCorpus))
}

