% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus.R, R/textfile.R
\name{docvars}
\alias{docvars}
\alias{docvars.corpus}
\alias{docvars.corpusSource}
\alias{docvars<-}
\title{get or set for document-level variables}
\usage{
docvars(x, ...)

\method{docvars}{corpus}(x, field = NULL, ...)

docvars(x, field = NULL) <- value

docvars(x, field = NULL) <- value

\method{docvars}{corpusSource}(x, ...)
}
\arguments{
\item{x}{corpus whose document-level variables will be read or set}

\item{...}{not used}

\item{field}{string containing the document-level variable name}

\item{value}{the new values of the document-level variable}
}
\value{
\code{docvars} returns a data.frame of the document-level variables

\code{docvars<-} assigns \code{value} to the named \code{field}
}
\description{
Get or set variables for the documents in a corpus
}
\examples{
head(docvars(inaugCorpus))
docvars(inaugCorpus, "President") <- paste("prez", 1:ndoc(inaugCorpus), sep="")
head(docvars(inaugCorpus))
}

