% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/corpus.R
\name{encoding}
\alias{encoding}
\alias{encoding<-}
\title{get the encoding of documents in a corpus}
\usage{
encoding(x, drop = TRUE)

encoding(x) <- value
}
\arguments{
\item{x}{a corpus object}

\item{drop}{return as a vector if \code{TRUE}, otherwise return a \code{data.frame}}

\item{value}{a character vector or scalar representing the new value of the encoding (see Note)}
}
\description{
Get or set the \code{_encoding} document-level metadata field(s) in a corpus.
}
\details{
This function modifies the \code{_encoding} value set by
  \code{\link{metadoc}}.  It is a wrapper for \code{metadoc(corp, "encoding")}.
}
\note{
This function differs from R's built-in \link{Encoding} function, which
  only allows the four values of "latin1", "UTF-8", "bytes", and "unknown"
  (and which assigns "unknown" to any text that contains only ASCII characters).
  Legal values for encodings must be from \link{iconvlist}.  Note that encoding
  does not convert or set encodings, it simply records a user declaration of a
  valid encoding.  (We hope to implement checking and conversion later.)
}

