% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataDocs.R
\docType{data}
\name{exampleString}
\alias{exampleString}
\title{A paragraph of text for testing various text-based functions}
\format{character vector with one element}
\description{
This is a long paragraph (2,914 characters) of text taken from an Irish budget speech by Joe Higgins
}
\examples{
tokenize(exampleString, removePunct = TRUE)
}

