% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern2fixed.R
\name{expand}
\alias{expand}
\title{Simpler and faster version of expand.grid() in base package}
\usage{
expand(elem)
}
\arguments{
\item{elem}{list of elements to be combined}
}
\description{
Simpler and faster version of expand.grid() in base package
}
\examples{
quanteda:::expand(list(c("a", "b", "c"), c("x", "y")))
}
\keyword{internal}
