% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcm.R
\docType{class}
\name{fcm-class}
\alias{fcm-class}
\title{Virtual class "fcm" for a feature co-occurrence matrix}
\description{
The fcm class of object is a special type of \link{dfm-class}
  object with additional slots, described below.
}
\section{Slots}{

\describe{
\item{\code{context}}{the context definition}

\item{\code{window}}{the size of the window, if \code{context = "window"}}

\item{\code{count}}{how co-occurrences are counted}

\item{\code{weights}}{context weighting for distance from target feature, equal in length to \code{window}}

\item{\code{tri}}{whether the lower triangle of the symmetric \eqn{V \times V} matrix is recorded}

\item{\code{ordered}}{whether a term appears before or after the target feature 
are counted seperately}
}}
\seealso{
\link{fcm}
}
\keyword{internal}

