% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfm_weight.R
\name{featfreq}
\alias{featfreq}
\title{Compute the frequencies of features}
\usage{
featfreq(x)
}
\arguments{
\item{x}{a \link{dfm}}
}
\value{
a (named) numeric vector of feature frequencies
}
\description{
For a \link{dfm} object, returns a frequency for each feature, computed
across all documents in the dfm. This is equivalent to \code{colSums(x)}.
}
\examples{
dfmat <- dfm(data_char_sampletext)
featfreq(dfmat)
}
\seealso{
\code{\link[=dfm_tfidf]{dfm_tfidf()}}, \code{\link[=dfm_weight]{dfm_weight()}}
}
\keyword{dfm}
\keyword{weighting}
