% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dfm.R
\name{features}
\alias{features}
\alias{features.dfm}
\title{#compute the tf-idf weights of a dfm
#
#Returns a matrix of tf-idf weights, as a \link{dfm} object
#
#data(inaugCorpus)
#dtm <- dfm(inaugCorpus)
#dtm[1:10, 100:110]
#tfidf(dtm)[1:10, 100:110]
#tfidf(dtm, normalize=FALSE)[1:10, 100:110]
#normalizes the term frequencies a dfm
#
#Returns a matrix of term weights, as a \link{dfm} object
#
#data(inaugCorpus)
#dtm <- dfm(inaugCorpus)
#dtm[1:10, 100:110]
#tf(dtm)[1:10, 100:110]
extract the feature labels from a \link{dfm}}
\usage{
features(x)

\method{features}{dfm}(x)
}
\arguments{
\item{x}{the object (dfm) whose features will be extracted}
}
\value{
Character vector of the features
}
\description{
Extract the features from a document-feature matrix, which are stored as the column names
of the \link{dfm} object.
}
\examples{
features(dfm(inaugTexts))[1:50]  # first 50 features (alphabetically sorted)
}

