% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfm-methods.R
\name{features}
\alias{features}
\alias{features.dfm}
\title{extract the feature labels from a dfm}
\usage{
features(x)

\method{features}{dfm}(x)
}
\arguments{
\item{x}{the object (dfm) whose features will be extracted}
}
\value{
Character vector of the features
}
\description{
Extract the features from a document-feature matrix, which are stored as the column names
of the \link{dfm} object.
}
\examples{
inaugDfm <- dfm(data_char_inaugural, verbose = FALSE)

# first 50 features (in original text order)
head(features(inaugDfm), 50)

# first 50 features alphabetically
head(sort(features(inaugDfm)), 50)

# contrast with descending total frequency order from topfeatures()
names(topfeatures(inaugDfm, 50))
}

