% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dfm-main.R
\name{features}
\alias{features}
\alias{features.dfm}
\title{extract the feature labels from a dfm}
\usage{
features(x)

\method{features}{dfm}(x)
}
\arguments{
\item{x}{the object (dfm) whose features will be extracted}
}
\value{
Character vector of the features
}
\description{
Extract the features from a document-feature matrix, which are stored as the column names
of the \link{dfm} object.
}
\examples{
features(dfm(inaugTexts))[1:50]  # first 50 features (alphabetically sorted)
}

