% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{field_system}
\alias{field_system}
\alias{field_system<-}
\alias{field_object}
\alias{field_object<-}
\alias{field_user}
\alias{field_user<-}
\title{Shortcut functions to access or assign metadata}
\usage{
field_system(x, field = NULL)

field_system(x, field = NULL) <- value

field_object(x, field = NULL)

field_object(x, field = NULL) <- value

field_user(x, field = NULL)

field_user(x, field = NULL) <- value
}
\arguments{
\item{x}{a list of attributes extracted from a \code{dfm}, \code{tokens}, or \code{corpus}
object by \code{attributes(x)}}

\item{field}{name of the sub-field to access or assign values}
}
\description{
Internal functions to access or replace an object metadata field without
going through attribute trees. \code{field_system()}, \code{field_object()} and
\code{field_user()} correspond to the system, object and user meta fields,
respectively.
}
\keyword{internal}
