% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fitmodel}
\alias{fitmodel}
\alias{fitmodel.dfm}
\title{fit a text model}
\usage{
fitmodel(x, ...)

\method{fitmodel}{dfm}(x, method = c("wordscores", "NB"), ...)
}
\arguments{
\item{x}{a dfm object}

\item{...}{additional parameters to be passed to other functions}

\item{method}{the model type to be fit.  Currently implemented methods are:
\describe{
  \item{\code{wordscores}}{Fits the "wordscores" model of Laver, Benoit, and
  Garry (2003). Options include the original linear scale of LBG or the logit
  scale proposed by Beauchamps (2001).  See \link{textmodel_wordscores}.}

  \item{\code{NB}}{Fits a Naive Bayes model to the dfm, with options for
  smoothing, setting class priors, and a choice of multinomial or binomial
  probabilities.}}}
}
\value{
a \code{textmodel} class list, containing the fitted model and
  additional information specific to the model class.  See the methods for
  specific models, e.g. \link{textmodel_wordscores}, \link{textmodel_NB},
  etc.
}
\description{
Fit a text model to a dfm.
}
\examples{
require(quantedaData)
data(ie2010Corpus)
ieDfm <- dfm(ie2010Corpus)
ws <- fitmodel(ieDfm, train=c("2010_BUDGET_06_Enda_Kenny_FG", "2010_BUDGET_05_Brian_Cowen_FF"),
               train.scores=c(-1,1), smooth=1)
bs <- fitmodel(ieDfm, train=c("2010_BUDGET_06_Enda_Kenny_FG", "2010_BUDGET_05_Brian_Cowen_FF"),
               train.scores=c(-1,1), scale="logit", smooth=1)
plot(ws$featureScores, bs$featureScores, xlim=c(-1, 1),
     xlab="Linear word score", ylab="Logit word score")

# prediction method for wordscores
predict(ws, ieDfm, rescaling="mv")
}
\references{
LBG (2003); Beauchamps (20XX)
}

