% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel_nb.R
\name{gain.textmodel_nb}
\alias{gain.textmodel_nb}
\title{Compute the information gain of features from a fitted \link{textmodel_nb}
object}
\usage{
gain.textmodel_nb(object, base = 2, class_specific = TRUE)
}
\arguments{
\item{object}{a \link{textmodel_nb} object}

\item{base}{the base of the logarithmic function to be used. Defaults to 2.}

\item{class_specific}{whether to return class specific information gain
scores or not. Defaults to \code{TRUE}.}
}
\description{
Computes the information gain of features from a fitted \link{textmodel_nb}
object.
\describe{
  \item{Formula:}{
  \deqn{G(w) = \sum_{i=1}^{C}{P(c_i) \log{P(c_i)}}
  + P(w) \sum_{i=1}^{C}{P(c_i | w) \log{P(c_i | w)}}
  + P(w^{c}) \sum_{i=1}^{C}{P(c_i | w^{c}) \log{P(c_i | w^{c})}}
  }
  where \itemize{
  \item \eqn{G(w)} is the information gain from word \eqn{w},
  \item \eqn{P(c_i)} is the prior probability of class \eqn{i}, 
  \item \eqn{P(w)} is the probability of word \eqn{w} being present in a document ,
  \item \eqn{P(w^c)} is the probability of word \eqn{w} being absent in a document,
  \item \eqn{P(c_i | w)} is the posterior probability of class \eqn{i} given that word \eqn{w} is present and 
  \item \eqn{P(c_i | w^c)} is the posterior probability  of class\eqn{i} given that word \eqn{w} is absent.
  }
 }
}
}
\examples{
txt <- c(d1 = "Chinese Beijing Chinese",
         d2 = "Chinese Chinese Shanghai",
         d3 = "Chinese Macao",
         d4 = "Tokyo Japan Chinese",
         d5 = "Chinese Chinese Chinese Tokyo Japan")
trainingset <- dfm(txt, tolower = FALSE)
trainingclass <- factor(c("Y", "Y", "Y", "N", NA), ordered = TRUE)
tmod <- textmodel_nb(trainingset, y = trainingclass, prior = "docfreq")
gain.textmodel_nb(tmod)
gain.textmodel_nb(tmod, class_specific = FALSE)
}
\references{
Jurafsky, Daniel and James H. Martin.  (2018).
\href{https://web.stanford.edu/~jurafsky/slp3/4.pdf}{\emph{Speech and
Language Processing}}.   Stanford University typescript.  Draft of September
23.
}
\keyword{textmodel}
\keyword{textmodel_nb}
