% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel_nb.R
\name{gain.textmodel_nb}
\alias{gain.textmodel_nb}
\title{Computes the information gain of features from a fitted \link{textmodel_nb} object}
\usage{
gain.textmodel_nb(model, base = 2, class_specific = TRUE)
}
\arguments{
\item{model}{a \link{textmodel_nb} object}

\item{base}{the base of the logarithmic function to be used. Defaults to 2.}

\item{class_specific}{whether to return class specific information gain scores or not. Defaults to TRUE.}
}
\description{
Computes the information gain of features from a fitted \link{textmodel_nb} object
}
\examples{
txt <- c(d1 = "Chinese Beijing Chinese",
d2 = "Chinese Chinese Shanghai",
d3 = "Chinese Macao",
d4 = "Tokyo Japan Chinese",
d5 = "Chinese Chinese Chinese Tokyo Japan")
trainingset <- dfm(txt, tolower = FALSE)
trainingclass <- factor(c("Y", "Y", "Y", "N", NA), ordered = TRUE)
nb <- textmodel_nb(trainingset, y = trainingclass, prior = "docfreq")
gain.textmodel_nb(nb)
 Chinese      Beijing     Shanghai        Macao       Tokyo       Japan
 N 0.019878397 0.0009238016 0.0009238016 0.0009238016 0.018915497 0.018915497
 Y 0.006033503 0.0002872741 0.0002872741 0.0002872741 0.008886182 0.008886182
gain.textmodel_nb(nb,class_specific = FALSE)
 Chinese     Beijing    Shanghai       Macao      Tokyo      Japan
 info_gain 0.0259119 0.001211076 0.001211076 0.001211076 0.02780168 0.02780168
}
\keyword{textmodel}
\keyword{textmodel_nb}
