% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docvars.R
\name{get_docvars}
\alias{get_docvars}
\title{Internal function to extract docvars}
\usage{
get_docvars(x, field = NULL, user = TRUE, system = FALSE, drop = FALSE)
}
\arguments{
\item{x}{an object from which docvars are extracted}

\item{field}{name of docvar fields}

\item{user}{if \code{TRUE}, return user variables}

\item{system}{if \code{TRUE}, return system variables}

\item{drop}{if \code{TRUE}, convert data.frame with one variable to a vector}
}
\description{
Internal function to extract docvars
}
\keyword{internal}
