% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_object_version}
\alias{get_object_version}
\alias{is_pre2}
\title{Get the package version that created an object}
\usage{
get_object_version(x)

is_pre2(x)
}
\value{
A three-element integer vector of class "package_version". For
versions of the package < 1.5 for which no version was recorded in the
object, \code{c(1, 4, 0)} is returned.

\code{ispre2()} returns \code{TRUE} if the object was created before
\pkg{quanteda} version 2, or \code{FALSE} otherwise
}
\description{
Return the the \pkg{quanteda} package version in which a \link{dfm},
\link{tokens}, or \link{corpus} object was created.
}
\keyword{internal}
\keyword{utils}
