\name{getTextDirGui}
\alias{getTextDirGui}
\title{provides a gui interface to choose a gui to load texts from}
\usage{
getTextDirGui()
}
\value{
character vector of texts read from disk
}
\description{
launches a GUI to allow the user to choose a directory from
which to load all files.
}
\examples{
getTextFiles('/home/paul/documents/libdem09.txt')
}

