% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quanteda-documentation.R
\name{groups}
\alias{groups}
\title{Grouping variable(s) for various functions}
\arguments{
\item{groups}{grouping variable for sampling, equal in length to the number
of documents. This will be evaluated in the docvars data.frame, so that
docvars may be referred to by name without quoting. This also changes
previous behaviours for \code{groups}. See \code{news(Version >= "2.9", package = "quanteda")} for details.}

\item{fill}{logical; if \code{TRUE} and \code{groups} is a factor, then use all levels
of the factor when forming the new documents of the grouped object.  This
will result in a new "document" with empty content, for levels not
observed.  Has no effect if the \code{groups} variable(s) are not factors.}
}
\description{
Groups for aggregation by various functions that take grouping options.
}
\keyword{internal}
