% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quanteda-documentation.R
\name{groups}
\alias{groups}
\title{Grouping variable(s) for various functions}
\arguments{
\item{groups}{grouping variable for sampling, equal in length to the number
of documents. This will be evaluated in the docvars data.frame, so that
docvars may be referred to by name without quoting. This also changes
previous behaviours for \code{groups}. See \code{news(Version >= "3.0", package = "quanteda")} for details.}

\item{fill}{logical; if \code{TRUE} and \code{groups} is a factor, then use all levels
of the factor when forming the new documents of the grouped object.  This
will result in a new "document" with empty content for levels not observed,
but for which an empty document may be needed.  If \code{groups} is a factor of
dates, for instance, then \code{fill = TRUE} ensures that the new object will
consist of one new "document" by date, regardless of whether any documents
previously existed with that date.  Has no effect if the \code{groups}
variable(s) are not factors.}
}
\description{
Groups for aggregation by various functions that take grouping options.
}
\seealso{
\code{\link[=corpus_group]{corpus_group()}}, \code{\link[=tokens_group]{tokens_group()}}, \code{\link[=dfm_group]{dfm_group()}}
}
\keyword{internal}
