% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataDocs.R
\docType{data}
\name{ie2010Corpus}
\alias{ie2010Corpus}
\alias{iebudgets}
\title{Irish budget speeches from 2010}
\format{The corpus object for the 2010 budget speeches, with document-level
  variables for year, debate, serial number, first and last name of the
  speaker, and the speaker's party.}
\source{
Lowe and Benoit (2013)
}
\description{
Speeches and document-level variables from the debate over the
  Irish budget of 2010.
}
\examples{
summary(ie2010Corpus)
}
\references{
Lowe, Will, and Kenneth R Benoit. 2013. "Validating Estimates of
  Latent Traits From Textual Data Using Human Judgment as a Benchmark."
  \emph{Political Analysis} 21: 298-313.
}

