% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message.R
\name{info}
\alias{info}
\title{Conditionally format and print messages}
\usage{
info(x, values, indices = NULL, warn = FALSE, ...)
}
\arguments{
\item{x}{message template to be pasted to \code{\link[=sprintf]{sprintf()}}}

\item{values}{list of values to be printed. Coerced to list if vector is given.}

\item{indices}{list of integer to specify which value to be printed}

\item{warn}{if \code{TRUE}, raise warning}

\item{...}{passed to underlying message functions}
}
\description{
Conditionally format and print messages
}
\examples{
\dontrun{
quanteda:::info("you cannot delete \%s", 
                c("a document", "documents"), indices = TRUE, warn = TRUE)
quanteda:::info("tokens has \%s", 
                c("sentences", "paragraphs", "documents"), indices = 2)

dfmt <- data_dfm_lbgexample
quanteda:::info("dfm has \%d \%s and \%d \%s", 
     list(ndoc(dfmt), c("document", "documents"),
          nfeat(dfmt), c("feature", "features")), 
     list(1, ndoc(dfmt) > 1, 
          1, nfeat(dfmt) > 1), warn = FALSE)
}      
}
