% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textstat.R
\name{is.collocations}
\alias{is.collocations}
\title{Check if an object is collocations}
\usage{
is.collocations(x)
}
\arguments{
\item{x}{object to be checked}
}
\value{
\code{TRUE} if the object is of class \code{collocations}, \code{FALSE} otherwise
}
\description{
Function to check if an object is a collocations object, created by
\code{quanteda.textstats::textstat_collocations()}.
}
