% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern2fixed.R
\name{is_indexed}
\alias{is_indexed}
\title{Check if a glob pattern is indexed by index_types}
\usage{
is_indexed(pattern)
}
\arguments{
\item{pattern}{a glob pattern to be tested}
}
\description{
Internal function for \code{select_types} to check if a glob pattern is
indexed by
\code{index_types}.
}
\keyword{internal}
