% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_regex}
\alias{is_regex}
\title{Internal function for \code{select_types()} to check if a string is a regular expression}
\usage{
is_regex(x)
}
\arguments{
\item{x}{a character string to be tested}
}
\description{
Internal function for \code{select_types()} to check if a string is a regular expression
}
\keyword{internal}
