% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinTokens.R, R/joinTokens2.R
\name{joinTokens}
\alias{joinTokens}
\alias{joinTokens.tokenizedTexts}
\alias{joinTokens.tokens}
\alias{joinTokens2.tokenizedTexts}
\title{join tokens function}
\usage{
joinTokens(x, ...)

\method{joinTokens}{tokenizedTexts}(x, ...)

\method{joinTokens}{tokens}(x, sequences, concatenator = "_",
  valuetype = c("glob", "fixed", "regex"), verbose = FALSE,
  case_insensitive = TRUE, ...)

\method{joinTokens2}{tokenizedTexts}(x, ...)
}
\arguments{
\item{x}{tokens or tokenizedTexts object}

\item{...}{additional arguments passed to other methods}

\item{sequences}{features to concatenate, a list of characters}

\item{concatenator}{character used for joining tokens}

\item{valuetype}{how to interpret sequences: \code{fixed} for words as
is; \code{"regex"} for regular expressions; or \code{"glob"} for
"glob"-style wildcard}

\item{verbose}{display progress}

\item{case_insensitive}{if \code{TRUE}, ignore case when matching}
}
\description{
For a set of tokens, given a set of token sequences, join the tokens matching the sequences.
}
\examples{
toks <- tokens(inaugCorpus, removePunct = TRUE)
seqs_token <- list(c('foreign', 'policy'), c('United', 'States'))
seqs_glob <- list(c('foreign', 'polic*'), c('United', 'States'))
seqs_regex <- list(c('^foreign', '^polic(ie|y)'), c('^United', '^States'))
toks2 <- joinTokens(toks, seqs_token, "_", 'fixed')
toks2 <- joinTokens(toks, seqs_glob, "_", 'glob')
toks2 <- joinTokens(toks, seqs_regex, "_", 'regex')
kwic(toks2, 'foreign_policy', window = 1) # joined
kwic(toks2, c('foreign', 'policy'), window = 1) # not joined
kwic(toks2, 'United_States', window = 1) # joined
               
}
\author{
Kohei Watanabe and Kenneth Benoit
}

