% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinTokens.R
\name{joinTokens}
\alias{joinTokens}
\title{join tokens function}
\usage{
joinTokens(x, seqs_token, concatenator = "-")
}
\arguments{
\item{x}{some object}

\item{seqs_token}{something}

\item{concatenator}{character used for joining tokens}
}
\description{
Needs some more explanation
}
\examples{
data(SOTUCorpus, package = "quantedaData")
toks <- tokenize(SOTUCorpus, removePunct = TRUE)
seqs_token <- list(c('foreign', 'policy'), c('United', 'States'))
toks2 <- joinTokens(toks, seqs_token, "_")
kwic(toks2, 'foreign_policy', window=1) # joined
kwic(toks2, c('foreign', 'policy'), window=1) # not joined
kwic(toks2, 'United_States', window=1) # joined
}

