% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinTokens.R
\name{joinTokens}
\alias{joinTokens}
\title{join tokens function}
\usage{
joinTokens(x, sequences, concatenator = "_", valuetype = c("glob", "fixed",
  "regex"), verbose = FALSE, case_insensitive = TRUE)
}
\arguments{
\item{x}{some object}

\item{sequences}{list of vector of features to concatenate}

\item{concatenator}{character used for joining tokens}

\item{valuetype}{how to interpret sequences: \code{fixed} for words as
is; \code{"regex"} for regular expressions; or \code{"glob"} for
"glob"-style wildcard}

\item{verbose}{display progress}

\item{case_insensitive}{if \code{TRUE}, ignore case when matching}
}
\description{
join tokens function
}
\examples{
# simple example
txt <- c("a b c d e f g", "A B C D E F G", "A b C d E f G", "aaa bbb ccc ddd eee fff ggg", "a_b b_c c_d d_e e_f f_g") 
toks_hash <- tokens(txt)
seqs <- tokens(c("a b", "C D", "aa* bb*", "eEE FFf", "d_e e_f"), hash=FALSE, what="fastestword")
joinTokens(toks_hash, seqs, valuetype = "glob", case_insensitive = TRUE, verbose=TRUE)
joinTokens(toks_hash, seqs, valuetype = "glob", case_insensitive = FALSE, verbose=TRUE)
}
\author{
Kohei Watanabe and Kenneth Benoit
}

