% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinTokens.R
\name{joinTokens}
\alias{joinTokens}
\alias{joinTokens.tokenizedTexts}
\alias{joinTokens.tokens}
\title{join tokens function}
\usage{
joinTokens(x, ...)

\method{joinTokens}{tokenizedTexts}(x, ...)

\method{joinTokens}{tokens}(x, sequences, concatenator = "_",
  valuetype = c("glob", "fixed", "regex"), verbose = FALSE,
  case_insensitive = TRUE, ...)
}
\arguments{
\item{x}{tokens or tokenizedTexts object}

\item{...}{additional arguments passed to other methods}

\item{sequences}{features to concatenate, a list of characters}

\item{concatenator}{character used for joining tokens}

\item{valuetype}{how to interpret sequences: \code{fixed} for words as
is; \code{"regex"} for regular expressions; or \code{"glob"} for
"glob"-style wildcard}

\item{verbose}{display progress}

\item{case_insensitive}{if \code{TRUE}, ignore case when matching}
}
\description{
For a set of tokens, given a set of token sequences, join the tokens matching the sequences.
}
\examples{

toks <- tokens(data_corpus_inaugural, removePunct = TRUE)
seqs <- list(c('foreign', 'polic*'), c('United', 'States'))
kwic(toks, 'fUnited_States', window = 1) # no exisit
kwic(toks, 'foreign_policy', window = 1) # no exisit
toks <- joinTokens(toks, seqs, "_", 'glob')
kwic(toks, 'United_States', window = 1)
kwic(toks, 'foreign_policy', window = 1)
kwic(toks, c('foreign', 'policy'), window = 1) # no longer exisit
kwic(toks, c('united', 'states'), window = 1) # no longer exisit
               
}
\author{
Kohei Watanabe and Kenneth Benoit
}

