% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinTokens.R
\name{joinTokens.tokenizedTextsHashed}
\alias{joinTokens.tokenizedTextsHashed}
\title{join tokens function}
\usage{
joinTokens.tokenizedTextsHashed(x, sequences, concatenator = "_",
  valuetype = c("glob", "fixed", "regex"), verbose = FALSE,
  case_insensitive = TRUE)
}
\description{
join tokens function
}
\examples{
# simple example
txt <- c("The United States is bordered by the Atlantic Ocean and the Pacific Ocean.",
         "The Supreme Court of the United States is seldom in a united state.",
         "It's Arsenal versus Manchester United, states the announcer.", 
         "luv the united states XXOO :-) atlantical oceania")
toks_hash <- hashTokens(tokenize(txt))
dfm(as.tokenizedTexts(toks_hash))
phrasesFixed <- tokenize(c("United States", "Supreme Court", "Atlantic* Ocean*", "Pacific Ocean"), what="fastestword")
joinTokens.tokenizedTextsHashed(toks_hash, phrasesFixed, valuetype = "glob", case_insensitive = TRUE, verbose=TRUE)

# For development
x <- toks_hash
sequences <- phrasesFixed
valuetype <- 'glob'
concatenator = "_"
verbose = TRUE
case_insensitive = TRUE
}

