% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/corpus-sources.R
\name{json}
\alias{json}
\title{Function to read files with tweets in JSON format}
\usage{
json(path = NULL, source = "twitter", enc = "unknown", ...)
}
\arguments{
\item{path}{string describing the full path of a directory that
contains files in json format, or a vector of file names in
in json format}

\item{source}{source of data in JSON format.}

\item{enc}{encoding of the input json file}

\item{...}{additional arguments passed to \code{\link[streamR]{parseTweets}}}
}
\description{
Function to read files with tweets in JSON format
}
\examples{
\dontrun{
# name a directory of files in json format
tweets <- json("~/Dropbox/QUANTESS/corpora/tweets")
corpus(tweets)

# read a single file in json format
tweets <- json("~/Dropbox/QUANTESS/corpora/tweets/BarackObama.json")
corpus(tweets)
}
}

