% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kwic2.R
\name{kwic_split}
\alias{kwic_split}
\title{split kwic results}
\usage{
kwic_split(x, start, end)
}
\arguments{
\item{start}{numeric vector for starting positions of keywords}

\item{char}{character vector}

\item{start}{numeric vector for ending positions of keywords}
}
\description{
Helper function for kwic designed to parse sequence matches into contexts and keywords.
}
\examples{
\dontrun{
kwic_split(rep(list(letters[1:5]), 3), c(2, 3, 4), c(2, 4, 5))
}
}
\author{
Kohei Watanabe
}
\keyword{internal}
