% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kwic2.R
\name{kwic_split}
\alias{kwic_split}
\title{split kwic results}
\usage{
kwic_split(char, mask, window)
}
\arguments{
\item{char}{character vector}

\item{mask}{numeric vector with the same length as char where 1 indicate appearance of keywords}

\item{window}{size of the pre and post contexts}
}
\description{
Helper function for kwic designed to parse sequence matches into contexts and keywords.
}
\examples{
\dontrun{
kwic_split(letters[1:5], c(0, 1, 1, 0, 1), window = 1)
kwic_split(letters[1:5], c(1, 1, 1, 0, 1), window = 1)
}
}
\author{
Kohei Watanabe
}
\keyword{internal}

