% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textstat_lexdiv.R
\name{lexdiv}
\alias{lexdiv}
\alias{compute_lexdiv}
\title{Compute lexical diversity (internal functions)}
\usage{
compute_lexdiv(x, measure, log.base)
}
\arguments{
\item{x}{a \link{dfm} object}

\item{measure}{a list of lexical diversity measures.}

\item{log.base}{a numeric value defining the base of the logarithm (for
measures using logs)}
}
\value{
returns a data.frame of documents and their lexical diversity scores.
}
\description{
Internal function used in \code{\link{textstat_lexdiv}}.
}
