% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfm-methods.R
\name{matrix2dfm}
\alias{matrix2dfm}
\title{Converts a Matrix to a dfm}
\usage{
matrix2dfm(x, docvars = NULL, meta = NULL)
}
\arguments{
\item{x}{a Matrix}

\item{meta}{a list of values to be assigned to slots}
}
\description{
Converts a Matrix to a dfm
}
\keyword{internal}
