% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message.R
\name{message_dfm}
\alias{message_dfm}
\title{Print messages in dfm methods}
\usage{
message_dfm(operation, before, after)
}
\arguments{
\item{before, after}{object statistics before and after the operation.}
}
\description{
Print messages in dfm methods
}
\keyword{internal}
\keyword{message}
