% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message.R
\name{message_tokens}
\alias{message_tokens}
\title{Print messages in tokens methods}
\usage{
message_tokens(operation, before, after)
}
\arguments{
\item{before, after}{object statistics before and after the operation.}
}
\description{
Print messages in tokens methods
}
\keyword{internal}
\keyword{message}
