% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message.R
\name{messages}
\alias{messages}
\title{Message parameter documentation}
\arguments{
\item{verbose}{if \code{TRUE} print the number of tokens and documents before and
after the function is applied. The number of tokens does not include paddings.}

\item{before, after}{object statistics before and after the operation.}
}
\description{
Used in printing verbose messages for message_tokens() and message_dfm()
}
\seealso{
message_tokens() message_dfm()
}
\keyword{internal}
