% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{meta}
\alias{meta}
\alias{metacorpus}
\alias{meta<-}
\alias{"metacorpus<-"}
\alias{metacorpus<-}
\title{Get or set corpus metadata}
\usage{
meta(x, field = NULL, type = c("user", "system", "all"))

meta(x, field = NULL) <- value

metacorpus(x, field = NULL, type = c("user", "system", "all"))

metacorpus(x, field = NULL) <- value
}
\arguments{
\item{x}{a \link{corpus} object}

\item{field}{metadata field name(s);  if \code{NULL} (default), return all 
metadata names}

\item{type}{\code{"user"} for user-provided corpus-level meta-data; \code{"system"} for 
meta-data set automatically when the corpus is created; or \code{"all"} for all meta-data.}

\item{value}{new value of the corpus metadata field}
}
\value{
For \code{meta}, a named list of the metadata fields in the corpus.

  For \code{meta <-}, the corpus with the updated user-level metadata.  Only
  user-level meta-data may be assigned.
}
\description{
Get or set the corpus-level metadata in a \link{corpus} object.
}
\details{
\code{metacorpus} and \code{metacorpus<-} are synonyms but are deprecated.
}
\examples{
meta(data_corpus_inaugural)
meta(data_corpus_inaugural, "source")
meta(data_corpus_inaugural, "citation") <- "Presidential Speeches Online Project (2014)."
meta(data_corpus_inaugural, "citation")
}
\keyword{corpus}
