% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadoc.R
\name{metadoc}
\alias{metadoc}
\alias{metadoc<-}
\title{Get or set document-level meta-data}
\usage{
metadoc(x, field = NULL)

metadoc(x, field = NULL) <- value
}
\arguments{
\item{x}{a \link{corpus} object}

\item{field}{character, the name of the metadata field(s) to be queried or}

\item{value}{the new value of the new meta-data field}
}
\description{
Get or set document-level meta-data
}
\keyword{corpus}
