% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message.R
\name{msg}
\alias{msg}
\title{Conditionally format messages}
\usage{
msg(format, ..., pretty = TRUE)
}
\arguments{
\item{pretty}{if \code{TRUE}, message is passed to \code{\link[base:formatc]{base::prettyNum()}}.}
}
\description{
Conditionally format messages
}
\examples{
quanteda:::msg("you cannot delete \%s \%s", 2000, "documents")
}
\keyword{development}
\keyword{internal}
