% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docnames.R
\name{names-quanteda}
\alias{names-quanteda}
\alias{names<-.corpus}
\alias{names<-.tokens}
\alias{rownames<-,dfm-method}
\alias{rownames<-.dfm}
\alias{rownames<-,fcm-method}
\alias{rownames<-.fcm}
\alias{docid}
\title{Special handling for names of quanteda objects}
\usage{
\method{names}{corpus}(x) <- value

\method{names}{tokens}(x) <- value

\S4method{rownames}{dfm}(x) <- value

\S4method{rownames}{fcm}(x) <- value

docid(x)
}
\arguments{
\item{x}{an \R object.}

\item{value}{a character vector of up to the same length as \code{x}, or
    \code{NULL}.}
}
\description{
Keeps the element names and rownames in sync with the system docvar
\code{docname_}.
}
\keyword{corpus}
\keyword{internal}
