% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus-methods-quanteda.R, R/dfm-methods.R, R/tokenize.R, R/tokens.R
\name{ndoc}
\alias{ndoc}
\alias{ndoc.corpus}
\alias{ndoc.dfm}
\alias{ndoc.tokenizedTexts}
\alias{ndoc.tokens}
\alias{nfeature}
\alias{nfeature.corpus}
\alias{nfeature.dfm}
\title{get the number of documents or features}
\usage{
ndoc(x)

\method{ndoc}{corpus}(x)

\method{ndoc}{dfm}(x)

nfeature(x)

\method{nfeature}{corpus}(x)

\method{nfeature}{dfm}(x)

\method{ndoc}{tokenizedTexts}(x)

\method{ndoc}{tokens}(x)
}
\arguments{
\item{x}{a corpus or dfm object}
}
\value{
an integer (count) of the number of documents or features in the corpus or dfm
}
\description{
\code{ndoc} returns the number of documents or features in a quanteda object, which can be 
a corpus, dfm, or tokenized texts.

\code{nfeature} is an alias for \code{ntype} when applied to dfm
  objects.  For a corpus or set of texts, "features" are only defined through
  tokenization, so you need to use \code{\link{ntoken}} to count these.
}
\examples{
ndoc(corpus_subset(data_corpus_inaugural, Year>1980))
ndoc(dfm(corpus_subset(data_corpus_inaugural, Year>1980), verbose=FALSE))
nfeature(dfm(data_corpus_inaugural))
nfeature(dfm_trim(dfm(data_corpus_inaugural), min_docfreq = 5, min_count = 10))
}

