% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokens_ngrams.R
\name{ngrams}
\alias{ngrams}
\alias{ngrams.default}
\alias{ngrams.tokenizedTexts}
\alias{skipgrams}
\title{deprecated function name for forming ngrams and skipgrams}
\usage{
ngrams(x, ...)

ngrams.default(x, ...)

ngrams.tokenizedTexts(x, ...)

skipgrams(x, ...)
}
\description{
Deprecated function names for forming ngrams and skipgrams; use 
\code{\link{tokens_ngrams}} and \code{\link{tokens_skipgrams}} instead.
}
\keyword{deprecated}
\keyword{internal}
\keyword{tokens}

