% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ngrams}
\alias{ngrams}
\title{Create ngrams}
\usage{
ngrams(text, n = 2, concatenator = "_", include.all = FALSE, ...)
}
\arguments{
\item{text}{character vector containing the texts from which ngrams will be extracted}

\item{n}{the number of tokens to concatenate. Default is 2 for bigrams.}

\item{window}{how many words to be counted for adjacency.  Default is 1 for only immediately
neighbouring words.}

\item{concatenator}{character for combining words, default is \code{_} (underscore) character}

\item{include.all}{if TRUE, add n-1...1 grams to the returned list}

\item{...}{additional arguments passed to \link{tokenize}}
}
\value{
a character vector of ngrams
}
\description{
Create a set of ngrams (words in sequence) from a text.
}
\examples{
ngrams("The quick brown fox jumped over the lazy dog.", n=2)
ngrams("The quick brown fox jumped over the lazy dog.", n=3)
ngrams("The quick brown fox jumped over the lazy dog.", n=3, concatenator="~")
ngrams("The quick brown fox jumped over the lazy dog.", n=3, include.all=TRUE)
}
\author{
Ken Benoit, Kohei Watanabe, Paul Nulty
}

