% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ngrams.R
\name{ngrams_c}
\alias{ngrams_c}
\alias{skipgrams_c}
\title{C++ version of ngrams and skipgrams for character vectors}
\usage{
ngrams_c(x, n = 2, window = 1, concatenator = "_")

skipgrams_c(x, n, k, concatenator = "_")
}
\description{
Form ngrams and skipgrams from character vectors using a faster C++ method
}
\note{
\strong{KOHEI:} See how the argument to \code{n} and \code{window} work for the existing
\code{ngrams()} -- would it be faster to implement this in C++, or call it as in the existing code multiple times when
the arguments are vectorized?  See examples.

\strong{KOHEI:} the output should be identical to that of \code{skipgrams()}
(see examples), perhaps by adding an option to \code{skipgrams_c()} that
includes all of the recursive (k-1):1 options?  See
\href{http://homepages.inf.ed.ac.uk/ballison/pdf/lrec_skipgrams.pdf}{Guthrie,
D., B. Allison, W. Liu, and L. Guthrie. 2006. "A Closer Look at Skip-Gram
Modelling."}
}
\examples{
## form ngrams
#' ngrams_c(head(LETTERS), n = 3, window = 1) #"A_B_C" "B_C_D" "C_D_E" "D_E_F"
ngrams(head(LETTERS), n = 3, window = 1) #"A_B_C" "B_C_D" "C_D_E" "D_E_F"

tokens <- tokenize("the quick brown fox jumped over the lazy dog.",
removePunct = TRUE, simplify = TRUE)
ngrams(tokens, n = c(2,4), window = 1:2, concatenator = " ")
## BUT THIS FAILS
\dontrun{ngrams_c(tokens, n = c(2,4), window = 1:2, concatenator = " ")}

ngrams_c(head(LETTERS), n = 3, window = 1) #"A_B_C" "B_C_D" "C_D_E" "D_E_F"
ngrams(head(LETTERS), n = 3, window = 1) #"A_B_C" "B_C_D" "C_D_E" "D_E_F"
ngrams_c(head(LETTERS), n = 3, window = 2) #"A_C_E" "B_D_F"
ngrams(head(LETTERS), n = 3, window = 2) #"A_C_E" "B_D_F"
ngrams_c(head(LETTERS), n = 3, window = 3) # None
ngrams(head(LETTERS), n = 3, window = 3) # None

## form skipgrams
skipgrams_c(head(LETTERS), n = 2, k = 1) #"A_B" "B_C" "C_D" "D_E" "E_F" "A_C" "B_D" "C_E" "D_F"
skipgrams(head(LETTERS), n = 2, k = 1)   #"A_B" "B_C" "C_D" "D_E" "E_F" "A_C" "B_D" "C_E" "D_F"

skipgrams_c(head(LETTERS), n = 2, k = 2) #"A_B" "B_C" "C_D" "D_E" "E_F" "A_C" "B_D" "C_E" "D_F" "A_D" "B_E" "C_F"
skipgrams(head(LETTERS), n = 2, k = 2)   #"A_B" "B_C" "C_D" "D_E" "E_F" "A_C" "B_D" "C_E" "D_F" "A_D" "B_E" "C_F"
tokens <- tokenize(toLower("Insurgents killed in ongoing fighting."),
removePunct = TRUE, simplify = TRUE)
skipgrams(tokens, n = 2, k = 2, concatenator = " ")
skipgrams_c(tokens, n = 3, k = 2, concatenator = " ")
}

