% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object-builder.R
\name{build_dfm}
\alias{build_dfm}
\alias{build_tokens}
\alias{build_corpus}
\title{Object compilers}
\usage{
build_dfm(x, features, docvars = data.frame(), meta = list(), ...)

build_tokens(
  x,
  types,
  padding = FALSE,
  docvars = data.frame(),
  meta = list(),
  ...
)

build_corpus(x, docvars = data.frame(), meta = list(), ...)
}
\arguments{
\item{docvars}{data.frame for docment level variables}

\item{meta}{list for meta fields}

\item{...}{added to object meta fields}

\item{types}{character for types of resulting `tokens`` object}

\item{padding}{logical indicating if the \code{tokens} object contains paddings}

\item{source}{character that indicating source object}

\item{feature}{character for feature of resulting \code{dfm}}
}
\description{
Object compilers
}
\keyword{internal}
