% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pattern2id}
\alias{pattern2id}
\title{Convert various input as pattern to a vector used in tokens_select, 
tokens_compound and kwic.}
\usage{
pattern2id(pattern, types, valuetype, case_insensitive, concatenator = "_",
  remove_unigram = FALSE)
}
\arguments{
\item{pattern}{a character vector, list of character vectors, \link{dictionary},
\link{collocations}, or \link{dfm}. See \link{pattern} for details.}

\item{valuetype}{the type of pattern matching: \code{"glob"} for 
"glob"-style wildcard expressions; \code{"regex"} for regular expressions;
or \code{"fixed"} for exact matching. See \link{valuetype} for details.}

\item{case_insensitive}{ignore the case of dictionary values if \code{TRUE}}

\item{concatenator}{concatenator that join multi-word expression in tokens object}

\item{remove_unigram}{ignore single-word patterns if \code{TRUE}}
}
\description{
Convert various input as pattern to a vector used in tokens_select, 
tokens_compound and kwic.
}
\seealso{
regex2id
}
\keyword{internal}
