% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phrases.R
\name{phrase}
\alias{phrase}
\title{declare a compound character to be a sequence of separate pattern matches}
\usage{
phrase(...)
}
\arguments{
\item{...}{the sequence, as a \code{character} object containing whitespace separating the patterns.}
}
\description{
Declares that a whitespace-separated expression consists of multiple patterns, separated by whitespace.
}
\examples{
# make phrases from characters
phrase(c("a b", "c d e", "f"))

# from a dictionary
phrase(dictionary(catone = c("a b"), cattwo = "c d e", catthree = "f"))
}
