% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.kwic}
\alias{plot.kwic}
\title{plot a dispersion plot of key word(s)}
\usage{
\method{plot}{kwic}(x, ...)
}
\arguments{
\item{x}{a \link{kwic} class object}

\item{...}{additional arguments passed to \code{\link{plot}}}
}
\description{
Plots a dispersion or "x-ray" plot of a selected word pattern across a text.
}
\note{
Currently works only for a a single document.  Because the dispersion plot is
designed to be a "strip", the height and width of your plotting object will determine
whether this plot looks correct or not.  See examples.
}
\examples{
\dontrun{
mobydick <- corpus(textfile(unzip(system.file("extdata", "pg2701.txt.zip", package = "quanteda"))))
plot(kwic(mobydick, "Ahab*"))
}
}

