% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.kwic}
\alias{plot.kwic}
\title{plot the dispersion of key word(s)}
\usage{
\method{plot}{kwic}(...)
}
\arguments{
\item{...}{any number of \link{kwic} class objects}
}
\value{
\code{plot.kwic} returns a ggplot object
}
\description{
Plots a dispersion or "x-ray" plot of selected word pattern(s) across one or more texts.
The format of the plot depends on the number of \link{kwic} class objects passed: if there
is only one document, keywords are plotted one below the other. If there are multiple documents
the documents are plotted one below the other, with keywords shown side-by-side.
}
\examples{
\dontrun{
inaugCorpusPost70 <- subset(inaugCorpus, Year > 1970)
plot(kwic(inaugCorpusPost70, "american"))
plot(kwic(inaugCorpusPost70, "american"), kwic(inaugCorpusPost70, "people"))
}
}
\author{
Adam Obeng
}

