% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textstat_sentiment.R
\name{polarity}
\alias{polarity}
\alias{polarity<-}
\title{Set or get polarities for dictionary keys}
\usage{
polarity(x)

polarity(x) <- value
}
\value{
\code{polarity()} returns the numeric valences of keys as a list.

\verb{polarity<-} sets the dictionary's polarity.
}
\description{
\code{polarity()} is a shortcut to valences that allows an easier method for
setting valences that are the same within a key, for example a key for
"positive" or "negative" whose values will have a uniform valence.  It
operates by setting the same valences for all values in the key.
}
\examples{
# setting polarities
dict <- dictionary(list(
    happy = c("happy", "jubilant", "exuberant"),
    sad = c("sad", "morose", "down")
))
polarity(dict)
polarity(dict) <- c(happy = 1, sad = -1)
polarity(dict)
valence(dict)

# can list multiple keys
polarity(data_dictionary_LSD2015) <- list(
    "positive" = 1, "neg_negative" = 1, "negative" = -1, "neg_positive" = -1
)
polarity(data_dictionary_LSD2015)
lapply(valence(data_dictionary_LSD2015), head)
}
\keyword{dictionary}
\keyword{internal}
\keyword{textstat}
\keyword{utility}
