% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/textmodel-S4generics.R, R/textmodel-naivebayes.R, R/textmodel-wordscores.R
\name{predict.textmodel}
\alias{predict.naivebayes}
\alias{predict.textmodel}
\alias{predict.textmodelfitted}
\alias{predict.wordscores}
\title{predict a text model on new data}
\usage{
\method{predict}{textmodel}(object, ...)

\method{predict}{textmodelfitted}(object, ...)

\method{predict}{naivebayes}(object, newdata = NULL, scores = c(-1, 1), ...)

\method{predict}{wordscores}(object, newdata = NULL, rescaling = "none",
  level = 0.95, verbose = TRUE, ...)
}
\arguments{
\item{object}{a fitted textmodel object (from \code{\link{textmodel}})}

\item{...}{further arguments passed to or from other methods}

\item{newdata}{A dfm or matrix object containing features found in the
training object.  If omitted, the original dfm on which the model was fit
will be used.}

\item{scores}{"reference" values when the wordscores equivalent implementation
of Naive Bayes prediction is used.  Default is \code{c(-1, 1)}.}

\item{rescaling}{\code{none} for "raw" scores; \code{lbg} for LBG (2003)
rescaling; or \code{mv} for the rescaling proposed by Martin and Vanberg
(2007).  (Note to authors: Provide full details here in documentation.)}

\item{level}{probability level for confidence interval width}

\item{verbose}{output status messages if \code{TRUE}}
}
\value{
A list of two data frames, named \code{docs} and \code{words} corresponding
to word- and document-level predicted quantities

\item{docs}{data frame with document-level predictive quantities:
nb.predicted, ws.predicted, bs.predicted, PcGw, wordscore.doc, bayesscore.doc,
posterior.diff, posterior.logdiff.  Note that the diff quantities are currently
implemented only for two-class solutions.}

\item{words}{data-frame with word-level predictive quantities:
wordscore.word, bayesscore.word}

\code{predict.wordscores} returns a data.frame whose rows are the
  documents fitted and whose columns contain the scored textvalues, with the
  number of columns depending on the options called (for instance, how many
  rescaled scores, and whether standard errors were requested.)  (Note: We
  may very well change this soon so that it is a list similar to other
  existing fitted objects.)
}
\description{
Apply a fitted text model to make predictions on test data.

implements class predictions using trained Naive Bayes examples
}
\author{
Kenneth Benoit

Ken Benoit, borrowed in places from Will Lowe, who probably borrowed from me at some early stage.
}
\references{
LBG (2003); Martin and Vanberg (2007)
}

