% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfm-classes.R, R/cfm.R
\docType{methods}
\name{print.dfm}
\alias{print,cfm-method}
\alias{print,dfm-method}
\alias{print,dfmDense-method}
\alias{print,dfmSparse-method}
\alias{print.dfm}
\alias{show,cfm-method}
\alias{show,dfmDense-method}
\alias{show,dfmSparse-method}
\title{print a dfm object}
\usage{
\S4method{print}{dfm}(x, show.values = FALSE, show.settings = FALSE,
  show.summary = TRUE, ndoc = 20L, nfeature = 20L, ...)

\S4method{print}{dfmSparse}(x, show.values = FALSE, show.settings = FALSE,
  show.summary = TRUE, ndoc = 20L, nfeature = 20L, ...)

\S4method{print}{dfmDense}(x, show.values = FALSE, show.settings = FALSE,
  show.summary = TRUE, ndoc = 20L, nfeature = 20L, ...)

\S4method{show}{dfmSparse}(object)

\S4method{show}{dfmDense}(object)

\method{print}{dfm}(x, show.values = FALSE, show.settings = FALSE,
  show.summary = TRUE, ndoc = 20L, nfeature = 20L, ...)

\S4method{print}{cfm}(x, show.values = FALSE, show.settings = FALSE,
  show.summary = TRUE, nfeature = 20L, ...)

\S4method{show}{cfm}(object)
}
\arguments{
\item{x}{the dfm to be printed}

\item{show.values}{print the dfm as a matrix or array (if resampled).}

\item{show.settings}{print the settings used to create the dfm.  See 
\link{settings}.}

\item{show.summary}{print a brief summary indicating the number of documents and features}

\item{ndoc}{max number of documents to print}

\item{nfeature}{max number of features to print}

\item{...}{further arguments passed to or from other methods}

\item{object}{the item to be printed}
}
\description{
print methods for document-feature matrices
}

