% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfm-print.R, R/fcm-print.R, R/fcm.R
\name{print.dfm}
\alias{print.dfm}
\alias{show,dfm-method}
\alias{show,fcm-method}
\alias{head.fcm}
\alias{tail.fcm}
\title{Print a dfm object}
\usage{
\S4method{show}{dfm}(object)

\S4method{show}{fcm}(object)

\method{head}{fcm}(x, n = 6L, nfeature = 6L, ...)

\method{tail}{fcm}(x, n = 6L, nfeature = 6L, ...)
}
\arguments{
\item{object}{the item to be printed}
}
\description{
Print methods for document-feature matrices
}
\keyword{dfm}
\keyword{internal}
