% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.dfm}
\alias{print.dfm}
\title{print a dfm object}
\usage{
print.dfm(x, show.values = FALSE, show.settings = FALSE, ...)
}
\arguments{
\item{x}{the dfm to be printed}

\item{show.values}{print the dfm as a matrix or array (if resampled).}

\item{show.settings}{Print the settings used to create the dfm.  See
\link{settings}.}

\item{...}{further arguments passed to or from other methods}
}
\description{
print method for dfm objects
}

