% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{print.dfm}
\alias{print,dfmDense-method}
\alias{print,dfmSparse-method}
\alias{print.dfm}
\alias{show,dfmDense-method}
\alias{show,dfmSparse-method}
\title{print a dfm object}
\usage{
\S4method{print}{dfmSparse}(x, show.values = FALSE, show.settings = FALSE,
  ...)

\S4method{print}{dfmDense}(x, show.values = FALSE, show.settings = FALSE,
  ...)

\S4method{show}{dfmSparse}(object)

\S4method{show}{dfmDense}(object)

\method{print}{dfm}(x, show.values = FALSE, show.settings = FALSE, ...)
}
\arguments{
\item{x}{the dfm to be printed}

\item{show.values}{print the dfm as a matrix or array (if resampled).}

\item{show.settings}{Print the settings used to create the dfm.  See
\link{settings}.}

\item{...}{further arguments passed to or from other methods}

\item{object}{the item to be printed}
}
\description{
print methods for document-feature matrices
}

