% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textstat_dist.R
\name{print.dist_selection}
\alias{print.dist_selection}
\title{print a dist_selection object}
\usage{
\method{print}{dist_selection}(x, ...)
}
\description{
Print method for a dist_selection object, to make it appear as a data.frame.
}
\keyword{internal}
\keyword{textstat}
