% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel-methods.R
\name{print.statistics_textmodel}
\alias{print.statistics_textmodel}
\title{Implements print methods for textmodel_statistics}
\usage{
\method{print}{statistics_textmodel}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{a textmodel_wordscore_statistics object}

\item{digits}{minimal number of \emph{significant digits}, see
\code{\link[=print.default]{print.default()}}}

\item{...}{further arguments passed to or from other methods}
}
\description{
Implements print methods for textmodel_statistics
}
\keyword{internal}
\keyword{textmodel}
