% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.textmodel}
\alias{print.naivebayes}
\alias{print.textmodel}
\alias{print.textmodelfitted}
\alias{print.textmodelpredicted}
\alias{print.wordscores}
\title{print a textmodel object}
\usage{
print.textmodel(x, ...)

print.textmodelfitted(x, ...)

print.textmodelpredicted(x, ...)

print.naivebayes(x, n = 30L, ...)

print.wordscores(x, n = 30L, ...)
}
\arguments{
\item{x}{textmodel object to be printed}

\item{...}{additional arguments passed to \code{print}}

\item{n}{max rows of dfm to print}
}
\description{
print a class of textmodel object
}

