% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/textmodel-S4generics.R
\name{print.textmodel}
\alias{print.textmodel}
\alias{print.textmodelfitted}
\alias{print.textmodelpredicted}
\title{print a textmodel object}
\usage{
\method{print}{textmodel}(x, ...)

\method{print}{textmodelfitted}(x, ...)

\method{print}{textmodelpredicted}(x, ...)
}
\arguments{
\item{x}{textmodel object to be printed}

\item{...}{additional arguments passed to \code{print}}
}
\description{
print a class of textmodel object
}

