% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quanteda-documentation.R
\name{quanteda-deprecations}
\alias{quanteda-deprecations}
\title{quanteda deprecations}
\description{
Deprecations in \strong{quanteda} v3 explained.
}
\examples{
txt <- c("This is a text example.", "With @usernames and #hashtag tokens.")
tokens(txt, remove_punct = TRUE) \%>\%
    dfm() \%>\%
    dfm_remove(pattern = c("@*", "#*"))
}
