\docType{package}
\name{quanteda}
\alias{quanteda}
\alias{quanteda-package}
\title{Quantitative Analysis of Textual Data}
\description{
Quanteda provides tools that make it easier to organize,
convert, and analyze textual data.  The package is part of
the project QUANTESS: Quantitative Analysis of Textual Data
for the Social Sciences, funded by European Research
Council grant ERC-2011-StG 283794-QUANTESS.
}
\details{
Something here

Quantess organizes texts into a basic corpus object...
Provides tools to converting the corpus into a
\emph{feature-value} matrix on which quantitative analysis
can be performed.
}
\references{
QUANTESS: Quantitative Analysis of Textual Data for the
Social Sciences, funded by European Research Council grant
ERC-2011-StG 283794-QUANTESS.
}

