\docType{package
Returns}
\name{quanteda}
\alias{quanteda}
\title{Quantitative Analysis of Textual Data}
\usage{
  countSyllables(sourceText)
}
\arguments{
  \item{text}{Text to be counted}
}
\description{
  Quantitative Analysis of Textual Data
}
\examples{
tokenize("This is an example sentence.")
}

