% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{quanteda}
\alias{quanteda}
\alias{quanteda-package}
\title{An R package for the quantitative analysis of textual data.}
\description{
A set of functions for creating and managing text corpora, extracting features from
text corpora, and analyzing those features using quantitative methods.
}
\author{
Ken Benoit and Paul Nulty
}

