\docType{package}
\name{quanteda}
\alias{quanteda}
\alias{quanteda-package}
\title{Quantitative Analysis of Textual Data}
\usage{
  translateChunk(sourceText, sourceLanguage,
    targetLanguage, key = NULL, verbose = TRUE)
}
\description{
  Quantitative Analysis of Textual Data

  Quantitative Analysis of Textual Data
}

