% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R
\name{read_dict_wordstat}
\alias{read_dict_wordstat}
\title{dict <- readWStatDict2('/home/kohei/Documents/Dictionary/Wordstat/ROGET.CAT', 'utf-8')
dict <- readWStatDict2('/home/kohei/Documents/Dictionary/Wordstat/WordStat Sentiments.CAT', 'iso-8859-1')}
\usage{
read_dict_wordstat(path, encoding = "auto")
}
\description{
dict <- readWStatDict2('/home/kohei/Documents/Dictionary/Wordstat/ROGET.CAT', 'utf-8')
dict <- readWStatDict2('/home/kohei/Documents/Dictionary/Wordstat/WordStat Sentiments.CAT', 'iso-8859-1')
}
