% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readability.R
\name{readability}
\alias{readability}
\alias{readability.character}
\title{calculate readability}
\usage{
readability(x, ...)

\method{readability}{character}(x, index)
}
\arguments{
\item{x}{a \link{corpus} object}

\item{index}{character vector defining the readability index to calculate}
}
\description{
Calculate the readability of text(s).
}
\examples{
readability(c("ABC.", "The fat, cat."), index = "Flesch.Kincaid")
}
\author{
Kenneth Benoit, adapted from the S4 class implementation written by
  Meik Michalke in the \pkg{koRpus} package.
}

