% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacyr-methods.R, R/stopwords.R, R/utils.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{spacy_parse}
\alias{stopwords}
\alias{\%>\%}
\title{Objects exported from other packages}
\section{Stopwords}{
 

Stopword lists were formerly built into \pkg{quanteda}, but have been moved to the 
\pkg{stopwords} package. See \code{\link[stopwords]{stopwords}}.
}

\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}

  \item{spacyr}{\code{\link[spacyr]{spacy_parse}}}

  \item{stopwords}{\code{\link[stopwords]{stopwords}}}
}}

