% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopwords.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{stopwords}
\title{Objects exported from other packages}
\section{Stopwords}{


Stopword lists were formerly built into \pkg{quanteda}, but have been moved to the
\pkg{stopwords} package. See \code{\link[stopwords:stopwords]{stopwords::stopwords()}}.
}

\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{stopwords}{\code{\link[stopwords]{stopwords}}}
}}

