% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectFeatures.R
\name{regex2fixed}
\alias{regex2fixed}
\title{Convert regex to fixed}
\usage{
regex2fixed(regex, types)
}
\arguments{
\item{regex}{regular expression}

\item{types}{unique types of tokens}
}
\description{
Convert regex to fixed
}

