% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R
\name{remove_empty_keys}
\alias{remove_empty_keys}
\title{Utility function to remove empty keys}
\usage{
remove_empty_keys(dict)
}
\arguments{
\item{dict}{a flat or hierarchical dictionary}
}
\description{
Utility function to remove empty keys
}
\keyword{internal}
