% Generated by roxygen2 (4.0.2): do not edit by hand
\name{removeFeatures}
\alias{removeFeatures}
\alias{removeFeatures.character}
\alias{removeFeatures.collocations}
\alias{removeFeatures.dfm}
\alias{stopwordsRemove}
\title{remove features from an object}
\usage{
removeFeatures(x, stopwords = NULL, verbose = TRUE)

\method{removeFeatures}{character}(x, stopwords = NULL, verbose = TRUE)

\method{removeFeatures}{dfm}(x, stopwords = NULL, verbose = TRUE)

\method{removeFeatures}{collocations}(x, stopwords = NULL, verbose = TRUE)

stopwordsRemove(x, stopwords = NULL, verbose = TRUE)
}
\arguments{
\item{x}{object from which stopwords will be removed}

\item{stopwords}{character vector of features to remove.  Now requires an explicit
list to be supplied, for instance \code{stopwords("english")}.}

\item{verbose}{if \code{TRUE} print message about how many features were removed}
}
\value{
an object with stopwords removed
}
\description{
This function removes features from a variety of objects, such as text, a dfm, or
a list of collocations.  The most common usage for \code{removeFeatures} will be
to eliminate stop words from a text or text-based object.  Some commonly used
built-in stop words can be accessed through \code{\link{stopwords}}.
}
\details{
Because we believe the user should take full responsibility for any features
that are removed, we do not provide a default list.  Use \code{\link{stopwords}} instead.
}
\examples{
## examples for character objects
someText <- "Here's some text containing words we want to remove."
removeFeatures(someText, stopwords("english", verbose=FALSE))
removeFeatures(someText, stopwords("SMART", verbose=FALSE))
removeFeatures(someText, c("some", "want"))
itText <- "Ecco alcuni di testo contenente le parole che vogliamo rimuovere."
removeFeatures(itText, stopwords("italian", verbose=FALSE))

## example for dfm objects
mydfm <- dfm(uk2010immig, verbose=FALSE)
removeFeatures(mydfm, stopwords("english", verbose=FALSE))

## example for collocations
(myCollocs <- collocations(inaugTexts, top=20))
removeFeatures(myCollocs, stopwords("english", verbose=FALSE))
}
\seealso{
\link{stopwords}
}

