% Generated by roxygen2 (4.0.2): do not edit by hand
\name{resample}
\alias{is.resampled}
\alias{is.resampled.corpus}
\alias{is.resampled.dfm}
\alias{resample}
\alias{resample.corpus}
\title{resampling methods for a corpus}
\usage{
resample(x, ...)

\method{resample}{corpus}(x, n = 100, unit = c("sentences", "paragraphs"),
  ...)

is.resampled(x)

\method{is.resampled}{corpus}(x)

\method{is.resampled}{dfm}(x)
}
\arguments{
\item{x}{corpus object containing the texts to be resampled}

\item{...}{additional arguments passed to \code{\link{segment}}}

\item{n}{number of resamples to be drawn}

\item{unit}{resampling unit for drawing the random samples, can be
\code{sentences} or \code{paragraphs}.}
}
\value{
a corpus object containing new resampled texts.
}
\description{
Draw a set of random resamples from a corpus object, at a specified level of
resampling, and record additional "resampled texts" as document-level
metadata, stored as \code{_resampleXXX} for the XXXth resample.
}
\details{
\code{is.resampled} checks a corpus or dfm object and returns
  \code{TRUE} if these contain resampled texts or the results of resampled texts
}
\note{
Additional resampling units to be added will include fixed length
  samples and random length samples.
}
\examples{
testCorp <- resample(subset(inaugCorpus, Year>2000), 10, "sentences")
testCorpPara <- resample(corpus(uk2010immig), 10, "paragraphs")
names(metadoc(testCorp))
x <- corpus(c("Sentence One C1.  Sentence Two C1.  Sentence Three C1.",
              "Sentence One C2.  Sentence Two C2.  Sentence Three C2. Sentence Four C2.  Sentence Five C2.  Sentence Six C2."),
            docnames=c("docTwo", "docOne"))
testRS <- resample(x, n=3)
metadoc(testRS)
testRS$documents

# tests to see if a corpus contains resampled texts
is.resampled(testCorp)
is.resampled(inaugCorpus)
}

