% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docvars.R
\name{reshape_docvars}
\alias{reshape_docvars}
\title{Internal function to subset or duplicate docvar rows}
\usage{
reshape_docvars(x, i = NULL, unique = FALSE, drop_docid = TRUE)
}
\arguments{
\item{x}{docvar data.frame}

\item{i}{numeric or logical vector for subsetting/duplicating rows}

\item{unique}{if \code{TRUE}, names must be all unique. If \code{FALSE}, documents with the same
names are treated as segments from the same document and given serial number.}

\item{drop_docid}{if \code{TRUE}, drop unused names of documents.}
}
\description{
Internal function to subset or duplicate docvar rows
}
\keyword{internal}
