% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docvars.R
\name{reshape_docvars}
\alias{reshape_docvars}
\title{Internal function to subset or duplicate docvar rows}
\usage{
reshape_docvars(x, i = NULL)
}
\arguments{
\item{x}{docvar data.frame}

\item{i}{numeric or logical vector for subsetting/duplicating rows}
}
\description{
Internal function to subset or duplicate docvar rows
}
\keyword{internal}
