% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus_sample.R, R/dfm-methods.R
\name{sample}
\alias{sample}
\alias{sample.corpus}
\alias{sample.default}
\alias{sample.dfm}
\title{randomly sample documents or features}
\usage{
sample(x, size, replace = FALSE, prob = NULL, ...)

\method{sample}{default}(x, size, replace = FALSE, prob = NULL, ...)

\method{sample}{corpus}(x, ...)

\method{sample}{dfm}(x, size = ndoc(x), replace = FALSE, prob = NULL,
  what = c("documents", "features"), ...)
}
\arguments{
\item{x}{a corpus or dfm object whose documents or features will be sampled}

\item{size}{a positive number, the number of documents to select}

\item{replace}{Should sampling be with replacement?}

\item{prob}{A vector of probability weights for obtaining the elements of the
vector being sampled.}

\item{...}{unused}

\item{what}{dimension (of a \link{dfm}) to sample: can be \code{documents} or
\code{features}}
}
\value{
A dfm object with number of documents equal to \code{size}, drawn 
  from the corpus \code{x}.  The returned corpus object will contain all of 
  the meta-data of the original corpus, and the same document variables for 
  the documents selected.
}
\description{
Deprecated function framework for randomly sampling from a corpus or dfm object.  Use
\code{\link{corpus_sample}} or \code{dfm_sample} instead
}
\examples{
# sampling from a dfm
myDfm <- dfm(data_char_inaugural[1:10], verbose = FALSE)
sample(myDfm)[, 1:10]
sample(myDfm, replace = TRUE)[, 1:10]
sample(myDfm, what = "features")[1:10, ]
}
\seealso{
corpus_sample
}
\keyword{internal}

