% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern2fixed.R
\name{search_index}
\alias{search_index}
\title{Internal function for \code{select_types} to search the index using
fastmatch.}
\usage{
search_index(pattern, index)
}
\arguments{
\item{index}{an index object created by \code{index_types}}

\item{regex}{a glob expression to search}
}
\description{
Internal function for \code{select_types} to search the index using
fastmatch.
}
\seealso{
\code{\link[=index_types]{index_types()}}
}
\keyword{internal}
