% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{segment}
\alias{segment}
\alias{segment.character}
\alias{segment.corpus}
\title{segment: deprecated function}
\usage{
segment(x, ...)

\method{segment}{character}(x, ...)

\method{segment}{corpus}(x, ...)
}
\arguments{
\item{x}{object to be segmented}

\item{...}{additional arguments passed to  \code{\link{corpus_segment}}.}
}
\description{
See \code{\link{corpus_segment}}
}
\keyword{deprecated}
\keyword{internal}

