% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textstat_sentiment.R
\name{sentiment-functions}
\alias{sentiment-functions}
\alias{sent_logit}
\alias{sent_abspropdiff}
\alias{sent_relpropdiff}
\title{Sentiment functions}
\usage{
sent_logit(x, smooth = 0.5, ...)

sent_abspropdiff(x, ...)

sent_relpropdiff(x, ...)
}
\arguments{
\item{x}{a \link{dfm} that has the following required feature names: \code{pos},
\code{neg}, \code{neut}, and \code{other}}

\item{smooth}{additional smoothing function added to \code{pos} and \code{neg} before
logarithmic transformation}

\item{...}{additional parameters as needed}
}
\description{
Functions for computing sentiment, for \code{\link[=textstat_sentiment]{textstat_sentiment()}}.  Each function
takes . Additional arguments may be passed via \code{...},
such as \code{smooth} for the logit scale.

\code{sent_logit} is \eqn{log(\frac{pos}{neg})}.

\code{sent_abspropdiff} is \eqn{\frac{pos - neg}{N}}, where \eqn{N}
is the total number of all features in a document.

\code{sent_relpropdiff} is \eqn{\frac{pos - neg}{pos + neg}}.
}
\details{
User supplied functions must take \code{x} and \code{...}, and refer to the required
feature names for the sentiment categories \code{pos},
\code{neg}, \code{neut}, and \code{other}.  (The \code{other} category is only required when
a scaling function needs the count of non-sentiment associated features.)
}
\references{
Lowe, W., Benoit, K. R., Mikhaylov, S., & Laver, M. (2011).
Scaling Policy Preferences from Coded Political Texts. \emph{Legislative Studies
Quarterly}, 36(1), 123–155.
\url{http://doi.org/10.1111/j.1939-9162.2010.00006.x}
}
\keyword{internal}
\keyword{textstat}
