% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequences.R
\name{sequences}
\alias{sequences}
\alias{is.sequences}
\title{Find variable-length collocations}
\usage{
sequences(x, min_count = 2, min_size = 2, max_size = 5,
  method = c("unigram", "all_subtuples"), nested = TRUE)

is.sequences(x)
}
\arguments{
\item{x}{a \link{tokens} object}

\item{min_count}{minimum frequency of sequences for which parameters are 
estimated}

\item{min_size}{minimum length of sequences which are collected}

\item{max_size}{maximum length of sequences which are collected}

\item{method}{default is "unigram" and option is "all_subtuples"}

\item{nested}{if \code{TRUE}, collect all the subsequences of a longer
sequence as separate entities. e.g. in a sequence of capitalized words
"United States Congress", "States Congress" is considered as a subsequence.
But "United States" is not a subsequence because it is followed by
"Congress".}
}
\value{
\code{sequences} returns \code{TRUE} if the object is of class
  sequences, \code{FALSE} otherwise.
}
\description{
This function automatically identifies contiguous collocations consisting of 
variable-length term sequences whose frequency is unlikely to have occurred 
by chance.  The algorithm is based on Blaheta and Johnson's (2001) 
"Unsupervised Learning of Multi-Word Verbs".
}
\examples{
toks <- tokens(corpus_segment(data_corpus_inaugural, what = "sentence"), remove_punct=TRUE)
toks <- tokens_select(toks, stopwords("english"), "remove", padding = TRUE)
# extracting multi-part proper nouns (capitalized terms)
toks <- tokens_select(toks, "^([A-Z][a-z\\\\-]{2,})", valuetype="regex", 
                     case_insensitive = FALSE, padding = TRUE)

seqs <- sequences(toks)
head(seqs, 10)
# to return only trigrams
seqs <- sequences(toks, min_size = 3, max_size = 3)
head(seqs, 10)
}
\references{
Blaheta, D., & Johnson, M. (2001). 
  \href{http://web.science.mq.edu.au/~mjohnson/papers/2001/dpb-colloc01.pdf}{Unsupervised
   learning of multi-word verbs}. Presented at the ACLEACL Workshop on the 
  Computational Extraction, Analysis and Exploitation of Collocations.
}
\author{
Kohei Watanabe and Haiyan Wang
}
\keyword{collocations}
\keyword{internal}
