% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequences.R
\name{sequences}
\alias{sequences}
\title{find sequences of tokens}
\usage{
sequences(x, features, valuetype = c("glob", "regex", "fixed"),
  case_insensitive = TRUE, count_min = 2, nested = TRUE, ...)
}
\arguments{
\item{x}{tokenizedTexts objects}

\item{count_min}{minimum frequency of sequences}

\item{nested}{collect nested sub-sequence}

\item{tokens}{types of token in sequuences}
}
\description{
This function automatically identify sequences of tokens. This algorithm is 
based on Blaheta and Johnson's “Unsupervised Learning of Multi-Word Verbs”.
}
\examples{

toks <- tokens(corpus_segment(data_corpus_inaugural, what = "sentence"), removePunct = TRUE)
toks <- tokens_select(toks, stopwords("english"), "remove", padding = TRUE)

# extracting multi-part nouns
seqs <- sequences(toks, "^([A-Z][a-z\\\\-]{2,})", valuetype="regex", count_min = 2)
head(seqs, 10)

# types can be any words
seqs2 <- sequences(tokens, "^([a-z]+)$", count_min = 3)
head(seqs2, 10)

}
\author{
Kohei Watanabe
}
\references{
Blaheta, D., & Johnson, M. (2001). Unsupervised learning of
  multi-word verbs. Presented at the ACLEACL Workshop on the Computational
  Extraction, Analysis and Exploitation of Collocations.
}
\keyword{collocations}
\keyword{internal}

