% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/settings.R
\name{settingsInitialize}
\alias{settingsInitialize}
\title{\code{settingsInitialize} returns a list of legal settings, set to their default values}
\usage{
settingsInitialize()
}
\description{
\code{settingsInitialize} returns a list of legal settings, set to their default values
}

