% Generated by roxygen2 (4.0.2): do not edit by hand
\name{similarity}
\alias{similarity}
\title{compute similarities between documents and/or features}
\usage{
similarity(x, selection, n = 10, margin = c("features", "documents"),
  method = "correlation", sort = TRUE, normalize = TRUE)
}
\arguments{
\item{x}{a \link{dfm} object}

\item{selection}{character or character vector of document names or feature
labels from the dfm}

\item{n}{the top \code{n} most similar items will be returned, sorted in
descending order.  If n is \code{NULL}, return all items.}

\item{margin}{identifies the margin of the dfm on which similarity will be
computed: \code{features} for word/term features or \code{documents} for
documents.}

\item{method}{a valid method for computing similarity from
\code{\link[proxy]{pr_DB}}}

\item{sort}{sort results in descending order if \code{TRUE}}

\item{normalize}{if \code{TRUE}, normalize the dfm by term frequency within
document (so that the dfm values will be relative term frequency within
each document)}
}
\value{
a named list of the selection labels, with a sorted named vector of
  similarity measures.
}
\description{
Compute similarities between documents and/or features from a
\code{\link{dfm}}. Uses the similarity measures defined in
\link[proxy]{simil}.  See \code{\link[proxy]{pr_DB}} for available distance
measures, or how to create your own.
}
\note{
The method for computing feature similarities can be quite slow when
  there are large numbers of feature types.  Future implementations will
  hopefully speed this up.
}
\examples{
# create a dfm from inaugural addresses from Reagan onwards
presDfm <- dfm(subset(inaugCorpus, Year>1980), stopwords=TRUE, stem=TRUE)

# compute some document similarities
similarity(presDfm, "1985-Reagan", n=5, margin="documents")
similarity(presDfm, c("2009-Obama" , "2013-Obama"), n=5, margin="documents")
similarity(presDfm, c("2009-Obama" , "2013-Obama"), n=NULL, margin="documents")
similarity(presDfm, c("2009-Obama" , "2013-Obama"), n=NULL, margin="documents", method="cosine")
similarity(presDfm, "2005-Bush", n=NULL, margin="documents", method="eJaccard", sort=FALSE)

\dontrun{
# compute some term similarities
similarity(presDfm, c("fair", "health", "terror"), method="cosine")

# compare to tm
require(tm)
data("crude")
tdm <- TermDocumentMatrix(crude)
findAssocs(tdm, c("oil", "opec", "xyz"), c(0.7, 0.75, 0.1))
crudeDfm <- dfm(corpus(crude))
similarity(crudeDfm, c("oil", "opec", "xyz"), normalize=FALSE)
}
}

