% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{slots<-}
\alias{slots<-}
\title{Function to assign multiple slots to a S4 object}
\usage{
slots(x, exceptions = c("Dim", "Dimnames", "i", "p", "x",
  "factors")) <- value
}
\arguments{
\item{x}{an S4 object}

\item{exceptions}{slots to ignore}

\item{value}{a list of attributes extracted by attributes()}
}
\description{
Function to assign multiple slots to a S4 object
}
\keyword{internal}
