% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfm-methods.R
\name{sparsity}
\alias{sparsity}
\title{Compute the sparsity of a document-feature matrix}
\usage{
sparsity(x)
}
\arguments{
\item{x}{the document-feature matrix}
}
\description{
Return the proportion of sparseness of a document-feature matrix, equal
to the proportion of cells that have zero counts.
}
\examples{
dfmat <- dfm(data_corpus_inaugural)
sparsity(dfmat)
sparsity(dfm_trim(dfmat, min_termfreq = 5))
}
