% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R
\name{split_values}
\alias{split_values}
\title{Internal function for special handling of multi-word dicitionary values}
\usage{
split_values(value, concatenator)
}
\arguments{
\item{value}{character vector for dictionary values to be split}

\item{concatenator}{concatenator from a dictionary object}
}
\description{
Internal function for special handling of multi-word dicitionary values
}
\keyword{internal}
