% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{stopwords}
\alias{stopwords}
\title{A named list containing common stopwords in 14 languages}
\description{
SMART English stopwords from the SMART information retrieval
system (obtained from
http://jmlr.csail.mit.edu/papers/volume5/lewis04a/a11-smart-stop-list/english.stop)
and a set of stopword lists from the Snowball stemmer project in
different languages (obtained from
http://svn.tartarus.org/snowball/trunk/website/algorithms/*/stop.txt). Supported
languages are danish, dutch, english, finnish, french, german,
hungarian, italian, norwegian, portuguese, russian, spanish, and
swedish. Language names are case sensitive. Alternatively, their IETF
language tags may be used.
}

