% Generated by roxygen2 (4.0.2): do not edit by hand
\name{stopwordsRemove}
\alias{stopwordsRemove}
\alias{stopwordsRemove.character}
\title{remove stopwords from a text or dfm}
\usage{
stopwordsRemove(text, stopwords = NULL)

\method{stopwordsRemove}{character}(text, stopwords = NULL)
}
\arguments{
\item{text}{Text from which stopwords will be removed}

\item{stopwords}{Character vector of stopwords to remove - if none is
supplied, a default set of English stopwords is used}
}
\value{
a character vector or dfm with stopwords removed
}
\description{
This function takes a character vector or \link{dfm} and removes words in the
remove common or 'semantically empty' words from a text. See \link{stopwordsGet}
for the information about the default lists.
}
\details{
This function takes a character vector 'text' and removes words in the list
provided in stopwords. If no list of stopwords is provided a default
list for English is used. The function \link{stopwordsGet} can load a default
set of stopwords for many languages.
}
\examples{
## examples for character objects
someText <- "Here is an example of text containing some stopwords we want to remove."
itText <- "Ecco un esempio di testo contenente alcune parole non significative che vogliamo rimuovere."
stopwordsRemove(someText)
stopwordsRemove(someText, stopwordsGet("SMART"))
stopwordsRemove(itText, stopwordsGet("italian"))
stopwordsRemove(someText, c("containing", "example"))

## example for dfm objects
docmat <- dfm(uk2010immig)
docmatNostopwords <- stopwordsRemove(docmat)
dim(docmat)
dim(docmatNostopwords)
dim(stopwordsRemove(docmat, stopwordsGet("SMART")))
}

