% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/remove-stopwords.R
\name{stopwordsRemove}
\alias{stopwordsRemove}
\alias{stopwordsRemove.character}
\alias{stopwordsRemove.collocations}
\alias{stopwordsRemove.dfm}
\alias{stopwordsRemove.dfmSnum}
\title{remove stopwords from a text or dfm}
\usage{
stopwordsRemove(x, stopwords = NULL, verbose = TRUE)

\method{stopwordsRemove}{character}(x, stopwords = NULL, verbose = TRUE)

\method{stopwordsRemove}{dfm}(x, stopwords = NULL, verbose = TRUE)

\method{stopwordsRemove}{dfmSnum}(x, stopwords = NULL, verbose = TRUE)

\method{stopwordsRemove}{collocations}(x, stopwords = NULL, verbose = TRUE)
}
\arguments{
\item{x}{Object from which stopwords will be removed}

\item{stopwords}{Character vector of stopwords to remove.  Now requires an explicit
list to be supplied, for instance \code{stopwordsGet("english")}.}

\item{verbose}{if \code{TRUE} print message about how many items were removed}
}
\value{
an object with stopwords removed
}
\description{
This function takes a character vector or \link{dfm} and removes words in the
remove common or 'semantically empty' words from a text. See \link{stopwordsGet}
for the information about the default lists.
}
\details{
This function takes a character vector 'text' and removes words in the list
provided in stopwords. If no list of stopwords is provided a default
list for English is used. The function \link{stopwordsGet} can load a default
set of stopwords for many languages.
}
\examples{
## examples for character objects
someText <- "Here's some text containing words we want to remove."
stopwordsRemove(someText, stopwordsGet("english"))
stopwordsRemove(someText, stopwordsGet("SMART"))
stopwordsRemove(someText, c("some", "want"))
itText <- "Ecco alcuni di testo contenente le parole che vogliamo rimuovere."
stopwordsRemove(itText, stopwordsGet("italian"))

## example for dfm objects
mydfm <- dfm(uk2010immig, verbose=FALSE)
stopwordsRemove(mydfm, stopwordsGet("english"))
mydfms <- dfm(uk2010immig, verbose=FALSE, matrixType="sparse")
dim(stopwordsRemove(mydfms, stopwordsGet("SMART")))

## example for collocations
(myCollocs <- collocations(inaugTexts, top=20))
stopwordsRemove(myCollocs, stopwordsGet("english"))
}

