% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textplot_network.R
\name{summary.character}
\alias{summary.character}
\title{summary.character method to override the network::summary.character()}
\usage{
\method{summary}{character}(object, ...)
}
\arguments{
\item{object}{the character input}

\item{...}{for additional passing of arguments to default method}
}
\description{
Necessary to prevent the \pkg{network} package's \code{summary.character} method
from causing inconsistent behaviour with other summary methods.
}
\keyword{internal}
