% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/describe-texts.R
\name{summary.character}
\alias{describeTexts}
\alias{summary.character}
\title{text summary}
\usage{
\method{summary}{character}(object, verbose = TRUE, ...)

describeTexts(object, verbose = TRUE, ...)
}
\arguments{
\item{object}{The texts to be described}

\item{verbose}{set to \code{FALSE} to suppress printing, for instance
if you simply want to assign the output to a \code{data.frame}}

\item{...}{additional arguments affecting the summary produced}
}
\description{
Prints to the console a desription of the texts, including
number of types, tokens, and sentences.
}
\examples{
summary(c("testing this text", "and this one"))
summary(uk2010immig)
myTextSummaryDF <- summary(uk2010immig, verbose=FALSE)
}

