% Generated by roxygen2 (4.0.2): do not edit by hand
\name{summary.corpus}
\alias{summary.corpus}
\title{Corpus summary}
\usage{
\method{summary}{corpus}(object, n = 100, verbose = TRUE,
  showmeta = FALSE, ...)
}
\arguments{
\item{object}{corpus to be summarized}

\item{n}{maximum number of texts to describe, default=100}

\item{verbose}{FALSE to turn off printed output}

\item{showmeta}{TRUE to include document-level meta-data}

\item{...}{additional arguments affecting the summary produced}
}
\description{
Displays information about a corpus object, including attributes and
metadata such as date of number of texts, creation and source.
}
\examples{
summary(inaugCorpus)
summary(inaugCorpus, n=10)
mycorpus <- corpus(uk2010immig, docvars=data.frame(party=names(uk2010immig)), enc="UTF-8")
summary(mycorpus, showmeta=TRUE)  # show the meta-data
mysummary <- summary(mycorpus, verbose=FALSE)  # (quietly) assign the results
mysummary$Types / mysummary$Tokens             # crude type-token ratio
}

