% Generated by roxygen2 (4.0.2): do not edit by hand
\name{summary.textmodel}
\alias{summary.textmodel}
\alias{summary.textmodelfitted}
\alias{summary.textmodelpredicted}
\alias{summary.wordscores}
\title{summary of a textmodel object}
\usage{
summary.textmodel(object, ...)

summary.textmodelfitted(object, ...)

summary.textmodelpredicted(object, ...)

summary.wordscores(object, ...)
}
\arguments{
\item{object}{textmodel object to be summarized}

\item{...}{additional arguments to \link{print}}
}
\description{
Summarize the results of a fitted or predicted textmodel object.
}

