% Generated by roxygen2 (4.0.2): do not edit by hand
\name{summary.textmodel}
\alias{summary.textmodel}
\alias{summary.wordscores}
\title{summay of a textmodel object}
\usage{
\method{summary}{textmodel}(object, ...)

\method{summary}{wordscores}(object, ...)
}
\arguments{
\item{object}{textmodel object to be summarized}

\item{...}{additional arguments to \link{print}}
}
\description{
Summarize the results of a fitted or predited textmodel object.
}

