% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{syllableCounts}
\alias{syllableCounts}
\title{A named list mapping words to counts of their syllables}
\description{
A named list mapping words to counts of their syllables, generated from
the CMU pronunciation dictionary
}
\examples{
data(syllableCounts)
syllableCounts["sixths"]
syllableCounts["onomatopeia"]
}
\references{
\url{http://www.speech.cs.cmu.edu/cgi-bin/cmudict}
}

