% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsyllable.R
\name{syllables}
\alias{syllables}
\title{deprecated name for nsyllable}
\usage{
syllables(x, ...)
}
\description{
Deprecated function name; use \code{\link{nsyllable}} instead.
}
\keyword{deprecated}
\keyword{internal}
