% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pos.R
\name{tagPOS}
\alias{tagPOS}
\title{tag parts of speech in a tokenized list}
\usage{
tagPOS(l, ...)

tagPOS(l, ...)
}
\arguments{
\item{...}{not currently used}

\item{x}{list of tokens to tag}
}
\description{
Add tags to a tokenized list of parts of speech
}
\examples{
sents <- tokenize(inaugTexts[1], what="sentence", simplify=TRUE)
tokens <- lapply(sents, function(x) tokenize(x, simplify=TRUE))
pos <- tagPOS(tokens) # tokens should be sentence
}

