% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/corpus-sources-S4.R
\docType{methods}
\name{textfile}
\alias{textfile}
\alias{textfile,character,index,missing-method}
\alias{textfile,character,missing,missing-method}
\title{read a text corpus source from a file}
\usage{
textfile(file, textField, directory, ...)

\S4method{textfile}{character,index,missing}(file, textField,
  directory = NULL, ...)

\S4method{textfile}{character,missing,missing}(file, textField = NULL,
  directory = NULL, ...)
}
\arguments{
\item{file}{the complete filename to be read.  Currently available file types
are:
\describe{
  \item{\code{txt}}{plain text files}
  \item{\code{json}}{data in JavaScript Object Notation, consisting of the texts and additional
  document-level variables and document-level meta-data.  The text key must be identified
  by specifying a \code{textField} value.}
  \item{\code{csv}}{comma separated value data, consisting of the texts and additional
  document-level variables and document-level meta-data.  The text file must be identified
  by specifying a \code{textField} value.}
  \item{a wildcard value}{any valid pathname with a wildcard ("glob") expression that can be
  expanded by the operating system.  This may consist of multiple file types.}
  \item{\code{doc, docx}:}{Word files coming soon.}
  \item{\code{pdf}:}{Adobe Portable Document Format files, coming soon.}
}}

\item{textField}{a variable (column) name or column number indicating where
to find the texts that form the documents for the corpus.  This must be
specified for file types \code{.csv} and \code{.json}.}

\item{directory}{not used yet, and may be removed (if I move this to a new
method called \code{textfiles})}

\item{...}{additional arguments passed through to other functions}
}
\value{
an object of class \link{corpusSource-class} that can be read by
  \link{corpus} to construct a corpus
}
\description{
Read a text corpus from a source file, where the single file will consist of
a set of texts in columns and document variables and document-level meta-data
in additional columns.  For spreadsheet-like files, the first row must be a
header.
}
\details{
The constructor does not store a copy of the texts, but rather reads
  in the texts and associated data, and saves them to a temporary R object
  whose location is specified in the \link{corpusSource-class} object.  This
  prevents a complete copy of the object from cluttering the global
  environment and consuming additional space.  This does mean however that
  the state of the file containing the source data will not be cross-platform
  and may not be persistent across sessions.  So the recommended usage is to
  load the data into a corpus in the same session in which \code{textfile} is
  called.
}
\examples{
# Twitter json
\donttest{mytf <- textfile("~/Dropbox/QUANTESS/corpora/misc/NinTANDO_Me.json")
summary(corpus(mytf))}
# generic json - needs a textField specifier
mytf2 <- textfile("~/Dropbox/QUANTESS/Manuscripts/Collocations/Corpora/sotu/sotu.json",
                  textField = "text")
summary(corpus(mytf2))
# text file
mytf3 <- textfile("~/Dropbox/QUANTESS/corpora/project_gutenberg/pg2701.txt")
summary(corpus(mytf3))
mytf4 <- textfile("~/Dropbox/QUANTESS/corpora/inaugural/*.txt")
summary(corpus(mytf4))
}

