% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readtext-methods.R
\name{textfile}
\alias{textfile}
\title{old function to read texts from files}
\usage{
textfile(...)
}
\arguments{
\item{...}{any collection of objects}
}
\description{
This function was removed from \pkg{quanteda} in v.0.9.9 and moved to a 
separate package, \pkg{readtext}.  Please use \code{readtext} instead
of \code{textfile}.
}
\keyword{deprecated}
\keyword{internal}

