% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel-generics.R, R/textmodel_affinity.R,
%   R/textmodel_ca.R, R/textmodel_lsa.R
\docType{class}
\name{textmodel-internal}
\alias{textmodel-internal}
\alias{print.textmodel_affinity}
\alias{print.predict.textmodel_affinity}
\alias{residuals.predict.textmodel_affinity}
\alias{resid}
\alias{rstandard.predict.textmodel_affinity}
\alias{print.affinity_influence}
\alias{summary.affinity_influence}
\alias{print.summary.affinity_influence}
\alias{textmodel_ca_fitted-class}
\alias{coef,textmodel_ca_fitted-method}
\alias{coefficients,textmodel_ca_fitted-method}
\alias{textmodel_lsa_fitted-class}
\alias{textmodel_lsa_predicted-class}
\alias{coef,textmodel_lsa_fitted-method}
\alias{coefficients,textmodel_lsa_fitted-method}
\title{Internal functions for textmodel objects}
\usage{
print.textmodel_affinity(x)

\method{print}{predict.textmodel_affinity}(x, ...)

\method{residuals}{predict.textmodel_affinity}(object, type = c("response",
  "pearson"), ...)

resid(object, ...)

\method{rstandard}{predict.textmodel_affinity}(model, ...)

\method{print}{affinity_influence}(x, n = 30, ...)

\method{summary}{affinity_influence}(object, ...)

\method{print}{summary.affinity_influence}(x, n = 30, ...)

\S4method{coef}{textmodel_ca_fitted}(object, doc_dim = 1, feat_dim = 1, ...)

\S4method{coefficients}{textmodel_ca_fitted}(object, ...)

\S4method{coef}{textmodel_lsa_fitted}(object, doc_dim = 1, feat_dim = 1,
  ...)

\S4method{coefficients}{textmodel_lsa_fitted}(object, ...)
}
\arguments{
\item{type}{see \link{residuals.lm}}

\item{doc_dim, feat_dim}{the document and feature dimension scores to be 
extracted as coefficients}

\item{doc_dim, feat_dim}{the document and feature dimension scores to be 
extracted as coefficients}
}
\description{
Internal function documentation for textmodel objects.
}
\keyword{internal}
\keyword{textmodel}
