% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel-generics.R, R/textmodel-methods.R,
%   R/textmodel_affinity.R, R/textmodel_ca.R, R/textmodel_lsa.R,
%   R/textmodel_wordfish.R, R/textmodel_wordscores.R
\docType{methods}
\name{textmodel-internal}
\alias{textmodel-internal}
\alias{show,textmodel_summary-method}
\alias{textmodel_affinity_fitted-class}
\alias{textmodel_affinity_predicted-class}
\alias{show,textmodel_affinity_fitted-method}
\alias{show,textmodel_affinity_predicted-method}
\alias{coef,textmodel_affinity_predicted-method}
\alias{coefficients,textmodel_affinity_predicted-method}
\alias{residuals,textmodel_affinity_predicted-method}
\alias{resid,textmodel_affinity_predicted-method}
\alias{rstandard.textmodel_affinity_predicted}
\alias{print.affinity_influence}
\alias{summary.affinity_influence}
\alias{print.summary_affinity_influence}
\alias{textmodel_ca_fitted-class}
\alias{coef,textmodel_ca_fitted-method}
\alias{coefficients,textmodel_ca_fitted-method}
\alias{textmodel_lsa_fitted-class}
\alias{textmodel_lsa_predicted-class}
\alias{coef,textmodel_lsa_fitted-method}
\alias{coefficients,textmodel_lsa_fitted-method}
\alias{textmodel_wordfish_fitted-class}
\alias{textmodel_wordfish_predicted-class}
\alias{print.textmodel_wordfish_fitted}
\alias{show,textmodel_wordfish_fitted-method}
\alias{show,textmodel_wordfish_predicted-method}
\alias{coef,textmodel_wordfish_fitted-method}
\alias{coefficients,textmodel_wordfish_fitted-method}
\alias{textmodel_wordscores-class}
\alias{predict.textmodel_wordscores}
\alias{print.textmodel_wordscores}
\alias{show,textmodel_wordscores-method}
\alias{coef,textmodel_wordscores-method}
\alias{coefficients,textmodel_wordscores-method}
\title{Internal functions for textmodel objects}
\usage{
\S4method{show}{textmodel_summary}(object)

\S4method{show}{textmodel_affinity_fitted}(object)

\S4method{show}{textmodel_affinity_predicted}(object)

\S4method{coef}{textmodel_affinity_predicted}(object, ...)

\S4method{coefficients}{textmodel_affinity_predicted}(object, ...)

\S4method{residuals}{textmodel_affinity_predicted}(object,
  type = c("response", "pearson"), ...)

\S4method{resid}{textmodel_affinity_predicted}(object, ...)

\method{rstandard}{textmodel_affinity_predicted}(model, ...)

\method{print}{affinity_influence}(x, n = 30, ...)

\method{summary}{affinity_influence}(object, ...)

\method{print}{summary_affinity_influence}(x, n = 30, ...)

\S4method{coef}{textmodel_ca_fitted}(object, doc_dim = 1, feat_dim = 1, ...)

\S4method{coefficients}{textmodel_ca_fitted}(object, ...)

\S4method{coef}{textmodel_lsa_fitted}(object, doc_dim = 1, feat_dim = 1,
  ...)

\S4method{coefficients}{textmodel_lsa_fitted}(object, ...)

\method{print}{textmodel_wordfish_fitted}(x, n = 30L, ...)

\S4method{show}{textmodel_wordfish_fitted}(object)

\S4method{show}{textmodel_wordfish_predicted}(object)

\S4method{coef}{textmodel_wordfish_fitted}(object, ...)

\S4method{coefficients}{textmodel_wordfish_fitted}(object, ...)

\method{predict}{textmodel_wordscores}(object, newdata = NULL,
  se.fit = FALSE, interval = c("none", "confidence"), level = 0.95,
  rescaling = c("none", "lbg", "mv"), verbose = quanteda_options("verbose"),
  ...)

\method{print}{textmodel_wordscores}(x, digits = 2, ...)

\S4method{show}{textmodel_wordscores}(object)

\S4method{coef}{textmodel_wordscores}(object, ...)

\S4method{coefficients}{textmodel_wordscores}(object, ...)
}
\arguments{
\item{object}{wordfish fitted or predicted object to be shown}

\item{...}{additional arguments passed to \code{\link{print}}}

\item{type}{see \link{residuals.lm}}

\item{x}{for print method, the object to be printed}

\item{n}{max rows of dfm to print}

\item{doc_dim, feat_dim}{the document and feature dimension scores to be 
extracted as coefficients}

\item{newdata}{dfm on which prediction should be made}

\item{level}{probability level for confidence interval width}

\item{rescaling}{\code{"none"} for "raw" scores; \code{"lbg"} for LBG (2003) 
rescaling; or \code{"mv"} for the rescaling proposed by Martin and Vanberg 
(2007).  (Note to authors: Provide full details here in documentation.)}

\item{verbose}{If \code{TRUE}, output status messages}

\item{digits}{number of decimal places to print for print methods}

\item{doc_dim, feat_dim}{the document and feature dimension scores to be 
extracted as coefficients}

\item{object}{a fitted Wordscores textmodel}

\item{...}{additional arguments passed to other functions}

\item{x}{for print method, the object to be printed}

\item{n}{max rows of dfm to print}
}
\value{
The \code{predict} method for a wordscores fitted object returns a 
  data.frame whose rows are the documents fitted and whose columns contain 
  the scored textvalues, with the number of columns depending on the options 
  called (for instance, how many rescaled scores, and whether standard errors
  were requested.)  (Note: We may very well change this soon so that it is a 
  list similar to other existing fitted objects.)
}
\description{
Internal function documentation for textmodel objects.
}
\references{
Laver, Michael, Kenneth R Benoit, and John Garry. 2003. 
"Extracting Policy Positions From Political Texts Using Words as Data." 
\emph{American Political Science Review} 97(02): 311-31.

Martin, L W, and G Vanberg. 2007. "A Robust Transformation Procedure for
Interpreting Political Text." \emph{Political Analysis} 16(1): 93-100.
}
\keyword{internal}
\keyword{textmodel}
