% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{textmodel}
\alias{textmodel}
\title{\code{textmodel} class of fitted models}
\description{
Here you can find mode documentation on the
implementation of the \code{textmodel} class of fitted models on \link{dfm} objects.
}
\seealso{
\code{\link{textmodel_wordscores}}, \code{\link{textmodel_NB}},
\code{\link{fitmodel}}
}

