% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel-ca_sparse.R
\name{textmodel_ca_sparse}
\alias{textmodel_ca_sparse}
\title{correspondence analysis of a document-feature matrix}
\usage{
textmodel_ca_sparse(obj, smooth = 0, nd = NA,
  threads = quanteda_options("threads"), residual_floor = 0.5)
}
\arguments{
\item{obj}{the dfm on which the model will be fit}

\item{smooth}{a smoothing parameter for word counts; defaults to zero.}

\item{nd}{Number of dimensions to be included (default: NA)}

\item{threads}{specifies the number of threads to use; set to 1 to override
the package settings and use a serial version of the function}

\item{residual_floor}{specifies the threshold for residual matrix when 
calculating the svds (default: 0.5)}
}
\description{
\code{textmodel_ca_sparse} implements correspondence analysis scaling on a
\link{dfm}.  The method is a sparse version of function \link[ca]{ca} in the \pkg{ca} package.
}
\examples{
ieDfm <- dfm(data_corpus_irishbudget2010)
wca <- textmodel_ca_sparse(ieDfm)
summary(wca) 
}
\references{
Nenadic, O. and Greenacre, M. (2007). Correspondence analysis in R, with two- and three-dimensional graphics: 
The ca package. \emph{Journal of Statistical Software}, 20 (3), \url{http://www.jstatsoft.org/v20/i03/}
}
\author{
Kenneth Benoit and Haiyan Wang
}
