% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel-generics.R, R/textmodel-wordshoal.R
\docType{class}
\name{textmodel_fitted-class}
\alias{textmodel_fitted-class}
\alias{textmodel_wordshoal_fitted-class}
\alias{textmodel_wordshoal_predicted-class}
\title{the fitted textmodel classes}
\description{
The \code{textmodel} virtual class is a parent class for more specific fitted text models,
which are the result of a quantitative text analysis applied to a document-feature
matrix.
}
\details{
Available types currently include...
}
\section{Slots}{

\describe{
\item{\code{x}}{the dfm on which the wordscores model was called}

\item{\code{y}}{supervised labels (for supervised methods only)}

\item{\code{call}}{the function call that fitted the model}

\item{\code{method}}{the type of textmodel, e.g. wordfish, wordscores, NB-}
}}
\keyword{internal}
\keyword{textmodel}

