% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/textmodel-generics.R, R/textmodel-wordfish.R, R/textmodel-wordscores.R
\docType{class}
\name{textmodel_fitted-class}
\alias{textmodel_fitted-class}
\alias{textmodel_wordfish_fitted-class}
\alias{textmodel_wordfish_predicted-class}
\alias{textmodel_wordscores_fitted-class}
\alias{textmodel_wordscores_predicted-class}
\title{the fitted textmodel classes}
\description{
The \code{textmodel} virtual class is a parent class for more specific fitted text models,
which are the result of a quantitative text analysis applied to a document-feature
matrix.
}
\details{
Available types currently include...
}
\section{Slots}{

\describe{
\item{\code{dfm}}{a \link{dfm-class} document-feature matrix}
}}

