% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel_lsa.R
\name{textmodel_lsa}
\alias{textmodel_lsa}
\title{Latent Semantic Analysis}
\usage{
textmodel_lsa(x, nd = 10)
}
\arguments{
\item{x}{the dfm on which the model will be fit}

\item{nd}{Number of dimensions to be included in output.}
}
\description{
\code{textmodel_lsa} implements Latent Semantic Analysis scaling on a 
\link{dfm}.
}
\details{
\link[RSpectra]{svds} in the \pkg{RSpectra} package is applied to 
  enable the fast computation of the SVD.
}
\note{
The number of dimensions \code{nd} retained in LSA is an empirical issue. While a reduction in k can remove much of the noise, keeping too few dimensions
 or factors may loose important information.
}
\examples{
ieDfm <- dfm(data_corpus_irishbudget2010)
mylsa <- textmodel_lsa(ieDfm)
}
\references{
Barbara Rosario.  2000. "Latent Semantic Indexing: An overview". 
\emph{Techn. rep. INFOSYS 240 Spring Paper, University of California, Berkeley.} 
\url{http://www.sims.berkeley.edu/rosario/projects/LSI.pdf}
}
\author{
Haiyan Wang and Kenneth Benoit
}
