% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel_lsa.R
\name{textmodel_lsa}
\alias{textmodel_lsa}
\alias{textmodel_lsa.dfm}
\alias{transform_lsa}
\title{Latent Semantic Analysis}
\usage{
textmodel_lsa(x, nd = 10)

\method{textmodel_lsa}{dfm}(x, nd = 10)

transform_lsa(newX, LSAspace)
}
\arguments{
\item{x}{the \link{dfm} on which the model will be fit}

\item{nd}{the number of dimensions to be included in output}

\item{newX}{new matrix to be transformed into the lsa space}

\item{LSAspace}{previously fitted lsa space}
}
\description{
Fit the Latent Semantic Analysis scaling model to a \link{dfm}, which may be
weighted (for instance using \code{\link{tfidf}}).
}
\details{
\link[RSpectra]{svds} in the \pkg{RSpectra} package is applied to 
  enable the fast computation of the SVD.
}
\note{
The number of dimensions \code{nd} retained in LSA is an empirical
  issue. While a reduction in \deqn{k} can remove much of the noise, keeping
  too few dimensions or factors may loose important information.
}
\examples{
ieDfm <- dfm(data_corpus_irishbudget2010)
# create an LSA space and return its truncated representation in the low-rank space
mylsa <- textmodel_lsa(ieDfm[1:10, ])
head(mylsa$docs)

#matrix in low_rank LSA space
mylsa$matrix_low_rank[,1:5]

# fold queries into the space generated by ieDfm[1:10,]
# and return its truncated versions of its representation in the new low-rank space
newlsa <- transform_lsa(ieDfm[11:14, ], mylsa)
newlsa$transDocs
}
\references{
Rosario, Barbara.  2000.
  "\href{http://www.cse.msu.edu/~cse960/Papers/LSI/LSI.pdf}{Latent 
  Semantic Indexing: An overview}". \emph{Technical report INFOSYS 240 Spring
  Paper, University of California, Berkeley.}
  
  Deerwester, S., Dumais, S. T., Furnas, G. W., Landauer, T. K., & Harshman, 
  R. 1990. 
  "\href{http://search.proquest.com/openview/a1907164bd88dfc38a4875b73a3f7b3d/1?pq-origsite=gscholar&cbl=1818555}{Indexing
  by latent semantic analysis}". \emph{Journal of the American society for 
  information science} 41(6), 391.
}
\author{
Haiyan Wang
}
