% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel_lsa.R
\name{textmodel_lsa}
\alias{textmodel_lsa}
\title{Latent Semantic Analysis}
\usage{
textmodel_lsa(x, nd = 10)
}
\arguments{
\item{x}{the dfm on which the model will be fit}

\item{nd}{Number of dimensions to be included in output.}
}
\description{
\code{textmodel_lsa} implements Latent Semantic Analysis scaling on a 
\link{dfm}.
}
\details{
\link[RSpectra]{svds} in the \pkg{RSpectra} package is applied to 
  enable the fast computation of the SVD.
}
\examples{
ieDfm <- dfm(data_corpus_irishbudget2010)
mylsa <- textmodel_lsa(ieDfm)
}
\author{
Haiyan Wang and Kenneth Benoit
}
