% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel_lsa.R
\name{textmodel_lsa}
\alias{textmodel_lsa}
\alias{predict.textmodel_lsa_fitted}
\alias{as.dfm.textmodel_lsa_fitted}
\title{Latent Semantic Analysis}
\usage{
textmodel_lsa(x, nd = 10, margin = c("both", "documents", "features"))

\method{predict}{textmodel_lsa_fitted}(object, newdata = NULL, ...)

\method{as.dfm}{textmodel_lsa_fitted}(x)
}
\arguments{
\item{x}{the \link{dfm} on which the model will be fit}

\item{nd}{the number of dimensions to be included in output}

\item{margin}{margin to be smoothes by the SVD}

\item{object}{previously fitted lsa space}

\item{newdata}{new matrix to be transformed into the lsa space}
}
\description{
Fit the Latent Semantic Analysis scaling model to a \link{dfm}, which may be
weighted (for instance using \code{\link{tfidf}}).

Transformation method for Latent Sementic Analysis objects
}
\details{
\link[RSpectra]{svds} in the \pkg{RSpectra} package is applied to 
  enable the fast computation of the SVD.
}
\note{
The number of dimensions \code{nd} retained in LSA is an empirical
  issue. While a reduction in \eqn{k} can remove much of the noise, keeping
  too few dimensions or factors may lose important information.
}
\examples{
ie_dfm <- dfm(data_corpus_irishbudget2010)
# create an LSA space and return its truncated representation in the low-rank space
ie_lsa <- textmodel_lsa(ie_dfm[1:10, ])
head(ie_lsa$docs)
#matrix in low_rank LSA space
ie_lsa$matrix_low_rank[,1:5]
# fold queries into the space generated by ie_dfm[1:10,]
# and return its truncated versions of its representation in the new low-rank space
new_lsa <- predict(ie_lsa, ie_dfm[11:14, ])
new_lsa$docs_newspace
}
\references{
Rosario, Barbara.  2000.
  "\href{http://www.cse.msu.edu/~cse960/Papers/LSI/LSI.pdf}{Latent 
  Semantic Indexing: An overview}". \emph{Technical report INFOSYS 240 Spring
  Paper, University of California, Berkeley.}
  
  Deerwester, S., Dumais, S. T., Furnas, G. W., Landauer, T. K., & Harshman, 
  R. 1990. 
  "\href{http://search.proquest.com/openview/a1907164bd88dfc38a4875b73a3f7b3d/1?pq-origsite=gscholar&cbl=1818555}{Indexing
  by latent semantic analysis}". \emph{Journal of the American society for 
  information science} 41(6), 391.
}
\author{
Haiyan Wang and Kohei Watanabe
}
\keyword{experimental}
\keyword{internal}
\keyword{textmodel}
