% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/textmodel-wordfish.R
\name{textmodel_wordfish}
\alias{textmodel_wordfish}
\title{Wordscores text model}
\usage{
textmodel_wordfish(wfm, dir = c(1, 10), control = list(priors = c(Inf, Inf,
  3, 1), tol = c(1e-06, 1e-08)))
}
\arguments{
\item{wfm}{the dfm on which the model will be fit.  Does not need to contain
only the training documents, since the index of these will be matched
automatically.}

\item{priors}{vector of training scores associated with each document
identified in \code{refData}}

\item{tolerances}{a smoothing parameter for word counts; defaults to zero for the
to match the LBG (2003) method.}
}
\description{
\code{textmodel_wordfish} implements Laver, Benoit and Garry's (2003)
wordscores method for scaling of a single dimension.  This can be called
directly, but the recommended method is through \code{\link{textmodel}}.
}

