% Generated by roxygen2 (4.0.2): do not edit by hand
\name{textmodel_wordfish}
\alias{textmodel_wordfish}
\title{Wordscores text model}
\usage{
textmodel_wordfish(wfm, dir = c(1, 10), control = list(priors = c(Inf, Inf,
  3, 1), tol = c(1e-06, 1e-08)))
}
\arguments{
\item{wfm}{the dfm on which the model will be fit.}

\item{dir}{}

\item{priors}{}

\item{tol}{}
}
\description{
\code{textmodel_wordfish} implements Laver, Benoit and Garry's (2003)
wordscores method for scaling of a single dimension.  This can be called
directly, but the recommended method is through \code{\link{textmodel}}.
}

