% Generated by roxygen2 (4.0.2): do not edit by hand
\name{textmodel_wordscores}
\alias{textmodel_wordscores}
\title{Wordscores text model}
\usage{
textmodel_wordscores(x, train, train.scores, scale = c("linear", "logit"),
  smooth = 0)
}
\arguments{
\item{x}{the dfm on which the model will be fit.  Does not need to contain
only the training documents, since the index of these will be identified in
\code{train}.}

\item{train}{integer index of documents, or character vector of document
names.  For \code{scale="logit"}, only two documents can be specified}

\item{train.scores}{vector of training scores associated with each document
identified in \code{train}}

\item{scale}{classic LBG linear posterior weighted word class differences, or
logit scale of log posterior differences}

\item{smooth}{a smoothing parameter for word counts}
}
\description{
\code{textmodel_wordscores} implements Laver, Benoit and Garry's (2003)
wordscores method for scaling of a single dimension
}
\author{
Kenneth Benoit
}
\references{
Laver, Benoit and Garry (2003); Martin and Vanberg (2007)
}

