% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/textmodel-wordscores.R
\name{textmodel_wordscores}
\alias{textmodel_wordscores}
\title{Wordscores text model}
\usage{
textmodel_wordscores(data, scores, scale = c("linear", "logit"), smooth = 0)
}
\arguments{
\item{data}{the dfm on which the model will be fit.  Does not need to
contain only the training documents, since the index of these will be
matched automatically.}

\item{scores}{vector of training scores associated with each document
identified in \code{refData}}

\item{scale}{classic LBG linear posterior weighted word class differences, or
logit scale of log posterior differences}

\item{smooth}{a smoothing parameter for word counts; defaults to zero for the
to match the LBG (2003) method.}
}
\description{
\code{textmodel_wordscores} implements Laver, Benoit and Garry's (2003)
wordscores method for scaling of a single dimension.  This can be called directly, but the
recommended method is through \code{\link{textmodel}}.
}
\examples{
library(quantedaData)
data(LBGexample)
ws <- textmodel(LBGexample, c(seq(-1.5, 1.5, .75), NA), model="wordscores")
ws
# same as:
textmodel_wordscores(LBGexample, c(seq(-1.5, 1.5, .75), NA))
predict(ws)
}
\author{
Kenneth Benoit
}
\references{
Laver, Michael, Kenneth R Benoit, and John Garry. 2003.
  "Extracting Policy Positions From Political Texts Using Words as Data."
  American Political Science Review 97(02): 311-31; Beauchamp, N. 2012.
  "Using Text to Scale Legislatures with Uninformative Voting." New York
  University Mimeo.; Martin, L W, and G Vanberg. 2007. "A Robust
  Transformation Procedure for Interpreting Political Text." Political
  Analysis 16(1): 93-100.
}

