% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textplot_keyness.R
\name{textplot_keyness}
\alias{textplot_keyness}
\title{plot word keyness}
\usage{
textplot_keyness(x, sort = TRUE, showBothDoc = FALSE, n = 20)
}
\arguments{
\item{x}{the keyness object to be plotted}

\item{sort}{if \code{TRUE} (the default), order points from low to high 
score.}

\item{showBothDoc}{if \code{TRUE} show dual plot for both documents}

\item{n}{number of terms to plot.}
}
\value{
a \pkg{ggplot2} object
}
\description{
Plot the results of textstat_keyness analysis
}
\examples{
\dontrun{
# compare pre- v. post-war terms using grouping
period <- ifelse(docvars(data_corpus_inaugural, "Year") < 1945, "pre-war", "post-war")
mydfm <- dfm(data_corpus_inaugural, groups = period)
head(mydfm) # make sure 'post-war' is in the first row
top20 <- head(result <- textstat_keyness(mydfm), 20)
tail10 <- tail(result, 10)
# plot estimated word keyness
textplot_keyness(top20) 
textplot_keyness(result, showBothDoc = TRUE)}

}
\seealso{
\code{\link{textstat_keyness}}
}
\author{
Kenneth Benoit and Haiyan Wang
}
\keyword{textplot}
