% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textplot_network.R
\name{textplot_network}
\alias{textplot_network}
\title{plot a semantic network of feature co-occurrences}
\usage{
textplot_network(x, width = 10, size = 15, shape = "circle",
  ignore = 0.5, ...)
}
\arguments{
\item{x}{a \link{dfm} object}

\item{width}{the maximum width of links that connects words}

\item{size}{size of the vertices}

\item{shape}{shape of the vertices}

\item{ignore}{a thread for infrequent collocations to be ignored}

\item{...}{additional parameters passed to \code{\link[igraph]{plot.igraph}}.
Note that the 'vertex.' prefix needs to be added.}
}
\description{
Plot an \link{fcm} object as a network, where vertices are features and edges
are co-occurrences, using \code{\link[igraph]{igraph}}.
}
\examples{
\dontrun{
toks <- corpus_subset(data_corpus_inaugural, President == "Obama") \%>\%
    tokens(remove_punct = TRUE) \%>\%
    tokens_remove(stopwords("english"), padding = FALSE)
myfcm <- fcm(toks, context = "window", tri = FALSE)
feat <- names(topfeatures(myfcm, 20))
fcm_select(myfcm, feat, verbose = FALSE) \%>\% textplot_network()
fcm_select(myfcm, feat, verbose = FALSE) \%>\% 
    textplot_network(size = 20, shape = "vrectangle")
}
}
\seealso{
\code{\link{fcm}}
}
\author{
Kohei Watanabe and Stefan Müller
}
\keyword{textplot}
