% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textplot_network.R
\name{textplot_network}
\alias{textplot_network}
\title{plot a semantic network of feature co-occurrences}
\usage{
textplot_network(x, min_freq = 0.5, omit_isolated = TRUE, col = "skyblue",
  alpha = 0.5, size = 2, width = 2, offset = NULL)
}
\arguments{
\item{x}{a \link{fcm} or \link{dfm}  object}

\item{min_freq}{a percentail or frequency count threashold for cooccurances
of features to be plotted.}

\item{omit_isolated}{if \code{TRUE}, features do not occuare more frequent
than \code{min_freq} will be omitted from the plot}

\item{col}{color of edges}

\item{alpha}{opacity of edges}

\item{size}{size of vertices}

\item{width}{width of edges}

\item{offset}{if \code{NULL}, distances between vertices and labels are
determined automatically as far as the number of vertices is less than 100.}
}
\description{
Plot an \link{fcm} object as a network, where vertices are features and edges
are co-occurrences.
}
\examples{
\dontrun{
toks <- corpus_subset(data_corpus_irishbudget2010) \%>\%
    tokens(remove_punct = TRUE) \%>\%
    tokens_tolower() \%>\%
    tokens_remove(stopwords("english"), padding = FALSE)
myfcm <- fcm(toks, context = "window", tri = FALSE)
feat <- names(topfeatures(myfcm, 30))
textplot_network(myfcm, min_freq = 10)
fcm_select(myfcm, feat, verbose = FALSE) \%>\% textplot_network(min_freq = 0.5)
fcm_select(myfcm, feat, verbose = FALSE) \%>\% textplot_network(min_freq = 0.8)
}
}
\seealso{
\code{\link{fcm}}
}
\author{
Kohei Watanabe and Stefan Müller
}
\keyword{textplot}
