% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textplot_network.R
\name{textplot_network}
\alias{textplot_network}
\title{plot semantic network based on collocations}
\usage{
textplot_network(x, width = 10, ignore = 0.99, ...)
}
\arguments{
\item{x}{a dfm object}

\item{width}{the maximum width of links that connects words}

\item{ignore}{a threadh for infrequentl collocations to be ignored}
}
\description{
plot semantic network based on collocations
}
\examples{
# plot the features (without stopwords) from Obama's two inaugural addresses
mydfm <- dfm(corpus_subset(data_corpus_inaugural, President=="Obama"), verbose = FALSE,
             remove = stopwords("english"))
textplot_network(mydfm)

}
\keyword{plot}
