% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textplot_network.R
\name{textplot_network}
\alias{textplot_network}
\title{plot semantic network based on collocations}
\usage{
textplot_network(x, width = 10, size = 15, shape = "circle",
  ignore = 0.5, ...)
}
\arguments{
\item{x}{a dfm object}

\item{width}{the maximum width of links that connects words}

\item{size}{size of the vertices}

\item{shape}{shape of the vertices}

\item{ignore}{a thread for infrequent collocations to be ignored}

\item{...}{additional parameters passed to \code{\link[igraph]{plot.igraph}}. Note that the 'vertex.' prefix needs to be added.}
}
\description{
Plot \link{fcm} as a network, where vertices are features and edges are cooccuraces 
using \code{\link[igraph]{igraph}}.
}
\examples{
\dontrun{
toks <- tokens(corpus_subset(data_corpus_inaugural, President=="Obama"), remove_punct = TRUE)
toks <- tokens_remove(toks, feature = stopwords("english"), padding = FALSE)
myfcm <- fcm(toks, context = 'window', tri = FALSE)
feats <- names(topfeatures(myfcm, 20))
textplot_network(myfcm[feats, feats], ignore = 0.5)
textplot_network(myfcm[feats, feats], ignore = 0.5, size = 20, shape = "vrectangle")
}
}
\seealso{
\code{\link{textmodel_wordfish}}, \code{\link{textmodel_wordscores}}, 
  \code{\link{coef.textmodel}}
}
\author{
Kohei Watanabe and Stefan Müller
}
\keyword{textplot}
