% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textplot_network.R
\name{textplot_network}
\alias{textplot_network}
\title{plot semantic network based on collocations}
\usage{
textplot_network(x, width = 10, ignore = 0.5, ...)
}
\arguments{
\item{x}{a dfm object}

\item{width}{the maximum width of links that connects words}

\item{ignore}{a threadh for infrequentl collocations to be ignored}

\item{...}{additional parameters passed to \code{\link[igraph]{plot.igraph}}}
}
\description{
Plot \link{fcm} as a network, where vertices are features and edges are cooccuraces 
using \code{\link[igraph]{igraph}}.
}
\examples{
\dontrun{
toks <- tokens(corpus_subset(data_corpus_inaugural, President=="Obama"), remove_punct = TRUE)
toks <- tokens_remove(toks, feature = stopwords("english"), padding = FALSE)
myfcm <- fcm(toks, context = 'window', tri = FALSE)
feats <- names(topfeatures(myfcm, 20))
textplot_network(myfcm[feats,feats], ignore = 0.5)
}
}
\keyword{plot}
