% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textplot_scale1d.R
\name{textplot_scale1d}
\alias{textplot_scale1d}
\title{plot a fitted scaling model}
\usage{
textplot_scale1d(x, margin = c("documents", "features"), doclabels = NULL,
  sort = TRUE, groups = NULL, highlighted = NULL, alpha = 0.7,
  highlighted_color = "black")
}
\arguments{
\item{x}{the fitted or predicted scaling model object to be plotted}

\item{margin}{\code{"documents"} to plot estimated document scores (the
default) or \code{"features"} to plot estimated feature scores by a measure
of relative frequency}

\item{doclabels}{a vector of names for document; if left NULL (the default), 
docnames will be used}

\item{sort}{if \code{TRUE} (the default), order points from low to high 
score. If a vector, order according to these values from low to high. Only 
applies when \code{margin = "documents"}.}

\item{groups}{optional grouping variable for sorting categories of the documents.
Only applies when \code{margin = "documents"}.}

\item{highlighted}{a vector of feature names to draw attention to in a 
feature plot; only applies if \code{margin = "features"}}

\item{alpha}{A number between 0 and 1 (default 0.5) representing the level of
alpha transparency used to overplot feature names in a feature plot; only 
applies if \code{margin = "features"}}

\item{highlighted_color}{color for highlighted terms in \code{highlighted}}
}
\value{
a \pkg{ggplot2} object
}
\description{
Plot the results of a fitted scaling model, from (e.g.) a predicted 
\link{textmodel_wordscores} model or a fitted \link{textmodel_wordfish} 
or \link{textmodel_ca}
model. Either document or feature parameters may be plotted: an ideal
point-style plot (estimated document position plus confidence interval on the
x-axis, document labels on the y-axis) with optional renaming and sorting, or
as a plot of estimated feature-level parameters (estimated feature positions
on the x-axis, and a measure of relative frequency or influence on the
y-axis, with feature names replacing plotting points with some being chosen
by the user to be highlighted).
}
\examples{
ie_dfm <- dfm(data_corpus_irishbudget2010)
doclab <- apply(docvars(data_corpus_irishbudget2010, c("name", "party")), 
                1, paste, collapse = " ")

## wordscores
refscores <- c(rep(NA, 4), -1, 1, rep(NA, 8))
ws <- textmodel(ie_dfm, refscores, model="wordscores", smooth = 1)
pred <- predict(ws)
\donttest{# plot estimated word positions
textplot_scale1d(pred, margin = "features", 
                 highlighted = c("minister", "have", "our", "budget"))}
# plot estimated document positions
textplot_scale1d(pred, margin = "documents",
                 doclabels = doclab,
                 groups = docvars(data_corpus_irishbudget2010, "party"))

## wordfish
wfm <- textmodel_wordfish(dfm(data_corpus_irishbudget2010), dir = c(6,5))
# plot estimated document positions
textplot_scale1d(wfm, doclabels = doclab)
textplot_scale1d(wfm, doclabels = doclab,
                 groups = docvars(data_corpus_irishbudget2010, "party"))
\donttest{# plot estimated word positions
textplot_scale1d(wfm, margin = "features", 
                 highlighted = c("government", "global", "children", 
                                 "bank", "economy", "the", "citizenship",
                                 "productivity", "deficit"))}
}
\seealso{
\code{\link{textmodel_wordfish}}, \code{\link{textmodel_wordscores}}, 
  \code{\link{coef.textmodel}}
}
\author{
Kenneth Benoit, Stefan Müller, and Adam Obeng
}
\keyword{textplot}
