% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus-methods-quanteda.R
\name{texts}
\alias{as.character.corpus}
\alias{texts}
\alias{texts<-}
\title{get corpus texts}
\usage{
texts(x, groups = NULL, ...)

texts(x) <- value

\method{as.character}{corpus}(x, ...)
}
\arguments{
\item{x}{a quanteda \link{corpus} or character object}

\item{groups}{character vector containing the names of document variables in
a corpus, or a factor equal in length to the number of documents, used for 
aggregating the texts through concatenation.  If \code{x} is of type character,
then \code{groups} must be a factor.}

\item{...}{unused}

\item{value}{character vector of the new texts}
}
\value{
For \code{texts}, a character vector of the texts in the corpus.
  
  For \code{texts <-}, the corpus with the updated texts.
}
\description{
Get the texts in a quanteda corpus object, with grouping options.  Works for plain character
vectors too, if \code{groups} is a factor.
}
\details{
\code{as.character(x)} where \code{x} is a corpus is equivalent to
calling \code{texts(x)}
}
\note{
You are strongly encouraged as a good practice of text analysis 
  workflow \emph{not} to modify the substance of the texts in a corpus. 
  Rather, this sort of processing is better performed through downstream 
  operations.  For instance, do not lowercase the texts in a corpus, or you 
  will never be able to recover the original case.  Rather, apply 
  \code{\link{toLower}} to the corpus and use the result as an input, e.g. to
  \code{\link{tokenize}}.
}
\examples{
nchar(texts(corpus_subset(data_corpus_inaugural, Year < 1806)))

# grouping on a document variable
nchar(texts(corpus_subset(data_corpus_inaugural, Year < 1806), groups = "President"))

# grouping a character vector using a factor
nchar(data_char_inaugural[1:5])
nchar(texts(data_char_inaugural[1:5], 
            groups = as.factor(data_corpus_inaugural[1:5, "President"])))

BritCorpus <- corpus(c("We must prioritise honour in our neighbourhood.", 
                       "Aluminium is a valourous metal."))
texts(BritCorpus) <- 
    stringi::stri_replace_all_regex(texts(BritCorpus),
                                   c("ise", "([nlb])our", "nium"),
                                   c("ize", "$1or", "num"),
                                   vectorize_all = FALSE)
texts(BritCorpus)
texts(BritCorpus)[2] <- "New text number 2."
texts(BritCorpus)
}

