% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus.R, R/textfile.R
\name{texts}
\alias{texts}
\alias{texts.corpus}
\alias{texts.corpusSource}
\title{get corpus texts}
\usage{
texts(x, ...)

\method{texts}{corpus}(x, groups = NULL, ...)

\method{texts}{corpusSource}(x, groups = NULL, ...)
}
\arguments{
\item{x}{A quanteda corpus object}

\item{...}{not currently used}

\item{groups}{character vector containing the names of document variables for
aggregating documents}
}
\value{
For \code{texts}, a character vector of the texts in the corpus.

For \code{texts <-}, the corpus with the updated texts.
}
\description{
Get the texts in a quanteda corpus object, with grouping options
}
\examples{
texts(inaugCorpus)[1]
sapply(texts(inaugCorpus), nchar)  # length in characters of the inaugual corpus texts
str(texts(ie2010Corpus, groups = "party"))
}

