% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textstat_collocations.R
\name{textstat_collocations}
\alias{textstat_collocations}
\alias{is.collocations}
\title{identify and score multi-word expressions}
\usage{
textstat_collocations(x, method = "all", size = 2, min_count = 2,
  smoothing = 0.5, tolower = TRUE, show_counts = FALSE, ...)

is.collocations(x)
}
\arguments{
\item{x}{a character, \link{corpus}, or \link{tokens} object whose
collocations will be scored.  The tokens object should include punctuation,
and if any words have been removed, these should have been removed with
\code{padding = TRUE}}

\item{method}{association measure for detecting collocations. Let \eqn{i}
index documents, and \eqn{j} index features, \eqn{n_{ij}} refers to 
observed counts, and \eqn{m_{ij}} the expected counts in a collocations 
frequency table of dimensions \eqn{(J - size + 1)^2}.

Available measures are computed as: 

\describe{
\item{\code{"lambda"}}{log-linear coefficient for the full-ngram, similar to Blaheta and Johnson's (2001) method}
\item{\code{"lambda1"}}{unigram subtuples, alternative score from Blaheta and Johnson (2001)}
\item{\code{"lr"}}{The likelihood ratio statistic \eqn{G^2}, computed as: \deqn{2 * \sum_i \sum_j ( n_{ij} * log 
\frac{n_{ij}}{m_{ij}} )} } 
\item{\code{"chi2"}}{Pearson's \eqn{\chi^2} statistic, computed as: 
\deqn{\sum_i \sum_j \frac{(n_{ij} - m_{ij})^2}{m_{ij}}} } 
\item{\code{"pmi"}}{point-wise mutual information 
score, computed as log \eqn{n_{11}/m_{11}}} 
\item{\code{"dice"}}{the Dice coefficient, computed as \eqn{n_{11}/n_{1.} + n_{.1}}} 
 }}

\item{size}{integer; the length of the collocations
to be scored}

\item{min_count}{numeric; minimum frequency of collocations that will be scored}

\item{smoothing}{numeric; a smoothing parameter added to the observed counts
(default is 0.5)}

\item{tolower}{logical; if \code{TRUE}, form collocations as lower-cased combinations}

\item{show_counts}{logical; if \code{TRUE}, output observed and expected counts}

\item{...}{additional arguments passed to \code{\link{tokens}}, if \code{x}
is not a \link{tokens} object already}
}
\value{
\code{textstat_collocations} returns a data.frame of collocations and their
  scores and statistsics.

\code{is.collocation} returns \code{TRUE} if the object is of class
  \code{collocations}, \code{FALSE} otherwise.
}
\description{
Identify and score multi-word expressions, or adjacent fixed-length collocations, from text.
}
\details{
Documents are grouped for the purposes of scoring, but collocations will not span sentences.
If \code{x} is a \link{tokens} object and some tokens have been removed, this should be done
using \code{\link{tokens_remove}(x, pattern, padding = TRUE)} so that counts will still be
accurate, but the pads will prevent those collocations from being scored.
}
\note{
This function is under active development, and we aim to improve both its operation and 
efficiency in the next release of \pkg{quanteda}.
}
\examples{
txts <- data_corpus_inaugural[1:2]
head(cols <- textstat_collocations(txts, size = 2, min_count = 2), 10)
head(cols <- textstat_collocations(txts, size = 3, min_count = 2), 10)

# extracting multi-part proper nouns (capitalized terms)
toks2 <- tokens(data_corpus_inaugural)
toks2 <- tokens_remove(toks2, stopwords("english"), padding = TRUE)
toks2 <- tokens_select(toks2, "^([A-Z][a-z\\\\-]{2,})", valuetype = "regex", 
                       case_insensitive = FALSE, padding = TRUE)
seqs <- textstat_collocations(toks2, size = 3, tolower = FALSE)
head(seqs, 10)
}
\references{
Blaheta, D., & Johnson, M. (2001). 
  \href{http://web.science.mq.edu.au/~mjohnson/papers/2001/dpb-colloc01.pdf}{Unsupervised
   learning of multi-word verbs}. Presented at the ACLEACL Workshop on the 
  Computational Extraction, Analysis and Exploitation of Collocations.
  
  McInnes, B T. (2004). "Extending the Log Likelihood Measure to Improve 
  Collocation Identification."  M.Sc. Thesis, University of Minnesota.
  
  A. Thanopoulos et al(2002), 
  \href{http://www.lrec-conf.org/proceedings/lrec2002/pdf/128.pdf}{Comparative evaluation of collocations extraction metrics}
  
  gensim score,  
  \href{https://radimrehurek.com/gensim/models/phrases.html#gensim.models.phrases.Phrases}{gensim models}
}
\keyword{collocations}
\keyword{experimental}
\keyword{textstat}
