% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textstat_collocations.R
\name{textstat_collocations}
\alias{textstat_collocations}
\title{calculate collocation statistics}
\usage{
textstat_collocations(x, method = c("lr", "chi2", "pmi", "dice", "bj"), ...)
}
\arguments{
\item{x}{a character, \link{corpus}, or \link{tokens} object to be mined 
for collocations}

\item{method}{either one of the methods in \code{\link{collocations}} or 
\code{"bj"} for Blaheta and Johnson's method (called through 
\code{\link{sequences}})}

\item{...}{additional arguments passed to \code{\link{collocations}}}
}
\description{
Identify and score collocations from a corpus, character, or tokens object,
with targeted selection.
}
\examples{
txts <- c("This is software testing: looking for (word) pairs!  
           This [is] a software testing again. For.",
          "Here: this is more Software Testing, looking again for word pairs.")
toks <- tokens(txts)
textstat_collocations(toks, method = "lr")
textstat_collocations(toks, method = "lr", size = 3, min_count = 2)
textstat_collocations(toks, method = "lr", size = 3, min_count = 1)
(cols <- textstat_collocations(toks, method = "lr", "^([a-z]+)$", 
                               valuetype = "regex", min_count = 1))
as.tokens(cols)

toks <- tokens(corpus_segment(data_corpus_inaugural, what = "sentence"))
toks <- tokens_select(toks, stopwords("english"), "remove", padding = TRUE)

# extracting multi-part proper nouns (capitalized terms)
seqs <- textstat_collocations(toks, method = "bj", 
                              features = "^([A-Z][a-z\\\\-]{2,})", 
                              valuetype = "regex", case_insensitive = FALSE)
head(seqs, 10)

# more efficient when applied to the same tokens object 
toks_comp <- tokens_compound(toks, seqs)

# types can be any words
seqs2 <- textstat_collocations(toks, "bj", "^([a-z]+)$", valuetype="regex", 
                               case_insensitive = FALSE, min_count = 2, ordered = TRUE)
head(seqs2, 10)
}
\keyword{textstat}
