% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textstat_collocations.R
\name{textstat_collocations}
\alias{textstat_collocations}
\title{calculate collocation statistics}
\usage{
textstat_collocations(x, y, ...)
}
\arguments{
\item{x}{a \link{dfm} containing the features to be examined for keyness}

\item{target}{the document index (numeric, character or logical) identifying the 
document forming the "target" for computing keyness; all other documents' 
feature frequencies will be combined for use as a reference}

\item{measure}{(signed) association measure to be used for computing keyness.
Currenly available: \code{"chi2"} (\eqn{chi^2} with Yates correction); 
\code{"exact"} (Fisher's exact test); \code{"lr"} for the likelihood ratio
\eqn{G} statistic with Yates correction.}

\item{sort}{logical; if \code{TRUE} sort features scored in descending order 
of the measure, otherwise leave in original feature order}
}
\description{
calculate collocation statistics
}
\examples{
toks <- tokens(data_char_inaugural)
toks <- tokens_remove(toks, stopwords())
coxs <- contexts(toks, "clinton", window = 10, valuetype = "glob")
textstat_collocations(coxs, toks)
}
\keyword{textstat}
