% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textstat_entropy.R
\name{textstat_entropy}
\alias{textstat_entropy}
\title{Compute entropies of documents or features}
\usage{
textstat_entropy(x, margin = c("documents", "features"), base = 2)
}
\arguments{
\item{x}{a \code{dfm}}

\item{margin}{character indicating for which margin to compute entropy}

\item{base}{base for logarithm function}
}
\value{
a data.frame of entropies for the given document or feature
}
\description{
Compute entropies of documents or features
}
\examples{
textstat_entropy(data_dfm_lbgexample)
textstat_entropy(data_dfm_lbgexample, "features")
}
