% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textstat_keyness.R
\name{textstat_keyness}
\alias{textstat_keyness}
\title{Calculate keyness statistics}
\usage{
textstat_keyness(
  x,
  target = 1L,
  measure = c("chi2", "exact", "lr", "pmi"),
  sort = TRUE,
  correction = c("default", "yates", "williams", "none"),
  ...
)
}
\arguments{
\item{x}{a \link{dfm} containing the features to be examined for keyness}

\item{target}{the document index (numeric, character or logical) identifying
the document forming the "target" for computing keyness; all other
documents' feature frequencies will be combined for use as a reference}

\item{measure}{(signed) association measure to be used for computing keyness.
Currently available: \code{"chi2"}; \code{"exact"} (Fisher's exact test); \code{"lr"} for
the likelihood ratio; \code{"pmi"} for pointwise mutual information.  Note that
the "exact" test is very computationally intensive and therefore much
slower than the other methods.}

\item{sort}{logical; if \code{TRUE} sort features scored in descending order
of the measure, otherwise leave in original feature order}

\item{correction}{if \code{"default"}, Yates correction is applied to
\code{"chi2"}; William's correction is applied to \code{"lr"}; and no
correction is applied for the \code{"exact"} and \code{"pmi"} measures.
Specifying a value other than the default can be used to override the
defaults, for instance to apply the Williams correction to the chi2
measure.  Specifying a correction for the \code{"exact"} and \code{"pmi"}
measures has no effect and produces a warning.}

\item{...}{not used}
}
\value{
a data.frame of computed statistics and associated p-values, where
the features scored name each row, and the number of occurrences for both
the target and reference groups. For \code{measure = "chi2"} this is the
chi-squared value, signed positively if the observed value in the target
exceeds its expected value; for \code{measure = "exact"} this is the
estimate of the odds ratio; for \code{measure = "lr"} this is the
likelihood ratio \eqn{G2} statistic; for \code{"pmi"} this is the pointwise
mutual information statistics.

\code{textstat_keyness} returns a data.frame of features and
their keyness scores and frequency counts.
}
\description{
Calculate "keyness", a score for features that occur differentially across
different categories.  Here, the categories are defined by reference to a
"target" document index in the \link{dfm}, with the reference group
consisting of all other documents.
}
\examples{
# compare pre- v. post-war terms using grouping
period <- ifelse(docvars(data_corpus_inaugural, "Year") < 1945, "pre-war", "post-war")
dfmat1 <- dfm(data_corpus_inaugural, groups = period)
head(dfmat1) # make sure 'post-war' is in the first row
head(tstat1 <- textstat_keyness(dfmat1), 10)
tail(tstat1, 10)

# compare pre- v. post-war terms using logical vector
dfmat2 <- dfm(data_corpus_inaugural)
head(textstat_keyness(dfmat2, docvars(data_corpus_inaugural, "Year") >= 1945), 10)

# compare Trump 2017 to other post-war preseidents
dfmat3 <- dfm(corpus_subset(data_corpus_inaugural, period == "post-war"))
head(textstat_keyness(dfmat3, target = "2017-Trump"), 10)

# using the likelihood ratio method
head(textstat_keyness(dfm_smooth(dfmat3), measure = "lr", target = "2017-Trump"), 10)
}
\references{
Bondi, M. & Scott, M. (eds) (2010). \emph{Keyness in
Texts}. Amsterdam, Philadelphia: John Benjamins.

Stubbs, M. (2010). Three Concepts of Keywords. In \emph{Keyness in
Texts}, Bondi, M. & Scott, M. (eds): 1--42. Amsterdam, Philadelphia:
John Benjamins.

Scott, M. & Tribble, C. (2006). \emph{Textual Patterns: Keyword and Corpus
Analysis in Language Education}. Amsterdam: Benjamins: 55.

Dunning, T. (1993). \href{https://dl.acm.org/doi/10.5555/972450.972454}{Accurate Methods for the Statistics of Surprise and Coincidence}. \emph{Computational
Linguistics}, 19(1): 61--74.
}
\keyword{textstat}
