% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textstat_readability.R
\name{textstat_readability}
\alias{textstat_readability}
\title{Calculate readability}
\usage{
textstat_readability(x, measure = c("all", "ARI", "ARI.simple",
  "Bormuth", "Bormuth.GP", "Coleman", "Coleman.C2", "Coleman.Liau",
  "Coleman.Liau.grade", "Coleman.Liau.short", "Dale.Chall",
  "Dale.Chall.old", "Dale.Chall.PSK", "Danielson.Bryan",
  "Danielson.Bryan.2", "Dickes.Steiwer", "DRP", "ELF",
  "Farr.Jenkins.Paterson", "Flesch", "Flesch.PSK", "Flesch.Kincaid", "FOG",
  "FOG.PSK", "FOG.NRI", "FORCAST", "FORCAST.RGL", "Fucks", "Linsear.Write",
  "LIW", "nWS", "nWS.2", "nWS.3", "nWS.4", "RIX", "Scrabble", "SMOG",
  "SMOG.C", "SMOG.simple",      "SMOG.de", "Spache", "Spache.old",
  "Strain", "Traenkle.Bailer", "Traenkle.Bailer.2", "Wheeler.Smith",
  "meanSentenceLength", "meanWordSyllables"), remove_hyphens = TRUE,
  min_sentence_length = 1, max_sentence_length = 10000,
  intermediate = FALSE, ...)
}
\arguments{
\item{x}{a character or \link{corpus} object containing the texts}

\item{measure}{character vector defining the readability measure to calculate.
Matches are case-insensitive.}

\item{remove_hyphens}{if \code{TRUE}, treat constituent words in hyphenated as
separate terms, for purposes of computing word lengths, e.g.
"decision-making" as two terms of lengths 8 and 6 characters respectively,
rather than as a single word of 15 characters}

\item{min_sentence_length, max_sentence_length}{set the minimum and maximum 
sentence lengths (in tokens, excluding punctuation) to include in the
computation of readability.  This makes it easy to exclude "sentences" that
may not really be sentences, such as section titles, table elements, and 
other cruft that might be in the texts following conversion.

For finer-grained control, consider filtering sentences prior first, 
including through pattern-matching, using \code{\link{corpus_trim}}.}

\item{intermediate}{if \code{TRUE}, include intermediate quantities in the output}

\item{...}{not used}
}
\value{
\code{textstat_readability} returns a data.frame of documents and
  their readability scores.
}
\description{
Calculate the readability of text(s) using one of a variety of computed 
indexes.
}
\details{
\code{textstat_readability} calculates the readability of documents 
  using a variety of indices

In the following formulas, we define  Average Sentence Length \deqn{ASL = No. Of
  Tokens / No. of Sentences} , Average Word Length \deqn{AWL = No. of 
  Characters / No. of Tokens}, Average Familiar Words \deqn{AFL = No. of Tokens
   in Dale-Chall List of 3000 Simple Words /  No. of Tokens}, Difficult Words = \deqn{Tokens 
   not in the Dale-Chall List of 3000 Simple Words}.
  \describe{
  
  \item{\code{"ARI"}:}{\emph{Automated Readability Index}: \deqn{ARI = 0.5 ASL  +
  4.71 AWL - 21.34}{ARI = 0.5 ASL  + 4.71 AWL - 21.34}}
  
  \item{\code{"ARI.Simple"}:}{\emph{Simplified Automated Readability Index}: \deqn{
  ARI.Simple =  ASL + 9 AWL}{ARI.Simple =  ASL + 9 AWL}}
  
  \item{\code{"Bormuth"}:}{\emph{Bormuth Readability Index}: \deqn{Bormuth = 
  0.886593 - 0.03640 x AWL + 0.161911 x AFW  - 0.21401 x ASL - 0.000577 x ASL - 
  0.000005 x ASL }{Bormuth = 0.886593 - 0.03640 x AWL + 0.161911 x AFW  - 0.21401 x 
  ASL - 0.000577 x ASL -  0.000005 x ASL }}
  
  \item{\code{"Coleman.Liau"}:}{\emph{Coleman-Liau Index}: \deqn{Coleman-Liau = 
  0.0588 * \frac{No. of Characters}{100 Tokens} + 0.296 \frac{No. of Sentences}{100 Tokens} 
  - 1.58}{Coleman-Liau =  0.0588 x No. of Characters Per 100 Tokens + 0.296 x No. of Sentences Per
   100 Tokens - 1.58}}
   
  \item{\code{"Dale.Chall.Old"}:}{\emph{Dale.Chall.Old}: \deqn{
  Dale.Chall.Old =  0.1579 x 100 x \frac{No. of Difficult Words}{No. of Tokens} + 
  0.0496 x ASL (+ 3.6365)}{Dale.Chall.Old = 0.1579 x 100 x No. of Difficult Words / No.
   of Tokens  + 0.0496 x ASL (+3.6365 if No. of Difficult WOrds / No. of Tokens is > 0.05)}}
  
   
  }
}
\examples{
txt <- c(doc1 = "Readability zero one. Ten, Eleven.", 
         doc2 = "The cat in a dilapidated tophat.")
textstat_readability(txt, measure = "Flesch")
textstat_readability(txt, measure = c("FOG", "FOG.PSK", "FOG.NRI"))

textstat_readability(data_corpus_inaugural[48:58], 
                     measure = c("Flesch.Kincaid", "Dale.Chall.old"))
}
\references{
Senter, R. J., & Smith, E. A. (1967). Automated readability index. CINCINNATI UNIV OH.
}
\author{
Kenneth Benoit, re-engineered from Meik Michalke's \pkg{koRpus}
  package.
}
