% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textstat_readability.R
\name{textstat_readability}
\alias{textstat_readability}
\title{Calculate readability}
\usage{
textstat_readability(x, measure, remove_hyphens = TRUE,
  min_sentence_length = 1, max_sentence_length = 10000,
  intermediate = FALSE, ...)
}
\arguments{
\item{x}{a character or \link{corpus} object containing the texts}

\item{measure}{character vector defining the readability measure to calculate.
Matches are case-insensitive.}

\item{remove_hyphens}{if \code{TRUE}, treat constituent words in hyphenated as
separate terms, for purposes of computing word lengths, e.g.
"decision-making" as two terms of lengths 8 and 6 characters respectively,
rather than as a single word of 15 characters}

\item{min_sentence_length, max_sentence_length}{set the minimum and maximum 
sentence lengths (in tokens, excluding punctuation) to include in the
computation of readability.  This makes it easy to exclude "sentences" that
may not really be sentences, such as section titles, table elements, and 
other cruft that might be in the texts following conversion.

For finer-grained control, consider filtering sentences prior first, 
including through pattern-matching, using \code{\link{corpus_trim}}.}

\item{intermediate}{if \code{TRUE}, include intermediate quantities in the output}

\item{...}{not used}
}
\value{
\code{textstat_readability} returns a data.frame of documents and
  their readability scores.
}
\description{
Calculate the readability of text(s) using one of a variety of computed 
indexes.
}
\details{
The following readability formulas have been implemented, where
\itemize{
  \item Nw = \eqn{n_w} = number of words
  \item Nc = \eqn{n_{c}} = number of characters
  \item Nst = \eqn{n_{st}} = number of sentences
  \item Nsy = \eqn{n_{sy}} = number of syllables
  \item Nwf = \eqn{n_{wf}} = number of words matching the Dale-Chall List
        of 3000 "familiar words"
  \item ASL = Average Sentence Length: number of words / number of sentences
  \item AWL = Average Word Length: number of characters / number of words
  \item AFW = Average Familiar Words: count of words matching the Dale-Chall 
        list of 3000 "familiar words" / number of all words
  \item Nwd = \eqn{n_{wd}} = number of "difficult" words not matching the
        Dale-Chall list of "familiar" words
}

\describe{
  \item{\code{"ARI"}:}{Senter and Smith's (1967) Automated Readability Index.
  \deqn{0.5 ASL  + 4.71 AWL - 21.34}}
  
  \item{\code{"ARI.Simple"}:}{A simplified version of Senter and Smith's
  (1967) Automated Readability Index. Automated Readability Index.
  \deqn{ASL + 9 AWL}}
  
  \item{\code{"Bormuth.MC"}:}{Bormuth's (1969) Mean Cloze Formula. 
  \deqn{0.886593 - 0.03640 \times AWL + 0.161911 \times AFW  - 0.21401 \times
  ASL - 0.000577 \times ASL^2 - 0.000005 \times ASL^3}{
  0.886593 - 0.03640 * AWL + 0.161911 * AFW  - 0.21401 *
  ASL - 0.000577 * ASL^2 - 0.000005 * ASL^3}}
  
  \item{\code{"Bormuth.GP"}:}{Bormuth's (1969) Grade Placement score.
  \deqn{4.275 + 12.881M - 34.934M^2 + 20.388 M^3 + 26.194 CCS -
  2.046 CCS^2 - 11.767 CCS^3 - 42.285(M \times CCS) + 97.620(M \times CCS)^2 -
  59.538(M \times CCS)^2}{
  4.275 + 12.881M - 34.934M^2 + 20.388 M^3 + 26.194 CCS -
  2.046 CCS^2 - 11.767 CCS^3 - 42.285(M * CCS) + 97.620(M * CCS)^2 -
  59.538(M * CCS)^2}
  where \eqn{M} is the Bormuth Mean Cloze Formula as in
  \code{"Bormuth"} above, and \eqn{CCS} is the Cloze Criterion Score (Bormuth,
  1968).}
 
  \item{\code{"Coleman"}:}{Coleman's (1971) Readability Formula 1.
  \deqn{1.29 \times \frac{100 \times n_{wsy=1}}{n_w} - 38.45}{
  1.29 x (100 x Nwsy1 / Nw) - 38.45}
  
  where \eqn{n_{wsy=1}} = Nwsy1 = the number of one-syllable words.  The
  scaling by 100 in this and the other Coleman-derived measures arises
  because the Coleman measures are calculated on a per 100 words basis.}
  
  \item{\code{"Coleman.C2"}:}{Coleman's (1971) Readability Formula 2.
  \deqn{1.16 \times \frac{100 \times n_{wsy=1}}{
  Nw + 1.48 \times \frac{100 \times n_{st}}{n_w} - 37.95}}{
  1.16 * (100 * Nwsy1 / Nw) + 1.48 * (100 * Nst / Nw) - 37.95}}

  \item{\code{"Coleman.Liau.ECP"}:}{Coleman-Liau Estimated Cloze Percent
  (ECP) (Coleman and Liau 1975).
  \deqn{141.8401 - 0.214590 \times 100
  \times AWL + 1.079812 \times \frac{n_{st} \times 100}{n_w}}{
  141.8401 - (0.214590 * 100 * AWL) + (1.079812 * Nst x 100 / Nw)}}
  
  \item{\code{"Coleman.Liau.Grade"}:}{Coleman-Liau Grade Level (Coleman
  and Liau 1975).
  \deqn{-27.4004 \times \mathtt{Coleman.Liau.ECP} \times 100 +
  23.06395}{-27.4004 * Coleman.Liau.ECP / 100 + 23.06395}}

  \item{\code{"Coleman.Liau.Short"}:}{Coleman-Liau Index (Coleman and Liau 1975).
  \deqn{5.88 \times AWL + 29.6 \times \frac{n_{st}}{n_w} - 15.8}{
  5.88 * AWL + 0.296 * Nst / Nw) - 15.8}}
  
  \item{\code{"Dale.Chall"}:}{The New Dale-Chall Readability formula (Chall
  and Dale 1995).
  \deqn{64 - (0.95 \times 100 \times \frac{n_{wd}}{n_w}) - (0.69 \times ASL)}{
  64 - (0.95 * 100 * Nwd / Nw) - (0.69 * ASL)}}

  \item{\code{"Dale.Chall.Old"}:}{The original Dale-Chall Readability formula
  (Dale and Chall (1948). 
  \deqn{0.1579 \times 100 \times \frac{n_{wd}}{n_w} + 0.0496 \times ASL [+ 3.6365]}{
  0.1579 x 100 x Nwd / Nw  + 0.0496 x ASL [+ 3.6365]}
  
  The additional 3.6365 is only added if (Nwd / Nw) > 0.05.}

  \item{\code{"Dale.Chall.PSK"}:}{The Powers-Sumner-Kearl Variation of the 
  Dale and Chall Readability formula (Powers, Sumner, and Kearl 1958).
  \deqn{0.1155 \times
  100 \frac{n_{wd}}{n_w}) + (0.0596 \times ASL) + 3.2672 }{
  (0.1155 x 100 x Nwd / Nw) + (0.0596 x ASL) + 3.2672}}

  \item{\code{"Danielson.Bryan"}:}{\emph{Danielson-Bryan (1963)}: \deqn{
  Danielson-Bryan = (1.0364 \times \frac{No. of Characters}{No. of Blanks}) +
  (0.0194 \times \frac{No. of Characters}{No. of Sentences}) -
  0.6059}{Danielson-Bryan = (1.0364 x No. of Characters / No. of Blanks) +
  (0.0194 x No. of Characters / No. of Sentences) - 0.6059}}

  \item{\code{"Danielson.Bryan2"}:}{\emph{Danielson-Bryan2 (1963)}: \deqn{
  Danielson-Bryan = 131.059- (10.364 \times \frac{No. of Characters}{No. of
  Blanks}) + (0.0194 \times \frac{No. of Characters}{No. of
  Sentences})}{Danielson-Bryan2 = 131.059 - (10.364 x No. of Characters / No.
  of Blanks) + (0.0194 x No. of Characters / No. of Sentences)}}

  \item{\code{"Dickes.Steiwer"}:}{\emph{Dickes-Steiwer Index (1977)}: \deqn{
  Dickes-Steiwer = 235.95993 - (7.3021 \times AWL)  - (12.56438 \times ASL) -
  (50.03293 \times Type-Token Ratio)}{Dickes-Steiwer = 235.95993 - (73.021 x
  AWL) - (12.56438 x ASL) - (50.03293 x TTR)}}

  \item{\code{"DRP"}:}{\emph{Degrees of Reading Power (Relies on Bormuth.MC
  (1969) as in "Bormuth" above)}: \deqn{ DRP = (1 - Bormuth.MC) \times
  100}{DRP = (1 - Bormuth.MC) x 100}}

  \item{\code{"ELF"}:}{\emph{Easy Listening Formula, Fang (1966)}: \deqn{ ELF
  = \frac{No. of Words with >= 2 Syllables}{No. of Sentences}}{ELF = (No. of
  Words with >= 2 Syllables / No. of Sentences)}}

  \item{\code{"Farr.Jenkins.Paterson"}:}{\emph{Farr-Jenkins-Paterson's
  Simplification of Flesch's Reading Ease Score (1951)}: \deqn{
  Farr.Jenkins.Paterson = -31.517 - (1.015 \times ASL) + (1.599 \times
  \frac{No. of 1-Syllable Words}{No. of Words}}{Farr.Jenkins.Paterson -31.517
  - (1.015 x ASL) + (1.599 x No. of 1-Syllable Words / No. of Words)}}

  \item{\code{"Flesch"}:}{\emph{Flesch Reading Ease Score (FRES) (1948)}:
  \deqn{ FRES = 206.835 - (1.015 \times ASL) - (84.6 \times \frac{No. of
  Syllables}{No. of Words})}{FRES = 206.835 - (1.015 x ASL) - (84.6 x (No. of
  Syllables / No. of Words))}}

  \item{\code{"Flesch.PSK"}:}{\emph{Flesch Reading Ease Score,
  Powers-Sumner-Kearl's Variation (1958)}: \deqn{Flesch.PSK = (0.0778 \times
  ASL) + (4.55 \times \frac{No. of Syllables}{No. of Words}) -
  2.2029}{Flesch.PSK = (0.0078 x ASL) + (4.55 x No. of Syllables / No. of
  Words) - 2.2029}}

  \item{\code{"Flesch.Kincaid"}:}{\emph{Flesch-Kincaid Score (1975)}: \deqn{
  Flesch-Kincaid = 0.39 \times ASL + 11.8  \times \frac{No. of Syllables}{No.
  of Words} - 15.59}{Flesch-Kincaid = 0.39 x ASL + 11.8  x(No. of Syllables /
  No. of Words) - 15.59}}

  \item{\code{"FOG"}:}{\emph{Gunning's Fog Index (1952)}: \deqn{FOG = 0.4
  \times (ASL + 100 \times (No. of Words with >= 3 Syllables  / No. of
  Words))}{FOG = 0.4 x (ASL + 100 x (No. of Words with >= 3 Syllables / No.
  of Words))
  
  (The scaling by 100 arises because the initial FOG index is based on
  just a sample of 100 words)}}

  \item{\code{"FOG.PSK"}:}{\emph{FOG, Powers-Sumner-Kearl's Variation
  (1958)}: \deqn{FOG.PSK = 3.0680 \times (0.0877 \times ASL) +(0.0984 \times
  100 \times \frac{No. of Words with >=3 Syllables}{No. of Words})}{FOG.PSK =
  3.0680 x (0.0877 x ASL) +(0.0984 x 100 x No. of Words with >=3 Syllables /
  No. of Words)
  
  (The scaling by 100 arises because the initial FOG index is based on
  just a sample of 100 words)}}

  \item{\code{"FOG.NRI"}:}{\emph{FOG, Navy Readability Index (1975)}:
  \deqn{FOG.NRI = ( \frac{(No. of Words with < 3 Syllables + 3 \times No. of
  3-Syllables Words)}{(100 \times \frac{No. of Sentences}{No. of Words})}  -
  3) / 2 }{FOG = (((No. of Words with <3 Syllables + 3 x No. of 3-Syllables
  Words) / (100 x No. of Sentences / No. of Words))-3) / 2
  
  (The scaling by 100 arises because the initial FOG index is based on
  just a sample of 100 words)}}

  \item{\code{"FORCAST"}:}{\emph{FORCAST (Simplified FORCAST.RGL) (Caylor &
  Sticht, 1973)}: \deqn{FORCAST = 20 - \frac{(No. of 1-Syllable Words \times
  150)}{(No. of Words \times 10)}}{FORCAST = 20 - (No. of 1-Syllable Words x
  150) / (No. of Words x 10)

  (The scaling by 150 arises because the initial FORCAST index is based on
  just a sample of 150 words)}}

  \item{\code{"FORCAST.RGL"}:}{\emph{FORCAST.RGL (Caylor & Sticht, 1973)}:
  \deqn{FORCAST.RGL= 20.43 - 0.11 \times \frac{(No. of 1-Syllable Words
  \times 150)}{No. of Words}}{FORCAST.RGL = 20.43 - 0.11 x (No. of 1-Syllable
  Words x 150) / (No. of Words)

  (The scaling by 150 arises because the initial FORCAST index is based on
  just a sample of 150 words)}}

  \item{\code{"Fucks"}:}{\emph{Fucks' Stilcharakteristik (Style
  Characteristic)}: \deqn{Fucks = AWL  \times ASL }{Fucks = AWL x ASL}}

  \item{\code{"Linsear.Write"}:}{\emph{Linsear Write (Klare, 1975)}:
  \deqn{Linsear Write = \frac{( (100 - (\frac{100 \times No. of Words with <3
  Syllables}{No. of Words})) + (3 \times \frac{100 \times No. of Words with
  >= 3 Syllables}{No. of Words}) )}{(100 \times \frac{No. of Sentences}{No.
  of Words})}}{((100 - (100 x No. of Words with <3 Syllables / No. of Words))
  + (3 x 100 x No. of Words with >= 3 Syllables / No. of Words)) / (100 x No.
  of Sentences / No. of Words)
  
  (The scaling by 100 arises because the initial Linsear.Write measure is based on
  just a sample of 100 words)}}


  \item{\code{"LIW"}:}{\emph{Björnsson's Läsbarhetsindex (1968) (For Swedish
  Texts)}: \deqn{LIW = ASL + \frac{100 \times No. of Words with >= 7
  Syllables}{No. of Words}}{LIW = ASL + (100 * No. of Words with >= 7
  Syllables / No. of Words)
  
  (The scaling by 100 arises because the Läsbarhetsindex index is based on
  just a sample of 100 words)}}

  \item{\code{"nWS"}:}{\emph{Neue Wiener Sachtextformeln 1 (Bamberger &
  Vanecek, 1984)}: \deqn{nwS = 19.35 \times \frac{No. of Words with >=3
  Syllables}{No. of Words} + 0.1672 \times ASL + 12.97 \times \frac{No. of
  Words with >= 6 Characters}{No. of Words} - 3.27 \times \frac{No. of
  1-Syllable Words}{No. of Words} - 0.875}{ nwS = (19.35 x No. of Words with
  >=3 Syllables / No. of Words) + (0.1672 x ASL) + (12.97 x No. of Words with
  >=6 Characters / No. of Words) - (3.27 x No. of 1-Syllable Words / No. of
  Words)- 0.875}}

  \item{\code{"nWS.2"}:}{\emph{Neue Wiener Sachtextformeln 2 (Bamberger &
  Vanecek, 1984)}: \deqn{nwS.2 = 20.07 \times \frac{No. of Words with >=3
  Syllables}{No. of Words} + 0.1682 \times ASL + 13.73 \times \frac{No. of
  Words with >= 6 Characters}{No. of Words} - 2.779}{ nWs.2 = (20.07 x No. of
  Words with >=3 Syllables / No. of Words) + (0.1682 x ASL) + (13.73 x No. of
  Words with >=6 Characters / No. of Words) - 2.779}}

  \item{\code{"nWS.3"}:}{\emph{Neue Wiener Sachtextformeln 3 (Bamberger &
  Vanecek, 1984)}: \deqn{nwS.3 = 29.63 \times \frac{No. of Words with >=3
  Syllables}{No. of Words} + 0.1905 \times ASL - 1.1144}{ nWs.3 = (29.63 x
  No. of Words with >=3 Syllables / No. of Words) + (0.1905 x ASL) - 1.1144}}

  \item{\code{"nWS.4"}:}{\emph{Neue Wiener Sachtextformeln 4 (Bamberger &
  Vanecek, 1984)}: \deqn{nwS.4 = 27.44 \times \frac{No. of Words with >=3
  Syllables}{No. of Words} + 0.2656 \times ASL - 1.693}{ nWs.4 = (27.44 x No.
  of Words with >=3 Syllables / No. of Words) + (0.2656 x ASL) - 1.693}}

  \item{\code{"RIX"}:}{\emph{Anderson's Readability Index (1983)}: \deqn{RIX
  = \frac{No. of Words with >= 7 Syllables}{No. of Sentences}}{RIX = No. of
  Words with >= 7 Syllables / No. of Sentences)}}

  \item{\code{"Scrabble"}:}{\emph{Scrabble Measure}: {Scrabble = Mean
  Scrabble Letter Values of All Words}}

  \item{\code{"SMOG"}:}{\emph{Simple Measure of Gobbledygook (McLaughlin,
  1969)}: \deqn{SMOG = 1.043 \times \sqrt{No. of  Words with >=3 Syllables
  \times \frac{30}{No. of Sentences}} + 3.1291}{SMOG = 1.043 x sqrt(No. of
  Words with >=3 Syllables x 30 / No. Of Sentences) + 3.1291}}

  \item{\code{"SMOG.C"}:}{\emph{SMOG.C}: \deqn{SMOG.C = 0.9986 \times
  \sqrt{No. of Words with >=3 Syllables \times \frac{30}{No. of Sentences} +
  5} +  2.8795}{SMOG.C = 0.9986 x sqrt(No. of Words with >=3 Syllables x (30
  / No. Of Sentences) + 5) + 2.8795}}

  \item{\code{"SMOG.Simple"}:}{\emph{SMOG Simplified}: \deqn{SMOG.Simple =
  \sqrt{No. of Words with >=3 Syllables \times \frac{30}{No. of Sentences}} +
  3}{SMOG.Simple = sqrt(No. of Words with >=3 Syllables x 30 / No. of
  Sentences) + 3}}

  \item{\code{"SMOG.de"}:}{\emph{SMOG.de (Adapted for German Texts)}:
  \deqn{SMOG.de = \sqrt{No. of Words with >=3 Syllables \times \frac{30}{No.
  of Sentences}}-2}{SMOG.de = sqrt(No. of Words with >=3 Syllables x 30 / No.
  Of Sentences) - 2 }}

  \item{\code{"Spache"}:}{\emph{Spache (1952)}: \deqn{Spache = 0.121 \times
  ASL + 0.082 \times (Unique Words not in Spache Word List / No. of Words) +
  0.659}{Spache = 0.121 x ASL + 0.082 x (Unique Words not in Spache Word List
  / No. of Words) + 0.659}}

  \item{\code{"Spache.old"}:}{\emph{Spache (1952)}: \deqn{Spache.old = 0.141
  \times ASL + 0.086 \times Unique Words not in Spache Word List / No. of
  Words) + 0.839}{Spache.old = 0.141 x ASL + 0.086 x (Unique Words not in
  Spache Word List/ No. of Words) + 0.839}}

  \item{\code{"Strain"}:}{\emph{Strain Index (Solomon, 2006)}: \deqn{ Strain
  = Number of Syllables / (Number of Sentences / 3) /10}{Strain = Number of
  Syllables / (Number of Sentences / 3) / 10 (The scaling by 3 arises because
  the initial Strain index is based on just the first 3 sentences )}}

  \item{\code{"Traenkle.Bailer"}:}{\emph{Tränkle & Bailer (1984)}:
  \deqn{Tränkle.Bailer = 224.6814 - (79.8304 \times AWL) - (12.24032 \times
  ASL) - (1.292857 \times 100 \times \frac{No. of Prepositions}{No. of
  Words}}{Tränkle.Bailer = 224.6814 - (79.8304 x AWL) + (12.24032 x ASL) -
  (1.292857 x 100 x No. of Prepositions / No. of Words)
  
  (The scaling by 100 arises because the initial Tränkle & Bailer index is based on
  just a sample of 100 words)}}

  \item{\code{"Traenkle.Bailer2"}:}{\emph{Tränkle & Bailer (1984)}:
  \deqn{Tränkle.Bailer2 = Tränkle.Bailer2 =  234.1063 - (96.11069 \times AWL
  ) - (2.05444 \times 100 \times \frac{No. of Prepositions}{No. of Words}) -
  (1.02805 \times 100 \times \frac{No. of Conjucntions}{No. of
  Words}}{Tränkle.Bailer2 = 234.1063 - 96.11069 x AWL  - 2.05444 x 100 x (No.
  of Prepositions / No. of Words) - 1.02805 x 100 x (No. of Conjunctions/No.
  of Words)
  
  (The scaling by 100 arises because the initial Tränkle & Bailer index is based on
  just a sample of 100 words)}}

  \item{\code{"Wheeler.Smith"}:}{\emph{Wheeler & Smith (1954)}:
  \deqn{Wheeler.Smith = ASL \times 10 \times \frac{No. of Words with >=
  2Syllables}{No. of Words}}{Wheeler.Smith = ASL x 10 x (No. of Words with >=
  2Syllables /No. of Words)}}

  \item{\code{"meanSentenceLength"}:}{\emph{Average Sentence Length}:
  \deqn{Average Sentence Length = \frac{No. of Words}{No. of
  Sentences}}{Average Sentence Length = No. of Words / No. of Sentences}}

  \item{\code{"meanWordSyllables"}:}{\emph{Average Word Syllables}:
  \deqn{Average Word Syllables = \frac{No. of Syllables}{No. of
  Words}}{Average Word Syllables = No. of Syllables / No. of Words}}
  
}
}
\examples{
txt <- c(doc1 = "Readability zero one. Ten, Eleven.", 
         doc2 = "The cat in a dilapidated tophat.")
textstat_readability(txt, measure = "Flesch")
textstat_readability(txt, measure = c("FOG", "FOG.PSK", "FOG.NRI"))

textstat_readability(data_corpus_inaugural[48:58], 
                     measure = c("Flesch.Kincaid", "Dale.Chall.old"))
}
\references{
Anderson, J. (1983). \href{https://www.jstor.org/stable/40031755}{"Lix and
  rix: Variations on a little-known readability index."} 
  Journal of Reading, 26(6), 490-496.
  
  Bamberger, R. & Vanecek, E. (1984). Lesen–Verstehen–Lernen–Schreiben. Wien: Jugend und Volk.
  
  Björnsson, C. H. (1968). \emph{Läsbarhet}. Stockholm: Liber.
  
  Bormuth, J. R. (1969). \href{https://files.eric.ed.gov/fulltext/ED029166.pdf}{Development 
  of Readability Analysis.}  
  
  Bormuth, J. R. (1968). \href{https://www.jstor.org/stable/1433978}{"Cloze test readability:
  Criterion reference scores."} Journal of educational measurement, 5(3), 189-196.  
  
  Caylor, J. S. (1973). \href{https://eric.ed.gov/?id=ED074343}{"Methodologies for Determining 
  Reading Requirements of Military Occupational Specialities."}
  
  Caylor, J. S., & Sticht, T. G. (1973). \href{https://archive.org/details/ERIC_ED076707}{"Development 
  of a Simple Readability Index for Job Reading Material."}
  
  Coleman, E. B. (1971). Developing a technology of written instruction: Some determiners 
  of the complexity of prose. Verbal learning research and the technology of written 
  instruction, 155-204.
  
  Coleman, M., & Liau, T. L. (1975). \href{https://psycnet.apa.org/record/1975-22007-001}{"A computer 
  readability formula designed for machine scoring."} Journal of Applied Psychology, 60(2), 283.
  
  Dale, E. and Chall, J. S. (1948).
  \href{https://www.jstor.org/stable/pdf/1473669}{"A formula for predicting
  readability: Instructions."} \emph{Educational Research Bulletin} 37-54.
  
  Chall, J. S. and Dale, E. (1995). \emph{Readability revisited: The new
  Dale-Chall readability formula}. Brookline Books.
  
  Dickes, P. & Steiwer, L. (1977). "Ausarbeitung von Lesbarkeitsformeln
  für die Deutsche Sprache." \emph{Zeitschrift für Entwicklungspsychologie und
  Pädagogische Psychologie} 9(1), 20--28.
  
  Danielson, W. A., & Bryan, S. D. (1963). \href{https://journals.sagepub.com/doi/abs/10.1177/107769906304000207}{"Computer 
  automation of two readability formulas."} \emph{Journalism Quarterly} 40(2), 201-206.
  
  DuBay, W. H. (2004). \href{http://www.impact-information.com/impactinfo/readability02.pdf}{\emph{The 
  Principles of Readability.}}
  
  Fang, I. E. (1966). \href{https://www.tandfonline.com/doi/abs/10.1080/08838156609363529?journalCode=hbem19}{"The
  "Easy listening formula""}. \emph{Journal of Broadcasting & Electronic Media} 11(1): 63-68.
  
  Farr, J. N., Jenkins, J. J., & Paterson, D. G. (1951).  \href{https://psycnet.apa.org/record/1952-03973-001}{"Simplification of
  Flesch Reading Ease Formula."} \emph{Journal of applied psychology} 35(5):
  333.
  
  Flesch, R. (1948). \href{https://psycnet.apa.org/record/1949-01274-001}{"A new readability yardstick."} \emph{Journal of applied
  psychology} 32(3): 221.
  
  Fucks, W. (1955). Der Unterschied des Prosastils von Dichtern und anderen Schriftstellern. 
  Sprachforum, 1, 233–244.
  
  Gunning, R. (1952). The technique of clear writing.
  
  Klare, G.R. (1975). \href{https://www.jstor.org/stable/pdf/747086.pdf}{"Assessing readability."} \emph{Reading Research Quarterly}, 10(1), 62--102.
  
  Kincaid, J. P., Fishburne Jr, R. P., Rogers, R. L., & Chissom, B. S.
  (1975). \href{https://stars.library.ucf.edu/istlibrary/56/}{"Derivation of new readability formulas (automated readability
  index, FOG count and Flesch reading ease formula) for navy enlisted
  personnel."}
  
  McLaughlin, G. H. (1969). \href{https://ogg.osu.edu/media/documents/health_lit/WRRSMOG_Readability_Formula_G._Harry_McLaughlin__1969_.pdf}{"SMOG grading-a new readability formula."} 
  \emph{Journal of reading}, 12(8), 639-646.

  Powers, R. D., Sumner, W. A., and Kearl, B. E. (1958).
  \href{https://psycnet.apa.org/record/1962-03617-001}{"A recalculation of
  four adult readability formulas."} \emph{Journal of Educational
  Psychology}, 49(2), 99.

  Senter, R. J., & Smith, E. A. (1967). \href{https://apps.dtic.mil/dtic/tr/fulltext/u2/667273.pdf}{Automated readability index.}
  CINCINNATI UNIV OH.

  *Solomon, N. W. (2006). Qualitative Analysis of Media Language. India.

  Spache, G. (1953). \href{https://www.jstor.org/stable/998915}{"A new readability formula for primary-grade reading
  materials."} \emph{The Elementary School Journal}, 53, 410–413.

  Tränkle, U. & Bailer, H. (1984). Kreuzvalidierung und Neuberechnung von
  Lesbarkeitsformeln für die deutsche Sprache. Zeitschrift für
  Entwicklungspsychologie und Pädagogische Psychologie, 16(3), 231–244.

  Wheeler, L.R. & Smith, E.H. (1954). \href{https://www.jstor.org/stable/pdf/41384251}{A practical readability formula for the
  classroom teacher in the primary grades.} \emph{Elementary English}, 31, 397–399.

  *Nimaldasan is the pen name of N. Watson Solomon, Assistant Professor of
  Journalism, School of Media Studies, SRM University, India.
}
\author{
Kenneth Benoit, re-engineered from Meik Michalke's \pkg{koRpus}
  package.
}
