% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textstat_sentiment.R
\name{textstat_sentiment}
\alias{textstat_sentiment}
\title{Compute sentiment using a dictionary method}
\usage{
textstat_sentiment(x, dictionary, ...)
}
\arguments{
\item{x}{a character, \link{corpus}, \link{tokens}, or \link{dfm} object containing
text, tokens, or features whose sentiment will be scored}

\item{dictionary}{a \link{dictionary} that has \link{valence} set, in the form of
numerical valences associated with sentiment}
}
\value{
a data.frame of sentiment scores
}
\description{
Compute sentiment scores from tokens or document-feature matrices, based on
the valences of dictionary keys and values.
}
\examples{
corp <- tail(data_corpus_inaugural, n = 5)
toks <- tokens(corp)
dfmat <- dfm(toks)

valence(data_dictionary_LSD2015) <- list(positive = 1, neg_negative = 1, 
                                         negative = -1, neg_positive = -1)
textstat_sentiment(toks, dictionary = data_dictionary_LSD2015)
# slightly different because does not account for phrases
textstat_sentiment(dfmat, dictionary = data_dictionary_LSD2015)

# Lowe et al (2011) log(pos / neg)
as.dfm(log(dfmat + 0.5)) \%>\%
    textstat_sentiment(dictionary = data_dictionary_LSD2015)

\dontrun{

# AFINN
afinn <- tidytext::get_sentiments(lexicon = c("afinn"))
data_dictionary_afinn <- dictionary(list(afinn = afinn$word))
valence(data_dictionary_afinn) <- list(afinn = afinn$value)
textstat_sentiment(toks, dictionary = data_dictionary_afinn)

# ANEW
anew <- read.delim(url("https://bit.ly/2zZ44w0"))
anew <- anew[!duplicated(anew$Word), ] # because some words repeat
data_dictionary_anew <- dictionary(list(pleasure = anew$Word, 
                                        arousal = anew$Word, 
                                        dominance = anew$Word))
valence(data_dictionary_anew) <- list(pleasure = anew$ValMn, 
                                      arousal = anew$AroMn, 
                                      dominance = anew$DomMn)
textstat_sentiment(toks, data_dictionary_anew["pleasure"])
textstat_sentiment(toks, data_dictionary_anew["arousal"])}

}
\references{
For a discussion of how to aggregate sentiment scores to the document
level, see:

Lowe, W., Benoit, K. R., Mikhaylov, S., & Laver, M. (2011).
Scaling Policy Preferences from Coded Political Texts. \emph{Legislative Studies
Quarterly}, 36(1), 123–155.
\url{http://doi.org/10.1111/j.1939-9162.2010.00006.x}
}
\seealso{
\code{\link[=valence]{valence()}}
}
