% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textstat_simil.R
\name{textstat_simil}
\alias{textstat_simil}
\alias{textstat_dist}
\alias{as.matrix.textstat_simil}
\alias{as.list.textstat_simil}
\title{Similarity and distance computation between documents or features}
\usage{
textstat_simil(x, selection = NULL, margin = c("documents",
  "features"), method = c("correlation", "cosine", "jaccard", "ejaccard",
  "dice", "edice", "hamman", "simple matching"), upper = TRUE,
  diag = TRUE, min_simil = 0)

textstat_dist(x, selection = NULL, margin = c("documents", "features"),
  method = c("euclidean", "kullback", "manhattan", "maximum", "canberra",
  "minkowski"), upper = FALSE, diag = FALSE, p = 2)

\method{as.matrix}{textstat_simil}(x, ...)

\method{as.list}{textstat_simil}(x, sorted = TRUE, n = NULL, ...)
}
\arguments{
\item{x}{a \link{dfm} object}

\item{selection}{a valid index for document or feature names (depending on
\code{margin}) from \code{x}, to be selected for comparison.  The selected
document(s) or feature(s) will form the second of the pairs returned.  
(See Value.)}

\item{margin}{identifies the margin of the dfm on which similarity or
difference will be computed:  \code{"documents"} for documents or
\code{"features"} for word/term features.}

\item{method}{character; the method identifying the similarity or distance
measure to be used; see Details.}

\item{upper}{logical; if \code{TRUE}, record both (A, B) and (B, A) pairs.
Similar to returning just the lower triangle from a \link{dist} object, or
the upper triangle when transforming the output into a matrix.}

\item{diag}{logical; whether to record the similarity of an item with itself.
(similar to the "diagonal" were the output in matrix form).}

\item{min_simil}{numeric; a threshold for the similarity values below which similarity
values will not be returned}

\item{p}{The power of the Minkowski distance.}

\item{sorted}{sort results in descending order if \code{TRUE}}

\item{n}{the top \code{n} highest-ranking items will be returned.  If n is 
\code{NULL}, return all items.}

\item{upper}{whether the upper triangle of the symmetric \eqn{V \times V}
matrix is recorded. Only used when \code{value = "dist"}.}

\item{diag}{whether the diagonal of the distance matrix should be recorded. .
Only used when \code{value = "dist"}.}
}
\value{
A \link{data.frame} containing the pairs of documents or features
  whose similarity or distance will be compared.  
     
  These can be transformed easily into a list format using
  \code{as.list()}, which returns a list for each unique element of the
  second of the pairs.
  
  You can also transform the output in matrix object using
  \code{as.matrix()}, which could (for a \code{textstat_dist} object) be
  further transformed into a \link{dist} object.)
}
\description{
These functions compute matrixes of distances and similarities between
documents or features from a \code{\link{dfm}} and return a
\code{\link[stats]{dist}} object (or a matrix if specific targets are
selected).  They are fast and robust because they operate directly on the
sparse \link{dfm} objects.
}
\details{
\code{textstat_simil} options are: \code{"correlation"} (default),
  \code{"cosine"}, \code{"jaccard"}, \code{"ejaccard"}, \code{"dice"},
  \code{"edice"}, \code{"simple matching"}, and \code{"hamman"}.

\code{textstat_dist} options are: \code{"euclidean"} (default), 
  \code{"kullback"}. \code{"manhattan"}, \code{"maximum"}, \code{"canberra"},
  and \code{"minkowski"}.
}
\note{
If you want to compute similarity on a "normalized" dfm object
  (controlling for variable document lengths, for methods such as correlation
  for which different document lengths matter), then wrap the input dfm in
  \code{\link{dfm_weight}(x, "prop")}.
}
\examples{
# similarities for documents
dfmat <- dfm(corpus_subset(data_corpus_inaugural, Year > 2000), 
             remove_punct = TRUE, remove = stopwords("english"))
(tstat1 <- textstat_simil(dfmat, method = "cosine", margin = "documents"))
as.matrix(tstat1)
as.list(tstat1)

textstat_simil(dfmat, method = "cosine", margin = "documents", diag = FALSE)
textstat_simil(dfmat, method = "cosine", margin = "documents", upper = FALSE)
textstat_simil(dfmat, method = "cosine", margin = "documents", diag = FALSE, upper = FALSE)
 
# min_simil
(tstat2 <- textstat_simil(dfmat, method = "cosine", margin = "documents", min_simil = 0.5))
as.matrix(tstat2)

# similarities for for specific documents
textstat_simil(dfmat, selection = "2017-Trump", margin = "documents")
textstat_simil(dfmat, selection = "2017-Trump", method = "cosine", margin = "documents")
textstat_simil(dfmat, selection = c("2009-Obama" , "2013-Obama"), margin = "documents")

# compute some term similarities
tstat2 <- textstat_simil(dfmat, selection = c("fair", "health", "terror"), method = "cosine",
                         margin = "features")
head(as.matrix(tstat2), 10)
as.list(tstat2, n = 3)

# create a dfm from inaugural addresses from Reagan onwards
dfmat <- dfm(corpus_subset(data_corpus_inaugural, Year > 1990), 
               remove = stopwords("english"), stem = TRUE, remove_punct = TRUE)
               
# distances for documents 
(tstat1 <- textstat_dist(dfmat, margin = "documents"))
as.matrix(tstat1)

# distances for specific documents
textstat_dist(dfmat, "2017-Trump", margin = "documents")
(tstat2 <- textstat_dist(dfmat, c("2009-Obama" , "2013-Obama"), margin = "documents"))
as.list(tstat2)

}
\references{
\code{"kullback"} is the Kullback-Leibler distance, which assumes that
  \eqn{P(x_i) = 0} implies \eqn{P(y_i)=0}, and in case either \eqn{P(x_i)} or
  \eqn{P(y_i)} equals to zero, then \eqn{P(x_i) * log(p(x_i)/p(y_i))} is
  assumed to be zero as the limit value.  The formula is:
   \deqn{\sum{P(x)*log(P(x)/p(y))}}
   
  All other measures are described in the \pkg{proxy} package.
}
\seealso{
\code{\link[stats]{as.dist}}
}
\keyword{internal}
\keyword{textstat}
