% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textstat_simil.R
\docType{class}
\name{textstat_simildist-class}
\alias{textstat_simildist-class}
\alias{textstat_dist-class}
\alias{textstat_dist_sel-class}
\alias{textstat_simil-class}
\alias{textstat_simil_sel-class}
\alias{textstat_simil_sparse-class}
\alias{textstat_simil_sel_sparse-class}
\alias{head.textstat_simildist}
\alias{tail.textstat_simildist}
\title{textstat_simil/dist classes}
\usage{
head.textstat_simildist(x, n = 6L, ...)

tail.textstat_simildist(x, n = 6L, ...)
}
\arguments{
\item{x}{an object}

\item{n}{a single integer.  If positive or zero, size for the resulting
    object: number of elements for a vector (including lists), rows for
    a matrix or data frame or lines for a function.  If negative, all but
    the \code{n} last/first number of elements of \code{x}.}

\item{...}{arguments to be passed to or from other methods.}
}
\description{
Sparse classes for similarity and distance matrices created by
  \code{\link{textstat_simil}} and \code{\link{textstat_dist}}.
}
\section{Slots}{

\describe{
\item{\code{.Data}}{a sparse \pkg{Matrix} object, symmetric if selection is
\code{NULL}}

\item{\code{method}}{the method used for computing similarity or distance}

\item{\code{min_simil}}{numeric; a threshold for the similarity values below which similarity
values are not computed}

\item{\code{margin}}{identifies the margin of the dfm on which similarity or
difference was computed:  \code{"documents"} for documents or
\code{"features"} for word/term features.}

\item{\code{type}}{either \code{"textstat_simil"} or \code{"textstat_dist"}}

\item{\code{selection}}{target units, if any}
}}

\seealso{
\code{\link{textstat_simil}}
}
\keyword{internal}
\keyword{textstat}
