% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textstat.R
\name{textstats}
\alias{textstats}
\title{Statistics for textual data}
\description{
The \verb{textstat_*()} functions formerly in \pkg{quanteda} have now been moved
to the \pkg{quanteda.textstats} package.
}
\seealso{
\code{quanteda.textstats::quanteda.textstats-package}
}
