% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tf}
\alias{tf}
\title{normalizes the term frequencies a dfm}
\usage{
tf(x)
}
\arguments{
\item{x}{Document-feature matrix created by \code{\link{dfm}}}
}
\value{
A dfm matrix object where values are relative term proportions within the document
}
\description{
Returns a matrix of term weights, as a \link{dfm} object
}
\examples{
data(inaugCorpus)
dtm <- dfm(inaugCorpus)
dtm[1:10, 100:110]
tf(dtm)[1:10, 100:110]
}
\author{
Ken Benoit
}

