% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfm-methods.R
\name{tf}
\alias{idf}
\alias{idf.dfm}
\alias{tf}
\alias{tfidf}
\alias{tfidf.dfm}
\title{compute tf-idf weights from a dfm}
\usage{
tf(x)

tfidf(x, ...)

\method{tfidf}{dfm}(x, normalize = TRUE, smoothing = 0L, k = 1, ...)

idf(x, ...)

\method{idf}{dfm}(x, k = 1, USE.NAMES = TRUE, ...)
}
\arguments{
\item{x}{object for which idf or tf-idf will be computed (a document-feature 
matrix)}

\item{...}{not used}

\item{normalize}{if \code{TRUE}, use relative term frequency}

\item{smoothing}{amount to apply as additive smoothing to the 
document-feature matrix prior to weighting, default is \code{0} for no 
smoothing.  Another sensible value would be 0.5.}

\item{k}{additional constant to add to the denominator in \emph{idf} 
computation, default is \code{k = 1}}

\item{USE.NAMES}{logical; if \code{TRUE} attach feature labels as names of 
the resulting numeric vector}
}
\description{
Compute tf-idf, inverse document frequency, and relative term frequency on
document-feature matrices.  See also \code{\link{weight}}.
}
\details{
\code{tf} is a shortcut to compute relative term frequecies (sasme as 
\code{\link{weight}(x, "relFreq")}).
}

