% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighting.R
\docType{methods}
\name{tfidf}
\alias{idf}
\alias{idf.dfm}
\alias{tf}
\alias{tf,dfmDense-method}
\alias{tf,dfmSparse-method}
\alias{tfidf}
\alias{tfidf.dfm}
\title{compute tf-idf weights from a dfm}
\usage{
tfidf(x, ...)

\method{tfidf}{dfm}(x, normalize = TRUE, smoothing = 0L, ...)

idf(x, k = 1, USE.NAMES = TRUE, base = 10)

\method{idf}{dfm}(x, k = 1, USE.NAMES = TRUE, base = 10)

tf(x)

\S4method{tf}{dfmSparse}(x)

\S4method{tf}{dfmDense}(x)
}
\arguments{
\item{x}{object for which idf or tf-idf will be computed (a document-feature 
matrix)}

\item{...}{additional arguments passed to \code{\link{idf}} when calling
\code{tdidf}}

\item{normalize}{if \code{TRUE}, use relative term frequency}

\item{smoothing}{amount to apply as additive smoothing to the 
document-feature matrix prior to weighting, default is \code{0} for no 
smoothing.  Another sensible value would be 0.5.}

\item{k}{additional constant to add to the denominator in \emph{idf} 
computation, default is \code{k = 1}}

\item{USE.NAMES}{logical; if \code{TRUE} attach feature labels as names of 
the resulting numeric vector}

\item{base}{the base with respect to which logarithms in \emph{idf}
computation are computed, default is 10 (see Manning, Raghavan, and Schutze
2008, p123).}
}
\description{
Compute tf-idf, inverse document frequency, and relative term frequency on 
document-feature matrices.  See also \code{\link{weight}}.
}
\details{
\code{tfidf} computes term frequency-inverse document frequency 
  weighting.  The default is to normalize term frequency (by computing 
  relative term frequency within document) but this is not performed if 
  \code{normalize = FALSE}.

\code{idf} computes inverse document frequency with a constant 
  \code{k} added to the denominator of log document frequency.

\code{tf} is a shortcut to compute relative term frequecies (identical to 
\code{\link{weight}(x, "relFreq")}).
}
\examples{
head(LBGexample[, 5:10])
head(tfidf(LBGexample)[, 5:10])
idf(LBGexample)[5:15]
head(tf(LBGexample)[, 5:10])

# replication of worked example from
# https://en.wikipedia.org/wiki/Tf–idf#Example_of_tf.E2.80.93idf
(wikiDfm <- new("dfmSparse", 
                Matrix::Matrix(c(1,1,2,1,0,0, 1,1,0,0,2,3),
                   byrow = TRUE, nrow = 2,  
                   dimnames = list(docs = c("document1", "document2"), 
                     features = c("this", "is", "a", "sample", "another",
                                  "example")), sparse = TRUE)))
idf(wikiDfm, k = 0)
tfidf(wikiDfm, normalize = FALSE, k = 0)
}
\references{
Manning, C. D., Raghavan, P., & Schutze, H. (2008). 
  \emph{Introduction to Information Retrieval}. Cambridge University Press.
}
\seealso{
\code{\link{docfreq}}, \code{\link{weight}}
}

