% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenizers.R
\name{tokenize_custom}
\alias{tokenize_custom}
\alias{breakrules}
\title{Customizable tokenizer}
\source{
\url{https://raw.githubusercontent.com/unicode-org/icu/main/icu4c/source/data/brkitr/rules/word.txt}

\url{https://raw.githubusercontent.com/unicode-org/icu/main/icu4c/source/data/brkitr/rules/sent.txt}
}
\usage{
tokenize_custom(x, rules)

breakrules(what = c("word", "sentence"))
}
\arguments{
\item{x}{character vector for texts to tokenize}

\item{rules}{a list of rules for rule-based boundary detection}

\item{what}{which set of rules to return, one of \code{"word"} or \code{"sentence"}}
}
\value{
a list of characters containing tokens
}
\description{
Allows users to tokenize texts using customized boundary rules. See the \href{https://unicode-org.github.io/icu/userguide/boundaryanalysis/break-rules.html}{ICU website}
for how to define boundary rules.
}
\details{
The package contains internal sets of rules for word and sentence
breaks, which are lists
of rules for word and sentence boundary detection. \code{base} is copied from
the ICU library. Other rules are created by the package maintainers in
\code{system.file("breakrules/breakrules_custom.yml")}.

This function allows modification of those rules, and applies them as a new
tokenizer.

Custom word rules:
\describe{
\item{\code{base}}{ICU's rules for detecting word/sentence boundaries}
\item{\code{keep_hyphens}}{quanteda's rule for preserving hyphens}
\item{\code{keep_url}}{quanteda's rule for preserving URLs}
\item{\code{keep_email}}{quanteda's rule for preserving emails}
\item{\code{keep_tags}}{quanteda's rule for preserving tags}
\item{\code{split_elisions}}{quanteda's rule for splitting elisions}
\item{\code{split_tags}}{quanteda's rule for splitting tags}
}
}
\examples{
lis <- tokenize_custom("a well-known http://example.com", rules = breakrules("word"))
tokens(lis, remove_separators = TRUE)
breakrules("word")
breakrules("sentence")
}
\keyword{internal}
