% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenizers.R
\docType{data}
\name{tokenize_custom}
\alias{tokenize_custom}
\alias{data_breakrules_word}
\alias{data_breakrules_sentence}
\title{Customizable tokenizer}
\format{
\code{data_breakrules_word} and \code{data_breakrules_sentence} are lists of
pre-define word/sentence break rules. \code{base} is copied from the ICU
library. Other rules are created by the package maintainers.
\describe{
\item{\code{base}}{ICU's rules for detecting word/sentence boundaries}
\item{\code{hyphen}}{quanteda's rule for preserving hyphens}
\item{\code{url}}{quanteda's rule for preserving URLs}
\item{\code{email}}{quanteda's rule for preserving emails}
\item{\code{elision}}{quanteda's rule for splitting elisions}
\item{\code{tag}}{quanteda's rule for splitting tags}
}

An object of class \code{list} of length 1.
}
\source{
\url{https://raw.githubusercontent.com/unicode-org/icu/main/icu4c/source/data/brkitr/rules/word.txt}

\url{https://raw.githubusercontent.com/unicode-org/icu/main/icu4c/source/data/brkitr/rules/sent.txt}
}
\usage{
tokenize_custom(x, rules)

data_breakrules_word

data_breakrules_sentence
}
\description{
Allows users to tokenize texts using customized boundary rules. See the \href{https://unicode-org.github.io/icu/userguide/boundaryanalysis/break-rules.html}{ICU website}
for how to define boundary rules.
}
\examples{
tokenize_custom("a well-known website http://example.com", data_breakrules_word["base"])
tokenize_custom("a well-known http://example.com", data_breakrules_word)
}
\keyword{data}
