% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R
\name{tokenize_dictionary_values}
\alias{tokenize_dictionary_values}
\title{Internal function to tokenize dictionary values}
\usage{
tokenize_dictionary_values(dict, separator, ...)
}
\arguments{
\item{dict}{list of object}
}
\description{
Internal function to tokenize dictionary values
}
\examples{
dict <- dictionary(list(ASIA = list("IN" = "印度", 
                                    "ID" = "印度尼西亚")))
quanteda:::tokenize_dictionary_values(dict, " ")
}
\keyword{internal}
