% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokens_chunk.R
\name{tokens_chunk}
\alias{tokens_chunk}
\title{Segment tokens object by chunks of a given size}
\usage{
tokens_chunk(x, size, overlap = FALSE, keep_remainder = FALSE)
}
\arguments{
\item{x}{\link{tokens} object whose token elements will be segmented into
chunks}

\item{size}{integer; the size of the chunks in tokens}

\item{overlap}{logical; if \code{TRUE}, the last token of a chunk will be the
first token of the next chunk}

\item{keep_remainder}{logical; if \code{TRUE}, keep any ending chunk whose
length is less than \code{size}, otherwise this will be discarded}
}
\value{
A \link{tokens} object whose documents have been split into chunks of
  length \code{size}, similar to \code{\link{tokens_segment}}.
}
\description{
Segment tokens by splitting into equally sized chunks.
}
\examples{
txts <- c(doc1 = "Fellow citizens, I am again called upon by the voice of 
                  my country to execute the functions of its Chief Magistrate.",
          doc2 = "When the occasion proper for it shall arrive, I shall 
                  endeavor to express the high sense I entertain of this
                  distinguished honor.")
toks <- tokens(txts)
tokens_chunk(toks, size = 5)
tokens_chunk(toks, size = 5, overlap = TRUE)
tokens_chunk(toks, size = 5, keep_remainder = TRUE)
}
\keyword{tokens}
