% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokens_chunk.R
\name{tokens_chunk}
\alias{tokens_chunk}
\title{[Experimental] Chunk documents in equal sizes}
\usage{
tokens_chunk(x, size, overlap = FALSE)
}
\arguments{
\item{x}{\link{tokens} object whose token elements will be segmented}

\item{size}{size of each chunk}

\item{overlap}{if \code{TRUE}, chunks overlaps with next chunks}
}
\value{
returns a \link{tokens} object whose documents are in the same length
}
\description{
Segment tokens by splitting documents in equal sizes.
}
\examples{
txt <- "Fellow citizens, I am again called upon by the voice of my country to
execute the functions of its Chief Magistrate. When the occasion proper for
it shall arrive, I shall endeavor to express the high sense I entertain of
this distinguished honor."
toks <- tokens(txt)
tokens_chunk(toks, 5)
tokens_chunk(toks, 5, TRUE)
}
\keyword{tokens}
