% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokens_group.R
\name{tokens_group}
\alias{tokens_group}
\title{recombine documents tokens by groups}
\usage{
tokens_group(x, groups)
}
\arguments{
\item{x}{\link{tokens} object}

\item{groups}{a character or numeric vector that speficy grouping of documents}
}
\description{
recombine documents tokens by groups
}
