% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokens_match.R
\name{tokens_match}
\alias{tokens_match}
\title{Match the tokens IDs with given types}
\usage{
tokens_match(x, types, verbose = quanteda_options("verbose"))
}
\arguments{
\item{x}{the \link{tokens} object.}

\item{types}{character vector for the token types to be matched in the resulting
\link{tokens}. Tokens not included in \code{types} are all removed.}

\item{verbose}{if \code{TRUE} print the number of tokens and documents before and
after the function is applied. The number of tokens does not include paddings.}
}
\description{
Match the token types and IDs of multiple \link{tokens} objects using a character vector.
}
\examples{
txt <- c("This is text one", "The text two", "This is text three")
(toks1 <- tokens(txt[1:2]))
(toks2 <- tokens(txt[2:3]))
(toks3 <- tokens_match(toks1, types(toks2)))
identical(types(toks2), types(toks3))
}
\seealso{
dfm_match
}
