% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokens_parallel.R
\name{tokens_parallel}
\alias{tokens_parallel}
\title{Experimental function for parallel tokenization}
\usage{
tokens_parallel(x, docnames = NULL, block_size = 10000, FUN = lapply, ...)
}
\arguments{
\item{x}{a character vector}

\item{block_size}{number of documents in a block (only for development)}

\item{...}{passed to tokens.tokens()}

\item{lapply}{family function for itteration}
}
\description{
Experimental function for parallel tokenization
}
