% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokens_recompile.R
\name{tokens_recompile}
\alias{tokens_recompile}
\title{Remove redundant or unused tokens IDs}
\usage{
tokens_recompile(x, force = FALSE)
}
\arguments{
\item{x}{the \link{tokens_xptr} object.}

\item{force}{if \code{TRUE}, update tokens IDs even when it may not be necessary.}
}
\description{
This function reassign tokens IDs after changes in removal or transformation of
tokens. It update token IDs to ensure that they are unique and dense.
Usually, this function is only applied to \link{tokens_xptr} objects at the end of
text pre-processing.
}
\examples{
toks <- tokens(c(one = "a b c d A B C D",
                 two = "A B C d"), xptr = TRUE)
toks <- tokens_tolower(toks)
types(toks)
types(tokens_recompile(toks))
}
\keyword{internal}
\keyword{tokens}
