% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokens_restore.R
\name{tokens_restore}
\alias{tokens_restore}
\title{Restore special tokens
Compounds a segment of tokens marked by the private Unicode characters. The
segment should start with uE001 and ends with uE002.}
\usage{
tokens_restore(x)
}
\arguments{
\item{x}{tokens object}
}
\description{
Restore special tokens
Compounds a segment of tokens marked by the private Unicode characters. The
segment should start with uE001 and ends with uE002.
}
\examples{
txt <- c(d1 = "オリンピック延期決定！ #politics @abe #政治# #政治 #安部政権 @安部政権 ！")
txt <- stri_replace_all_regex(txt, "@[a-zA-Z0-9_]+|#[\\\\p{L}\\\\p{N}]+#?", "\uE001$0\uE002")
toks <- as.tokens(stri_split_boundaries(txt, type = "word"))
tokens_restore(toks)
}
