% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokens_restore.R
\name{tokens_restore}
\alias{tokens_restore}
\title{Restore special tokens}
\usage{
tokens_restore(x)
}
\arguments{
\item{x}{tokens object}
}
\description{
Compounds a sequence of tokens marked by special markers. The beginning and
the end of the sequence should be marked by U+E001 and U+E002 respectively.
}
\keyword{internal}
