% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokens_sample.R
\name{tokens_sample}
\alias{tokens_sample}
\title{Randomly sample documents from a tokens object}
\usage{
tokens_sample(x, size = NULL, replace = FALSE, prob = NULL, by = NULL)
}
\arguments{
\item{x}{the \link{tokens} object whose documents will be sampled}

\item{size}{a positive number, the number of documents to select; when used
with groups, the number to select from each group or a vector equal in
length to the number of groups defining the samples to be chosen in each
group category.  By defining a size larger than the number of documents, it
is possible to \emph{over}sample groups.}

\item{replace}{if \code{TRUE}, sample be with replacement.}

\item{prob}{a vector of probability weights for obtaining the elements of the
vector being sampled.  May not be applied when \code{by} is used.}

\item{by}{a grouping variable for sampling.  Useful for resampling
sub-document units such as sentences, for instance by specifying \code{by = "document"}}
}
\value{
A \link{tokens} object with number of documents or features equal to
\code{size}, drawn from the tokens \code{x}.
}
\description{
Sample tokenized documents randomly from a tokens object, with or without
replacement. Works just as \code{\link[=sample]{sample()}} works, for document-level
units (and their associated document-level variables).
}
\examples{
set.seed(10)
toks <- tokens(data_corpus_inaugural[1:10])
head(toks)
head(tokens_sample(toks))
head(tokens_sample(toks, replace = TRUE))
}
\seealso{
\link{sample}
}
\keyword{tokens}
