% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokens_sample.R
\name{tokens_sample}
\alias{tokens_sample}
\title{Randomly sample documents or features from a tokens}
\usage{
tokens_sample(x, size = ndoc(x), replace = FALSE, prob = NULL)
}
\arguments{
\item{x}{the tokens object whose documents or features will be sampled}

\item{size}{a positive number, the number of documents or features to select}

\item{replace}{logical; should sampling be with replacement?}

\item{prob}{a vector of probability weights for obtaining the elements of the
vector being sampled.}
}
\value{
A tokens object with number of documents or features equal to \code{size}, drawn 
  from the tokens \code{x}.
}
\description{
Sample documents randomly from a tokens object.
}
\examples{
set.seed(10)
toks <- tokens(data_corpus_inaugural[1:10])
head(toks)
head(tokens_sample(toks))
head(tokens_sample(toks, replace = TRUE))
}
\seealso{
\link{sample}
}
