% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokens_segment.R
\name{tokens_segment}
\alias{tokens_segment}
\title{segment tokens object by patterns}
\usage{
tokens_segment(x, pattern, valuetype = c("glob", "regex", "fixed"),
  case_insensitive = TRUE, extract_pattern = FALSE,
  pattern_position = c("before", "after"),
  verbose = quanteda_options("verbose"))
}
\arguments{
\item{x}{\link{tokens} object whose token elements will be segmented}

\item{pattern}{a character vector, list of character vectors, \link{dictionary},
\link{collocations}, or \link{dfm}. See \link{pattern} for details.}

\item{valuetype}{the type of pattern matching: \code{"glob"} for 
"glob"-style wildcard expressions; \code{"regex"} for regular expressions;
or \code{"fixed"} for exact matching. See \link{valuetype} for details.}

\item{case_insensitive}{ignore case when matching, if \code{TRUE}}

\item{extract_pattern}{remove deimiter from the texts and save in \link{docvars}, if \code{TRUE}}

\item{verbose}{if \code{TRUE} print messages about how many tokens were 
selected or removed}
}
\value{
\code{tokens_segment} returns a \link{tokens} object whose documents
  have been redefined according to the segmentation type
}
\description{
segment tokens object by patterns
}
\examples{
txts <- "Fellow citizens, I am again called upon by the voice of my country to
execute the functions of its Chief Magistrate. When the occasion proper for
it shall arrive, I shall endeavor to express the high sense I entertain of
this distinguished honor."
toks <- tokens(txts)

# split by any punctuation
toks_punc <- tokens_segment(toks, c(".", "?", "!"), valuetype = 'fixed', pattern_position = 'after')
toks_punc <- tokens_segment(toks, "^\\\\p{P}$", valuetype = 'regex', extract_pattern = TRUE, pattern_position = 'after')

}
\keyword{internal}
\keyword{tokens}
