% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokens_split.R
\name{tokens_split}
\alias{tokens_split}
\title{Split tokens by separator}
\usage{
tokens_split(x, separator = " ", valuetype = c("fixed", "regex"),
  remove_separator = FALSE)
}
\arguments{
\item{x}{tokens object}

\item{separator}{a character by which tokens are seperated}

\item{valuetype}{the type of pattern matching: \code{"glob"} for 
"glob"-style wildcard expressions; \code{"regex"} for regular expressions;
or \code{"fixed"} for exact matching. See \link{valuetype} for details.}

\item{remove_separator}{if \code{TRUE}, remove separator from new tokens}
}
\description{
Split tokens by separator
}
\examples{
toks <- tokens("UK-EU negotiation is not going anywhere as of 2019-12-24", 
               separator = "-")
tokens_split(toks) 

}
