% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tolower-misc.R
\name{tokens_tolower}
\alias{tokens_tolower}
\alias{tokens_toupper}
\title{convert the case of tokens}
\usage{
tokens_tolower(x, ...)

tokens_toupper(x, ...)
}
\arguments{
\item{x}{a \link{tokens} object}

\item{...}{additional arguments passed to \code{\link{char_tolower}}
(currently this is limited to \code{keep_acronyms})}
}
\description{
\code{tokens_tolower} and \code{tokens_toupper} convert the features of a
\link{tokens} object and reindex the types.
}
\examples{
# for a document-feature matrix
toks <- tokens(c(txt1 = "b A A", txt2 = "C C a b B"))
tokens_tolower(toks) 
tokens_toupper(toks)
}
\seealso{
\code{\link{char_tolower}}, \code{\link{char_toupper}}
}
