% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokens_xptr.R
\name{tokens_xptr}
\alias{tokens_xptr}
\alias{is.tokens_xptr}
\alias{as.tokens_xptr}
\alias{as.tokens_xptr.tokens}
\alias{as.tokens_xptr.tokens_xptr}
\title{Methods for tokens_xptr objects}
\usage{
is.tokens_xptr(x)

as.tokens_xptr(x)

\method{as.tokens_xptr}{tokens}(x)

\method{as.tokens_xptr}{tokens_xptr}(x)
}
\arguments{
\item{x}{a \link{tokens} object to convert or a \code{tokens_xptr} class object to deep
copy.}
}
\value{
\code{is.tokens_xptr()} returns \code{TRUE} if the object is a external
pointer-based tokens object, \code{FALSE} otherwise.

\code{as.tokens_xptr()} returns a (deep copy of a) \code{tokens_xtpr} class
object.
}
\description{
Methods for creating and testing for \code{tokens_xptr} objects, which are
\link{tokens} objects containing pointers to memory locations that can be passed
by reference for efficient processing in \verb{tokens_*()} functions that modify
them, or for constructing a document-feature matrix without requiring a deep
copy to be passed to \code{\link[=dfm]{dfm()}}.

\code{is.tokens_xptr()} tests whether an object is of class
\code{tokens_xtpr}.

\code{as.tokens_xptr()} coerces a \code{tokens} object to an external
pointer-based tokens object, or returns a deep copy of a \code{tokens_xtpr} when
\code{x} is already a \code{tokens_xtpr} object.
}
\keyword{tokens}
