% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/toLower.R
\name{toLower}
\alias{toLower}
\title{Convert texts to lower case}
\usage{
toLower(x, ...)
}
\arguments{
\item{x}{The text to be lower-cased}

\item{...}{additional arguments passed to \code{\link{clean}}}
}
\value{
A list of length \code{\link{ndoc}(x)} of the tokens found in each text.
#'
}
\description{
Convert texts to lower case
}
\examples{
# same for character vectors and for lists
tokensFromChar <- tokenize(inaugTexts[1:3])
tokensFromCorp <- tokenize(subset(inaugCorpus, Year<1798))
identical(tokensFromChar, tokensFromCorp)
str(tokensFromChar)
}

