% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfm-methods.R
\name{topfeatures}
\alias{topfeatures}
\title{identify the most frequent features in a dfm}
\usage{
topfeatures(x, n = 10, decreasing = TRUE, groups = NULL, ci = 0.95)
}
\arguments{
\item{x}{the object whose features will be returned}

\item{n}{how many top features should be returned}

\item{decreasing}{If \code{TRUE}, return the \code{n} most frequent features; 
otherwise return the \code{n} least frequent features}

\item{groups}{either a character or integer vector to group documents. If given, 
return a list of top features of the document groups. See \link{dfm_group} for 
detail.}

\item{ci}{confidence interval from 0-1.0 for use if dfm is resampled
(not currently used)}
}
\value{
A named numeric vector of feature counts, where the names are the
  feature labels, or a list of these if \code{groups} is given.
}
\description{
List the most (or least) frequently occuring features in a \link{dfm}, either
as a whole or separated by document.
}
\examples{
mydfm <- dfm(corpus_subset(data_corpus_inaugural, Year > 1980), remove_punct = TRUE)
mydfm_nostopw <- dfm_remove(mydfm, stopwords("english"))

# most frequent features
topfeatures(mydfm)
topfeatures(mydfm_nostopw)

# least frequent features
topfeatures(mydfm_nostopw, decreasing = FALSE)

# top features of individual documents  
topfeatures(mydfm_nostopw, n = 5, groups = docnamses(mydfm_nostopw))
}
