% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokens.R
\name{types}
\alias{types}
\title{get types of tokens from a tokens object}
\usage{
types(x)
}
\description{
Get unique types of tokens from a \link{tokens} object.
}
\examples{
toks <- tokens(data_corpus_inaugural)
types(toks)
}
\seealso{
\link{featnames}
}
