% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dataDocs.R
\docType{data}
\name{ukimmigrTexts}
\alias{ukimmigrTexts}
\title{Immigration-related sections of 2010 UK party manifestos}
\format{A named character vector of plain ASCII texts}
\description{
Extracts from the election manifestos of 9 UK political parties from 2010, related
to immigration or asylum-seekers.
}
\examples{
data(ukimmigrTexts)
ukimmigrCorpus <- corpus(ukimmigrTexts, docvars=list(party=names(ukimmigrTexts)))
language(ukimmigrTexts) <- "english"
encoding(ukimmigrTexts) <- "UTF-8"
summary(ukimmigrTexts)
}

