% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataDocs.R
\docType{data}
\name{ukimmigTexts}
\alias{ukimmigTexts}
\title{Immigration-related sections of 2010 UK party manifestos}
\format{A named character vector of plain ASCII texts}
\description{
Extracts from the election manifestos of 9 UK political parties from 2010, related
to immigration or asylum-seekers.
}
\examples{
data(ukimmigTexts)
ukimmigCorpus <- corpus(ukimmigTexts, docvars=list(party=names(ukimmigTexts)))
language(ukimmigCorpus) <- "english"
encoding(ukimmigCorpus) <- "UTF-8"
summary(ukimmigCorpus)
}

