% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfm-deprecated.R
\name{weight}
\alias{weight}
\alias{smoother}
\title{weight or smooth a dfm}
\usage{
weight(x, ...)

smoother(x, ...)
}
\arguments{
\item{x}{document-feature matrix created by \link{dfm}}

\item{...}{arguments passed to \code{\link{dfm_weight}} and and \code{\link{dfm_smooth}}}
}
\description{
Deprecated weighting and smoothing functions for dfm objects.  See instead
\code{\link{dfm_weight}} and \code{\link{dfm_smooth}}.
}
\keyword{deprecated}
\keyword{internal}
