% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfm-deprecated.R, R/dfm_weight.R
\docType{methods}
\name{weight}
\alias{dfm_weight,dfm,character-method}
\alias{smoother}
\alias{weight}
\title{weight or smooth a dfm}
\usage{
weight(x, ...)

smoother(x, ...)

\S4method{dfm_weight}{dfm,character}(x, type = c("frequency", "relFreq",
  "relMaxFreq", "logFreq", "tfidf"), ...)
}
\arguments{
\item{x}{document-feature matrix created by \link{dfm}}

\item{...}{arguments passed to \code{\link{dfm_weight}} and and \code{\link{dfm_smooth}}}
}
\description{
Deprecated weighting and smoothing functions for dfm objects.  See instead
\code{\link{dfm_weight}} and \code{\link{dfm_smooth}}.
}
\examples{
\dontshow{
testdfm <- dfm(data_char_inaugural[1:5], verbose = FALSE)
for (w in c("frequency", "relFreq", "relMaxFreq", "logFreq", "tfidf")) {
    testw <- dfm_weight(testdfm, w)
    cat("\\n\\n=== weight() TEST for:", w, "; class:", class(testw), "\\n")
    head(testw)
}}
}
\keyword{deprecated}
\keyword{internal}

