% Generated by roxygen2 (4.0.2): do not edit by hand
\name{weight}
\alias{weight}
\title{Weight the feature frequencies in a dfm by various methods}
\usage{
weight(x, type = c("normTf", "maxTf", "logTf", "tfidf", "ppmi"),
  smooth = FALSE)
}
\arguments{
\item{x}{document-feature matrix created by \link{dfm}}

\item{type}{The weighting function to aapply to the dfm. One of:
\itemize{
  \item normTf - Length normalization: dividing the frequency of the feature
  by the length of the document)
  \item tf-idf - Term-frequency * inverse
  document frequency. For a full explanation, see, for example,
  \url{(http://nlp.stanford.edu/IR-book/html/htmledition/term-frequency-and-weighting-1.html)}.
   This implementation will not return negative values.
   \item logTf - The
  natural log of the term frequency
  \item maxTf - The term frequency divided
  by the frequency of the most frequent term in the document
  \item ppmi -   Positive Pointwise Mutual Information }}

\item{smooth}{Apply additivee smoothing to the matrix before weighting. TRUE
by default, adding 0.5 to counts.}
}
\value{
The original dfm, with values weighted according to type function.
}
\description{
Returns a document by feature matrix with the feature frequencies weighted
according to one of several common methods.
}
\examples{
dtm <- dfm(inaugCorpus)
x <- apply(dtm, 1, function(tf) tf/max(tf))
topfeatures(dtm)
normDtm <- weight(dtm)
topfeatures(normDtm)
maxTfDtm <- weight(dtm, type="maxTf")
topfeatures(maxTfDtm)
logTfDtm <- weight(dtm, type="logTf")
topfeatures(logTfDtm)
tfidfDtm <- weight(dtm, type="tfidf")
topfeatures(tfidfDtm)
pmiDtm <- weight(dtm+1, type="ppmi")
topfeatures(pmiDtm)

# combine these methods for more complex weightings, e.g. as in Section 6.4 of
# Introduction to Information Retrieval
logTfDtm <- weight(dtm, type="logTf")
wfidfDtm <- weight(logTfDtm, type="tfidf")
}
\author{
Paul Nulty
}
\references{
Manning, Christopher D., Prabhakar Raghavan, and Hinrich Schutze.
  Introduction to information retrieval. Vol. 1. Cambridge: Cambridge
  university press, 2008.
}

