% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dfm-main.R, R/dfm_main.R
\docType{methods}
\name{weight}
\alias{smoother}
\alias{tf}
\alias{tfidf}
\alias{weight}
\alias{weight,dfm-method}
\alias{weighting}
\alias{weighting,dfm-method}
\title{Weight the feature frequencies in a dfm by various methods}
\usage{
weight(x, ...)

\S4method{weight}{dfm}(x, type = c("frequency", "relFreq", "relMaxFreq",
  "logFreq", "tfidf"), smooth = 0, normalize = TRUE, verbose = TRUE, ...)

tf(x)

tfidf(x)

smoother(x, smooth)

weighting(object)

\S4method{weighting}{dfm}(object)

weight(x, ...)

\S4method{weight}{dfm}(x, type = c("frequency", "relFreq", "relMaxFreq",
  "logFreq", "tfidf"), smooth = 0, normalize = TRUE, verbose = TRUE, ...)

tf(x)

tfidf(x)

smoother(x, smooth)

weighting(object)

\S4method{weighting}{dfm}(object)
}
\arguments{
\item{x}{document-feature matrix created by \link{dfm}}

\item{...}{not currently used}

\item{type}{The weighting function to aapply to the dfm. One of:
\itemize{
  \item normTf - Length normalization: dividing the frequency of the feature
  by the length of the document)
  \item logTf - The natural log of the term frequency
  \item tf-idf - Term-frequency * inverse
  document frequency. For a full explanation, see, for example,
  \url{http://nlp.stanford.edu/IR-book/html/htmledition/term-frequency-and-weighting-1.html}.
   This implementation will not return negative values.
  \item maxTf - The term frequency divided
  by the frequency of the most frequent term in the document
  \item ppmi -   Positive Pointwise Mutual Information }}

\item{smooth}{amount to apply as additive smoothing to the document-feature matrix prior to
weighting, default is 0.5, set to \code{smooth=0} for no smoothing.}

\item{normalize}{if \code{TRUE} (default) then normalize the dfm by relative
term frequency prior to computing tfidf}

\item{verbose}{if \code{TRUE} output status messages}

\item{object}{the dfm object for accessing the weighting setting}

\item{x}{document-feature matrix created by \link{dfm}}

\item{type}{The weighting function to aapply to the dfm. One of:
\itemize{
  \item normTf - Length normalization: dividing the frequency of the feature
  by the length of the document)
  \item logTf - The natural log of the term frequency
  \item tf-idf - Term-frequency * inverse
  document frequency. For a full explanation, see, for example,
  \url{http://nlp.stanford.edu/IR-book/html/htmledition/term-frequency-and-weighting-1.html}.
   This implementation will not return negative values.
  \item maxTf - The term frequency divided
  by the frequency of the most frequent term in the document
  \item ppmi -   Positive Pointwise Mutual Information }}

\item{smooth}{amount to apply as additive smoothing to the document-feature matrix prior to
weighting, default is 0.5, set to \code{smooth=0} for no smoothing.}

\item{normalize}{if \code{TRUE} (default) then normalize the dfm by relative
term frequency prior to computing tfidf}

\item{verbose}{if \code{TRUE} output status messages}

\item{...}{not currently used}

\item{object}{the dfm object for accessing the weighting setting}
}
\value{
The dfm with weighted values

\code{weighting} returns a character object describing the type of weighting applied to the dfm.

The dfm with weighted values

\code{weighting} returns a character object describing the type of weighting applied to the dfm.
}
\description{
Returns a document by feature matrix with the feature frequencies weighted
according to one of several common methods.

Returns a document by feature matrix with the feature frequencies weighted
according to one of several common methods.
}
\details{
\code{tf} is a shortcut for \code{weight(x, "relFreq")}

\code{tfidf} is a shortcut for \code{weight(x, "tfidf")}

\code{smoother} is a shortcut for \code{weight(x, "frequency", smooth)}

\code{weighting} queries (but cannot set) the weighting applied to the dfm.

\code{tf} is a shortcut for \code{weight(x, "relFreq")}

\code{tfidf} is a shortcut for \code{weight(x, "tfidf")}

\code{smoother} is a shortcut for \code{weight(x, "frequency", smooth)}

\code{weighting} queries (but cannot set) the weighting applied to the dfm.
}
\examples{
dtm <- dfm(inaugCorpus)
x <- apply(dtm, 1, function(tf) tf/max(tf))
topfeatures(dtm)
normDtm <- weight(dtm)
topfeatures(normDtm)
maxTfDtm <- weight(dtm, type="relMaxFreq")
topfeatures(maxTfDtm)
logTfDtm <- weight(dtm, type="logFreq")
topfeatures(logTfDtm)
tfidfDtm <- weight(dtm, type="tfidf")
topfeatures(tfidfDtm)

# combine these methods for more complex weightings, e.g. as in Section 6.4 of
# Introduction to Information Retrieval
logTfDtm <- weight(dtm, type="logFreq")
wfidfDtm <- weight(logTfDtm, type="tfidf", normalize=FALSE)
\dontshow{
testdfm <- dfm(inaugTexts[1:5])
print(testdfm[, 1:5])
for (w in c("frequency", "relFreq", "relMaxFreq", "logFreq", "tfidf")) {
    testw <- weight(testdfm, w)
    cat("\\nweight test for:", w, "; class:", class(testw), "\\n")
    print(testw[, 1:5])
}
}
dtm <- dfm(inaugCorpus)
x <- apply(dtm, 1, function(tf) tf/max(tf))
topfeatures(dtm)
normDtm <- weight(dtm)
topfeatures(normDtm)
maxTfDtm <- weight(dtm, type="relMaxFreq")
topfeatures(maxTfDtm)
logTfDtm <- weight(dtm, type="logFreq")
topfeatures(logTfDtm)
tfidfDtm <- weight(dtm, type="tfidf")
topfeatures(tfidfDtm)

# combine these methods for more complex weightings, e.g. as in Section 6.4 of
# Introduction to Information Retrieval
logTfDtm <- weight(dtm, type="logFreq")
wfidfDtm <- weight(logTfDtm, type="tfidf", normalize=FALSE)
\dontshow{
testdfm <- dfm(inaugTexts[1:5])
print(testdfm[, 1:5])
for (w in c("frequency", "relFreq", "relMaxFreq", "logFreq", "tfidf")) {
    testw <- weight(testdfm, w)
    cat("\\nweight test for:", w, "; class:", class(testw), "\\n")
    print(testw[, 1:5])
}
}
}
\author{
Paul Nulty and Kenneth Benoit

Paul Nulty and Kenneth Benoit
}
\references{
Manning, Christopher D., Prabhakar Raghavan, and Hinrich Schutze.
  Introduction to information retrieval. Vol. 1. Cambridge: Cambridge
  university press, 2008.

Manning, Christopher D., Prabhakar Raghavan, and Hinrich Schutze.
  Introduction to information retrieval. Vol. 1. Cambridge: Cambridge
  university press, 2008.
}

