% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textplot_wordcloud.R
\name{wordcloud}
\alias{wordcloud}
\title{Internal function for textplot_wordcloud}
\usage{
wordcloud(
  x,
  min_size,
  max_size,
  min_count,
  max_words,
  color,
  font,
  adjust,
  rotation,
  random_order,
  random_color,
  ordered_color,
  labelcolor,
  labelsize,
  labeloffset,
  fixed_aspect,
  colors,
  scale,
  min.freq,
  max.words,
  random.order,
  random.color,
  rot.per,
  ordered.colors,
  use.r.layout,
  fixed.asp,
  ...
)
}
\arguments{
\item{x}{a \link{dfm} or \link{textstat_keyness} object}

\item{min_size}{size of the smallest word}

\item{max_size}{size of the largest word}

\item{min_count}{words with frequency below min_count will not be plotted}

\item{max_words}{maximum number of words to be plotted. The least frequent
terms dropped.  The maximum frequency will be split evenly across
categories when \code{comparison = TRUE}.}

\item{color}{color of words from least to most frequent}

\item{font}{font-family of words and labels. Use default font if \code{NULL}.}

\item{adjust}{adjust sizes of words by a constant. Useful for non-English
words for which R fails to obtain correct sizes.}

\item{rotation}{proportion of words with 90 degree rotation}

\item{random_order}{plot words in random order. If \code{FALSE}, they will be
plotted in decreasing frequency.}

\item{random_color}{choose colors randomly from the colors. If \code{FALSE},
the color is chosen based on the frequency}

\item{ordered_color}{if \code{TRUE}, then colors are assigned to words in
order.}

\item{labelcolor}{color of group labels. Only used when \code{comparison = TRUE}.}

\item{labelsize}{size of group labels. Only used when \code{comparison = TRUE}.}

\item{labeloffset}{position of group labels. Only used when
\code{comparison = TRUE}.}

\item{fixed_aspect}{logical; if \code{TRUE}, the aspect ratio is fixed. Variable
aspect ratio only supported if rotation = 0.}

\item{scale}{deprecated argument}

\item{min.freq}{deprecated argument}

\item{max.words}{deprecated argument}

\item{random.order}{deprecated argument}

\item{random.color}{deprecated argument}

\item{rot.per}{deprecated argument}

\item{ordered.colors}{deprecated argument}

\item{use.r.layout}{deprecated argument}

\item{fixed.asp}{deprecated argument}

\item{...}{additional parameters. Only used to make it compatible with
\pkg{wordcloud}}
}
\description{
This function implements wordcloud without dependencies. Code is adopted from
\code{\link[wordcloud:wordcloud]{wordcloud::wordcloud()}}.
}
\author{
Kohei Watanabe, building on code from Ian Fellows's \pkg{wordcloud} package.
}
\keyword{internal}
