% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordstem.R
\name{wordstem}
\alias{wordstem}
\alias{wordstem.character}
\alias{wordstem.dfm}
\alias{wordstem.tokenizedTexts}
\alias{wordstem.tokens}
\title{stem words}
\usage{
wordstem(x, language = "porter")

\method{wordstem}{dfm}(x, language = "porter")

\method{wordstem}{tokens}(x, language = "porter")

\method{wordstem}{tokenizedTexts}(x, language = "porter")

\method{wordstem}{character}(x, language = "porter")
}
\description{
Deprecated function names to stem words.  See \code{\link{char_wordstem}}, 
\code{\link{tokens_wordstem}}, and \code{\link{dfm_wordstem}} instead.
}
\keyword{deprecated}
\keyword{internal}

